% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_approval_services_api.R
\docType{class}
\name{AccessApprovalServicesApi}
\alias{AccessApprovalServicesApi}
\title{AccessApprovalServices operations}
\format{
An \code{R6Class} generator object
}
\description{
AccessApprovalServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ deleteRepoV1AccessApproval } \emph{  }

\itemize{
\item \emph{ @param } requirement_id character
\item \emph{ @param } accessor_id character


\item status code : 200 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1AccessApprovalApprovalId } \emph{  }

\itemize{
\item \emph{ @param } approval_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAccessApproval} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAccessApproval
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1AccessApproval } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_access_approval \link{OrgSagebionetworksRepoModelAccessApproval}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAccessApproval} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAccessApproval
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1AccessApprovalGroup } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_accessor_group_request \link{OrgSagebionetworksRepoModelDataaccessAccessorGroupRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessAccessorGroupResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessAccessorGroupResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1AccessApprovalInformation } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_batch_access_approval_info_request \link{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1AccessApprovalNotifications } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_access_approval_notification_request \link{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1AccessApprovalSearch } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_access_approval_search_request \link{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1AccessApprovalGroupRevoke } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_accessor_group_revoke_request \link{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest}


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  deleteRepoV1AccessApproval  ####################

library(synr)
var_requirement_id <- "requirement_id_example" # character | 
var_accessor_id <- "accessor_id_example" # character | - The user whose access is being revoked

api_instance <- AccessApprovalServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1AccessApproval(var_requirement_id, var_accessor_id)


####################  getRepoV1AccessApprovalApprovalId  ####################

library(synr)
var_approval_id <- "approval_id_example" # character | 

api_instance <- AccessApprovalServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1AccessApprovalApprovalId(var_approval_iddata_file = "result.txt")
result <- api_instance$getRepoV1AccessApprovalApprovalId(var_approval_id)
dput(result)


####################  postRepoV1AccessApproval  ####################

library(synr)
var_org_sagebionetworks_repo_model_access_approval <- org.sagebionetworks.repo.model.AccessApproval$new(123, "etag_example", "createdOn_example", "modifiedOn_example", "createdBy_example", "modifiedBy_example", 123, 123, "submitterId_example", "accessorId_example", "expiredOn_example", "state_example") # OrgSagebionetworksRepoModelAccessApproval | 

api_instance <- AccessApprovalServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1AccessApproval(var_org_sagebionetworks_repo_model_access_approvaldata_file = "result.txt")
result <- api_instance$postRepoV1AccessApproval(var_org_sagebionetworks_repo_model_access_approval)
dput(result)


####################  postRepoV1AccessApprovalGroup  ####################

library(synr)
var_org_sagebionetworks_repo_model_dataaccess_accessor_group_request <- org.sagebionetworks.repo.model.dataaccess.AccessorGroupRequest$new("accessRequirementId_example", "submitterId_example", "accessorId_example", "expireBefore_example", "nextPageToken_example") # OrgSagebionetworksRepoModelDataaccessAccessorGroupRequest | 

api_instance <- AccessApprovalServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1AccessApprovalGroup(var_org_sagebionetworks_repo_model_dataaccess_accessor_group_requestdata_file = "result.txt")
result <- api_instance$postRepoV1AccessApprovalGroup(var_org_sagebionetworks_repo_model_dataaccess_accessor_group_request)
dput(result)


####################  postRepoV1AccessApprovalInformation  ####################

library(synr)
var_org_sagebionetworks_repo_model_batch_access_approval_info_request <- org.sagebionetworks.repo.model.BatchAccessApprovalInfoRequest$new("userId_example", c("accessRequirementIds_example")) # OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest | 

api_instance <- AccessApprovalServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1AccessApprovalInformation(var_org_sagebionetworks_repo_model_batch_access_approval_info_requestdata_file = "result.txt")
result <- api_instance$postRepoV1AccessApprovalInformation(var_org_sagebionetworks_repo_model_batch_access_approval_info_request)
dput(result)


####################  postRepoV1AccessApprovalNotifications  ####################

library(synr)
var_org_sagebionetworks_repo_model_dataaccess_access_approval_notification_request <- org.sagebionetworks.repo.model.dataaccess.AccessApprovalNotificationRequest$new(123, c(123)) # OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationRequest | 

api_instance <- AccessApprovalServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1AccessApprovalNotifications(var_org_sagebionetworks_repo_model_dataaccess_access_approval_notification_requestdata_file = "result.txt")
result <- api_instance$postRepoV1AccessApprovalNotifications(var_org_sagebionetworks_repo_model_dataaccess_access_approval_notification_request)
dput(result)


####################  postRepoV1AccessApprovalSearch  ####################

library(synr)
var_org_sagebionetworks_repo_model_dataaccess_access_approval_search_request <- org.sagebionetworks.repo.model.dataaccess.AccessApprovalSearchRequest$new("accessorId_example", "accessRequirementId_example", c(org.sagebionetworks.repo.model.dataaccess.AccessApprovalSearchSort$new("field_example", "direction_example")), "nextPageToken_example") # OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchRequest | 

api_instance <- AccessApprovalServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1AccessApprovalSearch(var_org_sagebionetworks_repo_model_dataaccess_access_approval_search_requestdata_file = "result.txt")
result <- api_instance$postRepoV1AccessApprovalSearch(var_org_sagebionetworks_repo_model_dataaccess_access_approval_search_request)
dput(result)


####################  putRepoV1AccessApprovalGroupRevoke  ####################

library(synr)
var_org_sagebionetworks_repo_model_dataaccess_accessor_group_revoke_request <- org.sagebionetworks.repo.model.dataaccess.AccessorGroupRevokeRequest$new("accessRequirementId_example", "submitterId_example") # OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest | 

api_instance <- AccessApprovalServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$putRepoV1AccessApprovalGroupRevoke(var_org_sagebionetworks_repo_model_dataaccess_accessor_group_revoke_request)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AccessApprovalServicesApi-new}{\code{AccessApprovalServicesApi$new()}}
\item \href{#method-AccessApprovalServicesApi-deleteRepoV1AccessApproval}{\code{AccessApprovalServicesApi$deleteRepoV1AccessApproval()}}
\item \href{#method-AccessApprovalServicesApi-deleteRepoV1AccessApprovalWithHttpInfo}{\code{AccessApprovalServicesApi$deleteRepoV1AccessApprovalWithHttpInfo()}}
\item \href{#method-AccessApprovalServicesApi-getRepoV1AccessApprovalApprovalId}{\code{AccessApprovalServicesApi$getRepoV1AccessApprovalApprovalId()}}
\item \href{#method-AccessApprovalServicesApi-getRepoV1AccessApprovalApprovalIdWithHttpInfo}{\code{AccessApprovalServicesApi$getRepoV1AccessApprovalApprovalIdWithHttpInfo()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApproval}{\code{AccessApprovalServicesApi$postRepoV1AccessApproval()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApprovalWithHttpInfo}{\code{AccessApprovalServicesApi$postRepoV1AccessApprovalWithHttpInfo()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApprovalGroup}{\code{AccessApprovalServicesApi$postRepoV1AccessApprovalGroup()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApprovalGroupWithHttpInfo}{\code{AccessApprovalServicesApi$postRepoV1AccessApprovalGroupWithHttpInfo()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApprovalInformation}{\code{AccessApprovalServicesApi$postRepoV1AccessApprovalInformation()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApprovalInformationWithHttpInfo}{\code{AccessApprovalServicesApi$postRepoV1AccessApprovalInformationWithHttpInfo()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApprovalNotifications}{\code{AccessApprovalServicesApi$postRepoV1AccessApprovalNotifications()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApprovalNotificationsWithHttpInfo}{\code{AccessApprovalServicesApi$postRepoV1AccessApprovalNotificationsWithHttpInfo()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApprovalSearch}{\code{AccessApprovalServicesApi$postRepoV1AccessApprovalSearch()}}
\item \href{#method-AccessApprovalServicesApi-postRepoV1AccessApprovalSearchWithHttpInfo}{\code{AccessApprovalServicesApi$postRepoV1AccessApprovalSearchWithHttpInfo()}}
\item \href{#method-AccessApprovalServicesApi-putRepoV1AccessApprovalGroupRevoke}{\code{AccessApprovalServicesApi$putRepoV1AccessApprovalGroupRevoke()}}
\item \href{#method-AccessApprovalServicesApi-putRepoV1AccessApprovalGroupRevokeWithHttpInfo}{\code{AccessApprovalServicesApi$putRepoV1AccessApprovalGroupRevokeWithHttpInfo()}}
\item \href{#method-AccessApprovalServicesApi-clone}{\code{AccessApprovalServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AccessApprovalServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-deleteRepoV1AccessApproval"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-deleteRepoV1AccessApproval}{}}}
\subsection{Method \code{deleteRepoV1AccessApproval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$deleteRepoV1AccessApproval(
  requirement_id,
  accessor_id,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{accessor_id}}{- The user whose access is being revoked}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-deleteRepoV1AccessApprovalWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-deleteRepoV1AccessApprovalWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1AccessApprovalWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$deleteRepoV1AccessApprovalWithHttpInfo(
  requirement_id,
  accessor_id,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{accessor_id}}{- The user whose access is being revoked}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-getRepoV1AccessApprovalApprovalId"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-getRepoV1AccessApprovalApprovalId}{}}}
\subsection{Method \code{getRepoV1AccessApprovalApprovalId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$getRepoV1AccessApprovalApprovalId(
  approval_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{approval_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAccessApproval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-getRepoV1AccessApprovalApprovalIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-getRepoV1AccessApprovalApprovalIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1AccessApprovalApprovalIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$getRepoV1AccessApprovalApprovalIdWithHttpInfo(
  approval_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{approval_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAccessApproval) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApproval"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApproval}{}}}
\subsection{Method \code{postRepoV1AccessApproval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApproval(
  org_sagebionetworks_repo_model_access_approval,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_access_approval}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAccessApproval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApprovalWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApprovalWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1AccessApprovalWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApprovalWithHttpInfo(
  org_sagebionetworks_repo_model_access_approval,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_access_approval}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAccessApproval) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApprovalGroup"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApprovalGroup}{}}}
\subsection{Method \code{postRepoV1AccessApprovalGroup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApprovalGroup(
  org_sagebionetworks_repo_model_dataaccess_accessor_group_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_accessor_group_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessorGroupResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApprovalGroupWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApprovalGroupWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1AccessApprovalGroupWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApprovalGroupWithHttpInfo(
  org_sagebionetworks_repo_model_dataaccess_accessor_group_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_accessor_group_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessAccessorGroupResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApprovalInformation"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApprovalInformation}{}}}
\subsection{Method \code{postRepoV1AccessApprovalInformation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApprovalInformation(
  org_sagebionetworks_repo_model_batch_access_approval_info_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_batch_access_approval_info_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApprovalInformationWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApprovalInformationWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1AccessApprovalInformationWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApprovalInformationWithHttpInfo(
  org_sagebionetworks_repo_model_batch_access_approval_info_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_batch_access_approval_info_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApprovalNotifications"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApprovalNotifications}{}}}
\subsection{Method \code{postRepoV1AccessApprovalNotifications()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApprovalNotifications(
  org_sagebionetworks_repo_model_dataaccess_access_approval_notification_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_access_approval_notification_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApprovalNotificationsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApprovalNotificationsWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1AccessApprovalNotificationsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApprovalNotificationsWithHttpInfo(
  org_sagebionetworks_repo_model_dataaccess_access_approval_notification_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_access_approval_notification_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApprovalSearch"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApprovalSearch}{}}}
\subsection{Method \code{postRepoV1AccessApprovalSearch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApprovalSearch(
  org_sagebionetworks_repo_model_dataaccess_access_approval_search_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_access_approval_search_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-postRepoV1AccessApprovalSearchWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-postRepoV1AccessApprovalSearchWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1AccessApprovalSearchWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$postRepoV1AccessApprovalSearchWithHttpInfo(
  org_sagebionetworks_repo_model_dataaccess_access_approval_search_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_access_approval_search_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-putRepoV1AccessApprovalGroupRevoke"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-putRepoV1AccessApprovalGroupRevoke}{}}}
\subsection{Method \code{putRepoV1AccessApprovalGroupRevoke()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$putRepoV1AccessApprovalGroupRevoke(
  org_sagebionetworks_repo_model_dataaccess_accessor_group_revoke_request,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_accessor_group_revoke_request}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-putRepoV1AccessApprovalGroupRevokeWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-putRepoV1AccessApprovalGroupRevokeWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1AccessApprovalGroupRevokeWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$putRepoV1AccessApprovalGroupRevokeWithHttpInfo(
  org_sagebionetworks_repo_model_dataaccess_accessor_group_revoke_request,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_accessor_group_revoke_request}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessApprovalServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AccessApprovalServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessApprovalServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
