% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_services_api.R
\docType{class}
\name{ActivityServicesApi}
\alias{ActivityServicesApi}
\title{ActivityServices operations}
\format{
An \code{R6Class} generator object
}
\description{
ActivityServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ deleteRepoV1ActivityId } \emph{  }

\itemize{
\item \emph{ @param } id character


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1ActivityId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelProvenanceActivity} \cr


\item status code : 200 | The requested Activity if it exists.

\item return type : OrgSagebionetworksRepoModelProvenanceActivity
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1ActivityIdGenerated } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResultsOfReference} \cr


\item status code : 200 | Auto-generated description

\item return type : PaginatedResultsOfReference
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1Activity } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_provenance_activity \link{OrgSagebionetworksRepoModelProvenanceActivity}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelProvenanceActivity} \cr


\item status code : 201 | The new ,<a href=\"${org.sagebionetworks.repo.model.provenance.Activity}\">,Activity,</a>

\item return type : OrgSagebionetworksRepoModelProvenanceActivity
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1ActivityId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } org_sagebionetworks_repo_model_provenance_activity \link{OrgSagebionetworksRepoModelProvenanceActivity}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelProvenanceActivity} \cr


\item status code : 200 | the updated activity

\item return type : OrgSagebionetworksRepoModelProvenanceActivity
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  deleteRepoV1ActivityId  ####################

library(synr)
var_id <- "id_example" # character | The id of activity to delete.

api_instance <- ActivityServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1ActivityId(var_id)


####################  getRepoV1ActivityId  ####################

library(synr)
var_id <- "id_example" # character | The ID of the activity to fetch.

api_instance <- ActivityServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1ActivityId(var_iddata_file = "result.txt")
result <- api_instance$getRepoV1ActivityId(var_id)
dput(result)


####################  getRepoV1ActivityIdGenerated  ####################

library(synr)
var_id <- "id_example" # character | 
var_offset <- 56 # integer |  (Optional)
var_limit <- 56 # integer |  (Optional)

api_instance <- ActivityServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1ActivityIdGenerated(var_id, offset = var_offset, limit = var_limitdata_file = "result.txt")
result <- api_instance$getRepoV1ActivityIdGenerated(var_id, offset = var_offset, limit = var_limit)
dput(result)


####################  postRepoV1Activity  ####################

library(synr)
var_org_sagebionetworks_repo_model_provenance_activity <- org.sagebionetworks.repo.model.provenance.Activity$new("id_example", "name_example", "description_example", "etag_example", "createdOn_example", "modifiedOn_example", "createdBy_example", "modifiedBy_example", c(org.sagebionetworks.repo.model.provenance.Used$new("concreteType_example", "wasExecuted_example", "name_example", "url_example", org.sagebionetworks.repo.model.Reference$new("targetId_example", 123)))) # OrgSagebionetworksRepoModelProvenanceActivity | 

api_instance <- ActivityServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1Activity(var_org_sagebionetworks_repo_model_provenance_activitydata_file = "result.txt")
result <- api_instance$postRepoV1Activity(var_org_sagebionetworks_repo_model_provenance_activity)
dput(result)


####################  putRepoV1ActivityId  ####################

library(synr)
var_id <- "id_example" # character | The id of the activity to update.
var_org_sagebionetworks_repo_model_provenance_activity <- org.sagebionetworks.repo.model.provenance.Activity$new("id_example", "name_example", "description_example", "etag_example", "createdOn_example", "modifiedOn_example", "createdBy_example", "modifiedBy_example", c(org.sagebionetworks.repo.model.provenance.Used$new("concreteType_example", "wasExecuted_example", "name_example", "url_example", org.sagebionetworks.repo.model.Reference$new("targetId_example", 123)))) # OrgSagebionetworksRepoModelProvenanceActivity | 

api_instance <- ActivityServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1ActivityId(var_id, var_org_sagebionetworks_repo_model_provenance_activitydata_file = "result.txt")
result <- api_instance$putRepoV1ActivityId(var_id, var_org_sagebionetworks_repo_model_provenance_activity)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ActivityServicesApi-new}{\code{ActivityServicesApi$new()}}
\item \href{#method-ActivityServicesApi-deleteRepoV1ActivityId}{\code{ActivityServicesApi$deleteRepoV1ActivityId()}}
\item \href{#method-ActivityServicesApi-deleteRepoV1ActivityIdWithHttpInfo}{\code{ActivityServicesApi$deleteRepoV1ActivityIdWithHttpInfo()}}
\item \href{#method-ActivityServicesApi-getRepoV1ActivityId}{\code{ActivityServicesApi$getRepoV1ActivityId()}}
\item \href{#method-ActivityServicesApi-getRepoV1ActivityIdWithHttpInfo}{\code{ActivityServicesApi$getRepoV1ActivityIdWithHttpInfo()}}
\item \href{#method-ActivityServicesApi-getRepoV1ActivityIdGenerated}{\code{ActivityServicesApi$getRepoV1ActivityIdGenerated()}}
\item \href{#method-ActivityServicesApi-getRepoV1ActivityIdGeneratedWithHttpInfo}{\code{ActivityServicesApi$getRepoV1ActivityIdGeneratedWithHttpInfo()}}
\item \href{#method-ActivityServicesApi-postRepoV1Activity}{\code{ActivityServicesApi$postRepoV1Activity()}}
\item \href{#method-ActivityServicesApi-postRepoV1ActivityWithHttpInfo}{\code{ActivityServicesApi$postRepoV1ActivityWithHttpInfo()}}
\item \href{#method-ActivityServicesApi-putRepoV1ActivityId}{\code{ActivityServicesApi$putRepoV1ActivityId()}}
\item \href{#method-ActivityServicesApi-putRepoV1ActivityIdWithHttpInfo}{\code{ActivityServicesApi$putRepoV1ActivityIdWithHttpInfo()}}
\item \href{#method-ActivityServicesApi-clone}{\code{ActivityServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ActivityServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-deleteRepoV1ActivityId"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-deleteRepoV1ActivityId}{}}}
\subsection{Method \code{deleteRepoV1ActivityId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$deleteRepoV1ActivityId(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of activity to delete.}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-deleteRepoV1ActivityIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-deleteRepoV1ActivityIdWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1ActivityIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$deleteRepoV1ActivityIdWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of activity to delete.}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-getRepoV1ActivityId"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-getRepoV1ActivityId}{}}}
\subsection{Method \code{getRepoV1ActivityId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$getRepoV1ActivityId(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The ID of the activity to fetch.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelProvenanceActivity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-getRepoV1ActivityIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-getRepoV1ActivityIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1ActivityIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$getRepoV1ActivityIdWithHttpInfo(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The ID of the activity to fetch.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelProvenanceActivity) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-getRepoV1ActivityIdGenerated"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-getRepoV1ActivityIdGenerated}{}}}
\subsection{Method \code{getRepoV1ActivityIdGenerated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$getRepoV1ActivityIdGenerated(
  id,
  offset = NULL,
  limit = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{offset}}{(optional) No description}

\item{\code{limit}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PaginatedResultsOfReference
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-getRepoV1ActivityIdGeneratedWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-getRepoV1ActivityIdGeneratedWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1ActivityIdGeneratedWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$getRepoV1ActivityIdGeneratedWithHttpInfo(
  id,
  offset = NULL,
  limit = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{offset}}{(optional) No description}

\item{\code{limit}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PaginatedResultsOfReference) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-postRepoV1Activity"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-postRepoV1Activity}{}}}
\subsection{Method \code{postRepoV1Activity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$postRepoV1Activity(
  org_sagebionetworks_repo_model_provenance_activity,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_provenance_activity}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelProvenanceActivity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-postRepoV1ActivityWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-postRepoV1ActivityWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1ActivityWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$postRepoV1ActivityWithHttpInfo(
  org_sagebionetworks_repo_model_provenance_activity,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_provenance_activity}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelProvenanceActivity) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-putRepoV1ActivityId"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-putRepoV1ActivityId}{}}}
\subsection{Method \code{putRepoV1ActivityId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$putRepoV1ActivityId(
  id,
  org_sagebionetworks_repo_model_provenance_activity,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the activity to update.}

\item{\code{org_sagebionetworks_repo_model_provenance_activity}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelProvenanceActivity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-putRepoV1ActivityIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-putRepoV1ActivityIdWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1ActivityIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$putRepoV1ActivityIdWithHttpInfo(
  id,
  org_sagebionetworks_repo_model_provenance_activity,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the activity to update.}

\item{\code{org_sagebionetworks_repo_model_provenance_activity}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelProvenanceActivity) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
