% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnyType.R
\docType{class}
\name{AnyType}
\alias{AnyType}
\title{AnyType}
\description{
AnyType object
}
\details{
API response object `AnyType`.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnyType-clone}{\code{AnyType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnyType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
