% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_services_api.R
\docType{class}
\name{AuthenticationServicesApi}
\alias{AuthenticationServicesApi}
\title{AuthenticationServices operations}
\format{
An \code{R6Class} generator object
}
\description{
AuthenticationServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ deleteAuthV1Oauth2Alias } \emph{  }

\itemize{
\item \emph{ @param } provider character
\item \emph{ @param } alias character


\item status code : 200 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ deleteAuthV1PersonalAccessTokenId } \emph{  }

\itemize{
\item \emph{ @param } id numeric


\item status code : 200 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ deleteAuthV1SecretKey } \emph{  }

\itemize{


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ getAuthV1AuthenticatedOn } \emph{  }

\itemize{
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAuthAuthenticatedOn} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAuthAuthenticatedOn
\item response headers :

\tabular{ll}{
}
}

\strong{ getAuthV1PersonalAccessToken } \emph{  }

\itemize{
\item \emph{ @param } next_page_token character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAuthAccessTokenRecordList} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAuthAccessTokenRecordList
\item response headers :

\tabular{ll}{
}
}

\strong{ getAuthV1PersonalAccessTokenId } \emph{  }

\itemize{
\item \emph{ @param } id numeric
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAuthAccessTokenRecord} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAuthAccessTokenRecord
\item response headers :

\tabular{ll}{
}
}

\strong{ getAuthV1SecretKey } \emph{  }

\itemize{
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAuthSecretKey} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAuthSecretKey
\item response headers :

\tabular{ll}{
}
}

\strong{ postAuthV1Login2 } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_auth_login_request \link{OrgSagebionetworksRepoModelAuthLoginRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAuthLoginResponse} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAuthLoginResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postAuthV1Oauth2Account2 } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_oauth_o_auth_account_creation_request \link{OrgSagebionetworksRepoModelOauthOAuthAccountCreationRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAuthLoginResponse} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAuthLoginResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postAuthV1Oauth2Alias } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_oauth_o_auth_validation_request \link{OrgSagebionetworksRepoModelOauthOAuthValidationRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelPrincipalPrincipalAlias} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelPrincipalPrincipalAlias
\item response headers :

\tabular{ll}{
}
}

\strong{ postAuthV1Oauth2Authurl } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_oauth_o_auth_url_request \link{OrgSagebionetworksRepoModelOauthOAuthUrlRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelOauthOAuthUrlResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelOauthOAuthUrlResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postAuthV1Oauth2Session2 } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_oauth_o_auth_validation_request \link{OrgSagebionetworksRepoModelOauthOAuthValidationRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAuthLoginResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAuthLoginResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postAuthV1PersonalAccessToken } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_auth_access_token_generation_request \link{OrgSagebionetworksRepoModelAuthAccessTokenGenerationRequest}
\item \emph{ @param } synapse_authorization character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAuthAccessTokenGenerationResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAuthAccessTokenGenerationResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postAuthV1TermsOfUse2 } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_auth_access_token \link{OrgSagebionetworksRepoModelAuthAccessToken}


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ postAuthV1UserChangePassword } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_auth_change_password_interface \link{OrgSagebionetworksRepoModelAuthChangePasswordInterface}


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ postAuthV1UserPasswordReset } \emph{  }

\itemize{
\item \emph{ @param } password_reset_endpoint character
\item \emph{ @param } org_sagebionetworks_repo_model_auth_username \link{OrgSagebionetworksRepoModelAuthUsername}


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  deleteAuthV1Oauth2Alias  ####################

library(synr)
var_provider <- "provider_example" # character | the OAuth provider through which the alias was associated
var_alias <- "alias_example" # character | the alias for the user given by the provider

api_instance <- AuthenticationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteAuthV1Oauth2Alias(var_provider, var_alias)


####################  deleteAuthV1PersonalAccessTokenId  ####################

library(synr)
var_id <- 3.4 # numeric | The unique ID of the token, which is the unique ID (the \"jti\" claim) contained in the JWT

api_instance <- AuthenticationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteAuthV1PersonalAccessTokenId(var_id)


####################  deleteAuthV1SecretKey  ####################

library(synr)

api_instance <- AuthenticationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteAuthV1SecretKey()


####################  getAuthV1AuthenticatedOn  ####################

library(synr)

api_instance <- AuthenticationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getAuthV1AuthenticatedOn(data_file = "result.txt")
result <- api_instance$getAuthV1AuthenticatedOn()
dput(result)


####################  getAuthV1PersonalAccessToken  ####################

library(synr)
var_next_page_token <- "next_page_token_example" # character |  (Optional)

api_instance <- AuthenticationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getAuthV1PersonalAccessToken(next_page_token = var_next_page_tokendata_file = "result.txt")
result <- api_instance$getAuthV1PersonalAccessToken(next_page_token = var_next_page_token)
dput(result)


####################  getAuthV1PersonalAccessTokenId  ####################

library(synr)
var_id <- 3.4 # numeric | The unique ID of the token, which is the unique ID (the \"jti\" claim) contained in the JWT

api_instance <- AuthenticationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getAuthV1PersonalAccessTokenId(var_iddata_file = "result.txt")
result <- api_instance$getAuthV1PersonalAccessTokenId(var_id)
dput(result)


####################  getAuthV1SecretKey  ####################

library(synr)

api_instance <- AuthenticationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getAuthV1SecretKey(data_file = "result.txt")
result <- api_instance$getAuthV1SecretKey()
dput(result)


####################  postAuthV1Login2  ####################

library(synr)
var_org_sagebionetworks_repo_model_auth_login_request <- org.sagebionetworks.repo.model.auth.LoginRequest$new("username_example", "password_example", "authenticationReceipt_example") # OrgSagebionetworksRepoModelAuthLoginRequest | 

api_instance <- AuthenticationServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postAuthV1Login2(var_org_sagebionetworks_repo_model_auth_login_requestdata_file = "result.txt")
result <- api_instance$postAuthV1Login2(var_org_sagebionetworks_repo_model_auth_login_request)
dput(result)


####################  postAuthV1Oauth2Account2  ####################

library(synr)
var_org_sagebionetworks_repo_model_oauth_o_auth_account_creation_request <- org.sagebionetworks.repo.model.oauth.OAuthAccountCreationRequest$new("provider_example", "authenticationCode_example", "redirectUrl_example", "userName_example") # OrgSagebionetworksRepoModelOauthOAuthAccountCreationRequest | 

api_instance <- AuthenticationServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postAuthV1Oauth2Account2(var_org_sagebionetworks_repo_model_oauth_o_auth_account_creation_requestdata_file = "result.txt")
result <- api_instance$postAuthV1Oauth2Account2(var_org_sagebionetworks_repo_model_oauth_o_auth_account_creation_request)
dput(result)


####################  postAuthV1Oauth2Alias  ####################

library(synr)
var_org_sagebionetworks_repo_model_oauth_o_auth_validation_request <- org.sagebionetworks.repo.model.oauth.OAuthValidationRequest$new("provider_example", "authenticationCode_example", "redirectUrl_example") # OrgSagebionetworksRepoModelOauthOAuthValidationRequest | 

api_instance <- AuthenticationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postAuthV1Oauth2Alias(var_org_sagebionetworks_repo_model_oauth_o_auth_validation_requestdata_file = "result.txt")
result <- api_instance$postAuthV1Oauth2Alias(var_org_sagebionetworks_repo_model_oauth_o_auth_validation_request)
dput(result)


####################  postAuthV1Oauth2Authurl  ####################

library(synr)
var_org_sagebionetworks_repo_model_oauth_o_auth_url_request <- org.sagebionetworks.repo.model.oauth.OAuthUrlRequest$new("provider_example", "redirectUrl_example", "state_example") # OrgSagebionetworksRepoModelOauthOAuthUrlRequest | 

api_instance <- AuthenticationServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postAuthV1Oauth2Authurl(var_org_sagebionetworks_repo_model_oauth_o_auth_url_requestdata_file = "result.txt")
result <- api_instance$postAuthV1Oauth2Authurl(var_org_sagebionetworks_repo_model_oauth_o_auth_url_request)
dput(result)


####################  postAuthV1Oauth2Session2  ####################

library(synr)
var_org_sagebionetworks_repo_model_oauth_o_auth_validation_request <- org.sagebionetworks.repo.model.oauth.OAuthValidationRequest$new("provider_example", "authenticationCode_example", "redirectUrl_example") # OrgSagebionetworksRepoModelOauthOAuthValidationRequest | 

api_instance <- AuthenticationServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postAuthV1Oauth2Session2(var_org_sagebionetworks_repo_model_oauth_o_auth_validation_requestdata_file = "result.txt")
result <- api_instance$postAuthV1Oauth2Session2(var_org_sagebionetworks_repo_model_oauth_o_auth_validation_request)
dput(result)


####################  postAuthV1PersonalAccessToken  ####################

library(synr)
var_org_sagebionetworks_repo_model_auth_access_token_generation_request <- org.sagebionetworks.repo.model.auth.AccessTokenGenerationRequest$new(c("scope_example"), c(key = org.sagebionetworks.repo.model.oauth.OIDCClaimsRequestDetails$new("essential_example", "value_example", c("values_example"))), "name_example") # OrgSagebionetworksRepoModelAuthAccessTokenGenerationRequest | 
var_synapse_authorization <- "synapse_authorization_example" # character |  (Optional)

api_instance <- AuthenticationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postAuthV1PersonalAccessToken(var_org_sagebionetworks_repo_model_auth_access_token_generation_request, synapse_authorization = var_synapse_authorizationdata_file = "result.txt")
result <- api_instance$postAuthV1PersonalAccessToken(var_org_sagebionetworks_repo_model_auth_access_token_generation_request, synapse_authorization = var_synapse_authorization)
dput(result)


####################  postAuthV1TermsOfUse2  ####################

library(synr)
var_org_sagebionetworks_repo_model_auth_access_token <- org.sagebionetworks.repo.model.auth.AccessToken$new("accessToken_example") # OrgSagebionetworksRepoModelAuthAccessToken | 

api_instance <- AuthenticationServicesApi$new()

api_instance$postAuthV1TermsOfUse2(var_org_sagebionetworks_repo_model_auth_access_token)


####################  postAuthV1UserChangePassword  ####################

library(synr)
var_org_sagebionetworks_repo_model_auth_change_password_interface <- org.sagebionetworks.repo.model.auth.ChangePasswordInterface$new("concreteType_example", "newPassword_example", org.sagebionetworks.repo.model.auth.PasswordResetSignedToken$new("org.sagebionetworks.repo.model.auth.PasswordResetSignedToken", "hmac_example", 123, "expiresOn_example", "createdOn_example", "userId_example", "validity_example"), "username_example", "currentPassword_example", "authenticationReceipt_example", 123, "twoFaToken_example", "otpCode_example", "otpType_example") # OrgSagebionetworksRepoModelAuthChangePasswordInterface | 

api_instance <- AuthenticationServicesApi$new()

api_instance$postAuthV1UserChangePassword(var_org_sagebionetworks_repo_model_auth_change_password_interface)


####################  postAuthV1UserPasswordReset  ####################

library(synr)
var_password_reset_endpoint <- "password_reset_endpoint_example" # character | the Portal's url prefix for handling password resets.
var_org_sagebionetworks_repo_model_auth_username <- org.sagebionetworks.repo.model.auth.Username$new("email_example") # OrgSagebionetworksRepoModelAuthUsername | 

api_instance <- AuthenticationServicesApi$new()

api_instance$postAuthV1UserPasswordReset(var_password_reset_endpoint, var_org_sagebionetworks_repo_model_auth_username)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthenticationServicesApi-new}{\code{AuthenticationServicesApi$new()}}
\item \href{#method-AuthenticationServicesApi-deleteAuthV1Oauth2Alias}{\code{AuthenticationServicesApi$deleteAuthV1Oauth2Alias()}}
\item \href{#method-AuthenticationServicesApi-deleteAuthV1Oauth2AliasWithHttpInfo}{\code{AuthenticationServicesApi$deleteAuthV1Oauth2AliasWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-deleteAuthV1PersonalAccessTokenId}{\code{AuthenticationServicesApi$deleteAuthV1PersonalAccessTokenId()}}
\item \href{#method-AuthenticationServicesApi-deleteAuthV1PersonalAccessTokenIdWithHttpInfo}{\code{AuthenticationServicesApi$deleteAuthV1PersonalAccessTokenIdWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-deleteAuthV1SecretKey}{\code{AuthenticationServicesApi$deleteAuthV1SecretKey()}}
\item \href{#method-AuthenticationServicesApi-deleteAuthV1SecretKeyWithHttpInfo}{\code{AuthenticationServicesApi$deleteAuthV1SecretKeyWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-getAuthV1AuthenticatedOn}{\code{AuthenticationServicesApi$getAuthV1AuthenticatedOn()}}
\item \href{#method-AuthenticationServicesApi-getAuthV1AuthenticatedOnWithHttpInfo}{\code{AuthenticationServicesApi$getAuthV1AuthenticatedOnWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-getAuthV1PersonalAccessToken}{\code{AuthenticationServicesApi$getAuthV1PersonalAccessToken()}}
\item \href{#method-AuthenticationServicesApi-getAuthV1PersonalAccessTokenWithHttpInfo}{\code{AuthenticationServicesApi$getAuthV1PersonalAccessTokenWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-getAuthV1PersonalAccessTokenId}{\code{AuthenticationServicesApi$getAuthV1PersonalAccessTokenId()}}
\item \href{#method-AuthenticationServicesApi-getAuthV1PersonalAccessTokenIdWithHttpInfo}{\code{AuthenticationServicesApi$getAuthV1PersonalAccessTokenIdWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-getAuthV1SecretKey}{\code{AuthenticationServicesApi$getAuthV1SecretKey()}}
\item \href{#method-AuthenticationServicesApi-getAuthV1SecretKeyWithHttpInfo}{\code{AuthenticationServicesApi$getAuthV1SecretKeyWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Login2}{\code{AuthenticationServicesApi$postAuthV1Login2()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Login2WithHttpInfo}{\code{AuthenticationServicesApi$postAuthV1Login2WithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Oauth2Account2}{\code{AuthenticationServicesApi$postAuthV1Oauth2Account2()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Oauth2Account2WithHttpInfo}{\code{AuthenticationServicesApi$postAuthV1Oauth2Account2WithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Oauth2Alias}{\code{AuthenticationServicesApi$postAuthV1Oauth2Alias()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Oauth2AliasWithHttpInfo}{\code{AuthenticationServicesApi$postAuthV1Oauth2AliasWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Oauth2Authurl}{\code{AuthenticationServicesApi$postAuthV1Oauth2Authurl()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Oauth2AuthurlWithHttpInfo}{\code{AuthenticationServicesApi$postAuthV1Oauth2AuthurlWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Oauth2Session2}{\code{AuthenticationServicesApi$postAuthV1Oauth2Session2()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1Oauth2Session2WithHttpInfo}{\code{AuthenticationServicesApi$postAuthV1Oauth2Session2WithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1PersonalAccessToken}{\code{AuthenticationServicesApi$postAuthV1PersonalAccessToken()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1PersonalAccessTokenWithHttpInfo}{\code{AuthenticationServicesApi$postAuthV1PersonalAccessTokenWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1TermsOfUse2}{\code{AuthenticationServicesApi$postAuthV1TermsOfUse2()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1TermsOfUse2WithHttpInfo}{\code{AuthenticationServicesApi$postAuthV1TermsOfUse2WithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1UserChangePassword}{\code{AuthenticationServicesApi$postAuthV1UserChangePassword()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1UserChangePasswordWithHttpInfo}{\code{AuthenticationServicesApi$postAuthV1UserChangePasswordWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1UserPasswordReset}{\code{AuthenticationServicesApi$postAuthV1UserPasswordReset()}}
\item \href{#method-AuthenticationServicesApi-postAuthV1UserPasswordResetWithHttpInfo}{\code{AuthenticationServicesApi$postAuthV1UserPasswordResetWithHttpInfo()}}
\item \href{#method-AuthenticationServicesApi-clone}{\code{AuthenticationServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AuthenticationServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-deleteAuthV1Oauth2Alias"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-deleteAuthV1Oauth2Alias}{}}}
\subsection{Method \code{deleteAuthV1Oauth2Alias()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$deleteAuthV1Oauth2Alias(provider, alias, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{provider}}{the OAuth provider through which the alias was associated}

\item{\code{alias}}{the alias for the user given by the provider}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-deleteAuthV1Oauth2AliasWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-deleteAuthV1Oauth2AliasWithHttpInfo}{}}}
\subsection{Method \code{deleteAuthV1Oauth2AliasWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$deleteAuthV1Oauth2AliasWithHttpInfo(
  provider,
  alias,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{provider}}{the OAuth provider through which the alias was associated}

\item{\code{alias}}{the alias for the user given by the provider}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-deleteAuthV1PersonalAccessTokenId"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-deleteAuthV1PersonalAccessTokenId}{}}}
\subsection{Method \code{deleteAuthV1PersonalAccessTokenId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$deleteAuthV1PersonalAccessTokenId(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The unique ID of the token, which is the unique ID (the \"jti\" claim) contained in the JWT}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-deleteAuthV1PersonalAccessTokenIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-deleteAuthV1PersonalAccessTokenIdWithHttpInfo}{}}}
\subsection{Method \code{deleteAuthV1PersonalAccessTokenIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$deleteAuthV1PersonalAccessTokenIdWithHttpInfo(
  id,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The unique ID of the token, which is the unique ID (the \"jti\" claim) contained in the JWT}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-deleteAuthV1SecretKey"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-deleteAuthV1SecretKey}{}}}
\subsection{Method \code{deleteAuthV1SecretKey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$deleteAuthV1SecretKey(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-deleteAuthV1SecretKeyWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-deleteAuthV1SecretKeyWithHttpInfo}{}}}
\subsection{Method \code{deleteAuthV1SecretKeyWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$deleteAuthV1SecretKeyWithHttpInfo(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-getAuthV1AuthenticatedOn"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-getAuthV1AuthenticatedOn}{}}}
\subsection{Method \code{getAuthV1AuthenticatedOn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$getAuthV1AuthenticatedOn(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAuthAuthenticatedOn
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-getAuthV1AuthenticatedOnWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-getAuthV1AuthenticatedOnWithHttpInfo}{}}}
\subsection{Method \code{getAuthV1AuthenticatedOnWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$getAuthV1AuthenticatedOnWithHttpInfo(
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAuthAuthenticatedOn) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-getAuthV1PersonalAccessToken"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-getAuthV1PersonalAccessToken}{}}}
\subsection{Method \code{getAuthV1PersonalAccessToken()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$getAuthV1PersonalAccessToken(
  next_page_token = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{next_page_token}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAuthAccessTokenRecordList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-getAuthV1PersonalAccessTokenWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-getAuthV1PersonalAccessTokenWithHttpInfo}{}}}
\subsection{Method \code{getAuthV1PersonalAccessTokenWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$getAuthV1PersonalAccessTokenWithHttpInfo(
  next_page_token = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{next_page_token}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAuthAccessTokenRecordList) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-getAuthV1PersonalAccessTokenId"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-getAuthV1PersonalAccessTokenId}{}}}
\subsection{Method \code{getAuthV1PersonalAccessTokenId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$getAuthV1PersonalAccessTokenId(
  id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The unique ID of the token, which is the unique ID (the \"jti\" claim) contained in the JWT}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAuthAccessTokenRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-getAuthV1PersonalAccessTokenIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-getAuthV1PersonalAccessTokenIdWithHttpInfo}{}}}
\subsection{Method \code{getAuthV1PersonalAccessTokenIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$getAuthV1PersonalAccessTokenIdWithHttpInfo(
  id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The unique ID of the token, which is the unique ID (the \"jti\" claim) contained in the JWT}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAuthAccessTokenRecord) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-getAuthV1SecretKey"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-getAuthV1SecretKey}{}}}
\subsection{Method \code{getAuthV1SecretKey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$getAuthV1SecretKey(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAuthSecretKey
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-getAuthV1SecretKeyWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-getAuthV1SecretKeyWithHttpInfo}{}}}
\subsection{Method \code{getAuthV1SecretKeyWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$getAuthV1SecretKeyWithHttpInfo(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAuthSecretKey) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Login2"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Login2}{}}}
\subsection{Method \code{postAuthV1Login2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Login2(
  org_sagebionetworks_repo_model_auth_login_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_auth_login_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAuthLoginResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Login2WithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Login2WithHttpInfo}{}}}
\subsection{Method \code{postAuthV1Login2WithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Login2WithHttpInfo(
  org_sagebionetworks_repo_model_auth_login_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_auth_login_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAuthLoginResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Oauth2Account2"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Oauth2Account2}{}}}
\subsection{Method \code{postAuthV1Oauth2Account2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Oauth2Account2(
  org_sagebionetworks_repo_model_oauth_o_auth_account_creation_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_oauth_o_auth_account_creation_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAuthLoginResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Oauth2Account2WithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Oauth2Account2WithHttpInfo}{}}}
\subsection{Method \code{postAuthV1Oauth2Account2WithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Oauth2Account2WithHttpInfo(
  org_sagebionetworks_repo_model_oauth_o_auth_account_creation_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_oauth_o_auth_account_creation_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAuthLoginResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Oauth2Alias"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Oauth2Alias}{}}}
\subsection{Method \code{postAuthV1Oauth2Alias()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Oauth2Alias(
  org_sagebionetworks_repo_model_oauth_o_auth_validation_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_oauth_o_auth_validation_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalPrincipalAlias
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Oauth2AliasWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Oauth2AliasWithHttpInfo}{}}}
\subsection{Method \code{postAuthV1Oauth2AliasWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Oauth2AliasWithHttpInfo(
  org_sagebionetworks_repo_model_oauth_o_auth_validation_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_oauth_o_auth_validation_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelPrincipalPrincipalAlias) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Oauth2Authurl"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Oauth2Authurl}{}}}
\subsection{Method \code{postAuthV1Oauth2Authurl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Oauth2Authurl(
  org_sagebionetworks_repo_model_oauth_o_auth_url_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_oauth_o_auth_url_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthUrlResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Oauth2AuthurlWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Oauth2AuthurlWithHttpInfo}{}}}
\subsection{Method \code{postAuthV1Oauth2AuthurlWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Oauth2AuthurlWithHttpInfo(
  org_sagebionetworks_repo_model_oauth_o_auth_url_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_oauth_o_auth_url_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelOauthOAuthUrlResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Oauth2Session2"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Oauth2Session2}{}}}
\subsection{Method \code{postAuthV1Oauth2Session2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Oauth2Session2(
  org_sagebionetworks_repo_model_oauth_o_auth_validation_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_oauth_o_auth_validation_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAuthLoginResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1Oauth2Session2WithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1Oauth2Session2WithHttpInfo}{}}}
\subsection{Method \code{postAuthV1Oauth2Session2WithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1Oauth2Session2WithHttpInfo(
  org_sagebionetworks_repo_model_oauth_o_auth_validation_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_oauth_o_auth_validation_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAuthLoginResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1PersonalAccessToken"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1PersonalAccessToken}{}}}
\subsection{Method \code{postAuthV1PersonalAccessToken()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1PersonalAccessToken(
  org_sagebionetworks_repo_model_auth_access_token_generation_request,
  synapse_authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_auth_access_token_generation_request}}{}

\item{\code{synapse_authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAuthAccessTokenGenerationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1PersonalAccessTokenWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1PersonalAccessTokenWithHttpInfo}{}}}
\subsection{Method \code{postAuthV1PersonalAccessTokenWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1PersonalAccessTokenWithHttpInfo(
  org_sagebionetworks_repo_model_auth_access_token_generation_request,
  synapse_authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_auth_access_token_generation_request}}{}

\item{\code{synapse_authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAuthAccessTokenGenerationResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1TermsOfUse2"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1TermsOfUse2}{}}}
\subsection{Method \code{postAuthV1TermsOfUse2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1TermsOfUse2(
  org_sagebionetworks_repo_model_auth_access_token,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_auth_access_token}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1TermsOfUse2WithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1TermsOfUse2WithHttpInfo}{}}}
\subsection{Method \code{postAuthV1TermsOfUse2WithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1TermsOfUse2WithHttpInfo(
  org_sagebionetworks_repo_model_auth_access_token,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_auth_access_token}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1UserChangePassword"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1UserChangePassword}{}}}
\subsection{Method \code{postAuthV1UserChangePassword()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1UserChangePassword(
  org_sagebionetworks_repo_model_auth_change_password_interface,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_auth_change_password_interface}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1UserChangePasswordWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1UserChangePasswordWithHttpInfo}{}}}
\subsection{Method \code{postAuthV1UserChangePasswordWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1UserChangePasswordWithHttpInfo(
  org_sagebionetworks_repo_model_auth_change_password_interface,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_auth_change_password_interface}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1UserPasswordReset"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1UserPasswordReset}{}}}
\subsection{Method \code{postAuthV1UserPasswordReset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1UserPasswordReset(
  password_reset_endpoint,
  org_sagebionetworks_repo_model_auth_username,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{password_reset_endpoint}}{the Portal's url prefix for handling password resets.}

\item{\code{org_sagebionetworks_repo_model_auth_username}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-postAuthV1UserPasswordResetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-postAuthV1UserPasswordResetWithHttpInfo}{}}}
\subsection{Method \code{postAuthV1UserPasswordResetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$postAuthV1UserPasswordResetWithHttpInfo(
  password_reset_endpoint,
  org_sagebionetworks_repo_model_auth_username,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{password_reset_endpoint}}{the Portal's url prefix for handling password resets.}

\item{\code{org_sagebionetworks_repo_model_auth_username}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
