% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_distance_from_centroid.R
\name{avg_distance_from_centroid}
\alias{avg_distance_from_centroid}
\title{Calculate average distance from centroid}
\usage{
avg_distance_from_centroid(point_matrix)
}
\arguments{
\item{point_matrix}{An n-by-3 numeric matrix where each
row corresponds to a single point in 3D space.}
}
\description{
Calculates average distance from centroid
of passed 3D points.
}
\seealso{
\code{\link[grDevices]{convertColor}}
}
