% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_services_api.R
\docType{class}
\name{DataAccessServicesApi}
\alias{DataAccessServicesApi}
\title{DataAccessServices operations}
\format{
An \code{R6Class} generator object
}
\description{
DataAccessServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ deleteRepoV1DataAccessSubmissionSubmissionId } \emph{  }

\itemize{
\item \emph{ @param } submission_id character


\item status code : 200 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate } \emph{  }

\itemize{
\item \emph{ @param } requirement_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessRequestInterface} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessRequestInterface
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate } \emph{  }

\itemize{
\item \emph{ @param } requirement_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessResearchProject} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessResearchProject
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1AccessRequirementRequirementIdStatus } \emph{  }

\itemize{
\item \emph{ @param } requirement_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessAccessRequirementStatus} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessAccessRequirementStatus
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1DataAccessSubmissionOpenSubmissions } \emph{  }

\itemize{
\item \emph{ @param } next_page_token character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1DataAccessSubmissionSubmissionId } \emph{  }

\itemize{
\item \emph{ @param } submission_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessSubmission} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessSubmission
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo } \emph{  }

\itemize{
\item \emph{ @param } requirement_id character
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_submission_info_page_request \link{OrgSagebionetworksRepoModelDataaccessSubmissionInfoPageRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessSubmissionInfoPage} \cr


\item status code : 200 | in order of modifiedOn, ascending

\item return type : OrgSagebionetworksRepoModelDataaccessSubmissionInfoPage
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1AccessRequirementRequirementIdSubmissions } \emph{  }

\itemize{
\item \emph{ @param } requirement_id character
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_submission_page_request \link{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessSubmissionPage} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessSubmissionPage
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1DataAccessRequest } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_request_interface \link{OrgSagebionetworksRepoModelDataaccessRequestInterface}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessRequestInterface} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessRequestInterface
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1DataAccessRequestRequestIdSubmission } \emph{  }

\itemize{
\item \emph{ @param } request_id character
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_create_submission_request \link{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessSubmissionStatus} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessSubmissionStatus
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1DataAccessSubmissionSearch } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_submission_search_request \link{OrgSagebionetworksRepoModelDataaccessSubmissionSearchRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessSubmissionSearchResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1ResearchProject } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_research_project \link{OrgSagebionetworksRepoModelDataaccessResearchProject}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessResearchProject} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessResearchProject
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1RestrictionInformation } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_restriction_information_request \link{OrgSagebionetworksRepoModelRestrictionInformationRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelRestrictionInformationResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelRestrictionInformationResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1DataAccessSubmissionSubmissionId } \emph{  }

\itemize{
\item \emph{ @param } submission_id character
\item \emph{ @param } org_sagebionetworks_repo_model_dataaccess_submission_state_change_request \link{OrgSagebionetworksRepoModelDataaccessSubmissionStateChangeRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessSubmission} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessSubmission
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1DataAccessSubmissionSubmissionIdCancellation } \emph{  }

\itemize{
\item \emph{ @param } submission_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDataaccessSubmissionStatus} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDataaccessSubmissionStatus
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  deleteRepoV1DataAccessSubmissionSubmissionId  ####################

library(synr)
var_submission_id <- "submission_id_example" # character | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1DataAccessSubmissionSubmissionId(var_submission_id)


####################  getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate  ####################

library(synr)
var_requirement_id <- "requirement_id_example" # character | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate(var_requirement_iddata_file = "result.txt")
result <- api_instance$getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate(var_requirement_id)
dput(result)


####################  getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate  ####################

library(synr)
var_requirement_id <- "requirement_id_example" # character | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate(var_requirement_iddata_file = "result.txt")
result <- api_instance$getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate(var_requirement_id)
dput(result)


####################  getRepoV1AccessRequirementRequirementIdStatus  ####################

library(synr)
var_requirement_id <- "requirement_id_example" # character | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1AccessRequirementRequirementIdStatus(var_requirement_iddata_file = "result.txt")
result <- api_instance$getRepoV1AccessRequirementRequirementIdStatus(var_requirement_id)
dput(result)


####################  getRepoV1DataAccessSubmissionOpenSubmissions  ####################

library(synr)
var_next_page_token <- "next_page_token_example" # character |  (Optional)

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1DataAccessSubmissionOpenSubmissions(next_page_token = var_next_page_tokendata_file = "result.txt")
result <- api_instance$getRepoV1DataAccessSubmissionOpenSubmissions(next_page_token = var_next_page_token)
dput(result)


####################  getRepoV1DataAccessSubmissionSubmissionId  ####################

library(synr)
var_submission_id <- "submission_id_example" # character | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1DataAccessSubmissionSubmissionId(var_submission_iddata_file = "result.txt")
result <- api_instance$getRepoV1DataAccessSubmissionSubmissionId(var_submission_id)
dput(result)


####################  postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo  ####################

library(synr)
var_requirement_id <- "requirement_id_example" # character | 
var_org_sagebionetworks_repo_model_dataaccess_submission_info_page_request <- org.sagebionetworks.repo.model.dataaccess.SubmissionInfoPageRequest$new("accessRequirementId_example", "nextPageToken_example") # OrgSagebionetworksRepoModelDataaccessSubmissionInfoPageRequest | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo(var_requirement_id, var_org_sagebionetworks_repo_model_dataaccess_submission_info_page_requestdata_file = "result.txt")
result <- api_instance$postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo(var_requirement_id, var_org_sagebionetworks_repo_model_dataaccess_submission_info_page_request)
dput(result)


####################  postRepoV1AccessRequirementRequirementIdSubmissions  ####################

library(synr)
var_requirement_id <- "requirement_id_example" # character | 
var_org_sagebionetworks_repo_model_dataaccess_submission_page_request <- org.sagebionetworks.repo.model.dataaccess.SubmissionPageRequest$new("accessRequirementId_example", "nextPageToken_example", "filterBy_example", "accessorId_example", "orderBy_example", "isAscending_example") # OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1AccessRequirementRequirementIdSubmissions(var_requirement_id, var_org_sagebionetworks_repo_model_dataaccess_submission_page_requestdata_file = "result.txt")
result <- api_instance$postRepoV1AccessRequirementRequirementIdSubmissions(var_requirement_id, var_org_sagebionetworks_repo_model_dataaccess_submission_page_request)
dput(result)


####################  postRepoV1DataAccessRequest  ####################

library(synr)
var_org_sagebionetworks_repo_model_dataaccess_request_interface <- org.sagebionetworks.repo.model.dataaccess.RequestInterface$new("concreteType_example", "id_example", "accessRequirementId_example", "researchProjectId_example", "createdOn_example", "modifiedOn_example", "createdBy_example", "modifiedBy_example", "ducFileHandleId_example", "irbFileHandleId_example", c("attachments_example"), c(org.sagebionetworks.repo.model.dataaccess.AccessorChange$new("userId_example", "type_example")), "etag_example", "publication_example", "summaryOfUse_example") # OrgSagebionetworksRepoModelDataaccessRequestInterface | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1DataAccessRequest(var_org_sagebionetworks_repo_model_dataaccess_request_interfacedata_file = "result.txt")
result <- api_instance$postRepoV1DataAccessRequest(var_org_sagebionetworks_repo_model_dataaccess_request_interface)
dput(result)


####################  postRepoV1DataAccessRequestRequestIdSubmission  ####################

library(synr)
var_request_id <- "request_id_example" # character | - The ID of the request object.
var_org_sagebionetworks_repo_model_dataaccess_create_submission_request <- org.sagebionetworks.repo.model.dataaccess.CreateSubmissionRequest$new("requestId_example", "requestEtag_example", "subjectId_example", "subjectType_example") # OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1DataAccessRequestRequestIdSubmission(var_request_id, var_org_sagebionetworks_repo_model_dataaccess_create_submission_requestdata_file = "result.txt")
result <- api_instance$postRepoV1DataAccessRequestRequestIdSubmission(var_request_id, var_org_sagebionetworks_repo_model_dataaccess_create_submission_request)
dput(result)


####################  postRepoV1DataAccessSubmissionSearch  ####################

library(synr)
var_org_sagebionetworks_repo_model_dataaccess_submission_search_request <- org.sagebionetworks.repo.model.dataaccess.SubmissionSearchRequest$new("accessorId_example", "accessRequirementId_example", "submissionState_example", "reviewerId_example", "reviewerFilterType_example", c(org.sagebionetworks.repo.model.dataaccess.SubmissionSearchSort$new("field_example", "direction_example")), "nextPageToken_example") # OrgSagebionetworksRepoModelDataaccessSubmissionSearchRequest | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1DataAccessSubmissionSearch(var_org_sagebionetworks_repo_model_dataaccess_submission_search_requestdata_file = "result.txt")
result <- api_instance$postRepoV1DataAccessSubmissionSearch(var_org_sagebionetworks_repo_model_dataaccess_submission_search_request)
dput(result)


####################  postRepoV1ResearchProject  ####################

library(synr)
var_org_sagebionetworks_repo_model_dataaccess_research_project <- org.sagebionetworks.repo.model.dataaccess.ResearchProject$new("id_example", "accessRequirementId_example", "institution_example", "projectLead_example", "intendedDataUseStatement_example", "createdOn_example", "modifiedOn_example", "createdBy_example", "modifiedBy_example", "etag_example") # OrgSagebionetworksRepoModelDataaccessResearchProject | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1ResearchProject(var_org_sagebionetworks_repo_model_dataaccess_research_projectdata_file = "result.txt")
result <- api_instance$postRepoV1ResearchProject(var_org_sagebionetworks_repo_model_dataaccess_research_project)
dput(result)


####################  postRepoV1RestrictionInformation  ####################

library(synr)
var_org_sagebionetworks_repo_model_restriction_information_request <- org.sagebionetworks.repo.model.RestrictionInformationRequest$new("restrictableObjectType_example", "objectId_example") # OrgSagebionetworksRepoModelRestrictionInformationRequest | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1RestrictionInformation(var_org_sagebionetworks_repo_model_restriction_information_requestdata_file = "result.txt")
result <- api_instance$postRepoV1RestrictionInformation(var_org_sagebionetworks_repo_model_restriction_information_request)
dput(result)


####################  putRepoV1DataAccessSubmissionSubmissionId  ####################

library(synr)
var_submission_id <- "submission_id_example" # character | 
var_org_sagebionetworks_repo_model_dataaccess_submission_state_change_request <- org.sagebionetworks.repo.model.dataaccess.SubmissionStateChangeRequest$new("submissionId_example", "newState_example", "rejectedReason_example") # OrgSagebionetworksRepoModelDataaccessSubmissionStateChangeRequest | 

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1DataAccessSubmissionSubmissionId(var_submission_id, var_org_sagebionetworks_repo_model_dataaccess_submission_state_change_requestdata_file = "result.txt")
result <- api_instance$putRepoV1DataAccessSubmissionSubmissionId(var_submission_id, var_org_sagebionetworks_repo_model_dataaccess_submission_state_change_request)
dput(result)


####################  putRepoV1DataAccessSubmissionSubmissionIdCancellation  ####################

library(synr)
var_submission_id <- "submission_id_example" # character | - The ID of the submission to cancel.

api_instance <- DataAccessServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1DataAccessSubmissionSubmissionIdCancellation(var_submission_iddata_file = "result.txt")
result <- api_instance$putRepoV1DataAccessSubmissionSubmissionIdCancellation(var_submission_id)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataAccessServicesApi-new}{\code{DataAccessServicesApi$new()}}
\item \href{#method-DataAccessServicesApi-deleteRepoV1DataAccessSubmissionSubmissionId}{\code{DataAccessServicesApi$deleteRepoV1DataAccessSubmissionSubmissionId()}}
\item \href{#method-DataAccessServicesApi-deleteRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo}{\code{DataAccessServicesApi$deleteRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate}{\code{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate()}}
\item \href{#method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdateWithHttpInfo}{\code{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdateWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate}{\code{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate()}}
\item \href{#method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdResearchProjectForUpdateWithHttpInfo}{\code{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdResearchProjectForUpdateWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdStatus}{\code{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdStatus()}}
\item \href{#method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdStatusWithHttpInfo}{\code{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdStatusWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-getRepoV1DataAccessSubmissionOpenSubmissions}{\code{DataAccessServicesApi$getRepoV1DataAccessSubmissionOpenSubmissions()}}
\item \href{#method-DataAccessServicesApi-getRepoV1DataAccessSubmissionOpenSubmissionsWithHttpInfo}{\code{DataAccessServicesApi$getRepoV1DataAccessSubmissionOpenSubmissionsWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-getRepoV1DataAccessSubmissionSubmissionId}{\code{DataAccessServicesApi$getRepoV1DataAccessSubmissionSubmissionId()}}
\item \href{#method-DataAccessServicesApi-getRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo}{\code{DataAccessServicesApi$getRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo}{\code{DataAccessServicesApi$postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo()}}
\item \href{#method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfoWithHttpInfo}{\code{DataAccessServicesApi$postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfoWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdSubmissions}{\code{DataAccessServicesApi$postRepoV1AccessRequirementRequirementIdSubmissions()}}
\item \href{#method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdSubmissionsWithHttpInfo}{\code{DataAccessServicesApi$postRepoV1AccessRequirementRequirementIdSubmissionsWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-postRepoV1DataAccessRequest}{\code{DataAccessServicesApi$postRepoV1DataAccessRequest()}}
\item \href{#method-DataAccessServicesApi-postRepoV1DataAccessRequestWithHttpInfo}{\code{DataAccessServicesApi$postRepoV1DataAccessRequestWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-postRepoV1DataAccessRequestRequestIdSubmission}{\code{DataAccessServicesApi$postRepoV1DataAccessRequestRequestIdSubmission()}}
\item \href{#method-DataAccessServicesApi-postRepoV1DataAccessRequestRequestIdSubmissionWithHttpInfo}{\code{DataAccessServicesApi$postRepoV1DataAccessRequestRequestIdSubmissionWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-postRepoV1DataAccessSubmissionSearch}{\code{DataAccessServicesApi$postRepoV1DataAccessSubmissionSearch()}}
\item \href{#method-DataAccessServicesApi-postRepoV1DataAccessSubmissionSearchWithHttpInfo}{\code{DataAccessServicesApi$postRepoV1DataAccessSubmissionSearchWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-postRepoV1ResearchProject}{\code{DataAccessServicesApi$postRepoV1ResearchProject()}}
\item \href{#method-DataAccessServicesApi-postRepoV1ResearchProjectWithHttpInfo}{\code{DataAccessServicesApi$postRepoV1ResearchProjectWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-postRepoV1RestrictionInformation}{\code{DataAccessServicesApi$postRepoV1RestrictionInformation()}}
\item \href{#method-DataAccessServicesApi-postRepoV1RestrictionInformationWithHttpInfo}{\code{DataAccessServicesApi$postRepoV1RestrictionInformationWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionId}{\code{DataAccessServicesApi$putRepoV1DataAccessSubmissionSubmissionId()}}
\item \href{#method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo}{\code{DataAccessServicesApi$putRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionIdCancellation}{\code{DataAccessServicesApi$putRepoV1DataAccessSubmissionSubmissionIdCancellation()}}
\item \href{#method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionIdCancellationWithHttpInfo}{\code{DataAccessServicesApi$putRepoV1DataAccessSubmissionSubmissionIdCancellationWithHttpInfo()}}
\item \href{#method-DataAccessServicesApi-clone}{\code{DataAccessServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DataAccessServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-deleteRepoV1DataAccessSubmissionSubmissionId"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-deleteRepoV1DataAccessSubmissionSubmissionId}{}}}
\subsection{Method \code{deleteRepoV1DataAccessSubmissionSubmissionId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$deleteRepoV1DataAccessSubmissionSubmissionId(
  submission_id,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_id}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-deleteRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-deleteRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$deleteRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo(
  submission_id,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_id}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate}{}}}
\subsection{Method \code{getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdate(
  requirement_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessRequestInterface
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdateWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdateWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdateWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdDataAccessRequestForUpdateWithHttpInfo(
  requirement_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessRequestInterface) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate}{}}}
\subsection{Method \code{getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdResearchProjectForUpdate(
  requirement_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessResearchProject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdResearchProjectForUpdateWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdResearchProjectForUpdateWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1AccessRequirementRequirementIdResearchProjectForUpdateWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdResearchProjectForUpdateWithHttpInfo(
  requirement_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessResearchProject) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdStatus"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdStatus}{}}}
\subsection{Method \code{getRepoV1AccessRequirementRequirementIdStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdStatus(
  requirement_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessRequirementStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdStatusWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1AccessRequirementRequirementIdStatusWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1AccessRequirementRequirementIdStatusWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1AccessRequirementRequirementIdStatusWithHttpInfo(
  requirement_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessAccessRequirementStatus) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1DataAccessSubmissionOpenSubmissions"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1DataAccessSubmissionOpenSubmissions}{}}}
\subsection{Method \code{getRepoV1DataAccessSubmissionOpenSubmissions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1DataAccessSubmissionOpenSubmissions(
  next_page_token = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{next_page_token}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1DataAccessSubmissionOpenSubmissionsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1DataAccessSubmissionOpenSubmissionsWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1DataAccessSubmissionOpenSubmissionsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1DataAccessSubmissionOpenSubmissionsWithHttpInfo(
  next_page_token = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{next_page_token}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1DataAccessSubmissionSubmissionId"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1DataAccessSubmissionSubmissionId}{}}}
\subsection{Method \code{getRepoV1DataAccessSubmissionSubmissionId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1DataAccessSubmissionSubmissionId(
  submission_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-getRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-getRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$getRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo(
  submission_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessSubmission) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo}{}}}
\subsection{Method \code{postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfo(
  requirement_id,
  org_sagebionetworks_repo_model_dataaccess_submission_info_page_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{org_sagebionetworks_repo_model_dataaccess_submission_info_page_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionInfoPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfoWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfoWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfoWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1AccessRequirementRequirementIdApprovedSubmissionInfoWithHttpInfo(
  requirement_id,
  org_sagebionetworks_repo_model_dataaccess_submission_info_page_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{org_sagebionetworks_repo_model_dataaccess_submission_info_page_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessSubmissionInfoPage) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdSubmissions"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdSubmissions}{}}}
\subsection{Method \code{postRepoV1AccessRequirementRequirementIdSubmissions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1AccessRequirementRequirementIdSubmissions(
  requirement_id,
  org_sagebionetworks_repo_model_dataaccess_submission_page_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{org_sagebionetworks_repo_model_dataaccess_submission_page_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdSubmissionsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1AccessRequirementRequirementIdSubmissionsWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1AccessRequirementRequirementIdSubmissionsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1AccessRequirementRequirementIdSubmissionsWithHttpInfo(
  requirement_id,
  org_sagebionetworks_repo_model_dataaccess_submission_page_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirement_id}}{}

\item{\code{org_sagebionetworks_repo_model_dataaccess_submission_page_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessSubmissionPage) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1DataAccessRequest"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1DataAccessRequest}{}}}
\subsection{Method \code{postRepoV1DataAccessRequest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1DataAccessRequest(
  org_sagebionetworks_repo_model_dataaccess_request_interface,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_request_interface}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessRequestInterface
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1DataAccessRequestWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1DataAccessRequestWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1DataAccessRequestWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1DataAccessRequestWithHttpInfo(
  org_sagebionetworks_repo_model_dataaccess_request_interface,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_request_interface}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessRequestInterface) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1DataAccessRequestRequestIdSubmission"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1DataAccessRequestRequestIdSubmission}{}}}
\subsection{Method \code{postRepoV1DataAccessRequestRequestIdSubmission()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1DataAccessRequestRequestIdSubmission(
  request_id,
  org_sagebionetworks_repo_model_dataaccess_create_submission_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request_id}}{- The ID of the request object.}

\item{\code{org_sagebionetworks_repo_model_dataaccess_create_submission_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1DataAccessRequestRequestIdSubmissionWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1DataAccessRequestRequestIdSubmissionWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1DataAccessRequestRequestIdSubmissionWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1DataAccessRequestRequestIdSubmissionWithHttpInfo(
  request_id,
  org_sagebionetworks_repo_model_dataaccess_create_submission_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request_id}}{- The ID of the request object.}

\item{\code{org_sagebionetworks_repo_model_dataaccess_create_submission_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessSubmissionStatus) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1DataAccessSubmissionSearch"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1DataAccessSubmissionSearch}{}}}
\subsection{Method \code{postRepoV1DataAccessSubmissionSearch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1DataAccessSubmissionSearch(
  org_sagebionetworks_repo_model_dataaccess_submission_search_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_submission_search_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1DataAccessSubmissionSearchWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1DataAccessSubmissionSearchWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1DataAccessSubmissionSearchWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1DataAccessSubmissionSearchWithHttpInfo(
  org_sagebionetworks_repo_model_dataaccess_submission_search_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_submission_search_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessSubmissionSearchResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1ResearchProject"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1ResearchProject}{}}}
\subsection{Method \code{postRepoV1ResearchProject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1ResearchProject(
  org_sagebionetworks_repo_model_dataaccess_research_project,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_research_project}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessResearchProject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1ResearchProjectWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1ResearchProjectWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1ResearchProjectWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1ResearchProjectWithHttpInfo(
  org_sagebionetworks_repo_model_dataaccess_research_project,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_dataaccess_research_project}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessResearchProject) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1RestrictionInformation"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1RestrictionInformation}{}}}
\subsection{Method \code{postRepoV1RestrictionInformation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1RestrictionInformation(
  org_sagebionetworks_repo_model_restriction_information_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_restriction_information_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictionInformationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-postRepoV1RestrictionInformationWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-postRepoV1RestrictionInformationWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1RestrictionInformationWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$postRepoV1RestrictionInformationWithHttpInfo(
  org_sagebionetworks_repo_model_restriction_information_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_restriction_information_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelRestrictionInformationResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionId"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionId}{}}}
\subsection{Method \code{putRepoV1DataAccessSubmissionSubmissionId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$putRepoV1DataAccessSubmissionSubmissionId(
  submission_id,
  org_sagebionetworks_repo_model_dataaccess_submission_state_change_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_id}}{}

\item{\code{org_sagebionetworks_repo_model_dataaccess_submission_state_change_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$putRepoV1DataAccessSubmissionSubmissionIdWithHttpInfo(
  submission_id,
  org_sagebionetworks_repo_model_dataaccess_submission_state_change_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_id}}{}

\item{\code{org_sagebionetworks_repo_model_dataaccess_submission_state_change_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessSubmission) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionIdCancellation"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionIdCancellation}{}}}
\subsection{Method \code{putRepoV1DataAccessSubmissionSubmissionIdCancellation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$putRepoV1DataAccessSubmissionSubmissionIdCancellation(
  submission_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_id}}{- The ID of the submission to cancel.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionIdCancellationWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-putRepoV1DataAccessSubmissionSubmissionIdCancellationWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1DataAccessSubmissionSubmissionIdCancellationWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$putRepoV1DataAccessSubmissionSubmissionIdCancellationWithHttpInfo(
  submission_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_id}}{- The ID of the submission to cancel.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDataaccessSubmissionStatus) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAccessServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataAccessServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAccessServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
