% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_authorization_services_api.R
\docType{class}
\name{DockerAuthorizationServicesApi}
\alias{DockerAuthorizationServicesApi}
\title{DockerAuthorizationServices operations}
\format{
An \code{R6Class} generator object
}
\description{
DockerAuthorizationServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ getDockerV1BearerToken } \emph{  }

\itemize{
\item \emph{ @param } service character
\item \emph{ @param } synapse_authorization character
\item \emph{ @param } scope list( character )
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelDockerDockerAuthorizationToken
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  getDockerV1BearerToken  ####################

library(synr)
var_service <- "service_example" # character | 
var_synapse_authorization <- "synapse_authorization_example" # character |  (Optional)
var_scope <- c("inner_example") # array[character] |  (Optional)

api_instance <- DockerAuthorizationServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getDockerV1BearerToken(var_service, synapse_authorization = var_synapse_authorization, scope = var_scopedata_file = "result.txt")
result <- api_instance$getDockerV1BearerToken(var_service, synapse_authorization = var_synapse_authorization, scope = var_scope)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DockerAuthorizationServicesApi-new}{\code{DockerAuthorizationServicesApi$new()}}
\item \href{#method-DockerAuthorizationServicesApi-getDockerV1BearerToken}{\code{DockerAuthorizationServicesApi$getDockerV1BearerToken()}}
\item \href{#method-DockerAuthorizationServicesApi-getDockerV1BearerTokenWithHttpInfo}{\code{DockerAuthorizationServicesApi$getDockerV1BearerTokenWithHttpInfo()}}
\item \href{#method-DockerAuthorizationServicesApi-clone}{\code{DockerAuthorizationServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerAuthorizationServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DockerAuthorizationServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DockerAuthorizationServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerAuthorizationServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerAuthorizationServicesApi-getDockerV1BearerToken"></a>}}
\if{latex}{\out{\hypertarget{method-DockerAuthorizationServicesApi-getDockerV1BearerToken}{}}}
\subsection{Method \code{getDockerV1BearerToken()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerAuthorizationServicesApi$getDockerV1BearerToken(
  service,
  synapse_authorization = NULL,
  scope = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service}}{}

\item{\code{synapse_authorization}}{(optional) No description}

\item{\code{scope}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDockerDockerAuthorizationToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerAuthorizationServicesApi-getDockerV1BearerTokenWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DockerAuthorizationServicesApi-getDockerV1BearerTokenWithHttpInfo}{}}}
\subsection{Method \code{getDockerV1BearerTokenWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerAuthorizationServicesApi$getDockerV1BearerTokenWithHttpInfo(
  service,
  synapse_authorization = NULL,
  scope = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service}}{}

\item{\code{synapse_authorization}}{(optional) No description}

\item{\code{scope}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDockerDockerAuthorizationToken) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerAuthorizationServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DockerAuthorizationServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerAuthorizationServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
