% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_commit_services_api.R
\docType{class}
\name{DockerCommitServicesApi}
\alias{DockerCommitServicesApi}
\title{DockerCommitServices operations}
\format{
An \code{R6Class} generator object
}
\description{
DockerCommitServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ getRepoV1EntityIdDockerTag } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } sort character
\item \emph{ @param } ascending character
\item \emph{ @param } limit numeric
\item \emph{ @param } offset numeric
\item \emph{ @returnType } \link{PaginatedResultsOfDockerCommit} \cr


\item status code : 200 | Auto-generated description

\item return type : PaginatedResultsOfDockerCommit
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1EntityIdDockerCommit } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } org_sagebionetworks_repo_model_docker_docker_commit \link{OrgSagebionetworksRepoModelDockerDockerCommit}


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  getRepoV1EntityIdDockerTag  ####################

library(synr)
var_id <- "id_example" # character | 
var_sort <- "sort_example" # character |  (Optional)
var_ascending <- "ascending_example" # character | , optional (default is false) (Optional)
var_limit <- 3.4 # numeric | pagination parameter, optional (default is 20) (Optional)
var_offset <- 3.4 # numeric | pagination parameter, optional (default is 0) (Optional)

api_instance <- DockerCommitServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1EntityIdDockerTag(var_id, sort = var_sort, ascending = var_ascending, limit = var_limit, offset = var_offsetdata_file = "result.txt")
result <- api_instance$getRepoV1EntityIdDockerTag(var_id, sort = var_sort, ascending = var_ascending, limit = var_limit, offset = var_offset)
dput(result)


####################  postRepoV1EntityIdDockerCommit  ####################

library(synr)
var_id <- "id_example" # character | 
var_org_sagebionetworks_repo_model_docker_docker_commit <- org.sagebionetworks.repo.model.docker.DockerCommit$new("tag_example", "digest_example", "createdOn_example") # OrgSagebionetworksRepoModelDockerDockerCommit | 

api_instance <- DockerCommitServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$postRepoV1EntityIdDockerCommit(var_id, var_org_sagebionetworks_repo_model_docker_docker_commit)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DockerCommitServicesApi-new}{\code{DockerCommitServicesApi$new()}}
\item \href{#method-DockerCommitServicesApi-getRepoV1EntityIdDockerTag}{\code{DockerCommitServicesApi$getRepoV1EntityIdDockerTag()}}
\item \href{#method-DockerCommitServicesApi-getRepoV1EntityIdDockerTagWithHttpInfo}{\code{DockerCommitServicesApi$getRepoV1EntityIdDockerTagWithHttpInfo()}}
\item \href{#method-DockerCommitServicesApi-postRepoV1EntityIdDockerCommit}{\code{DockerCommitServicesApi$postRepoV1EntityIdDockerCommit()}}
\item \href{#method-DockerCommitServicesApi-postRepoV1EntityIdDockerCommitWithHttpInfo}{\code{DockerCommitServicesApi$postRepoV1EntityIdDockerCommitWithHttpInfo()}}
\item \href{#method-DockerCommitServicesApi-clone}{\code{DockerCommitServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerCommitServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DockerCommitServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DockerCommitServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCommitServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerCommitServicesApi-getRepoV1EntityIdDockerTag"></a>}}
\if{latex}{\out{\hypertarget{method-DockerCommitServicesApi-getRepoV1EntityIdDockerTag}{}}}
\subsection{Method \code{getRepoV1EntityIdDockerTag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCommitServicesApi$getRepoV1EntityIdDockerTag(
  id,
  sort = NULL,
  ascending = NULL,
  limit = NULL,
  offset = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{sort}}{(optional) No description}

\item{\code{ascending}}{(optional) , optional (default is false)}

\item{\code{limit}}{(optional) pagination parameter, optional (default is 20)}

\item{\code{offset}}{(optional) pagination parameter, optional (default is 0)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PaginatedResultsOfDockerCommit
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerCommitServicesApi-getRepoV1EntityIdDockerTagWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DockerCommitServicesApi-getRepoV1EntityIdDockerTagWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1EntityIdDockerTagWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCommitServicesApi$getRepoV1EntityIdDockerTagWithHttpInfo(
  id,
  sort = NULL,
  ascending = NULL,
  limit = NULL,
  offset = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{sort}}{(optional) No description}

\item{\code{ascending}}{(optional) , optional (default is false)}

\item{\code{limit}}{(optional) pagination parameter, optional (default is 20)}

\item{\code{offset}}{(optional) pagination parameter, optional (default is 0)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PaginatedResultsOfDockerCommit) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerCommitServicesApi-postRepoV1EntityIdDockerCommit"></a>}}
\if{latex}{\out{\hypertarget{method-DockerCommitServicesApi-postRepoV1EntityIdDockerCommit}{}}}
\subsection{Method \code{postRepoV1EntityIdDockerCommit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCommitServicesApi$postRepoV1EntityIdDockerCommit(
  id,
  org_sagebionetworks_repo_model_docker_docker_commit,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{org_sagebionetworks_repo_model_docker_docker_commit}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerCommitServicesApi-postRepoV1EntityIdDockerCommitWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DockerCommitServicesApi-postRepoV1EntityIdDockerCommitWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1EntityIdDockerCommitWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCommitServicesApi$postRepoV1EntityIdDockerCommitWithHttpInfo(
  id,
  org_sagebionetworks_repo_model_docker_docker_commit,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{org_sagebionetworks_repo_model_docker_docker_commit}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerCommitServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DockerCommitServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerCommitServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
