% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_registry_event_services_api.R
\docType{class}
\name{DockerRegistryEventServicesApi}
\alias{DockerRegistryEventServicesApi}
\title{DockerRegistryEventServices operations}
\format{
An \code{R6Class} generator object
}
\description{
DockerRegistryEventServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ postDockerRegistryListenerV1Events } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_docker_docker_registry_event_list \link{OrgSagebionetworksRepoModelDockerDockerRegistryEventList}


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  postDockerRegistryListenerV1Events  ####################

library(synr)
var_org_sagebionetworks_repo_model_docker_docker_registry_event_list <- org.sagebionetworks.repo.model.docker.DockerRegistryEventList$new(c(org.sagebionetworks.repo.model.docker.DockerRegistryEvent$new("id_example", "timestamp_example", "action_example", org.sagebionetworks.repo.model.docker.RegistryEventTarget$new("mediaType_example", 123, "digest_example", 123, "repository_example", "tag_example", "url_example"), org.sagebionetworks.repo.model.docker.RegistryEventRequest$new("id_example", "addr_example", "host_example", "method_example", "useragent_example"), org.sagebionetworks.repo.model.docker.RegistryEventActor$new("name_example"), org.sagebionetworks.repo.model.docker.RegistryEventSource$new("addr_example", "instanceID_example")))) # OrgSagebionetworksRepoModelDockerDockerRegistryEventList | 

api_instance <- DockerRegistryEventServicesApi$new()

api_instance$postDockerRegistryListenerV1Events(var_org_sagebionetworks_repo_model_docker_docker_registry_event_list)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DockerRegistryEventServicesApi-new}{\code{DockerRegistryEventServicesApi$new()}}
\item \href{#method-DockerRegistryEventServicesApi-postDockerRegistryListenerV1Events}{\code{DockerRegistryEventServicesApi$postDockerRegistryListenerV1Events()}}
\item \href{#method-DockerRegistryEventServicesApi-postDockerRegistryListenerV1EventsWithHttpInfo}{\code{DockerRegistryEventServicesApi$postDockerRegistryListenerV1EventsWithHttpInfo()}}
\item \href{#method-DockerRegistryEventServicesApi-clone}{\code{DockerRegistryEventServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerRegistryEventServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DockerRegistryEventServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DockerRegistryEventServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerRegistryEventServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerRegistryEventServicesApi-postDockerRegistryListenerV1Events"></a>}}
\if{latex}{\out{\hypertarget{method-DockerRegistryEventServicesApi-postDockerRegistryListenerV1Events}{}}}
\subsection{Method \code{postDockerRegistryListenerV1Events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerRegistryEventServicesApi$postDockerRegistryListenerV1Events(
  org_sagebionetworks_repo_model_docker_docker_registry_event_list,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_docker_docker_registry_event_list}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerRegistryEventServicesApi-postDockerRegistryListenerV1EventsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DockerRegistryEventServicesApi-postDockerRegistryListenerV1EventsWithHttpInfo}{}}}
\subsection{Method \code{postDockerRegistryListenerV1EventsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerRegistryEventServicesApi$postDockerRegistryListenerV1EventsWithHttpInfo(
  org_sagebionetworks_repo_model_docker_docker_registry_event_list,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_docker_docker_registry_event_list}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DockerRegistryEventServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DockerRegistryEventServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DockerRegistryEventServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
