% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_services_api.R
\docType{class}
\name{DOIServicesApi}
\alias{DOIServicesApi}
\title{DOIServices operations}
\format{
An \code{R6Class} generator object
}
\description{
DOIServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ getRepoV1Doi } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } type \link{OrgSagebionetworksRepoModelObjectType}
\item \emph{ @param } version numeric
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDoiV2Doi} \cr


\item status code : 200 | The DOI and all its associated DOI metadata, if the DOI has been minted for the object

\item return type : OrgSagebionetworksRepoModelDoiV2Doi
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1DoiAssociation } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } type \link{OrgSagebionetworksRepoModelObjectType}
\item \emph{ @param } version numeric
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDoiV2DoiAssociation} \cr


\item status code : 200 | The DOI if the DOI has been minted for the object

\item return type : OrgSagebionetworksRepoModelDoiV2DoiAssociation
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1DoiAsyncGetAsyncToken } \emph{  }

\itemize{
\item \emph{ @param } async_token character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDoiV2DoiResponse} \cr


\item status code : 200 | The results of the call

\item return type : OrgSagebionetworksRepoModelDoiV2DoiResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1DoiLocate } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } type \link{OrgSagebionetworksRepoModelObjectType}
\item \emph{ @param } version numeric
\item \emph{ @param } redirect character
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Status 200 will be returned if the 'redirect' boolean param is false

\item return type : AnyType
\item response headers :

\tabular{ll}{
}
\item status code : 307 | Status 307 will be returned if the 'redirect' boolean param is true or null


\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1DoiAsyncStart } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_doi_v2_doi_request \link{OrgSagebionetworksRepoModelDoiV2DoiRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAsynchAsyncJobId} \cr


\item status code : 201 | The asynchronous job ID

\item return type : OrgSagebionetworksRepoModelAsynchAsyncJobId
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  getRepoV1Doi  ####################

library(synr)
var_id <- "id_example" # character | 
var_type <- org.sagebionetworks.repo.model.ObjectType$new() # OrgSagebionetworksRepoModelObjectType | 
var_version <- 3.4 # numeric |  (Optional)

api_instance <- DOIServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1Doi(var_id, var_type, version = var_versiondata_file = "result.txt")
result <- api_instance$getRepoV1Doi(var_id, var_type, version = var_version)
dput(result)


####################  getRepoV1DoiAssociation  ####################

library(synr)
var_id <- "id_example" # character | 
var_type <- org.sagebionetworks.repo.model.ObjectType$new() # OrgSagebionetworksRepoModelObjectType | 
var_version <- 3.4 # numeric |  (Optional)

api_instance <- DOIServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1DoiAssociation(var_id, var_type, version = var_versiondata_file = "result.txt")
result <- api_instance$getRepoV1DoiAssociation(var_id, var_type, version = var_version)
dput(result)


####################  getRepoV1DoiAsyncGetAsyncToken  ####################

library(synr)
var_async_token <- "async_token_example" # character | The async job token from the create/update call

api_instance <- DOIServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1DoiAsyncGetAsyncToken(var_async_tokendata_file = "result.txt")
result <- api_instance$getRepoV1DoiAsyncGetAsyncToken(var_async_token)
dput(result)


####################  getRepoV1DoiLocate  ####################

library(synr)
var_id <- "id_example" # character | 
var_type <- org.sagebionetworks.repo.model.ObjectType$new() # OrgSagebionetworksRepoModelObjectType | 
var_version <- 3.4 # numeric |  (Optional)
var_redirect <- "redirect_example" # character | Whether to return the URL or redirect to the URL (Optional)

api_instance <- DOIServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1DoiLocate(var_id, var_type, version = var_version, redirect = var_redirectdata_file = "result.txt")
result <- api_instance$getRepoV1DoiLocate(var_id, var_type, version = var_version, redirect = var_redirect)
dput(result)


####################  postRepoV1DoiAsyncStart  ####################

library(synr)
var_org_sagebionetworks_repo_model_doi_v2_doi_request <- org.sagebionetworks.repo.model.doi.v2.DoiRequest$new("org.sagebionetworks.repo.model.doi.v2.DoiRequest", org.sagebionetworks.repo.model.doi.v2.Doi$new(c(org.sagebionetworks.repo.model.doi.v2.DoiCreator$new("creatorName_example", c(org.sagebionetworks.repo.model.doi.v2.DoiNameIdentifier$new("identifier_example", "nameIdentifierScheme_example")))), c(org.sagebionetworks.repo.model.doi.v2.DoiTitle$new("title_example")), 123, org.sagebionetworks.repo.model.doi.v2.DoiResourceType$new("resourceTypeGeneral_example"), "associationId_example", "etag_example", "doiUri_example", "doiUrl_example", "objectId_example", "objectType_example", 123, "associatedBy_example", "associatedOn_example", "updatedBy_example", "updatedOn_example")) # OrgSagebionetworksRepoModelDoiV2DoiRequest | 

api_instance <- DOIServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1DoiAsyncStart(var_org_sagebionetworks_repo_model_doi_v2_doi_requestdata_file = "result.txt")
result <- api_instance$postRepoV1DoiAsyncStart(var_org_sagebionetworks_repo_model_doi_v2_doi_request)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DOIServicesApi-new}{\code{DOIServicesApi$new()}}
\item \href{#method-DOIServicesApi-getRepoV1Doi}{\code{DOIServicesApi$getRepoV1Doi()}}
\item \href{#method-DOIServicesApi-getRepoV1DoiWithHttpInfo}{\code{DOIServicesApi$getRepoV1DoiWithHttpInfo()}}
\item \href{#method-DOIServicesApi-getRepoV1DoiAssociation}{\code{DOIServicesApi$getRepoV1DoiAssociation()}}
\item \href{#method-DOIServicesApi-getRepoV1DoiAssociationWithHttpInfo}{\code{DOIServicesApi$getRepoV1DoiAssociationWithHttpInfo()}}
\item \href{#method-DOIServicesApi-getRepoV1DoiAsyncGetAsyncToken}{\code{DOIServicesApi$getRepoV1DoiAsyncGetAsyncToken()}}
\item \href{#method-DOIServicesApi-getRepoV1DoiAsyncGetAsyncTokenWithHttpInfo}{\code{DOIServicesApi$getRepoV1DoiAsyncGetAsyncTokenWithHttpInfo()}}
\item \href{#method-DOIServicesApi-getRepoV1DoiLocate}{\code{DOIServicesApi$getRepoV1DoiLocate()}}
\item \href{#method-DOIServicesApi-getRepoV1DoiLocateWithHttpInfo}{\code{DOIServicesApi$getRepoV1DoiLocateWithHttpInfo()}}
\item \href{#method-DOIServicesApi-postRepoV1DoiAsyncStart}{\code{DOIServicesApi$postRepoV1DoiAsyncStart()}}
\item \href{#method-DOIServicesApi-postRepoV1DoiAsyncStartWithHttpInfo}{\code{DOIServicesApi$postRepoV1DoiAsyncStartWithHttpInfo()}}
\item \href{#method-DOIServicesApi-clone}{\code{DOIServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DOIServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-getRepoV1Doi"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-getRepoV1Doi}{}}}
\subsection{Method \code{getRepoV1Doi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$getRepoV1Doi(id, type, version = NULL, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{type}}{}

\item{\code{version}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2Doi
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-getRepoV1DoiWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-getRepoV1DoiWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1DoiWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$getRepoV1DoiWithHttpInfo(
  id,
  type,
  version = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{type}}{}

\item{\code{version}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDoiV2Doi) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-getRepoV1DoiAssociation"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-getRepoV1DoiAssociation}{}}}
\subsection{Method \code{getRepoV1DoiAssociation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$getRepoV1DoiAssociation(
  id,
  type,
  version = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{type}}{}

\item{\code{version}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2DoiAssociation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-getRepoV1DoiAssociationWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-getRepoV1DoiAssociationWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1DoiAssociationWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$getRepoV1DoiAssociationWithHttpInfo(
  id,
  type,
  version = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{type}}{}

\item{\code{version}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDoiV2DoiAssociation) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-getRepoV1DoiAsyncGetAsyncToken"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-getRepoV1DoiAsyncGetAsyncToken}{}}}
\subsection{Method \code{getRepoV1DoiAsyncGetAsyncToken()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$getRepoV1DoiAsyncGetAsyncToken(
  async_token,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{async_token}}{The async job token from the create/update call}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2DoiResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-getRepoV1DoiAsyncGetAsyncTokenWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-getRepoV1DoiAsyncGetAsyncTokenWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1DoiAsyncGetAsyncTokenWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$getRepoV1DoiAsyncGetAsyncTokenWithHttpInfo(
  async_token,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{async_token}}{The async job token from the create/update call}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDoiV2DoiResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-getRepoV1DoiLocate"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-getRepoV1DoiLocate}{}}}
\subsection{Method \code{getRepoV1DoiLocate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$getRepoV1DoiLocate(
  id,
  type,
  version = NULL,
  redirect = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{type}}{}

\item{\code{version}}{(optional) No description}

\item{\code{redirect}}{(optional) Whether to return the URL or redirect to the URL}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-getRepoV1DoiLocateWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-getRepoV1DoiLocateWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1DoiLocateWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$getRepoV1DoiLocateWithHttpInfo(
  id,
  type,
  version = NULL,
  redirect = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{type}}{}

\item{\code{version}}{(optional) No description}

\item{\code{redirect}}{(optional) Whether to return the URL or redirect to the URL}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-postRepoV1DoiAsyncStart"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-postRepoV1DoiAsyncStart}{}}}
\subsection{Method \code{postRepoV1DoiAsyncStart()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$postRepoV1DoiAsyncStart(
  org_sagebionetworks_repo_model_doi_v2_doi_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_doi_v2_doi_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAsynchAsyncJobId
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-postRepoV1DoiAsyncStartWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-postRepoV1DoiAsyncStartWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1DoiAsyncStartWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$postRepoV1DoiAsyncStartWithHttpInfo(
  org_sagebionetworks_repo_model_doi_v2_doi_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_doi_v2_doi_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAsynchAsyncJobId) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DOIServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DOIServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DOIServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
