% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_services_api.R
\docType{class}
\name{DrsServicesApi}
\alias{DrsServicesApi}
\title{DrsServices operations}
\format{
An \code{R6Class} generator object
}
\description{
DrsServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ getGa4ghDrsV1ObjectsObjectId } \emph{  }

\itemize{
\item \emph{ @param } object_id character
\item \emph{ @param } expand character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDrsDrsObject} \cr


\item status code : 200 | the DRS object

\item return type : OrgSagebionetworksRepoModelDrsDrsObject
\item response headers :

\tabular{ll}{
}
}

\strong{ getGa4ghDrsV1ObjectsObjectIdAccessAccessId } \emph{  }

\itemize{
\item \emph{ @param } object_id character
\item \emph{ @param } access_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDrsAccessUrl} \cr


\item status code : 200 | the presigned url to download a file

\item return type : OrgSagebionetworksRepoModelDrsAccessUrl
\item response headers :

\tabular{ll}{
}
}

\strong{ getGa4ghDrsV1ServiceInfo } \emph{  }

\itemize{
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelDrsServiceInformation} \cr


\item status code : 200 | the drs service information

\item return type : OrgSagebionetworksRepoModelDrsServiceInformation
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  getGa4ghDrsV1ObjectsObjectId  ####################

library(synr)
var_object_id <- "object_id_example" # character | 
var_expand <- "expand_example" # character | 

api_instance <- DrsServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getGa4ghDrsV1ObjectsObjectId(var_object_id, var_expanddata_file = "result.txt")
result <- api_instance$getGa4ghDrsV1ObjectsObjectId(var_object_id, var_expand)
dput(result)


####################  getGa4ghDrsV1ObjectsObjectIdAccessAccessId  ####################

library(synr)
var_object_id <- "object_id_example" # character | 
var_access_id <- "access_id_example" # character | 

api_instance <- DrsServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getGa4ghDrsV1ObjectsObjectIdAccessAccessId(var_object_id, var_access_iddata_file = "result.txt")
result <- api_instance$getGa4ghDrsV1ObjectsObjectIdAccessAccessId(var_object_id, var_access_id)
dput(result)


####################  getGa4ghDrsV1ServiceInfo  ####################

library(synr)

api_instance <- DrsServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getGa4ghDrsV1ServiceInfo(data_file = "result.txt")
result <- api_instance$getGa4ghDrsV1ServiceInfo()
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DrsServicesApi-new}{\code{DrsServicesApi$new()}}
\item \href{#method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectId}{\code{DrsServicesApi$getGa4ghDrsV1ObjectsObjectId()}}
\item \href{#method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectIdWithHttpInfo}{\code{DrsServicesApi$getGa4ghDrsV1ObjectsObjectIdWithHttpInfo()}}
\item \href{#method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectIdAccessAccessId}{\code{DrsServicesApi$getGa4ghDrsV1ObjectsObjectIdAccessAccessId()}}
\item \href{#method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectIdAccessAccessIdWithHttpInfo}{\code{DrsServicesApi$getGa4ghDrsV1ObjectsObjectIdAccessAccessIdWithHttpInfo()}}
\item \href{#method-DrsServicesApi-getGa4ghDrsV1ServiceInfo}{\code{DrsServicesApi$getGa4ghDrsV1ServiceInfo()}}
\item \href{#method-DrsServicesApi-getGa4ghDrsV1ServiceInfoWithHttpInfo}{\code{DrsServicesApi$getGa4ghDrsV1ServiceInfoWithHttpInfo()}}
\item \href{#method-DrsServicesApi-clone}{\code{DrsServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrsServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DrsServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DrsServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrsServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectId"></a>}}
\if{latex}{\out{\hypertarget{method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectId}{}}}
\subsection{Method \code{getGa4ghDrsV1ObjectsObjectId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrsServicesApi$getGa4ghDrsV1ObjectsObjectId(
  object_id,
  expand,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_id}}{}

\item{\code{expand}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDrsDrsObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectIdWithHttpInfo}{}}}
\subsection{Method \code{getGa4ghDrsV1ObjectsObjectIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrsServicesApi$getGa4ghDrsV1ObjectsObjectIdWithHttpInfo(
  object_id,
  expand,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_id}}{}

\item{\code{expand}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDrsDrsObject) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectIdAccessAccessId"></a>}}
\if{latex}{\out{\hypertarget{method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectIdAccessAccessId}{}}}
\subsection{Method \code{getGa4ghDrsV1ObjectsObjectIdAccessAccessId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrsServicesApi$getGa4ghDrsV1ObjectsObjectIdAccessAccessId(
  object_id,
  access_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_id}}{}

\item{\code{access_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDrsAccessUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectIdAccessAccessIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DrsServicesApi-getGa4ghDrsV1ObjectsObjectIdAccessAccessIdWithHttpInfo}{}}}
\subsection{Method \code{getGa4ghDrsV1ObjectsObjectIdAccessAccessIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrsServicesApi$getGa4ghDrsV1ObjectsObjectIdAccessAccessIdWithHttpInfo(
  object_id,
  access_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_id}}{}

\item{\code{access_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDrsAccessUrl) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrsServicesApi-getGa4ghDrsV1ServiceInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DrsServicesApi-getGa4ghDrsV1ServiceInfo}{}}}
\subsection{Method \code{getGa4ghDrsV1ServiceInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrsServicesApi$getGa4ghDrsV1ServiceInfo(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelDrsServiceInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrsServicesApi-getGa4ghDrsV1ServiceInfoWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DrsServicesApi-getGa4ghDrsV1ServiceInfoWithHttpInfo}{}}}
\subsection{Method \code{getGa4ghDrsV1ServiceInfoWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrsServicesApi$getGa4ghDrsV1ServiceInfoWithHttpInfo(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelDrsServiceInformation) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrsServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DrsServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrsServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
