% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_graphemes.R
\name{filter_graphemes}
\alias{filter_graphemes}
\title{Filter graphemes of a single participant}
\usage{
filter_graphemes(graphemes, symbol_vector = NULL)
}
\arguments{
\item{graphemes}{A list of Grapheme objects.}

\item{symbol_vector}{A character vector of symbols to filter the
participant's graphemes by. Alternatively NULL (default), in which case
no filtering will be done and the full grapheme list is returned.}
}
\value{
A list of Grapheme objects.
}
\description{
Takes in a list of Grapheme objects and a character vector.
Returns a list of Grapheme objects, consisting of the participant's
graphemes which had a symbol included in the character vector to filter by.
}
\keyword{internal}
