% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_services_api.R
\docType{class}
\name{FormServicesApi}
\alias{FormServicesApi}
\title{FormServices operations}
\format{
An \code{R6Class} generator object
}
\description{
FormServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ deleteRepoV1FormDataId } \emph{  }

\itemize{
\item \emph{ @param } id character


\item status code : 200 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1FormGroupId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelFormFormGroup} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelFormFormGroup
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1FormGroupIdAcl } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAccessControlList} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAccessControlList
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1FormData } \emph{  }

\itemize{
\item \emph{ @param } group_id character
\item \emph{ @param } org_sagebionetworks_repo_model_form_form_change_request \link{OrgSagebionetworksRepoModelFormFormChangeRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelFormFormData} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelFormFormData
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1FormDataIdSubmit } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelFormFormData} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelFormFormData
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1FormDataList } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_form_list_request \link{OrgSagebionetworksRepoModelFormListRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelFormListResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelFormListResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1FormDataListReviewer } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_form_list_request \link{OrgSagebionetworksRepoModelFormListRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelFormListResponse} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelFormListResponse
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1FormGroup } \emph{  }

\itemize{
\item \emph{ @param } name character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelFormFormGroup} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelFormFormGroup
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1FormDataId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } org_sagebionetworks_repo_model_form_form_change_request \link{OrgSagebionetworksRepoModelFormFormChangeRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelFormFormData} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelFormFormData
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1FormDataIdAccept } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelFormFormData} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelFormFormData
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1FormDataIdReject } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } org_sagebionetworks_repo_model_form_form_rejection \link{OrgSagebionetworksRepoModelFormFormRejection}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelFormFormData} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelFormFormData
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1FormGroupIdAcl } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } org_sagebionetworks_repo_model_access_control_list \link{OrgSagebionetworksRepoModelAccessControlList}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAccessControlList} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAccessControlList
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  deleteRepoV1FormDataId  ####################

library(synr)
var_id <- "id_example" # character | Id of the FormData object to delete

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1FormDataId(var_id)


####################  getRepoV1FormGroupId  ####################

library(synr)
var_id <- "id_example" # character | 

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1FormGroupId(var_iddata_file = "result.txt")
result <- api_instance$getRepoV1FormGroupId(var_id)
dput(result)


####################  getRepoV1FormGroupIdAcl  ####################

library(synr)
var_id <- "id_example" # character | The identifier of the FormGroup.

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1FormGroupIdAcl(var_iddata_file = "result.txt")
result <- api_instance$getRepoV1FormGroupIdAcl(var_id)
dput(result)


####################  postRepoV1FormData  ####################

library(synr)
var_group_id <- "group_id_example" # character | The identifier of the group that manages this data.                          Required.
var_org_sagebionetworks_repo_model_form_form_change_request <- org.sagebionetworks.repo.model.form.FormChangeRequest$new("name_example", "fileHandleId_example") # OrgSagebionetworksRepoModelFormFormChangeRequest | 

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1FormData(var_group_id, var_org_sagebionetworks_repo_model_form_form_change_requestdata_file = "result.txt")
result <- api_instance$postRepoV1FormData(var_group_id, var_org_sagebionetworks_repo_model_form_form_change_request)
dput(result)


####################  postRepoV1FormDataIdSubmit  ####################

library(synr)
var_id <- "id_example" # character | 

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1FormDataIdSubmit(var_iddata_file = "result.txt")
result <- api_instance$postRepoV1FormDataIdSubmit(var_id)
dput(result)


####################  postRepoV1FormDataList  ####################

library(synr)
var_org_sagebionetworks_repo_model_form_list_request <- org.sagebionetworks.repo.model.form.ListRequest$new(c("filterByState_example"), "groupId_example", "nextPageToken_example") # OrgSagebionetworksRepoModelFormListRequest | 

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1FormDataList(var_org_sagebionetworks_repo_model_form_list_requestdata_file = "result.txt")
result <- api_instance$postRepoV1FormDataList(var_org_sagebionetworks_repo_model_form_list_request)
dput(result)


####################  postRepoV1FormDataListReviewer  ####################

library(synr)
var_org_sagebionetworks_repo_model_form_list_request <- org.sagebionetworks.repo.model.form.ListRequest$new(c("filterByState_example"), "groupId_example", "nextPageToken_example") # OrgSagebionetworksRepoModelFormListRequest | 

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1FormDataListReviewer(var_org_sagebionetworks_repo_model_form_list_requestdata_file = "result.txt")
result <- api_instance$postRepoV1FormDataListReviewer(var_org_sagebionetworks_repo_model_form_list_request)
dput(result)


####################  postRepoV1FormGroup  ####################

library(synr)
var_name <- "name_example" # character | A globally unique name for the group. Required. Between 3 and                256 characters.

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1FormGroup(var_namedata_file = "result.txt")
result <- api_instance$postRepoV1FormGroup(var_name)
dput(result)


####################  putRepoV1FormDataId  ####################

library(synr)
var_id <- "id_example" # character | The identifier of the FormData to update.
var_org_sagebionetworks_repo_model_form_form_change_request <- org.sagebionetworks.repo.model.form.FormChangeRequest$new("name_example", "fileHandleId_example") # OrgSagebionetworksRepoModelFormFormChangeRequest | 

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1FormDataId(var_id, var_org_sagebionetworks_repo_model_form_form_change_requestdata_file = "result.txt")
result <- api_instance$putRepoV1FormDataId(var_id, var_org_sagebionetworks_repo_model_form_form_change_request)
dput(result)


####################  putRepoV1FormDataIdAccept  ####################

library(synr)
var_id <- "id_example" # character | Identifier of the FormData to accept.

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1FormDataIdAccept(var_iddata_file = "result.txt")
result <- api_instance$putRepoV1FormDataIdAccept(var_id)
dput(result)


####################  putRepoV1FormDataIdReject  ####################

library(synr)
var_id <- "id_example" # character | Identifier of the FormData to accept.
var_org_sagebionetworks_repo_model_form_form_rejection <- org.sagebionetworks.repo.model.form.FormRejection$new("reason_example") # OrgSagebionetworksRepoModelFormFormRejection | 

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1FormDataIdReject(var_id, var_org_sagebionetworks_repo_model_form_form_rejectiondata_file = "result.txt")
result <- api_instance$putRepoV1FormDataIdReject(var_id, var_org_sagebionetworks_repo_model_form_form_rejection)
dput(result)


####################  putRepoV1FormGroupIdAcl  ####################

library(synr)
var_id <- "id_example" # character | The identifier of the FormGroup.
var_org_sagebionetworks_repo_model_access_control_list <- org.sagebionetworks.repo.model.AccessControlList$new("id_example", "createdBy_example", "creationDate_example", "modifiedBy_example", "modifiedOn_example", "etag_example", c(org.sagebionetworks.repo.model.ResourceAccess$new(123, c("accessType_example")))) # OrgSagebionetworksRepoModelAccessControlList | 

api_instance <- FormServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1FormGroupIdAcl(var_id, var_org_sagebionetworks_repo_model_access_control_listdata_file = "result.txt")
result <- api_instance$putRepoV1FormGroupIdAcl(var_id, var_org_sagebionetworks_repo_model_access_control_list)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FormServicesApi-new}{\code{FormServicesApi$new()}}
\item \href{#method-FormServicesApi-deleteRepoV1FormDataId}{\code{FormServicesApi$deleteRepoV1FormDataId()}}
\item \href{#method-FormServicesApi-deleteRepoV1FormDataIdWithHttpInfo}{\code{FormServicesApi$deleteRepoV1FormDataIdWithHttpInfo()}}
\item \href{#method-FormServicesApi-getRepoV1FormGroupId}{\code{FormServicesApi$getRepoV1FormGroupId()}}
\item \href{#method-FormServicesApi-getRepoV1FormGroupIdWithHttpInfo}{\code{FormServicesApi$getRepoV1FormGroupIdWithHttpInfo()}}
\item \href{#method-FormServicesApi-getRepoV1FormGroupIdAcl}{\code{FormServicesApi$getRepoV1FormGroupIdAcl()}}
\item \href{#method-FormServicesApi-getRepoV1FormGroupIdAclWithHttpInfo}{\code{FormServicesApi$getRepoV1FormGroupIdAclWithHttpInfo()}}
\item \href{#method-FormServicesApi-postRepoV1FormData}{\code{FormServicesApi$postRepoV1FormData()}}
\item \href{#method-FormServicesApi-postRepoV1FormDataWithHttpInfo}{\code{FormServicesApi$postRepoV1FormDataWithHttpInfo()}}
\item \href{#method-FormServicesApi-postRepoV1FormDataIdSubmit}{\code{FormServicesApi$postRepoV1FormDataIdSubmit()}}
\item \href{#method-FormServicesApi-postRepoV1FormDataIdSubmitWithHttpInfo}{\code{FormServicesApi$postRepoV1FormDataIdSubmitWithHttpInfo()}}
\item \href{#method-FormServicesApi-postRepoV1FormDataList}{\code{FormServicesApi$postRepoV1FormDataList()}}
\item \href{#method-FormServicesApi-postRepoV1FormDataListWithHttpInfo}{\code{FormServicesApi$postRepoV1FormDataListWithHttpInfo()}}
\item \href{#method-FormServicesApi-postRepoV1FormDataListReviewer}{\code{FormServicesApi$postRepoV1FormDataListReviewer()}}
\item \href{#method-FormServicesApi-postRepoV1FormDataListReviewerWithHttpInfo}{\code{FormServicesApi$postRepoV1FormDataListReviewerWithHttpInfo()}}
\item \href{#method-FormServicesApi-postRepoV1FormGroup}{\code{FormServicesApi$postRepoV1FormGroup()}}
\item \href{#method-FormServicesApi-postRepoV1FormGroupWithHttpInfo}{\code{FormServicesApi$postRepoV1FormGroupWithHttpInfo()}}
\item \href{#method-FormServicesApi-putRepoV1FormDataId}{\code{FormServicesApi$putRepoV1FormDataId()}}
\item \href{#method-FormServicesApi-putRepoV1FormDataIdWithHttpInfo}{\code{FormServicesApi$putRepoV1FormDataIdWithHttpInfo()}}
\item \href{#method-FormServicesApi-putRepoV1FormDataIdAccept}{\code{FormServicesApi$putRepoV1FormDataIdAccept()}}
\item \href{#method-FormServicesApi-putRepoV1FormDataIdAcceptWithHttpInfo}{\code{FormServicesApi$putRepoV1FormDataIdAcceptWithHttpInfo()}}
\item \href{#method-FormServicesApi-putRepoV1FormDataIdReject}{\code{FormServicesApi$putRepoV1FormDataIdReject()}}
\item \href{#method-FormServicesApi-putRepoV1FormDataIdRejectWithHttpInfo}{\code{FormServicesApi$putRepoV1FormDataIdRejectWithHttpInfo()}}
\item \href{#method-FormServicesApi-putRepoV1FormGroupIdAcl}{\code{FormServicesApi$putRepoV1FormGroupIdAcl()}}
\item \href{#method-FormServicesApi-putRepoV1FormGroupIdAclWithHttpInfo}{\code{FormServicesApi$putRepoV1FormGroupIdAclWithHttpInfo()}}
\item \href{#method-FormServicesApi-clone}{\code{FormServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FormServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-deleteRepoV1FormDataId"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-deleteRepoV1FormDataId}{}}}
\subsection{Method \code{deleteRepoV1FormDataId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$deleteRepoV1FormDataId(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id of the FormData object to delete}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-deleteRepoV1FormDataIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-deleteRepoV1FormDataIdWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1FormDataIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$deleteRepoV1FormDataIdWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id of the FormData object to delete}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-getRepoV1FormGroupId"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-getRepoV1FormGroupId}{}}}
\subsection{Method \code{getRepoV1FormGroupId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$getRepoV1FormGroupId(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelFormFormGroup
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-getRepoV1FormGroupIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-getRepoV1FormGroupIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1FormGroupIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$getRepoV1FormGroupIdWithHttpInfo(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelFormFormGroup) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-getRepoV1FormGroupIdAcl"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-getRepoV1FormGroupIdAcl}{}}}
\subsection{Method \code{getRepoV1FormGroupIdAcl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$getRepoV1FormGroupIdAcl(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The identifier of the FormGroup.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAccessControlList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-getRepoV1FormGroupIdAclWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-getRepoV1FormGroupIdAclWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1FormGroupIdAclWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$getRepoV1FormGroupIdAclWithHttpInfo(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The identifier of the FormGroup.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAccessControlList) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormData"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormData}{}}}
\subsection{Method \code{postRepoV1FormData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormData(
  group_id,
  org_sagebionetworks_repo_model_form_form_change_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_id}}{The identifier of the group that manages this data.                          Required.}

\item{\code{org_sagebionetworks_repo_model_form_form_change_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelFormFormData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormDataWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormDataWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1FormDataWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormDataWithHttpInfo(
  group_id,
  org_sagebionetworks_repo_model_form_form_change_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_id}}{The identifier of the group that manages this data.                          Required.}

\item{\code{org_sagebionetworks_repo_model_form_form_change_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelFormFormData) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormDataIdSubmit"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormDataIdSubmit}{}}}
\subsection{Method \code{postRepoV1FormDataIdSubmit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormDataIdSubmit(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelFormFormData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormDataIdSubmitWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormDataIdSubmitWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1FormDataIdSubmitWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormDataIdSubmitWithHttpInfo(
  id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelFormFormData) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormDataList"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormDataList}{}}}
\subsection{Method \code{postRepoV1FormDataList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormDataList(
  org_sagebionetworks_repo_model_form_list_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_form_list_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelFormListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormDataListWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormDataListWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1FormDataListWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormDataListWithHttpInfo(
  org_sagebionetworks_repo_model_form_list_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_form_list_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelFormListResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormDataListReviewer"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormDataListReviewer}{}}}
\subsection{Method \code{postRepoV1FormDataListReviewer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormDataListReviewer(
  org_sagebionetworks_repo_model_form_list_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_form_list_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelFormListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormDataListReviewerWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormDataListReviewerWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1FormDataListReviewerWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormDataListReviewerWithHttpInfo(
  org_sagebionetworks_repo_model_form_list_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_form_list_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelFormListResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormGroup"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormGroup}{}}}
\subsection{Method \code{postRepoV1FormGroup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormGroup(name, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A globally unique name for the group. Required. Between 3 and                256 characters.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelFormFormGroup
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-postRepoV1FormGroupWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-postRepoV1FormGroupWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1FormGroupWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$postRepoV1FormGroupWithHttpInfo(name, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A globally unique name for the group. Required. Between 3 and                256 characters.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelFormFormGroup) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-putRepoV1FormDataId"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-putRepoV1FormDataId}{}}}
\subsection{Method \code{putRepoV1FormDataId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$putRepoV1FormDataId(
  id,
  org_sagebionetworks_repo_model_form_form_change_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The identifier of the FormData to update.}

\item{\code{org_sagebionetworks_repo_model_form_form_change_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelFormFormData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-putRepoV1FormDataIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-putRepoV1FormDataIdWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1FormDataIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$putRepoV1FormDataIdWithHttpInfo(
  id,
  org_sagebionetworks_repo_model_form_form_change_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The identifier of the FormData to update.}

\item{\code{org_sagebionetworks_repo_model_form_form_change_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelFormFormData) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-putRepoV1FormDataIdAccept"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-putRepoV1FormDataIdAccept}{}}}
\subsection{Method \code{putRepoV1FormDataIdAccept()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$putRepoV1FormDataIdAccept(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Identifier of the FormData to accept.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelFormFormData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-putRepoV1FormDataIdAcceptWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-putRepoV1FormDataIdAcceptWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1FormDataIdAcceptWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$putRepoV1FormDataIdAcceptWithHttpInfo(
  id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Identifier of the FormData to accept.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelFormFormData) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-putRepoV1FormDataIdReject"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-putRepoV1FormDataIdReject}{}}}
\subsection{Method \code{putRepoV1FormDataIdReject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$putRepoV1FormDataIdReject(
  id,
  org_sagebionetworks_repo_model_form_form_rejection,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Identifier of the FormData to accept.}

\item{\code{org_sagebionetworks_repo_model_form_form_rejection}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelFormFormData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-putRepoV1FormDataIdRejectWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-putRepoV1FormDataIdRejectWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1FormDataIdRejectWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$putRepoV1FormDataIdRejectWithHttpInfo(
  id,
  org_sagebionetworks_repo_model_form_form_rejection,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Identifier of the FormData to accept.}

\item{\code{org_sagebionetworks_repo_model_form_form_rejection}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelFormFormData) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-putRepoV1FormGroupIdAcl"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-putRepoV1FormGroupIdAcl}{}}}
\subsection{Method \code{putRepoV1FormGroupIdAcl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$putRepoV1FormGroupIdAcl(
  id,
  org_sagebionetworks_repo_model_access_control_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The identifier of the FormGroup.}

\item{\code{org_sagebionetworks_repo_model_access_control_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAccessControlList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-putRepoV1FormGroupIdAclWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-putRepoV1FormGroupIdAclWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1FormGroupIdAclWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$putRepoV1FormGroupIdAclWithHttpInfo(
  id,
  org_sagebionetworks_repo_model_access_control_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The identifier of the FormGroup.}

\item{\code{org_sagebionetworks_repo_model_access_control_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAccessControlList) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FormServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FormServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
