% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grapheme.R
\docType{class}
\name{Grapheme-class}
\alias{Grapheme-class}
\alias{Grapheme}
\title{A Reference Class to represent consistency test graphemes.}
\description{
A Reference Class to represent consistency test graphemes.
}
\section{Fields}{

\describe{
\item{\code{symbol}}{A one-element character vector containing the symbol/set of symbols
that describe(s) the grapheme, e. g. '7' or 'Monday'. Set at class new() call or using set_symbol method.}

\item{\code{response_colors}}{A matrix where each row specifies color coordinates for each participant
response. Set using set_colors method.}

\item{\code{response_times}}{A numeric vector of response times. Set using set_times method.}

\item{\code{color_space}}{A one-element character vector which describes the color space
that response colors are coded in. Set when using set_colors method.}
}}

\section{Methods}{

\describe{
\item{\code{get_abbreviated_symbol()}}{Return a short (3 character) representation of
the grapheme's symbol.}

\item{\code{get_consistency_score(na.rm = FALSE, method = "euclidean")}}{Calculate the consistency score based on the
Grapheme instance's response colors. Throws an
error if no responses have been registered yet.
Always returns NA if all grapheme responses are NA.
If na.rm=FALSE, returns NA if any grapheme response
is NA. If na.rm=TRUE, returns the consistency score
for non-NA responses. This function relies on the
base/stats function dist() and so supports only
distance calculation methods implemented by dist()
(use help(dist) to learn more about it).}

\item{\code{get_mean_response_time(na.rm = FALSE)}}{Get the mean of the grapheme's associated
response times.}

\item{\code{get_num_non_na_colors()}}{Get the number of response colors that are non-NA, returned as
a one-element numeric vector.}

\item{\code{get_plot_data_list()}}{Get a list of the grapheme's data, bundled up in
a format ready for use in Participant.get_plot_data()
method as a row of plot
data.}

\item{\code{get_prop_color(color_label = NULL, r = NULL, g = NULL, b = NULL)}}{Get the proportion of response colors that are within a
specified color range. The range is specified using color_label
or the r/g/b arguments. Possible color_label specifications are:
"blue", "red", "green", "white", "black" or "hazy".
For r/g/b arguments, value ranges are specified using two-element
numeric vectors, with rgb values on a 0-1 scale. E. g. r=c(0, 0.3),
g=c(0, 0.3), b=c(0, 0.3) would code for a dark color range.}

\item{\code{has_only_non_na_colors()}}{Returns TRUE if the grapheme only has responses with valid colors,
FALSE if there are responses with nonvalid colors or there are
no responses at all.}

\item{\code{set_colors(hex_codes, color_space_spec)}}{Set response colors, using passed RGB hex codes. Converts
the hex codes to color coordinates in the specified
color space. Supports the following color spaces:
"XYZ", "sRGB", "Apple RGB", "Lab", and "Luv".
For all NA values passed, a row of NA values will be included
in the matrix (preserving order of responses). Returned/set
response colors are in the format of a matrix where each
row represents one response/color, and each
column represents one color coordinate axis (there are always
3 axes used for the currently supported color spaces)}

\item{\code{set_symbol(symbol_chars)}}{Set the grapheme's symbol attribute, using a passed
one-element character vector.}

\item{\code{set_times(times)}}{Add response times, using passed numeric vector.}
}}

\examples{
a <- synr::Grapheme$new(symbol='a')
a$set_colors(c("#101010", NA), "Luv")
a$set_times(c(5, 10))
a$get_num_non_na_colors()
}
