% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_color_range.R
\name{is_in_color_range}
\alias{is_in_color_range}
\title{Check if specified color falls within specified color range}
\usage{
is_in_color_range(
  color_to_check,
  color_label = NULL,
  r = NULL,
  g = NULL,
  b = NULL,
  color_space
)
}
\arguments{
\item{color_to_check}{A three-element numeric vector of color coordinates}

\item{color_label}{color specification (possible choices are: "blue", "red",
"green", "white", "black", "hazy")}

\item{r}{a two-value vector}

\item{g}{a two-value vector}

\item{b}{a two-value vector}

\item{color_space}{A one-element character vector describing the color space
that the passed color coordinates are in.}
}
\description{
Checks if specified color coordinates represent a color
within a specified color range. Color range is specified
using _either_ color_label _or_ r/g/b arguments. Possible
color_label specifications are: "blue", "red", "green",
"white", "black" or "hazy". r/g/b each take a two-value
numeric vector providing specifications of minimum/maximum
red/green/blue values within the range 0-1 (e.g. c(0.2, 0.8) ).

Supported color spaces are: "XYZ", "sRGB", "Apple RGB", "Lab",
and "Luv". Note that RGB values must be specified using the
range 0-1, not 0-255.
}
\seealso{
\code{\link[grDevices]{convertColor}}
}
\keyword{internal}
