% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_wrapper_of_team.R
\docType{class}
\name{ListWrapperOfTeam}
\alias{ListWrapperOfTeam}
\title{ListWrapperOfTeam}
\format{
An \code{R6Class} generator object
}
\description{
ListWrapperOfTeam Class

ListWrapperOfTeam Class
}
\details{
Create a new ListWrapperOfTeam
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{list}}{list(\link{OrgSagebionetworksRepoModelTeam}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListWrapperOfTeam-new}{\code{ListWrapperOfTeam$new()}}
\item \href{#method-ListWrapperOfTeam-toJSON}{\code{ListWrapperOfTeam$toJSON()}}
\item \href{#method-ListWrapperOfTeam-fromJSON}{\code{ListWrapperOfTeam$fromJSON()}}
\item \href{#method-ListWrapperOfTeam-toJSONString}{\code{ListWrapperOfTeam$toJSONString()}}
\item \href{#method-ListWrapperOfTeam-fromJSONString}{\code{ListWrapperOfTeam$fromJSONString()}}
\item \href{#method-ListWrapperOfTeam-validateJSON}{\code{ListWrapperOfTeam$validateJSON()}}
\item \href{#method-ListWrapperOfTeam-toString}{\code{ListWrapperOfTeam$toString()}}
\item \href{#method-ListWrapperOfTeam-isValid}{\code{ListWrapperOfTeam$isValid()}}
\item \href{#method-ListWrapperOfTeam-getInvalidFields}{\code{ListWrapperOfTeam$getInvalidFields()}}
\item \href{#method-ListWrapperOfTeam-print}{\code{ListWrapperOfTeam$print()}}
\item \href{#method-ListWrapperOfTeam-clone}{\code{ListWrapperOfTeam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListWrapperOfTeam class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$new(list = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list}}{list}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListWrapperOfTeam in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListWrapperOfTeam
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListWrapperOfTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListWrapperOfTeam in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListWrapperOfTeam
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListWrapperOfTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListWrapperOfTeam and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListWrapperOfTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-print"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
