% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_wrapper_of_team_member.R
\docType{class}
\name{ListWrapperOfTeamMember}
\alias{ListWrapperOfTeamMember}
\title{ListWrapperOfTeamMember}
\format{
An \code{R6Class} generator object
}
\description{
ListWrapperOfTeamMember Class

ListWrapperOfTeamMember Class
}
\details{
Create a new ListWrapperOfTeamMember
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{list}}{list(\link{OrgSagebionetworksRepoModelTeamMember}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListWrapperOfTeamMember-new}{\code{ListWrapperOfTeamMember$new()}}
\item \href{#method-ListWrapperOfTeamMember-toJSON}{\code{ListWrapperOfTeamMember$toJSON()}}
\item \href{#method-ListWrapperOfTeamMember-fromJSON}{\code{ListWrapperOfTeamMember$fromJSON()}}
\item \href{#method-ListWrapperOfTeamMember-toJSONString}{\code{ListWrapperOfTeamMember$toJSONString()}}
\item \href{#method-ListWrapperOfTeamMember-fromJSONString}{\code{ListWrapperOfTeamMember$fromJSONString()}}
\item \href{#method-ListWrapperOfTeamMember-validateJSON}{\code{ListWrapperOfTeamMember$validateJSON()}}
\item \href{#method-ListWrapperOfTeamMember-toString}{\code{ListWrapperOfTeamMember$toString()}}
\item \href{#method-ListWrapperOfTeamMember-isValid}{\code{ListWrapperOfTeamMember$isValid()}}
\item \href{#method-ListWrapperOfTeamMember-getInvalidFields}{\code{ListWrapperOfTeamMember$getInvalidFields()}}
\item \href{#method-ListWrapperOfTeamMember-print}{\code{ListWrapperOfTeamMember$print()}}
\item \href{#method-ListWrapperOfTeamMember-clone}{\code{ListWrapperOfTeamMember$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListWrapperOfTeamMember class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$new(list = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list}}{list}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListWrapperOfTeamMember in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListWrapperOfTeamMember
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListWrapperOfTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListWrapperOfTeamMember in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListWrapperOfTeamMember
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListWrapperOfTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListWrapperOfTeamMember and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListWrapperOfTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-print"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfTeamMember-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfTeamMember-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfTeamMember$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
