% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_wrapper_of_user_profile.R
\docType{class}
\name{ListWrapperOfUserProfile}
\alias{ListWrapperOfUserProfile}
\title{ListWrapperOfUserProfile}
\format{
An \code{R6Class} generator object
}
\description{
ListWrapperOfUserProfile Class

ListWrapperOfUserProfile Class
}
\details{
Create a new ListWrapperOfUserProfile
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{list}}{list(\link{OrgSagebionetworksRepoModelUserProfile}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListWrapperOfUserProfile-new}{\code{ListWrapperOfUserProfile$new()}}
\item \href{#method-ListWrapperOfUserProfile-toJSON}{\code{ListWrapperOfUserProfile$toJSON()}}
\item \href{#method-ListWrapperOfUserProfile-fromJSON}{\code{ListWrapperOfUserProfile$fromJSON()}}
\item \href{#method-ListWrapperOfUserProfile-toJSONString}{\code{ListWrapperOfUserProfile$toJSONString()}}
\item \href{#method-ListWrapperOfUserProfile-fromJSONString}{\code{ListWrapperOfUserProfile$fromJSONString()}}
\item \href{#method-ListWrapperOfUserProfile-validateJSON}{\code{ListWrapperOfUserProfile$validateJSON()}}
\item \href{#method-ListWrapperOfUserProfile-toString}{\code{ListWrapperOfUserProfile$toString()}}
\item \href{#method-ListWrapperOfUserProfile-isValid}{\code{ListWrapperOfUserProfile$isValid()}}
\item \href{#method-ListWrapperOfUserProfile-getInvalidFields}{\code{ListWrapperOfUserProfile$getInvalidFields()}}
\item \href{#method-ListWrapperOfUserProfile-print}{\code{ListWrapperOfUserProfile$print()}}
\item \href{#method-ListWrapperOfUserProfile-clone}{\code{ListWrapperOfUserProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListWrapperOfUserProfile class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$new(list = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list}}{list}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListWrapperOfUserProfile in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListWrapperOfUserProfile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListWrapperOfUserProfile
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListWrapperOfUserProfile in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListWrapperOfUserProfile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListWrapperOfUserProfile
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListWrapperOfUserProfile and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListWrapperOfUserProfile
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-print"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListWrapperOfUserProfile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListWrapperOfUserProfile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListWrapperOfUserProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
