% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_service_api.R
\docType{class}
\name{LogServiceApi}
\alias{LogServiceApi}
\title{LogService operations}
\format{
An \code{R6Class} generator object
}
\description{
LogServiceApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ postRepoV1Log } \emph{  }

\itemize{
\item \emph{ @param } user_agent character
\item \emph{ @param } org_sagebionetworks_repo_model_log_entry \link{OrgSagebionetworksRepoModelLogEntry}


\item status code : 200 | Void


\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  postRepoV1Log  ####################

library(synr)
var_user_agent <- "user_agent_example" # character | 
var_org_sagebionetworks_repo_model_log_entry <- org.sagebionetworks.repo.model.LogEntry$new("label_example", "message_example", "stacktrace_example") # OrgSagebionetworksRepoModelLogEntry | 

api_instance <- LogServiceApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$postRepoV1Log(var_user_agent, var_org_sagebionetworks_repo_model_log_entry)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogServiceApi-new}{\code{LogServiceApi$new()}}
\item \href{#method-LogServiceApi-postRepoV1Log}{\code{LogServiceApi$postRepoV1Log()}}
\item \href{#method-LogServiceApi-postRepoV1LogWithHttpInfo}{\code{LogServiceApi$postRepoV1LogWithHttpInfo()}}
\item \href{#method-LogServiceApi-clone}{\code{LogServiceApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogServiceApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogServiceApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new LogServiceApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogServiceApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogServiceApi-postRepoV1Log"></a>}}
\if{latex}{\out{\hypertarget{method-LogServiceApi-postRepoV1Log}{}}}
\subsection{Method \code{postRepoV1Log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogServiceApi$postRepoV1Log(
  user_agent,
  org_sagebionetworks_repo_model_log_entry,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_agent}}{}

\item{\code{org_sagebionetworks_repo_model_log_entry}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogServiceApi-postRepoV1LogWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-LogServiceApi-postRepoV1LogWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1LogWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogServiceApi$postRepoV1LogWithHttpInfo(
  user_agent,
  org_sagebionetworks_repo_model_log_entry,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_agent}}{}

\item{\code{org_sagebionetworks_repo_model_log_entry}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogServiceApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogServiceApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogServiceApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
