% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_services_api.R
\docType{class}
\name{MessageServicesApi}
\alias{MessageServicesApi}
\title{MessageServices operations}
\format{
An \code{R6Class} generator object
}
\description{
MessageServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ deleteRepoV1AdminMessageMessageId } \emph{  }

\itemize{
\item \emph{ @param } message_id character


\item status code : 200 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1MessageInbox } \emph{  }

\itemize{
\item \emph{ @param } inbox_filter character
\item \emph{ @param } order_by character
\item \emph{ @param } descending character
\item \emph{ @param } limit numeric
\item \emph{ @param } offset numeric
\item \emph{ @returnType } \link{PaginatedResultsOfMessageBundle} \cr


\item status code : 200 | Auto-generated description

\item return type : PaginatedResultsOfMessageBundle
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1MessageMessageId } \emph{  }

\itemize{
\item \emph{ @param } message_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelMessageMessageToUser} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelMessageMessageToUser
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1MessageMessageIdConversation } \emph{  }

\itemize{
\item \emph{ @param } message_id character
\item \emph{ @param } order_by character
\item \emph{ @param } descending character
\item \emph{ @param } limit numeric
\item \emph{ @param } offset numeric
\item \emph{ @returnType } \link{PaginatedResultsOfMessageToUser} \cr


\item status code : 200 | Auto-generated description

\item return type : PaginatedResultsOfMessageToUser
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1MessageMessageIdFile } \emph{  }

\itemize{
\item \emph{ @param } message_id character
\item \emph{ @param } redirect character
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Status 200 will be returned if the 'redirect' boolean param is false

\item return type : AnyType
\item response headers :

\tabular{ll}{
}
\item status code : 307 | Status 307 will be returned if the 'redirect' boolean param is true or null


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1MessageOutbox } \emph{  }

\itemize{
\item \emph{ @param } order_by character
\item \emph{ @param } descending character
\item \emph{ @param } limit numeric
\item \emph{ @param } offset numeric
\item \emph{ @returnType } \link{PaginatedResultsOfMessageToUser} \cr


\item status code : 200 | Auto-generated description

\item return type : PaginatedResultsOfMessageToUser
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1CloudMailInAuthorization } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_message_cloudmailin_authorization_check_header \link{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader}


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1CloudMailInMessage } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_message_cloudmailin_message \link{OrgSagebionetworksRepoModelMessageCloudmailinMessage}
\item \emph{ @param } notification_unsubscribe_endpoint character


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1EntityIdMessage } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } org_sagebionetworks_repo_model_message_message_to_user \link{OrgSagebionetworksRepoModelMessageMessageToUser}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelMessageMessageToUser} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelMessageMessageToUser
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1Message } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_message_message_to_user \link{OrgSagebionetworksRepoModelMessageMessageToUser}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelMessageMessageToUser} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelMessageMessageToUser
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1MessageMessageIdForward } \emph{  }

\itemize{
\item \emph{ @param } message_id character
\item \emph{ @param } org_sagebionetworks_repo_model_message_message_recipient_set \link{OrgSagebionetworksRepoModelMessageMessageRecipientSet}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelMessageMessageToUser} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelMessageMessageToUser
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1MessageStatus } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_message_message_status \link{OrgSagebionetworksRepoModelMessageMessageStatus}


\item status code : 200 | Void


\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  deleteRepoV1AdminMessageMessageId  ####################

library(synr)
var_message_id <- "message_id_example" # character | 

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1AdminMessageMessageId(var_message_id)


####################  getRepoV1MessageInbox  ####################

library(synr)
var_inbox_filter <- "inbox_filter_example" # character | 
var_order_by <- "order_by_example" # character | 
var_descending <- "descending_example" # character | 
var_limit <- 3.4 # numeric | 
var_offset <- 3.4 # numeric | 

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1MessageInbox(var_inbox_filter, var_order_by, var_descending, var_limit, var_offsetdata_file = "result.txt")
result <- api_instance$getRepoV1MessageInbox(var_inbox_filter, var_order_by, var_descending, var_limit, var_offset)
dput(result)


####################  getRepoV1MessageMessageId  ####################

library(synr)
var_message_id <- "message_id_example" # character | 

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1MessageMessageId(var_message_iddata_file = "result.txt")
result <- api_instance$getRepoV1MessageMessageId(var_message_id)
dput(result)


####################  getRepoV1MessageMessageIdConversation  ####################

library(synr)
var_message_id <- "message_id_example" # character | 
var_order_by <- "order_by_example" # character | 
var_descending <- "descending_example" # character | 
var_limit <- 3.4 # numeric | 
var_offset <- 3.4 # numeric | 

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1MessageMessageIdConversation(var_message_id, var_order_by, var_descending, var_limit, var_offsetdata_file = "result.txt")
result <- api_instance$getRepoV1MessageMessageIdConversation(var_message_id, var_order_by, var_descending, var_limit, var_offset)
dput(result)


####################  getRepoV1MessageMessageIdFile  ####################

library(synr)
var_message_id <- "message_id_example" # character | 
var_redirect <- "redirect_example" # character | When set to false, the URL will be returned as text/plain             instead of redirecting. (Optional)

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1MessageMessageIdFile(var_message_id, redirect = var_redirectdata_file = "result.txt")
result <- api_instance$getRepoV1MessageMessageIdFile(var_message_id, redirect = var_redirect)
dput(result)


####################  getRepoV1MessageOutbox  ####################

library(synr)
var_order_by <- "order_by_example" # character | 
var_descending <- "descending_example" # character | 
var_limit <- 3.4 # numeric | 
var_offset <- 3.4 # numeric | 

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1MessageOutbox(var_order_by, var_descending, var_limit, var_offsetdata_file = "result.txt")
result <- api_instance$getRepoV1MessageOutbox(var_order_by, var_descending, var_limit, var_offset)
dput(result)


####################  postRepoV1CloudMailInAuthorization  ####################

library(synr)
var_org_sagebionetworks_repo_model_message_cloudmailin_authorization_check_header <- org.sagebionetworks.repo.model.message.cloudmailin.AuthorizationCheckHeader$new("to_example", "from_example", 123, "helo_domain_example", "remote_ip_example") # OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader | 

api_instance <- MessageServicesApi$new()

api_instance$postRepoV1CloudMailInAuthorization(var_org_sagebionetworks_repo_model_message_cloudmailin_authorization_check_header)


####################  postRepoV1CloudMailInMessage  ####################

library(synr)
var_org_sagebionetworks_repo_model_message_cloudmailin_message <- org.sagebionetworks.repo.model.message.cloudmailin.Message$new(org.sagebionetworks.repo.model.message.cloudmailin.Envelope$new("to_example", c("recipients_example"), "from_example", "helo_domain_example", "remote_ip_example", "spf_example"), org.sagebionetworks.repo.model.message.cloudmailin.Headers$new("To_example", "Cc_example", "Bcc_example", "Subject_example", "From_example"), "plain_example", "html_example", "reply_plain_example", c(org.sagebionetworks.repo.model.message.cloudmailin.Attachment$new("content_example", "content_id_example", "url_example", "file_name_example", "content_type_example", "size_example", "disposition_example"))) # OrgSagebionetworksRepoModelMessageCloudmailinMessage | 
var_notification_unsubscribe_endpoint <- "notification_unsubscribe_endpoint_example" # character |  (Optional)

api_instance <- MessageServicesApi$new()

api_instance$postRepoV1CloudMailInMessage(var_org_sagebionetworks_repo_model_message_cloudmailin_message, notification_unsubscribe_endpoint = var_notification_unsubscribe_endpoint)


####################  postRepoV1EntityIdMessage  ####################

library(synr)
var_id <- "id_example" # character | 
var_org_sagebionetworks_repo_model_message_message_to_user <- org.sagebionetworks.repo.model.message.MessageToUser$new("id_example", "createdBy_example", "fileHandleId_example", "createdOn_example", c("recipients_example"), "subject_example", "inReplyTo_example", "inReplyToRoot_example", "notificationUnsubscribeEndpoint_example", "userProfileSettingEndpoint_example", "withUnsubscribeLink_example", "withProfileSettingLink_example", "isNotificationMessage_example", "to_example", "cc_example", "bcc_example") # OrgSagebionetworksRepoModelMessageMessageToUser | 

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1EntityIdMessage(var_id, var_org_sagebionetworks_repo_model_message_message_to_userdata_file = "result.txt")
result <- api_instance$postRepoV1EntityIdMessage(var_id, var_org_sagebionetworks_repo_model_message_message_to_user)
dput(result)


####################  postRepoV1Message  ####################

library(synr)
var_org_sagebionetworks_repo_model_message_message_to_user <- org.sagebionetworks.repo.model.message.MessageToUser$new("id_example", "createdBy_example", "fileHandleId_example", "createdOn_example", c("recipients_example"), "subject_example", "inReplyTo_example", "inReplyToRoot_example", "notificationUnsubscribeEndpoint_example", "userProfileSettingEndpoint_example", "withUnsubscribeLink_example", "withProfileSettingLink_example", "isNotificationMessage_example", "to_example", "cc_example", "bcc_example") # OrgSagebionetworksRepoModelMessageMessageToUser | 

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1Message(var_org_sagebionetworks_repo_model_message_message_to_userdata_file = "result.txt")
result <- api_instance$postRepoV1Message(var_org_sagebionetworks_repo_model_message_message_to_user)
dput(result)


####################  postRepoV1MessageMessageIdForward  ####################

library(synr)
var_message_id <- "message_id_example" # character | 
var_org_sagebionetworks_repo_model_message_message_recipient_set <- org.sagebionetworks.repo.model.message.MessageRecipientSet$new(c("recipients_example")) # OrgSagebionetworksRepoModelMessageMessageRecipientSet | 

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1MessageMessageIdForward(var_message_id, var_org_sagebionetworks_repo_model_message_message_recipient_setdata_file = "result.txt")
result <- api_instance$postRepoV1MessageMessageIdForward(var_message_id, var_org_sagebionetworks_repo_model_message_message_recipient_set)
dput(result)


####################  putRepoV1MessageStatus  ####################

library(synr)
var_org_sagebionetworks_repo_model_message_message_status <- org.sagebionetworks.repo.model.message.MessageStatus$new("messageId_example", "recipientId_example", "status_example") # OrgSagebionetworksRepoModelMessageMessageStatus | 

api_instance <- MessageServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$putRepoV1MessageStatus(var_org_sagebionetworks_repo_model_message_message_status)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MessageServicesApi-new}{\code{MessageServicesApi$new()}}
\item \href{#method-MessageServicesApi-deleteRepoV1AdminMessageMessageId}{\code{MessageServicesApi$deleteRepoV1AdminMessageMessageId()}}
\item \href{#method-MessageServicesApi-deleteRepoV1AdminMessageMessageIdWithHttpInfo}{\code{MessageServicesApi$deleteRepoV1AdminMessageMessageIdWithHttpInfo()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageInbox}{\code{MessageServicesApi$getRepoV1MessageInbox()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageInboxWithHttpInfo}{\code{MessageServicesApi$getRepoV1MessageInboxWithHttpInfo()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageMessageId}{\code{MessageServicesApi$getRepoV1MessageMessageId()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageMessageIdWithHttpInfo}{\code{MessageServicesApi$getRepoV1MessageMessageIdWithHttpInfo()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageMessageIdConversation}{\code{MessageServicesApi$getRepoV1MessageMessageIdConversation()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageMessageIdConversationWithHttpInfo}{\code{MessageServicesApi$getRepoV1MessageMessageIdConversationWithHttpInfo()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageMessageIdFile}{\code{MessageServicesApi$getRepoV1MessageMessageIdFile()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageMessageIdFileWithHttpInfo}{\code{MessageServicesApi$getRepoV1MessageMessageIdFileWithHttpInfo()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageOutbox}{\code{MessageServicesApi$getRepoV1MessageOutbox()}}
\item \href{#method-MessageServicesApi-getRepoV1MessageOutboxWithHttpInfo}{\code{MessageServicesApi$getRepoV1MessageOutboxWithHttpInfo()}}
\item \href{#method-MessageServicesApi-postRepoV1CloudMailInAuthorization}{\code{MessageServicesApi$postRepoV1CloudMailInAuthorization()}}
\item \href{#method-MessageServicesApi-postRepoV1CloudMailInAuthorizationWithHttpInfo}{\code{MessageServicesApi$postRepoV1CloudMailInAuthorizationWithHttpInfo()}}
\item \href{#method-MessageServicesApi-postRepoV1CloudMailInMessage}{\code{MessageServicesApi$postRepoV1CloudMailInMessage()}}
\item \href{#method-MessageServicesApi-postRepoV1CloudMailInMessageWithHttpInfo}{\code{MessageServicesApi$postRepoV1CloudMailInMessageWithHttpInfo()}}
\item \href{#method-MessageServicesApi-postRepoV1EntityIdMessage}{\code{MessageServicesApi$postRepoV1EntityIdMessage()}}
\item \href{#method-MessageServicesApi-postRepoV1EntityIdMessageWithHttpInfo}{\code{MessageServicesApi$postRepoV1EntityIdMessageWithHttpInfo()}}
\item \href{#method-MessageServicesApi-postRepoV1Message}{\code{MessageServicesApi$postRepoV1Message()}}
\item \href{#method-MessageServicesApi-postRepoV1MessageWithHttpInfo}{\code{MessageServicesApi$postRepoV1MessageWithHttpInfo()}}
\item \href{#method-MessageServicesApi-postRepoV1MessageMessageIdForward}{\code{MessageServicesApi$postRepoV1MessageMessageIdForward()}}
\item \href{#method-MessageServicesApi-postRepoV1MessageMessageIdForwardWithHttpInfo}{\code{MessageServicesApi$postRepoV1MessageMessageIdForwardWithHttpInfo()}}
\item \href{#method-MessageServicesApi-putRepoV1MessageStatus}{\code{MessageServicesApi$putRepoV1MessageStatus()}}
\item \href{#method-MessageServicesApi-putRepoV1MessageStatusWithHttpInfo}{\code{MessageServicesApi$putRepoV1MessageStatusWithHttpInfo()}}
\item \href{#method-MessageServicesApi-clone}{\code{MessageServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MessageServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-deleteRepoV1AdminMessageMessageId"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-deleteRepoV1AdminMessageMessageId}{}}}
\subsection{Method \code{deleteRepoV1AdminMessageMessageId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$deleteRepoV1AdminMessageMessageId(message_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-deleteRepoV1AdminMessageMessageIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-deleteRepoV1AdminMessageMessageIdWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1AdminMessageMessageIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$deleteRepoV1AdminMessageMessageIdWithHttpInfo(
  message_id,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageInbox"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageInbox}{}}}
\subsection{Method \code{getRepoV1MessageInbox()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageInbox(
  inbox_filter,
  order_by,
  descending,
  limit,
  offset,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inbox_filter}}{}

\item{\code{order_by}}{}

\item{\code{descending}}{}

\item{\code{limit}}{}

\item{\code{offset}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PaginatedResultsOfMessageBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageInboxWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageInboxWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1MessageInboxWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageInboxWithHttpInfo(
  inbox_filter,
  order_by,
  descending,
  limit,
  offset,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inbox_filter}}{}

\item{\code{order_by}}{}

\item{\code{descending}}{}

\item{\code{limit}}{}

\item{\code{offset}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PaginatedResultsOfMessageBundle) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageMessageId"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageMessageId}{}}}
\subsection{Method \code{getRepoV1MessageMessageId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageMessageId(message_id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelMessageMessageToUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageMessageIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageMessageIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1MessageMessageIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageMessageIdWithHttpInfo(
  message_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelMessageMessageToUser) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageMessageIdConversation"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageMessageIdConversation}{}}}
\subsection{Method \code{getRepoV1MessageMessageIdConversation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageMessageIdConversation(
  message_id,
  order_by,
  descending,
  limit,
  offset,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{order_by}}{}

\item{\code{descending}}{}

\item{\code{limit}}{}

\item{\code{offset}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PaginatedResultsOfMessageToUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageMessageIdConversationWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageMessageIdConversationWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1MessageMessageIdConversationWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageMessageIdConversationWithHttpInfo(
  message_id,
  order_by,
  descending,
  limit,
  offset,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{order_by}}{}

\item{\code{descending}}{}

\item{\code{limit}}{}

\item{\code{offset}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PaginatedResultsOfMessageToUser) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageMessageIdFile"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageMessageIdFile}{}}}
\subsection{Method \code{getRepoV1MessageMessageIdFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageMessageIdFile(
  message_id,
  redirect = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{redirect}}{(optional) When set to false, the URL will be returned as text/plain             instead of redirecting.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageMessageIdFileWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageMessageIdFileWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1MessageMessageIdFileWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageMessageIdFileWithHttpInfo(
  message_id,
  redirect = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{redirect}}{(optional) When set to false, the URL will be returned as text/plain             instead of redirecting.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageOutbox"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageOutbox}{}}}
\subsection{Method \code{getRepoV1MessageOutbox()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageOutbox(
  order_by,
  descending,
  limit,
  offset,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{order_by}}{}

\item{\code{descending}}{}

\item{\code{limit}}{}

\item{\code{offset}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PaginatedResultsOfMessageToUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-getRepoV1MessageOutboxWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-getRepoV1MessageOutboxWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1MessageOutboxWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$getRepoV1MessageOutboxWithHttpInfo(
  order_by,
  descending,
  limit,
  offset,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{order_by}}{}

\item{\code{descending}}{}

\item{\code{limit}}{}

\item{\code{offset}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PaginatedResultsOfMessageToUser) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1CloudMailInAuthorization"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1CloudMailInAuthorization}{}}}
\subsection{Method \code{postRepoV1CloudMailInAuthorization()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1CloudMailInAuthorization(
  org_sagebionetworks_repo_model_message_cloudmailin_authorization_check_header,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_message_cloudmailin_authorization_check_header}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1CloudMailInAuthorizationWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1CloudMailInAuthorizationWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1CloudMailInAuthorizationWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1CloudMailInAuthorizationWithHttpInfo(
  org_sagebionetworks_repo_model_message_cloudmailin_authorization_check_header,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_message_cloudmailin_authorization_check_header}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1CloudMailInMessage"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1CloudMailInMessage}{}}}
\subsection{Method \code{postRepoV1CloudMailInMessage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1CloudMailInMessage(
  org_sagebionetworks_repo_model_message_cloudmailin_message,
  notification_unsubscribe_endpoint = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_message_cloudmailin_message}}{}

\item{\code{notification_unsubscribe_endpoint}}{(optional) No description}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1CloudMailInMessageWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1CloudMailInMessageWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1CloudMailInMessageWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1CloudMailInMessageWithHttpInfo(
  org_sagebionetworks_repo_model_message_cloudmailin_message,
  notification_unsubscribe_endpoint = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_message_cloudmailin_message}}{}

\item{\code{notification_unsubscribe_endpoint}}{(optional) No description}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1EntityIdMessage"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1EntityIdMessage}{}}}
\subsection{Method \code{postRepoV1EntityIdMessage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1EntityIdMessage(
  id,
  org_sagebionetworks_repo_model_message_message_to_user,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{org_sagebionetworks_repo_model_message_message_to_user}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelMessageMessageToUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1EntityIdMessageWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1EntityIdMessageWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1EntityIdMessageWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1EntityIdMessageWithHttpInfo(
  id,
  org_sagebionetworks_repo_model_message_message_to_user,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{org_sagebionetworks_repo_model_message_message_to_user}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelMessageMessageToUser) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1Message"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1Message}{}}}
\subsection{Method \code{postRepoV1Message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1Message(
  org_sagebionetworks_repo_model_message_message_to_user,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_message_message_to_user}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelMessageMessageToUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1MessageWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1MessageWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1MessageWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1MessageWithHttpInfo(
  org_sagebionetworks_repo_model_message_message_to_user,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_message_message_to_user}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelMessageMessageToUser) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1MessageMessageIdForward"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1MessageMessageIdForward}{}}}
\subsection{Method \code{postRepoV1MessageMessageIdForward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1MessageMessageIdForward(
  message_id,
  org_sagebionetworks_repo_model_message_message_recipient_set,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{org_sagebionetworks_repo_model_message_message_recipient_set}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelMessageMessageToUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-postRepoV1MessageMessageIdForwardWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-postRepoV1MessageMessageIdForwardWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1MessageMessageIdForwardWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$postRepoV1MessageMessageIdForwardWithHttpInfo(
  message_id,
  org_sagebionetworks_repo_model_message_message_recipient_set,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{}

\item{\code{org_sagebionetworks_repo_model_message_message_recipient_set}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelMessageMessageToUser) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-putRepoV1MessageStatus"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-putRepoV1MessageStatus}{}}}
\subsection{Method \code{putRepoV1MessageStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$putRepoV1MessageStatus(
  org_sagebionetworks_repo_model_message_message_status,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_message_message_status}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-putRepoV1MessageStatusWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-putRepoV1MessageStatusWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1MessageStatusWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$putRepoV1MessageStatusWithHttpInfo(
  org_sagebionetworks_repo_model_message_message_status,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_message_message_status}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MessageServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
