% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_evaluation.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelEvaluation}
\alias{OrgSagebionetworksEvaluationModelEvaluation}
\title{OrgSagebionetworksEvaluationModelEvaluation}
\format{
An \code{R6Class} generator object
}
\description{
An Evaluation is the core object of the Evaluation API, used to support collaborative data analysis challenges in Synapse.

OrgSagebionetworksEvaluationModelEvaluation Class
}
\details{
Create a new OrgSagebionetworksEvaluationModelEvaluation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{ownerId}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{contentSource}}{character [optional]}

\item{\code{submissionInstructionsMessage}}{character [optional]}

\item{\code{submissionReceiptMessage}}{character [optional]}

\item{\code{quota}}{\link{OrgSagebionetworksEvaluationModelSubmissionQuota} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-new}{\code{OrgSagebionetworksEvaluationModelEvaluation$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-toJSON}{\code{OrgSagebionetworksEvaluationModelEvaluation$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-fromJSON}{\code{OrgSagebionetworksEvaluationModelEvaluation$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-toJSONString}{\code{OrgSagebionetworksEvaluationModelEvaluation$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-fromJSONString}{\code{OrgSagebionetworksEvaluationModelEvaluation$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-validateJSON}{\code{OrgSagebionetworksEvaluationModelEvaluation$validateJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-toString}{\code{OrgSagebionetworksEvaluationModelEvaluation$toString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-isValid}{\code{OrgSagebionetworksEvaluationModelEvaluation$isValid()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-getInvalidFields}{\code{OrgSagebionetworksEvaluationModelEvaluation$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-print}{\code{OrgSagebionetworksEvaluationModelEvaluation$print()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluation-clone}{\code{OrgSagebionetworksEvaluationModelEvaluation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelEvaluation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$new(
  id = NULL,
  etag = NULL,
  name = NULL,
  description = NULL,
  ownerId = NULL,
  createdOn = NULL,
  contentSource = NULL,
  submissionInstructionsMessage = NULL,
  submissionReceiptMessage = NULL,
  quota = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{ownerId}}{ownerId}

\item{\code{createdOn}}{createdOn}

\item{\code{contentSource}}{contentSource}

\item{\code{submissionInstructionsMessage}}{submissionInstructionsMessage}

\item{\code{submissionReceiptMessage}}{submissionReceiptMessage}

\item{\code{quota}}{quota}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelEvaluation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelEvaluation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelEvaluation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelEvaluation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelEvaluation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelEvaluation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksEvaluationModelEvaluation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksEvaluationModelEvaluation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
