% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_evaluation_round.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelEvaluationRound}
\alias{OrgSagebionetworksEvaluationModelEvaluationRound}
\title{OrgSagebionetworksEvaluationModelEvaluationRound}
\format{
An \code{R6Class} generator object
}
\description{
Defines the duration of a round and sets limits for maximum submissions per round

OrgSagebionetworksEvaluationModelEvaluationRound Class
}
\details{
Create a new OrgSagebionetworksEvaluationModelEvaluationRound
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{evaluationId}}{character}

\item{\code{roundStart}}{character}

\item{\code{roundEnd}}{character}

\item{\code{limits}}{Optional. Sets limits for maximum submissions in this round. list(\link{OrgSagebionetworksEvaluationModelEvaluationRoundLimit}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-new}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-toJSON}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-fromJSON}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-toJSONString}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-fromJSONString}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-validateJSON}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$validateJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-toString}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$toString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-isValid}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$isValid()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-getInvalidFields}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-print}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$print()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationRound-clone}{\code{OrgSagebionetworksEvaluationModelEvaluationRound$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelEvaluationRound class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$new(
  evaluationId,
  roundStart,
  roundEnd,
  id = NULL,
  etag = NULL,
  limits = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{evaluationId}}{evaluationId}

\item{\code{roundStart}}{roundStart}

\item{\code{roundEnd}}{roundEnd}

\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{limits}}{Optional. Sets limits for maximum submissions in this round.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelEvaluationRound in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelEvaluationRound
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelEvaluationRound
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelEvaluationRound in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelEvaluationRound
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelEvaluationRound
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksEvaluationModelEvaluationRound and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksEvaluationModelEvaluationRound
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationRound-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationRound-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationRound$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
