% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_evaluation_submissions.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelEvaluationSubmissions}
\alias{OrgSagebionetworksEvaluationModelEvaluationSubmissions}
\title{OrgSagebionetworksEvaluationModelEvaluationSubmissions}
\format{
An \code{R6Class} generator object
}
\description{
This object holds the metadata for a collection of Submissions belonging to an Evaluation.

OrgSagebionetworksEvaluationModelEvaluationSubmissions Class
}
\details{
Create a new OrgSagebionetworksEvaluationModelEvaluationSubmissions
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{evaluationId}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-new}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-toJSON}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-fromJSON}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-toJSONString}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-fromJSONString}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-validateJSON}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$validateJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-toString}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$toString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-isValid}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$isValid()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-getInvalidFields}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-print}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$print()}}
\item \href{#method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-clone}{\code{OrgSagebionetworksEvaluationModelEvaluationSubmissions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelEvaluationSubmissions class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$new(
  id = NULL,
  etag = NULL,
  evaluationId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{evaluationId}}{evaluationId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelEvaluationSubmissions in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelEvaluationSubmissions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelEvaluationSubmissions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelEvaluationSubmissions in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelEvaluationSubmissions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelEvaluationSubmissions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksEvaluationModelEvaluationSubmissions and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksEvaluationModelEvaluationSubmissions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelEvaluationSubmissions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelEvaluationSubmissions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
