% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_member_submission_eligibility.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility}
\alias{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility}
\title{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility}
\format{
An \code{R6Class} generator object
}
\description{
Describes the eligibility of an indivdiual to submit to an Evaluation queue, reflecting the queue's submission quotas and current submissions.

OrgSagebionetworksEvaluationModelMemberSubmissionEligibility Class
}
\details{
Create a new OrgSagebionetworksEvaluationModelMemberSubmissionEligibility
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isEligible}}{character [optional]}

\item{\code{isRegistered}}{character [optional]}

\item{\code{isQuotaFilled}}{character [optional]}

\item{\code{principalId}}{integer [optional]}

\item{\code{hasConflictingSubmission}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-new}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-toJSON}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-fromJSON}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-toJSONString}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-fromJSONString}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-validateJSON}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$validateJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-toString}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$toString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-isValid}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$isValid()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-getInvalidFields}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-print}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$print()}}
\item \href{#method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-clone}{\code{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelMemberSubmissionEligibility class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$new(
  isEligible = NULL,
  isRegistered = NULL,
  isQuotaFilled = NULL,
  principalId = NULL,
  hasConflictingSubmission = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isEligible}}{isEligible}

\item{\code{isRegistered}}{isRegistered}

\item{\code{isQuotaFilled}}{isQuotaFilled}

\item{\code{principalId}}{principalId}

\item{\code{hasConflictingSubmission}}{hasConflictingSubmission}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelMemberSubmissionEligibility in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelMemberSubmissionEligibility
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelMemberSubmissionEligibility
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelMemberSubmissionEligibility in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelMemberSubmissionEligibility
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelMemberSubmissionEligibility
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksEvaluationModelMemberSubmissionEligibility and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksEvaluationModelMemberSubmissionEligibility
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelMemberSubmissionEligibility-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelMemberSubmissionEligibility$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
