% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_submission.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelSubmission}
\alias{OrgSagebionetworksEvaluationModelSubmission}
\title{OrgSagebionetworksEvaluationModelSubmission}
\format{
An \code{R6Class} generator object
}
\description{
A Submission to a Synapse Evaluation is a pointer to a versioned Entity. Submissions are immutable, so we archive a copy of the EntityBundle at the time of submission.

OrgSagebionetworksEvaluationModelSubmission Class
}
\details{
Create a new OrgSagebionetworksEvaluationModelSubmission
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{userId}}{character [optional]}

\item{\code{submitterAlias}}{character [optional]}

\item{\code{evaluationId}}{character [optional]}

\item{\code{evaluationRoundId}}{character [optional]}

\item{\code{entityId}}{character [optional]}

\item{\code{entityBundleJSON}}{character [optional]}

\item{\code{versionNumber}}{integer [optional]}

\item{\code{dockerRepositoryName}}{character [optional]}

\item{\code{dockerDigest}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{teamId}}{character [optional]}

\item{\code{contributors}}{User ids of the submitter and (if a team submission) the team members involved in creating the submission. list(\link{OrgSagebionetworksEvaluationModelSubmissionContributor}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-new}{\code{OrgSagebionetworksEvaluationModelSubmission$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-toJSON}{\code{OrgSagebionetworksEvaluationModelSubmission$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-fromJSON}{\code{OrgSagebionetworksEvaluationModelSubmission$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-toJSONString}{\code{OrgSagebionetworksEvaluationModelSubmission$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-fromJSONString}{\code{OrgSagebionetworksEvaluationModelSubmission$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-validateJSON}{\code{OrgSagebionetworksEvaluationModelSubmission$validateJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-toString}{\code{OrgSagebionetworksEvaluationModelSubmission$toString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-isValid}{\code{OrgSagebionetworksEvaluationModelSubmission$isValid()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-getInvalidFields}{\code{OrgSagebionetworksEvaluationModelSubmission$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-print}{\code{OrgSagebionetworksEvaluationModelSubmission$print()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmission-clone}{\code{OrgSagebionetworksEvaluationModelSubmission$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelSubmission class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$new(
  id = NULL,
  userId = NULL,
  submitterAlias = NULL,
  evaluationId = NULL,
  evaluationRoundId = NULL,
  entityId = NULL,
  entityBundleJSON = NULL,
  versionNumber = NULL,
  dockerRepositoryName = NULL,
  dockerDigest = NULL,
  name = NULL,
  createdOn = NULL,
  teamId = NULL,
  contributors = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{userId}}{userId}

\item{\code{submitterAlias}}{submitterAlias}

\item{\code{evaluationId}}{evaluationId}

\item{\code{evaluationRoundId}}{evaluationRoundId}

\item{\code{entityId}}{entityId}

\item{\code{entityBundleJSON}}{entityBundleJSON}

\item{\code{versionNumber}}{versionNumber}

\item{\code{dockerRepositoryName}}{dockerRepositoryName}

\item{\code{dockerDigest}}{dockerDigest}

\item{\code{name}}{name}

\item{\code{createdOn}}{createdOn}

\item{\code{teamId}}{teamId}

\item{\code{contributors}}{User ids of the submitter and (if a team submission) the team members involved in creating the submission.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksEvaluationModelSubmission and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksEvaluationModelSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmission-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmission-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmission$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
