% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_submission_eligibility.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelSubmissionEligibility}
\alias{OrgSagebionetworksEvaluationModelSubmissionEligibility}
\title{OrgSagebionetworksEvaluationModelSubmissionEligibility}
\format{
An \code{R6Class} generator object
}
\description{
Describes the eligibility of a to submit to an Evaluation queue, reflecting the queue's submission quotas and current submissions.

OrgSagebionetworksEvaluationModelSubmissionEligibility Class
}
\details{
Create a new OrgSagebionetworksEvaluationModelSubmissionEligibility
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isEligible}}{character [optional]}

\item{\code{isRegistered}}{character [optional]}

\item{\code{isQuotaFilled}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-new}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-toJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-fromJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-toJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-fromJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-validateJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$validateJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-toString}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$toString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-isValid}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$isValid()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-getInvalidFields}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-print}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$print()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionEligibility-clone}{\code{OrgSagebionetworksEvaluationModelSubmissionEligibility$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelSubmissionEligibility class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$new(
  isEligible = NULL,
  isRegistered = NULL,
  isQuotaFilled = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isEligible}}{isEligible}

\item{\code{isRegistered}}{isRegistered}

\item{\code{isQuotaFilled}}{isQuotaFilled}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionEligibility in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionEligibility
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionEligibility
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionEligibility in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionEligibility
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionEligibility
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksEvaluationModelSubmissionEligibility and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksEvaluationModelSubmissionEligibility
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionEligibility-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionEligibility-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionEligibility$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
