% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_submission_quota.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelSubmissionQuota}
\alias{OrgSagebionetworksEvaluationModelSubmissionQuota}
\title{OrgSagebionetworksEvaluationModelSubmissionQuota}
\format{
An \code{R6Class} generator object
}
\description{
DEPRECATED: use EvaluationRound APIs instead. No SubmissionQuota will be returned with the Evaluation. Any SubmissionQuotas submitted will be instead be converted into similar EvaluationRounds.

OrgSagebionetworksEvaluationModelSubmissionQuota Class
}
\details{
Create a new OrgSagebionetworksEvaluationModelSubmissionQuota
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{firstRoundStart}}{character [optional]}

\item{\code{roundDurationMillis}}{integer [optional]}

\item{\code{numberOfRounds}}{integer [optional]}

\item{\code{submissionLimit}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-new}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-toJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-fromJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-toJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-fromJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-validateJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$validateJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-toString}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$toString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-isValid}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$isValid()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-getInvalidFields}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-print}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$print()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionQuota-clone}{\code{OrgSagebionetworksEvaluationModelSubmissionQuota$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelSubmissionQuota class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$new(
  firstRoundStart = NULL,
  roundDurationMillis = NULL,
  numberOfRounds = NULL,
  submissionLimit = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{firstRoundStart}}{firstRoundStart}

\item{\code{roundDurationMillis}}{roundDurationMillis}

\item{\code{numberOfRounds}}{numberOfRounds}

\item{\code{submissionLimit}}{submissionLimit}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionQuota in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionQuota
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionQuota
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionQuota in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionQuota
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionQuota
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksEvaluationModelSubmissionQuota and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksEvaluationModelSubmissionQuota
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionQuota-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionQuota-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionQuota$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
