% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_submission_status.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelSubmissionStatus}
\alias{OrgSagebionetworksEvaluationModelSubmissionStatus}
\title{OrgSagebionetworksEvaluationModelSubmissionStatus}
\format{
An \code{R6Class} generator object
}
\description{
A SubmissionStatus is a secondary, mutable object associated with a Submission. This object should be used to contain scoring data about the Submission.

OrgSagebionetworksEvaluationModelSubmissionStatus Class
}
\details{
Create a new OrgSagebionetworksEvaluationModelSubmissionStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{status}}{character [optional]}

\item{\code{score}}{numeric [optional]}

\item{\code{report}}{character [optional]}

\item{\code{annotations}}{\link{OrgSagebionetworksRepoModelAnnotationAnnotations} [optional]}

\item{\code{submissionAnnotations}}{\link{OrgSagebionetworksRepoModelAnnotationV2Annotations} [optional]}

\item{\code{entityId}}{character [optional]}

\item{\code{versionNumber}}{integer [optional]}

\item{\code{statusVersion}}{integer [optional]}

\item{\code{canCancel}}{character [optional]}

\item{\code{cancelRequested}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-new}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-toJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-fromJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-toJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-fromJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-validateJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-toString}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$toString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-isValid}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$isValid()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-getInvalidFields}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-print}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$print()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatus-clone}{\code{OrgSagebionetworksEvaluationModelSubmissionStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelSubmissionStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$new(
  id = NULL,
  etag = NULL,
  modifiedOn = NULL,
  status = NULL,
  score = NULL,
  report = NULL,
  annotations = NULL,
  submissionAnnotations = NULL,
  entityId = NULL,
  versionNumber = NULL,
  statusVersion = NULL,
  canCancel = NULL,
  cancelRequested = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{status}}{status}

\item{\code{score}}{score}

\item{\code{report}}{report}

\item{\code{annotations}}{annotations}

\item{\code{submissionAnnotations}}{submissionAnnotations}

\item{\code{entityId}}{entityId}

\item{\code{versionNumber}}{versionNumber}

\item{\code{statusVersion}}{statusVersion}

\item{\code{canCancel}}{canCancel}

\item{\code{cancelRequested}}{cancelRequested}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksEvaluationModelSubmissionStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksEvaluationModelSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
