% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_submission_status_batch.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelSubmissionStatusBatch}
\alias{OrgSagebionetworksEvaluationModelSubmissionStatusBatch}
\title{OrgSagebionetworksEvaluationModelSubmissionStatusBatch}
\format{
An \code{R6Class} generator object
}
\description{
A batch of status objects, to be updated en masse.

OrgSagebionetworksEvaluationModelSubmissionStatusBatch Class
}
\details{
Create a new OrgSagebionetworksEvaluationModelSubmissionStatusBatch
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{statuses}}{A collection of Submission Statuses list(\link{OrgSagebionetworksEvaluationModelSubmissionStatus}) [optional]}

\item{\code{batchToken}}{character [optional]}

\item{\code{isFirstBatch}}{character [optional]}

\item{\code{isLastBatch}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-new}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-toJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-fromJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-toJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-fromJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-validateJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$validateJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-toString}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$toString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-isValid}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$isValid()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-getInvalidFields}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-print}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$print()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-clone}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelSubmissionStatusBatch class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$new(
  statuses = NULL,
  batchToken = NULL,
  isFirstBatch = NULL,
  isLastBatch = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statuses}}{A collection of Submission Statuses}

\item{\code{batchToken}}{batchToken}

\item{\code{isFirstBatch}}{isFirstBatch}

\item{\code{isLastBatch}}{isLastBatch}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionStatusBatch in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionStatusBatch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionStatusBatch
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionStatusBatch in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionStatusBatch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionStatusBatch
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksEvaluationModelSubmissionStatusBatch and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksEvaluationModelSubmissionStatusBatch
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusBatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusBatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
