% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_evaluation_model_submission_status_enum.R
\docType{class}
\name{OrgSagebionetworksEvaluationModelSubmissionStatusEnum}
\alias{OrgSagebionetworksEvaluationModelSubmissionStatusEnum}
\title{OrgSagebionetworksEvaluationModelSubmissionStatusEnum}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksEvaluationModelSubmissionStatusEnum Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-new}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$new()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-toJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$toJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-fromJSON}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$fromJSON()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-toJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$toJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-fromJSONString}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$fromJSONString()}}
\item \href{#method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-clone}{\code{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksEvaluationModelSubmissionStatusEnum class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionStatusEnum in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionStatusEnum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionStatusEnum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksEvaluationModelSubmissionStatusEnum in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksEvaluationModelSubmissionStatusEnum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksEvaluationModelSubmissionStatusEnum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksEvaluationModelSubmissionStatusEnum-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksEvaluationModelSubmissionStatusEnum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
