% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_access_approval.R
\docType{class}
\name{OrgSagebionetworksRepoModelAccessApproval}
\alias{OrgSagebionetworksRepoModelAccessApproval}
\title{OrgSagebionetworksRepoModelAccessApproval}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for AccessApproval POJO

OrgSagebionetworksRepoModelAccessApproval Class
}
\details{
Create a new OrgSagebionetworksRepoModelAccessApproval
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{requirementId}}{integer [optional]}

\item{\code{requirementVersion}}{integer [optional]}

\item{\code{submitterId}}{character [optional]}

\item{\code{accessorId}}{character [optional]}

\item{\code{expiredOn}}{character [optional]}

\item{\code{state}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-new}{\code{OrgSagebionetworksRepoModelAccessApproval$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-toJSON}{\code{OrgSagebionetworksRepoModelAccessApproval$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-fromJSON}{\code{OrgSagebionetworksRepoModelAccessApproval$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-toJSONString}{\code{OrgSagebionetworksRepoModelAccessApproval$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-fromJSONString}{\code{OrgSagebionetworksRepoModelAccessApproval$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-validateJSON}{\code{OrgSagebionetworksRepoModelAccessApproval$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-toString}{\code{OrgSagebionetworksRepoModelAccessApproval$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-isValid}{\code{OrgSagebionetworksRepoModelAccessApproval$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-getInvalidFields}{\code{OrgSagebionetworksRepoModelAccessApproval$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-print}{\code{OrgSagebionetworksRepoModelAccessApproval$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApproval-clone}{\code{OrgSagebionetworksRepoModelAccessApproval$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAccessApproval class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$new(
  id = NULL,
  etag = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  requirementId = NULL,
  requirementVersion = NULL,
  submitterId = NULL,
  accessorId = NULL,
  expiredOn = NULL,
  state = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{requirementId}}{requirementId}

\item{\code{requirementVersion}}{requirementVersion}

\item{\code{submitterId}}{submitterId}

\item{\code{accessorId}}{accessorId}

\item{\code{expiredOn}}{expiredOn}

\item{\code{state}}{state}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAccessApproval in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAccessApproval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAccessApproval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAccessApproval in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAccessApproval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAccessApproval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAccessApproval and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAccessApproval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApproval-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApproval-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApproval$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
