% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_access_approval_info.R
\docType{class}
\name{OrgSagebionetworksRepoModelAccessApprovalInfo}
\alias{OrgSagebionetworksRepoModelAccessApprovalInfo}
\title{OrgSagebionetworksRepoModelAccessApprovalInfo}
\format{
An \code{R6Class} generator object
}
\description{
Information about whether or not a user has an AccessApproval for an AccessRequirement.

OrgSagebionetworksRepoModelAccessApprovalInfo Class
}
\details{
Create a new OrgSagebionetworksRepoModelAccessApprovalInfo
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{userId}}{character [optional]}

\item{\code{accessRequirementId}}{character [optional]}

\item{\code{hasAccessApproval}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-new}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-toJSON}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-fromJSON}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-toJSONString}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-fromJSONString}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-validateJSON}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-toString}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-isValid}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-getInvalidFields}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-print}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessApprovalInfo-clone}{\code{OrgSagebionetworksRepoModelAccessApprovalInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAccessApprovalInfo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$new(
  userId = NULL,
  accessRequirementId = NULL,
  hasAccessApproval = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userId}}{userId}

\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{hasAccessApproval}}{hasAccessApproval}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAccessApprovalInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAccessApprovalInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAccessApprovalInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAccessApprovalInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAccessApprovalInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAccessApprovalInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAccessApprovalInfo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAccessApprovalInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessApprovalInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessApprovalInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessApprovalInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
