% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_access_control_list.R
\docType{class}
\name{OrgSagebionetworksRepoModelAccessControlList}
\alias{OrgSagebionetworksRepoModelAccessControlList}
\title{OrgSagebionetworksRepoModelAccessControlList}
\format{
An \code{R6Class} generator object
}
\description{
Contains list of principals who can access the data with the allowed types of access for each.

OrgSagebionetworksRepoModelAccessControlList Class
}
\details{
Create a new OrgSagebionetworksRepoModelAccessControlList
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{creationDate}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{resourceAccess}}{The list of principals who can access the data with the allowed types of access for each. list(\link{OrgSagebionetworksRepoModelResourceAccess}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-new}{\code{OrgSagebionetworksRepoModelAccessControlList$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-toJSON}{\code{OrgSagebionetworksRepoModelAccessControlList$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-fromJSON}{\code{OrgSagebionetworksRepoModelAccessControlList$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-toJSONString}{\code{OrgSagebionetworksRepoModelAccessControlList$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-fromJSONString}{\code{OrgSagebionetworksRepoModelAccessControlList$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-validateJSON}{\code{OrgSagebionetworksRepoModelAccessControlList$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-toString}{\code{OrgSagebionetworksRepoModelAccessControlList$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-isValid}{\code{OrgSagebionetworksRepoModelAccessControlList$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-getInvalidFields}{\code{OrgSagebionetworksRepoModelAccessControlList$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-print}{\code{OrgSagebionetworksRepoModelAccessControlList$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAccessControlList-clone}{\code{OrgSagebionetworksRepoModelAccessControlList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAccessControlList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$new(
  id = NULL,
  createdBy = NULL,
  creationDate = NULL,
  modifiedBy = NULL,
  modifiedOn = NULL,
  etag = NULL,
  resourceAccess = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{createdBy}}{createdBy}

\item{\code{creationDate}}{creationDate}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{etag}}{etag}

\item{\code{resourceAccess}}{The list of principals who can access the data with the allowed types of access for each.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAccessControlList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAccessControlList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAccessControlList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAccessControlList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAccessControlList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAccessControlList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAccessControlList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAccessControlList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAccessControlList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAccessControlList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAccessControlList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
