% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_accesstype.R
\docType{class}
\name{OrgSagebionetworksRepoModelACCESSTYPE}
\alias{OrgSagebionetworksRepoModelACCESSTYPE}
\title{OrgSagebionetworksRepoModelACCESSTYPE}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelACCESSTYPE Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelACCESSTYPE-new}{\code{OrgSagebionetworksRepoModelACCESSTYPE$new()}}
\item \href{#method-OrgSagebionetworksRepoModelACCESSTYPE-toJSON}{\code{OrgSagebionetworksRepoModelACCESSTYPE$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelACCESSTYPE-fromJSON}{\code{OrgSagebionetworksRepoModelACCESSTYPE$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelACCESSTYPE-toJSONString}{\code{OrgSagebionetworksRepoModelACCESSTYPE$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelACCESSTYPE-fromJSONString}{\code{OrgSagebionetworksRepoModelACCESSTYPE$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelACCESSTYPE-clone}{\code{OrgSagebionetworksRepoModelACCESSTYPE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelACCESSTYPE-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelACCESSTYPE-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelACCESSTYPE class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelACCESSTYPE$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelACCESSTYPE-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelACCESSTYPE-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelACCESSTYPE$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelACCESSTYPE in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelACCESSTYPE-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelACCESSTYPE-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelACCESSTYPE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelACCESSTYPE$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelACCESSTYPE
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelACCESSTYPE-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelACCESSTYPE-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelACCESSTYPE$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelACCESSTYPE in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelACCESSTYPE-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelACCESSTYPE-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelACCESSTYPE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelACCESSTYPE$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelACCESSTYPE
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelACCESSTYPE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelACCESSTYPE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelACCESSTYPE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
