% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_acquisition_status_names.R
\docType{class}
\name{OrgSagebionetworksRepoModelAcquisitionStatusNames}
\alias{OrgSagebionetworksRepoModelAcquisitionStatusNames}
\title{OrgSagebionetworksRepoModelAcquisitionStatusNames}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAcquisitionStatusNames Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionStatusNames-new}{\code{OrgSagebionetworksRepoModelAcquisitionStatusNames$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionStatusNames-toJSON}{\code{OrgSagebionetworksRepoModelAcquisitionStatusNames$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionStatusNames-fromJSON}{\code{OrgSagebionetworksRepoModelAcquisitionStatusNames$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionStatusNames-toJSONString}{\code{OrgSagebionetworksRepoModelAcquisitionStatusNames$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionStatusNames-fromJSONString}{\code{OrgSagebionetworksRepoModelAcquisitionStatusNames$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionStatusNames-clone}{\code{OrgSagebionetworksRepoModelAcquisitionStatusNames$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionStatusNames-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionStatusNames-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAcquisitionStatusNames class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionStatusNames$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionStatusNames-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionStatusNames-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionStatusNames$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAcquisitionStatusNames in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionStatusNames-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionStatusNames-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAcquisitionStatusNames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionStatusNames$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAcquisitionStatusNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionStatusNames-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionStatusNames-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionStatusNames$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAcquisitionStatusNames in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionStatusNames-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionStatusNames-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAcquisitionStatusNames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionStatusNames$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAcquisitionStatusNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionStatusNames-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionStatusNames-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionStatusNames$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
