% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_acquisition_tracking_data.R
\docType{class}
\name{OrgSagebionetworksRepoModelAcquisitionTrackingData}
\alias{OrgSagebionetworksRepoModelAcquisitionTrackingData}
\title{OrgSagebionetworksRepoModelAcquisitionTrackingData}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for AcquisitionTrackingData

OrgSagebionetworksRepoModelAcquisitionTrackingData Class
}
\details{
Create a new OrgSagebionetworksRepoModelAcquisitionTrackingData
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{status}}{character [optional]}

\item{\code{dataAcquisitionReference}}{character [optional]}

\item{\code{requestor}}{character [optional]}

\item{\code{followupRequirements}}{character [optional]}

\item{\code{comments}}{character [optional]}

\item{\code{history}}{Status history list(\link{OrgSagebionetworksRepoModelStatusHistoryRecord}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-new}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-toJSON}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-fromJSON}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-toJSONString}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-fromJSONString}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-validateJSON}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-toString}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-isValid}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-getInvalidFields}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-print}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAcquisitionTrackingData-clone}{\code{OrgSagebionetworksRepoModelAcquisitionTrackingData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAcquisitionTrackingData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$new(
  status = NULL,
  dataAcquisitionReference = NULL,
  requestor = NULL,
  followupRequirements = NULL,
  comments = NULL,
  history = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{status}

\item{\code{dataAcquisitionReference}}{dataAcquisitionReference}

\item{\code{requestor}}{requestor}

\item{\code{followupRequirements}}{followupRequirements}

\item{\code{comments}}{comments}

\item{\code{history}}{Status history}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAcquisitionTrackingData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAcquisitionTrackingData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAcquisitionTrackingData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAcquisitionTrackingData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAcquisitionTrackingData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAcquisitionTrackingData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAcquisitionTrackingData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAcquisitionTrackingData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAcquisitionTrackingData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAcquisitionTrackingData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAcquisitionTrackingData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
