% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_admin_expire_quarantined_email_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest}
\alias{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest}
\title{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request body used to remove an email address from quarantine

OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-new}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-toJSON}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-fromJSON}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-toJSONString}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-validateJSON}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-toString}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-isValid}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-print}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-clone}{\code{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$new(
  email = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{email}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAdminExpireQuarantinedEmailRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
