% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_annotation_annotation_base.R
\docType{class}
\name{OrgSagebionetworksRepoModelAnnotationAnnotationBase}
\alias{OrgSagebionetworksRepoModelAnnotationAnnotationBase}
\title{OrgSagebionetworksRepoModelAnnotationAnnotationBase}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAnnotationAnnotationBase Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelAnnotationAnnotationBase.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-new}{\code{OrgSagebionetworksRepoModelAnnotationAnnotationBase$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-fromJSONString}{\code{OrgSagebionetworksRepoModelAnnotationAnnotationBase$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-fromJSON}{\code{OrgSagebionetworksRepoModelAnnotationAnnotationBase$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-toJSONString}{\code{OrgSagebionetworksRepoModelAnnotationAnnotationBase$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-toJSON}{\code{OrgSagebionetworksRepoModelAnnotationAnnotationBase$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-validateJSON}{\code{OrgSagebionetworksRepoModelAnnotationAnnotationBase$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-toString}{\code{OrgSagebionetworksRepoModelAnnotationAnnotationBase$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-print}{\code{OrgSagebionetworksRepoModelAnnotationAnnotationBase$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-clone}{\code{OrgSagebionetworksRepoModelAnnotationAnnotationBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAnnotationAnnotationBase.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotationBase$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelAnnotationDoubleAnnotation", "OrgSagebionetworksRepoModelAnnotationLongAnnotation", "OrgSagebionetworksRepoModelAnnotationStringAnnotation"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationAnnotationBase.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotationBase$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAnnotationAnnotationBase.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationAnnotationBase.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotationBase$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAnnotationAnnotationBase.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelAnnotationAnnotationBase to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotationBase$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelAnnotationAnnotationBase.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelAnnotationAnnotationBase to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotationBase$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelAnnotationAnnotationBase.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelAnnotationAnnotationBase and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotationBase$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotationBase$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotationBase$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotationBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotationBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
