% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_annotation_annotations.R
\docType{class}
\name{OrgSagebionetworksRepoModelAnnotationAnnotations}
\alias{OrgSagebionetworksRepoModelAnnotationAnnotations}
\title{OrgSagebionetworksRepoModelAnnotationAnnotations}
\format{
An \code{R6Class} generator object
}
\description{
Primary container object for Annotations on a Synapse object

OrgSagebionetworksRepoModelAnnotationAnnotations Class
}
\details{
Create a new OrgSagebionetworksRepoModelAnnotationAnnotations
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objectId}}{character [optional]}

\item{\code{scopeId}}{character [optional]}

\item{\code{version}}{integer [optional]}

\item{\code{stringAnnos}}{A list of StringAnnnotations list(\link{OrgSagebionetworksRepoModelAnnotationStringAnnotation}) [optional]}

\item{\code{longAnnos}}{A list of LongAnnnotations list(\link{OrgSagebionetworksRepoModelAnnotationLongAnnotation}) [optional]}

\item{\code{doubleAnnos}}{A list of DoubleAnnnotations list(\link{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-new}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-toJSON}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-fromJSON}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-toJSONString}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-fromJSONString}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-validateJSON}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-toString}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-isValid}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-getInvalidFields}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-print}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationAnnotations-clone}{\code{OrgSagebionetworksRepoModelAnnotationAnnotations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAnnotationAnnotations class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$new(
  objectId = NULL,
  scopeId = NULL,
  version = NULL,
  stringAnnos = NULL,
  longAnnos = NULL,
  doubleAnnos = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objectId}}{objectId}

\item{\code{scopeId}}{scopeId}

\item{\code{version}}{version}

\item{\code{stringAnnos}}{A list of StringAnnnotations}

\item{\code{longAnnos}}{A list of LongAnnnotations}

\item{\code{doubleAnnos}}{A list of DoubleAnnnotations}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationAnnotations in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationAnnotations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationAnnotations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationAnnotations in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationAnnotations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationAnnotations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAnnotationAnnotations and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAnnotationAnnotations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationAnnotations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationAnnotations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationAnnotations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
