% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_annotation_double_annotation.R
\docType{class}
\name{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation}
\alias{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation}
\title{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation}
\format{
An \code{R6Class} generator object
}
\description{
An Annotation containing a Double value

OrgSagebionetworksRepoModelAnnotationDoubleAnnotation Class
}
\details{
Create a new OrgSagebionetworksRepoModelAnnotationDoubleAnnotation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{key}}{character [optional]}

\item{\code{isPrivate}}{character [optional]}

\item{\code{value}}{numeric [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-new}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-toJSON}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-fromJSON}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-toJSONString}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-fromJSONString}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-validateJSON}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-toString}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-isValid}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-getInvalidFields}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-print}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-clone}{\code{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAnnotationDoubleAnnotation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$new(
  key = NULL,
  isPrivate = NULL,
  value = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{key}

\item{\code{isPrivate}}{isPrivate}

\item{\code{value}}{value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationDoubleAnnotation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationDoubleAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationDoubleAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationDoubleAnnotation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationDoubleAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationDoubleAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAnnotationDoubleAnnotation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAnnotationDoubleAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationDoubleAnnotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationDoubleAnnotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
