% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_annotation_v2_annotations_value.R
\docType{class}
\name{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue}
\alias{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue}
\title{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue}
\format{
An \code{R6Class} generator object
}
\description{
Contains a AnnotationV2's value and its type

OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue Class
}
\details{
Create a new OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{character [optional]}

\item{\code{value}}{List of values as strings. Clients should use the \"type\" field to determine how to interpret the strings. Single values Annotations are represented as a list of size 1, and multiple values size > 1. Empty list or null is not allowed list(character)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-new}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-toJSON}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-fromJSON}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-toJSONString}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-fromJSONString}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-validateJSON}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-toString}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-isValid}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-getInvalidFields}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-print}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-clone}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$new(
  value,
  type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{List of values as strings. Clients should use the \"type\" field to determine how to interpret the strings. Single values Annotations are represented as a list of size 1, and multiple values size > 1. Empty list or null is not allowed}

\item{\code{type}}{type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
