% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_annotation_v2_annotations_value_type.R
\docType{class}
\name{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType}
\alias{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType}
\title{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-new}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-toJSON}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-fromJSON}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-toJSONString}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-fromJSONString}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-clone}{\code{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2AnnotationsValueType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
