% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_annotation_v2_keys.R
\docType{class}
\name{OrgSagebionetworksRepoModelAnnotationV2Keys}
\alias{OrgSagebionetworksRepoModelAnnotationV2Keys}
\title{OrgSagebionetworksRepoModelAnnotationV2Keys}
\format{
An \code{R6Class} generator object
}
\description{
Array of annotation keys

OrgSagebionetworksRepoModelAnnotationV2Keys Class
}
\details{
Create a new OrgSagebionetworksRepoModelAnnotationV2Keys
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{keys}}{Array of annotation keys list(character)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-new}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-toJSON}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-fromJSON}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-toJSONString}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-fromJSONString}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-validateJSON}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-toString}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-isValid}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-getInvalidFields}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-print}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAnnotationV2Keys-clone}{\code{OrgSagebionetworksRepoModelAnnotationV2Keys$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAnnotationV2Keys class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$new(keys, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{Array of annotation keys}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationV2Keys in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationV2Keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationV2Keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAnnotationV2Keys in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAnnotationV2Keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAnnotationV2Keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAnnotationV2Keys and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAnnotationV2Keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAnnotationV2Keys-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAnnotationV2Keys-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAnnotationV2Keys$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
