% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_asynch_asynch_job_state.R
\docType{class}
\name{OrgSagebionetworksRepoModelAsynchAsynchJobState}
\alias{OrgSagebionetworksRepoModelAsynchAsynchJobState}
\title{OrgSagebionetworksRepoModelAsynchAsynchJobState}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAsynchAsynchJobState Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchJobState-new}{\code{OrgSagebionetworksRepoModelAsynchAsynchJobState$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchJobState-toJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchJobState$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchJobState-fromJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchJobState$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchJobState-toJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchJobState$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchJobState-fromJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchJobState$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchJobState-clone}{\code{OrgSagebionetworksRepoModelAsynchAsynchJobState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchJobState-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchJobState-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAsynchAsynchJobState class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchJobState$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchJobState-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchJobState-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchJobState$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAsynchAsynchJobState in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchJobState-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchJobState-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchJobState
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchJobState$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAsynchAsynchJobState
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchJobState-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchJobState-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchJobState$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAsynchAsynchJobState in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchJobState-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchJobState-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchJobState
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchJobState$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAsynchAsynchJobState
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchJobState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchJobState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchJobState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
