% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_asynch_asynchronous_admin_request_body.R
\docType{class}
\name{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody}
\alias{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody}
\title{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-new}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-fromJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-fromJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-toJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-toJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-validateJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-toString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-print}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-clone}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$new(
  instance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelFileFileHandleArchivalRequest", "OrgSagebionetworksRepoModelMigrationAsyncMigrationRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousAdminRequestBody$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
