% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_asynch_asynchronous_job_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus}
\alias{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus}
\title{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus}
\format{
An \code{R6Class} generator object
}
\description{
Object used to track the status of an Asynchronous Job.

OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{jobState}}{character [optional]}

\item{\code{jobCanceling}}{character [optional]}

\item{\code{requestBody}}{\link{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody} [optional]}

\item{\code{responseBody}}{\link{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody} [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{jobId}}{character [optional]}

\item{\code{startedByUserId}}{integer [optional]}

\item{\code{startedOn}}{character [optional]}

\item{\code{changedOn}}{character [optional]}

\item{\code{progressMessage}}{character [optional]}

\item{\code{progressCurrent}}{integer [optional]}

\item{\code{progressTotal}}{integer [optional]}

\item{\code{exception}}{character [optional]}

\item{\code{errorMessage}}{character [optional]}

\item{\code{errorDetails}}{character [optional]}

\item{\code{runtimeMS}}{integer [optional]}

\item{\code{callersContext}}{\link{OrgSagebionetworksRepoModelAuthCallersContext} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-new}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-toJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-fromJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-toJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-validateJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-toString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-isValid}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-print}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-clone}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$new(
  jobState = NULL,
  jobCanceling = NULL,
  requestBody = NULL,
  responseBody = NULL,
  etag = NULL,
  jobId = NULL,
  startedByUserId = NULL,
  startedOn = NULL,
  changedOn = NULL,
  progressMessage = NULL,
  progressCurrent = NULL,
  progressTotal = NULL,
  exception = NULL,
  errorMessage = NULL,
  errorDetails = NULL,
  runtimeMS = NULL,
  callersContext = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{jobState}}{jobState}

\item{\code{jobCanceling}}{jobCanceling}

\item{\code{requestBody}}{requestBody}

\item{\code{responseBody}}{responseBody}

\item{\code{etag}}{etag}

\item{\code{jobId}}{jobId}

\item{\code{startedByUserId}}{startedByUserId}

\item{\code{startedOn}}{startedOn}

\item{\code{changedOn}}{changedOn}

\item{\code{progressMessage}}{progressMessage}

\item{\code{progressCurrent}}{progressCurrent}

\item{\code{progressTotal}}{progressTotal}

\item{\code{exception}}{exception}

\item{\code{errorMessage}}{errorMessage}

\item{\code{errorDetails}}{errorDetails}

\item{\code{runtimeMS}}{runtimeMS}

\item{\code{callersContext}}{callersContext}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousJobStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
