% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_asynch_asynchronous_request_body.R
\docType{class}
\name{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody}
\alias{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody}
\title{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-new}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-fromJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-fromJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-toJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-toJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-validateJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-toString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-print}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-clone}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelDoiV2DoiRequest", "OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest", "OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest", "OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest", "OrgSagebionetworksRepoModelDownloadDownloadListQueryRequest", "OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest", "OrgSagebionetworksRepoModelFileBulkFileDownloadRequest", "OrgSagebionetworksRepoModelFileFileHandleArchivalRequest", "OrgSagebionetworksRepoModelFileFileHandleRestoreRequest", "OrgSagebionetworksRepoModelMigrationAsyncMigrationRequest", "OrgSagebionetworksRepoModelReportDownloadStorageReportRequest", "OrgSagebionetworksRepoModelSchemaCreateSchemaRequest", "OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest", "OrgSagebionetworksRepoModelTableAppendableRowSetRequest", "OrgSagebionetworksRepoModelTableDownloadFromTableRequest", "OrgSagebionetworksRepoModelTableQueryBundleRequest", "OrgSagebionetworksRepoModelTableQueryNextPageToken", "OrgSagebionetworksRepoModelTableTableSchemaChangeRequest", "OrgSagebionetworksRepoModelTableTableSearchChangeRequest", "OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest", "OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest", "OrgSagebionetworksRepoModelTableUploadToTableRequest", "OrgSagebionetworksRepoModelTableViewColumnModelRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousRequestBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
