% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_asynch_asynchronous_response_body.R
\docType{class}
\name{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody}
\alias{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody}
\title{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-new}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-fromJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-fromJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-toJSONString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-toJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-validateJSON}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-toString}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-print}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-clone}{\code{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelDoiV2DoiResponse", "OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse", "OrgSagebionetworksRepoModelDownloadDownloadListManifestResponse", "OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse", "OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse", "OrgSagebionetworksRepoModelFileAddFileToDownloadListResponse", "OrgSagebionetworksRepoModelFileBulkFileDownloadResponse", "OrgSagebionetworksRepoModelFileFileHandleArchivalResponse", "OrgSagebionetworksRepoModelFileFileHandleRestoreResponse", "OrgSagebionetworksRepoModelMigrationAsyncMigrationResponse", "OrgSagebionetworksRepoModelReportDownloadStorageReportResponse", "OrgSagebionetworksRepoModelSchemaCreateSchemaResponse", "OrgSagebionetworksRepoModelSchemaGetValidationSchemaResponse", "OrgSagebionetworksRepoModelTableDownloadFromTableResult", "OrgSagebionetworksRepoModelTableEntityUpdateResults", "OrgSagebionetworksRepoModelTableQueryResult", "OrgSagebionetworksRepoModelTableQueryResultBundle", "OrgSagebionetworksRepoModelTableRowReferenceSetResults", "OrgSagebionetworksRepoModelTableTableSchemaChangeResponse", "OrgSagebionetworksRepoModelTableTableSearchChangeResponse", "OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse", "OrgSagebionetworksRepoModelTableUploadToTablePreviewResult", "OrgSagebionetworksRepoModelTableUploadToTableResult", "OrgSagebionetworksRepoModelTableViewColumnModelResponse"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchAsynchronousResponseBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
