% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_asynch_cacheable_request_body.R
\docType{class}
\name{OrgSagebionetworksRepoModelAsynchCacheableRequestBody}
\alias{OrgSagebionetworksRepoModelAsynchCacheableRequestBody}
\title{OrgSagebionetworksRepoModelAsynchCacheableRequestBody}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAsynchCacheableRequestBody Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelAsynchCacheableRequestBody.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-new}{\code{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-fromJSONString}{\code{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-fromJSON}{\code{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-toJSONString}{\code{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-toJSON}{\code{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-validateJSON}{\code{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-toString}{\code{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-print}{\code{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-clone}{\code{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAsynchCacheableRequestBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelTableDownloadFromTableRequest", "OrgSagebionetworksRepoModelTableQueryBundleRequest", "OrgSagebionetworksRepoModelTableQueryNextPageToken"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchCacheableRequestBody.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchCacheableRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchCacheableRequestBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchCacheableRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelAsynchCacheableRequestBody to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelAsynchCacheableRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelAsynchCacheableRequestBody to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelAsynchCacheableRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelAsynchCacheableRequestBody and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchCacheableRequestBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchCacheableRequestBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
