% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_asynch_read_only_request_body.R
\docType{class}
\name{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody}
\alias{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody}
\title{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-new}{\code{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-fromJSONString}{\code{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-fromJSON}{\code{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-toJSONString}{\code{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-toJSON}{\code{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-validateJSON}{\code{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-toString}{\code{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-print}{\code{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-clone}{\code{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelMigrationAsyncMigrationRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAsynchReadOnlyRequestBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
