% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_attachment_presigned_url.R
\docType{class}
\name{OrgSagebionetworksRepoModelAttachmentPresignedUrl}
\alias{OrgSagebionetworksRepoModelAttachmentPresignedUrl}
\title{OrgSagebionetworksRepoModelAttachmentPresignedUrl}
\format{
An \code{R6Class} generator object
}
\description{
This object is deprecated and will be removed in future versions of Synapse.

OrgSagebionetworksRepoModelAttachmentPresignedUrl Class
}
\details{
Create a new OrgSagebionetworksRepoModelAttachmentPresignedUrl
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tokenID}}{character [optional]}

\item{\code{presignedUrl}}{character [optional]}

\item{\code{status}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-new}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-toJSON}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-fromJSON}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-toJSONString}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-fromJSONString}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-validateJSON}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-toString}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-isValid}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-getInvalidFields}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-print}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-clone}{\code{OrgSagebionetworksRepoModelAttachmentPresignedUrl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAttachmentPresignedUrl class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$new(
  tokenID = NULL,
  presignedUrl = NULL,
  status = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tokenID}}{tokenID}

\item{\code{presignedUrl}}{presignedUrl}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAttachmentPresignedUrl in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAttachmentPresignedUrl
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAttachmentPresignedUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAttachmentPresignedUrl in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAttachmentPresignedUrl
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAttachmentPresignedUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAttachmentPresignedUrl and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAttachmentPresignedUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAttachmentPresignedUrl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAttachmentPresignedUrl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
