% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_audit_access_record.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuditAccessRecord}
\alias{OrgSagebionetworksRepoModelAuditAccessRecord}
\title{OrgSagebionetworksRepoModelAuditAccessRecord}
\format{
An \code{R6Class} generator object
}
\description{
Data recored from a single web-services methods.

OrgSagebionetworksRepoModelAuditAccessRecord Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuditAccessRecord
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sessionId}}{character [optional]}

\item{\code{timestamp}}{integer [optional]}

\item{\code{userId}}{integer [optional]}

\item{\code{method}}{character [optional]}

\item{\code{requestURL}}{character [optional]}

\item{\code{userAgent}}{character [optional]}

\item{\code{host}}{character [optional]}

\item{\code{origin}}{character [optional]}

\item{\code{xForwardedFor}}{character [optional]}

\item{\code{via}}{character [optional]}

\item{\code{threadId}}{integer [optional]}

\item{\code{elapseMS}}{integer [optional]}

\item{\code{success}}{character [optional]}

\item{\code{stack}}{character [optional]}

\item{\code{instance}}{character [optional]}

\item{\code{date}}{character [optional]}

\item{\code{vmId}}{character [optional]}

\item{\code{returnObjectId}}{character [optional]}

\item{\code{queryString}}{character [optional]}

\item{\code{responseStatus}}{integer [optional]}

\item{\code{oauthClientId}}{character [optional]}

\item{\code{basicAuthUsername}}{character [optional]}

\item{\code{authenticationMethod}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-new}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-toJSON}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-fromJSON}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-toJSONString}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-fromJSONString}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-validateJSON}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-toString}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-isValid}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-print}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditAccessRecord-clone}{\code{OrgSagebionetworksRepoModelAuditAccessRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuditAccessRecord class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$new(
  sessionId = NULL,
  timestamp = NULL,
  userId = NULL,
  method = NULL,
  requestURL = NULL,
  userAgent = NULL,
  host = NULL,
  origin = NULL,
  xForwardedFor = NULL,
  via = NULL,
  threadId = NULL,
  elapseMS = NULL,
  success = NULL,
  stack = NULL,
  instance = NULL,
  date = NULL,
  vmId = NULL,
  returnObjectId = NULL,
  queryString = NULL,
  responseStatus = NULL,
  oauthClientId = NULL,
  basicAuthUsername = NULL,
  authenticationMethod = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sessionId}}{sessionId}

\item{\code{timestamp}}{timestamp}

\item{\code{userId}}{userId}

\item{\code{method}}{method}

\item{\code{requestURL}}{requestURL}

\item{\code{userAgent}}{userAgent}

\item{\code{host}}{host}

\item{\code{origin}}{origin}

\item{\code{xForwardedFor}}{xForwardedFor}

\item{\code{via}}{via}

\item{\code{threadId}}{threadId}

\item{\code{elapseMS}}{elapseMS}

\item{\code{success}}{success}

\item{\code{stack}}{stack}

\item{\code{instance}}{instance}

\item{\code{date}}{date}

\item{\code{vmId}}{vmId}

\item{\code{returnObjectId}}{returnObjectId}

\item{\code{queryString}}{queryString}

\item{\code{responseStatus}}{responseStatus}

\item{\code{oauthClientId}}{oauthClientId}

\item{\code{basicAuthUsername}}{basicAuthUsername}

\item{\code{authenticationMethod}}{authenticationMethod}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuditAccessRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuditAccessRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuditAccessRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuditAccessRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuditAccessRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuditAccessRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuditAccessRecord and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuditAccessRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditAccessRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditAccessRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditAccessRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
