% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_audit_file_handle_snapshot.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuditFileHandleSnapshot}
\alias{OrgSagebionetworksRepoModelAuditFileHandleSnapshot}
\title{OrgSagebionetworksRepoModelAuditFileHandleSnapshot}
\format{
An \code{R6Class} generator object
}
\description{
The FileHandleSnapshot captures all of the fields that are common to all FileHandle implementations.

OrgSagebionetworksRepoModelAuditFileHandleSnapshot Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuditFileHandleSnapshot
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{contentMd5}}{character [optional]}

\item{\code{contentType}}{character [optional]}

\item{\code{fileName}}{character [optional]}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{contentSize}}{integer [optional]}

\item{\code{key}}{character [optional]}

\item{\code{bucket}}{character [optional]}

\item{\code{previewId}}{character [optional]}

\item{\code{isPreview}}{character [optional]}

\item{\code{status}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-new}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-toJSON}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-fromJSON}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-toJSONString}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-fromJSONString}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-validateJSON}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-toString}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-isValid}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-print}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-clone}{\code{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuditFileHandleSnapshot class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$new(
  concreteType,
  id = NULL,
  createdBy = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  contentMd5 = NULL,
  contentType = NULL,
  fileName = NULL,
  storageLocationId = NULL,
  contentSize = NULL,
  key = NULL,
  bucket = NULL,
  previewId = NULL,
  isPreview = NULL,
  status = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{id}}{id}

\item{\code{createdBy}}{createdBy}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{contentMd5}}{contentMd5}

\item{\code{contentType}}{contentType}

\item{\code{fileName}}{fileName}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{contentSize}}{contentSize}

\item{\code{key}}{key}

\item{\code{bucket}}{bucket}

\item{\code{previewId}}{previewId}

\item{\code{isPreview}}{isPreview}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuditFileHandleSnapshot in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuditFileHandleSnapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuditFileHandleSnapshot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuditFileHandleSnapshot in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuditFileHandleSnapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuditFileHandleSnapshot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuditFileHandleSnapshot and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuditFileHandleSnapshot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuditFileHandleSnapshot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuditFileHandleSnapshot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
