% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_access_token_record.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthAccessTokenRecord}
\alias{OrgSagebionetworksRepoModelAuthAccessTokenRecord}
\title{OrgSagebionetworksRepoModelAuthAccessTokenRecord}
\format{
An \code{R6Class} generator object
}
\description{
Metadata related to an issued personal access token. After the token has been generated, it cannot be retrieved.

OrgSagebionetworksRepoModelAuthAccessTokenRecord Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthAccessTokenRecord
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{userId}}{character [optional]}

\item{\code{scopes}}{The scopes that have been granted to this token list(character) [optional]}

\item{\code{userInfoClaims}}{The OIDC claims that can be accessed using this token. named list(\link{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails}) [optional]}

\item{\code{name}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{lastUsed}}{character [optional]}

\item{\code{state}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-new}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-toJSON}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-fromJSON}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-toJSONString}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-validateJSON}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-toString}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-isValid}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-print}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-clone}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthAccessTokenRecord class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$new(
  id = NULL,
  userId = NULL,
  scopes = NULL,
  userInfoClaims = NULL,
  name = NULL,
  createdOn = NULL,
  lastUsed = NULL,
  state = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{userId}}{userId}

\item{\code{scopes}}{The scopes that have been granted to this token}

\item{\code{userInfoClaims}}{The OIDC claims that can be accessed using this token.}

\item{\code{name}}{name}

\item{\code{createdOn}}{createdOn}

\item{\code{lastUsed}}{lastUsed}

\item{\code{state}}{state}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthAccessTokenRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthAccessTokenRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthAccessTokenRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthAccessTokenRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthAccessTokenRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthAccessTokenRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthAccessTokenRecord and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthAccessTokenRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
