% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_access_token_record_list.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthAccessTokenRecordList}
\alias{OrgSagebionetworksRepoModelAuthAccessTokenRecordList}
\title{OrgSagebionetworksRepoModelAuthAccessTokenRecordList}
\format{
An \code{R6Class} generator object
}
\description{
Metadata for a paginated list of personal access token records.

OrgSagebionetworksRepoModelAuthAccessTokenRecordList Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthAccessTokenRecordList
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelAuthAccessTokenRecord}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-new}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-toJSON}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-fromJSON}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-toJSONString}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-validateJSON}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-toString}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-isValid}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-print}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-clone}{\code{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthAccessTokenRecordList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$new(
  results = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{results}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthAccessTokenRecordList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthAccessTokenRecordList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthAccessTokenRecordList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthAccessTokenRecordList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthAccessTokenRecordList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthAccessTokenRecordList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthAccessTokenRecordList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthAccessTokenRecordList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthAccessTokenRecordList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthAccessTokenRecordList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
