% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_change_password_interface.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthChangePasswordInterface}
\alias{OrgSagebionetworksRepoModelAuthChangePasswordInterface}
\title{OrgSagebionetworksRepoModelAuthChangePasswordInterface}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAuthChangePasswordInterface Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelAuthChangePasswordInterface.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-new}{\code{OrgSagebionetworksRepoModelAuthChangePasswordInterface$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordInterface$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-fromJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordInterface$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-toJSONString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordInterface$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-toJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordInterface$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-validateJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordInterface$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-toString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordInterface$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-print}{\code{OrgSagebionetworksRepoModelAuthChangePasswordInterface$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-clone}{\code{OrgSagebionetworksRepoModelAuthChangePasswordInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthChangePasswordInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordInterface$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword", "OrgSagebionetworksRepoModelAuthChangePasswordWithToken", "OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthChangePasswordInterface.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordInterface$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAuthChangePasswordInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthChangePasswordInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordInterface$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAuthChangePasswordInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelAuthChangePasswordInterface to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordInterface$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelAuthChangePasswordInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelAuthChangePasswordInterface to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordInterface$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelAuthChangePasswordInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelAuthChangePasswordInterface and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordInterface$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordInterface$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordInterface$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordInterface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordInterface$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
