% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_change_password_with_current_password.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword}
\alias{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword}
\title{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword}
\format{
An \code{R6Class} generator object
}
\description{
Change user account's password by verifying the current password

OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{newPassword}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{username}}{character [optional]}

\item{\code{currentPassword}}{character [optional]}

\item{\code{authenticationReceipt}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-new}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-toJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-fromJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-toJSONString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-validateJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-toString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-isValid}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-print}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-clone}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$new(
  concreteType,
  newPassword = NULL,
  username = NULL,
  currentPassword = NULL,
  authenticationReceipt = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{newPassword}}{newPassword}

\item{\code{username}}{username}

\item{\code{currentPassword}}{currentPassword}

\item{\code{authenticationReceipt}}{authenticationReceipt}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithCurrentPassword$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
