% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_change_password_with_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthChangePasswordWithToken}
\alias{OrgSagebionetworksRepoModelAuthChangePasswordWithToken}
\title{OrgSagebionetworksRepoModelAuthChangePasswordWithToken}
\format{
An \code{R6Class} generator object
}
\description{
Change user account's password by providing a token issued by the server

OrgSagebionetworksRepoModelAuthChangePasswordWithToken Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthChangePasswordWithToken
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{newPassword}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{passwordChangeToken}}{\link{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-new}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-toJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-fromJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-toJSONString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-validateJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-toString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-isValid}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-print}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-clone}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthChangePasswordWithToken class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$new(
  concreteType,
  newPassword = NULL,
  passwordChangeToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{newPassword}}{newPassword}

\item{\code{passwordChangeToken}}{passwordChangeToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthChangePasswordWithToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthChangePasswordWithToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthChangePasswordWithToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthChangePasswordWithToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthChangePasswordWithToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthChangePasswordWithToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthChangePasswordWithToken and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthChangePasswordWithToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
