% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_change_password_with_two_factor_auth_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken}
\alias{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken}
\title{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken}
\format{
An \code{R6Class} generator object
}
\description{
Used to change the user password when two factor authentication is enabled. The twoFaToken is generated by the server and included in the 401 response the user tries to update the password with either the old password or a token. Only TOTP and RECOVERY_CODE otp codes can be used for this operation.

OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{newPassword}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{userId}}{integer [optional]}

\item{\code{twoFaToken}}{character [optional]}

\item{\code{otpCode}}{character [optional]}

\item{\code{otpType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-new}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-toJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-fromJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-toJSONString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-validateJSON}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-toString}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-isValid}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-print}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-clone}{\code{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$new(
  concreteType,
  newPassword = NULL,
  userId = NULL,
  twoFaToken = NULL,
  otpCode = NULL,
  otpType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{newPassword}}{newPassword}

\item{\code{userId}}{userId}

\item{\code{twoFaToken}}{twoFaToken}

\item{\code{otpCode}}{otpCode}

\item{\code{otpType}}{otpType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
