% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_has_two_factor_auth_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken}
\alias{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken}
\title{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-new}{\code{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-fromJSON}{\code{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-toJSONString}{\code{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-toJSON}{\code{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-validateJSON}{\code{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-toString}{\code{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-print}{\code{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-clone}{\code{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelAuthChangePasswordWithTwoFactorAuthToken", "OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthHasTwoFactorAuthToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
