% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_new_integration_test_user.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser}
\alias{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser}
\title{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser}
\format{
An \code{R6Class} generator object
}
\description{
Administrative object.  Used to create users for integration tests.  All fields besides username can be null

OrgSagebionetworksRepoModelAuthNewIntegrationTestUser Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthNewIntegrationTestUser
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{username}}{character [optional]}

\item{\code{email}}{character [optional]}

\item{\code{password}}{character [optional]}

\item{\code{tou}}{character [optional]}

\item{\code{validatedUser}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-new}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-toJSON}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-fromJSON}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-toJSONString}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-validateJSON}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-toString}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-isValid}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-print}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-clone}{\code{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthNewIntegrationTestUser class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$new(
  username = NULL,
  email = NULL,
  password = NULL,
  tou = NULL,
  validatedUser = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{username}

\item{\code{email}}{email}

\item{\code{password}}{password}

\item{\code{tou}}{tou}

\item{\code{validatedUser}}{validatedUser}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthNewIntegrationTestUser in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthNewIntegrationTestUser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthNewIntegrationTestUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthNewIntegrationTestUser in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthNewIntegrationTestUser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthNewIntegrationTestUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthNewIntegrationTestUser and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthNewIntegrationTestUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewIntegrationTestUser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewIntegrationTestUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
