% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_auth_new_user.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthNewUser}
\alias{OrgSagebionetworksRepoModelAuthNewUser}
\title{OrgSagebionetworksRepoModelAuthNewUser}
\format{
An \code{R6Class} generator object
}
\description{
Fields required to create a new user account

OrgSagebionetworksRepoModelAuthNewUser Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthNewUser
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character [optional]}

\item{\code{firstName}}{character [optional]}

\item{\code{lastName}}{character [optional]}

\item{\code{userName}}{character [optional]}

\item{\code{oauthProvider}}{character [optional]}

\item{\code{subject}}{character [optional]}

\item{\code{encodedMembershipInvtnSignedToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-new}{\code{OrgSagebionetworksRepoModelAuthNewUser$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-toJSON}{\code{OrgSagebionetworksRepoModelAuthNewUser$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-fromJSON}{\code{OrgSagebionetworksRepoModelAuthNewUser$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-toJSONString}{\code{OrgSagebionetworksRepoModelAuthNewUser$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthNewUser$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-validateJSON}{\code{OrgSagebionetworksRepoModelAuthNewUser$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-toString}{\code{OrgSagebionetworksRepoModelAuthNewUser$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-isValid}{\code{OrgSagebionetworksRepoModelAuthNewUser$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthNewUser$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-print}{\code{OrgSagebionetworksRepoModelAuthNewUser$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthNewUser-clone}{\code{OrgSagebionetworksRepoModelAuthNewUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthNewUser class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$new(
  email = NULL,
  firstName = NULL,
  lastName = NULL,
  userName = NULL,
  oauthProvider = NULL,
  subject = NULL,
  encodedMembershipInvtnSignedToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{email}

\item{\code{firstName}}{firstName}

\item{\code{lastName}}{lastName}

\item{\code{userName}}{userName}

\item{\code{oauthProvider}}{oauthProvider}

\item{\code{subject}}{subject}

\item{\code{encodedMembershipInvtnSignedToken}}{encodedMembershipInvtnSignedToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthNewUser in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthNewUser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthNewUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthNewUser in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthNewUser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthNewUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthNewUser and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthNewUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthNewUser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthNewUser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthNewUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
