% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_password_reset_signed_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken}
\alias{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken}
\title{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken}
\format{
An \code{R6Class} generator object
}
\description{
Signed token containing the the information necessary to reset the password for a user.

OrgSagebionetworksRepoModelAuthPasswordResetSignedToken Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthPasswordResetSignedToken
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{hmac}}{character [optional]}

\item{\code{version}}{integer [optional]}

\item{\code{expiresOn}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{userId}}{character [optional]}

\item{\code{validity}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-new}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-toJSON}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-fromJSON}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-toJSONString}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-validateJSON}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-toString}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-isValid}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-print}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-clone}{\code{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthPasswordResetSignedToken class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$new(
  concreteType,
  hmac = NULL,
  version = NULL,
  expiresOn = NULL,
  createdOn = NULL,
  userId = NULL,
  validity = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{hmac}}{hmac}

\item{\code{version}}{version}

\item{\code{expiresOn}}{expiresOn}

\item{\code{createdOn}}{createdOn}

\item{\code{userId}}{userId}

\item{\code{validity}}{validity}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthPasswordResetSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthPasswordResetSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthPasswordResetSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthPasswordResetSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthPasswordResetSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthPasswordResetSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthPasswordResetSignedToken and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthPasswordResetSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthPasswordResetSignedToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthPasswordResetSignedToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
