% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_totp_secret.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthTotpSecret}
\alias{OrgSagebionetworksRepoModelAuthTotpSecret}
\title{OrgSagebionetworksRepoModelAuthTotpSecret}
\format{
An \code{R6Class} generator object
}
\description{
The generated shared secret to enable two factor authentication with TOTP compatible authenticator app. The secret is not active.

OrgSagebionetworksRepoModelAuthTotpSecret Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthTotpSecret
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{secretId}}{character [optional]}

\item{\code{secret}}{character [optional]}

\item{\code{alg}}{character [optional]}

\item{\code{digits}}{integer [optional]}

\item{\code{period}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-new}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-toJSON}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-fromJSON}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-toJSONString}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-validateJSON}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-toString}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-isValid}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-print}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecret-clone}{\code{OrgSagebionetworksRepoModelAuthTotpSecret$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthTotpSecret class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$new(
  secretId = NULL,
  secret = NULL,
  alg = NULL,
  digits = NULL,
  period = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{secretId}}{secretId}

\item{\code{secret}}{secret}

\item{\code{alg}}{alg}

\item{\code{digits}}{digits}

\item{\code{period}}{period}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTotpSecret in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTotpSecret
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTotpSecret
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTotpSecret in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTotpSecret
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTotpSecret
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthTotpSecret and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthTotpSecret
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecret-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecret-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecret$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
