% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_totp_secret_activation_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest}
\alias{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest}
\title{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest}
\format{
An \code{R6Class} generator object
}
\description{
Activate the TOTP secret for the user to be used in two factor authentication

OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{secretId}}{character [optional]}

\item{\code{totp}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-new}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-toJSON}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-fromJSON}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-toJSONString}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-validateJSON}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-toString}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-isValid}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-print}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-clone}{\code{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$new(
  secretId = NULL,
  totp = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{secretId}}{secretId}

\item{\code{totp}}{totp}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTotpSecretActivationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
