% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_two_factor_auth_login_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest}
\alias{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest}
\title{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest}
\format{
An \code{R6Class} generator object
}
\description{
Used to login when two factor authentication is enabled.

OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{userId}}{integer [optional]}

\item{\code{twoFaToken}}{character [optional]}

\item{\code{otpCode}}{character [optional]}

\item{\code{otpType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-new}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-toJSON}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-fromJSON}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-toJSONString}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-validateJSON}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-toString}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-isValid}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-print}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-clone}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$new(
  userId = NULL,
  twoFaToken = NULL,
  otpCode = NULL,
  otpType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userId}}{userId}

\item{\code{twoFaToken}}{twoFaToken}

\item{\code{otpCode}}{otpCode}

\item{\code{otpType}}{otpType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthLoginRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
