% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_two_factor_auth_otp_type.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType}
\alias{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType}
\title{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-new}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-toJSON}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-fromJSON}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-toJSONString}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-clone}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthOtpType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
