% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_auth_two_factor_auth_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus}
\alias{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus}
\title{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus}
\format{
An \code{R6Class} generator object
}
\description{
Status of 2FA for the user

OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{status}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-new}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-toJSON}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-fromJSON}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-toJSONString}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-validateJSON}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-toString}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-isValid}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-print}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-clone}{\code{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$new(status = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelAuthTwoFactorAuthStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
