% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_batch_access_approval_info_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest}
\alias{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest}
\title{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request for a batch of AccessApprovalInfo.

OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{userId}}{character [optional]}

\item{\code{accessRequirementIds}}{The ID of the access requirements. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-new}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-toJSON}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-fromJSON}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-toJSONString}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-validateJSON}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-toString}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-isValid}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-print}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-clone}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$new(
  userId = NULL,
  accessRequirementIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userId}}{userId}

\item{\code{accessRequirementIds}}{The ID of the access requirements.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
