% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_batch_access_approval_info_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse}
\alias{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse}
\title{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse}
\format{
An \code{R6Class} generator object
}
\description{
Response for an AccessApprovalInfoRequest.

OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{The list of resulting AccessApprovalInfo. list(\link{OrgSagebionetworksRepoModelAccessApprovalInfo}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-new}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-toJSON}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-fromJSON}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-toJSONString}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-validateJSON}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-toString}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-isValid}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-print}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-clone}{\code{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$new(
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{The list of resulting AccessApprovalInfo.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelBatchAccessApprovalInfoResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
