% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_challenge_team.R
\docType{class}
\name{OrgSagebionetworksRepoModelChallengeTeam}
\alias{OrgSagebionetworksRepoModelChallengeTeam}
\title{OrgSagebionetworksRepoModelChallengeTeam}
\format{
An \code{R6Class} generator object
}
\description{
Relationship between a Challenge and registered Team

OrgSagebionetworksRepoModelChallengeTeam Class
}
\details{
Create a new OrgSagebionetworksRepoModelChallengeTeam
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{challengeId}}{character [optional]}

\item{\code{teamId}}{character [optional]}

\item{\code{message}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-new}{\code{OrgSagebionetworksRepoModelChallengeTeam$new()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-toJSON}{\code{OrgSagebionetworksRepoModelChallengeTeam$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-fromJSON}{\code{OrgSagebionetworksRepoModelChallengeTeam$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-toJSONString}{\code{OrgSagebionetworksRepoModelChallengeTeam$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-fromJSONString}{\code{OrgSagebionetworksRepoModelChallengeTeam$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-validateJSON}{\code{OrgSagebionetworksRepoModelChallengeTeam$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-toString}{\code{OrgSagebionetworksRepoModelChallengeTeam$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-isValid}{\code{OrgSagebionetworksRepoModelChallengeTeam$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-getInvalidFields}{\code{OrgSagebionetworksRepoModelChallengeTeam$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-print}{\code{OrgSagebionetworksRepoModelChallengeTeam$print()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeam-clone}{\code{OrgSagebionetworksRepoModelChallengeTeam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelChallengeTeam class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$new(
  id = NULL,
  etag = NULL,
  challengeId = NULL,
  teamId = NULL,
  message = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{challengeId}}{challengeId}

\item{\code{teamId}}{teamId}

\item{\code{message}}{message}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelChallengeTeam in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelChallengeTeam
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelChallengeTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelChallengeTeam in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelChallengeTeam
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelChallengeTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelChallengeTeam and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelChallengeTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
