% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_challenge_team_paged_results.R
\docType{class}
\name{OrgSagebionetworksRepoModelChallengeTeamPagedResults}
\alias{OrgSagebionetworksRepoModelChallengeTeamPagedResults}
\title{OrgSagebionetworksRepoModelChallengeTeamPagedResults}
\format{
An \code{R6Class} generator object
}
\description{
A paginated list of ChallengeTeam objects

OrgSagebionetworksRepoModelChallengeTeamPagedResults Class
}
\details{
Create a new OrgSagebionetworksRepoModelChallengeTeamPagedResults
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{The list of results for this page list(\link{OrgSagebionetworksRepoModelChallengeTeam}) [optional]}

\item{\code{totalNumberOfResults}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-new}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$new()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-toJSON}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-fromJSON}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-toJSONString}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-fromJSONString}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-validateJSON}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-toString}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-isValid}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-getInvalidFields}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-print}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$print()}}
\item \href{#method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-clone}{\code{OrgSagebionetworksRepoModelChallengeTeamPagedResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelChallengeTeamPagedResults class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$new(
  results = NULL,
  totalNumberOfResults = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{The list of results for this page}

\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelChallengeTeamPagedResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelChallengeTeamPagedResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelChallengeTeamPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelChallengeTeamPagedResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelChallengeTeamPagedResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelChallengeTeamPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelChallengeTeamPagedResults and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelChallengeTeamPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelChallengeTeamPagedResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelChallengeTeamPagedResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
