% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_curation_status_names.R
\docType{class}
\name{OrgSagebionetworksRepoModelCurationStatusNames}
\alias{OrgSagebionetworksRepoModelCurationStatusNames}
\title{OrgSagebionetworksRepoModelCurationStatusNames}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelCurationStatusNames Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelCurationStatusNames-new}{\code{OrgSagebionetworksRepoModelCurationStatusNames$new()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationStatusNames-toJSON}{\code{OrgSagebionetworksRepoModelCurationStatusNames$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationStatusNames-fromJSON}{\code{OrgSagebionetworksRepoModelCurationStatusNames$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationStatusNames-toJSONString}{\code{OrgSagebionetworksRepoModelCurationStatusNames$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationStatusNames-fromJSONString}{\code{OrgSagebionetworksRepoModelCurationStatusNames$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationStatusNames-clone}{\code{OrgSagebionetworksRepoModelCurationStatusNames$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationStatusNames-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationStatusNames-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelCurationStatusNames class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationStatusNames$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationStatusNames-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationStatusNames-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationStatusNames$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelCurationStatusNames in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationStatusNames-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationStatusNames-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelCurationStatusNames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationStatusNames$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelCurationStatusNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationStatusNames-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationStatusNames-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationStatusNames$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelCurationStatusNames in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationStatusNames-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationStatusNames-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelCurationStatusNames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationStatusNames$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelCurationStatusNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationStatusNames-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationStatusNames-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationStatusNames$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
