% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_curation_tracking_data.R
\docType{class}
\name{OrgSagebionetworksRepoModelCurationTrackingData}
\alias{OrgSagebionetworksRepoModelCurationTrackingData}
\title{OrgSagebionetworksRepoModelCurationTrackingData}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for CurationTrackingData

OrgSagebionetworksRepoModelCurationTrackingData Class
}
\details{
Create a new OrgSagebionetworksRepoModelCurationTrackingData
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{status}}{character [optional]}

\item{\code{curator}}{character [optional]}

\item{\code{history}}{Status history list(\link{OrgSagebionetworksRepoModelStatusHistoryRecord}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-new}{\code{OrgSagebionetworksRepoModelCurationTrackingData$new()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-toJSON}{\code{OrgSagebionetworksRepoModelCurationTrackingData$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-fromJSON}{\code{OrgSagebionetworksRepoModelCurationTrackingData$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-toJSONString}{\code{OrgSagebionetworksRepoModelCurationTrackingData$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-fromJSONString}{\code{OrgSagebionetworksRepoModelCurationTrackingData$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-validateJSON}{\code{OrgSagebionetworksRepoModelCurationTrackingData$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-toString}{\code{OrgSagebionetworksRepoModelCurationTrackingData$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-isValid}{\code{OrgSagebionetworksRepoModelCurationTrackingData$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-getInvalidFields}{\code{OrgSagebionetworksRepoModelCurationTrackingData$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-print}{\code{OrgSagebionetworksRepoModelCurationTrackingData$print()}}
\item \href{#method-OrgSagebionetworksRepoModelCurationTrackingData-clone}{\code{OrgSagebionetworksRepoModelCurationTrackingData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelCurationTrackingData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$new(
  status = NULL,
  curator = NULL,
  history = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{status}

\item{\code{curator}}{curator}

\item{\code{history}}{Status history}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelCurationTrackingData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelCurationTrackingData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelCurationTrackingData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelCurationTrackingData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelCurationTrackingData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelCurationTrackingData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelCurationTrackingData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelCurationTrackingData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelCurationTrackingData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelCurationTrackingData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelCurationTrackingData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
