% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_daemon_backup_alias_type.R
\docType{class}
\name{OrgSagebionetworksRepoModelDaemonBackupAliasType}
\alias{OrgSagebionetworksRepoModelDaemonBackupAliasType}
\title{OrgSagebionetworksRepoModelDaemonBackupAliasType}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDaemonBackupAliasType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDaemonBackupAliasType-new}{\code{OrgSagebionetworksRepoModelDaemonBackupAliasType$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDaemonBackupAliasType-toJSON}{\code{OrgSagebionetworksRepoModelDaemonBackupAliasType$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDaemonBackupAliasType-fromJSON}{\code{OrgSagebionetworksRepoModelDaemonBackupAliasType$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDaemonBackupAliasType-toJSONString}{\code{OrgSagebionetworksRepoModelDaemonBackupAliasType$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDaemonBackupAliasType-fromJSONString}{\code{OrgSagebionetworksRepoModelDaemonBackupAliasType$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDaemonBackupAliasType-clone}{\code{OrgSagebionetworksRepoModelDaemonBackupAliasType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaemonBackupAliasType-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaemonBackupAliasType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDaemonBackupAliasType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaemonBackupAliasType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaemonBackupAliasType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaemonBackupAliasType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaemonBackupAliasType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDaemonBackupAliasType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaemonBackupAliasType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaemonBackupAliasType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDaemonBackupAliasType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaemonBackupAliasType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDaemonBackupAliasType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaemonBackupAliasType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaemonBackupAliasType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaemonBackupAliasType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDaemonBackupAliasType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaemonBackupAliasType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaemonBackupAliasType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDaemonBackupAliasType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaemonBackupAliasType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDaemonBackupAliasType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaemonBackupAliasType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaemonBackupAliasType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaemonBackupAliasType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
