% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dao_wiki_page_key.R
\docType{class}
\name{OrgSagebionetworksRepoModelDaoWikiPageKey}
\alias{OrgSagebionetworksRepoModelDaoWikiPageKey}
\title{OrgSagebionetworksRepoModelDaoWikiPageKey}
\format{
An \code{R6Class} generator object
}
\description{
Reference to a WikiPage

OrgSagebionetworksRepoModelDaoWikiPageKey Class
}
\details{
Create a new OrgSagebionetworksRepoModelDaoWikiPageKey
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{wikiPageId}}{character [optional]}

\item{\code{ownerObjectId}}{character [optional]}

\item{\code{ownerObjectType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-new}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-toJSON}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-fromJSON}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-toJSONString}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-fromJSONString}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-validateJSON}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-toString}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-isValid}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-getInvalidFields}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-print}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDaoWikiPageKey-clone}{\code{OrgSagebionetworksRepoModelDaoWikiPageKey$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDaoWikiPageKey class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$new(
  wikiPageId = NULL,
  ownerObjectId = NULL,
  ownerObjectType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wikiPageId}}{wikiPageId}

\item{\code{ownerObjectId}}{ownerObjectId}

\item{\code{ownerObjectType}}{ownerObjectType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDaoWikiPageKey in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDaoWikiPageKey
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDaoWikiPageKey
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDaoWikiPageKey in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDaoWikiPageKey
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDaoWikiPageKey
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDaoWikiPageKey and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDaoWikiPageKey
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDaoWikiPageKey-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDaoWikiPageKey-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDaoWikiPageKey$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
