% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_access_approval_notification_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse}
\alias{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse}
\title{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse}
\format{
An \code{R6Class} generator object
}
\description{
Contains the response for an <a href=\"${org.sagebionetworks.repo.model.dataaccess.AccessApprovalNotificationRequest}\">AccessApprovalNotificationRequest</a>.

OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{requirementId}}{integer [optional]}

\item{\code{results}}{The list of notifications, sorted by the recipient and the sentOn date. list(\link{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotification}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-new}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-toString}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-isValid}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-print}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-clone}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$new(
  requirementId = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requirementId}}{requirementId}

\item{\code{results}}{The list of notifications, sorted by the recipient and the sentOn date.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$toJSON(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$isValid(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalNotificationResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
