% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_access_approval_search_sort.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort}
\alias{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort}
\title{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort}
\format{
An \code{R6Class} generator object
}
\description{
Identifies both the field and direction for a single sort operation

OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field}}{character [optional]}

\item{\code{direction}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-new}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-toString}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-isValid}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-print}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-clone}{\code{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$new(
  field = NULL,
  direction = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{field}

\item{\code{direction}}{direction}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessApprovalSearchSort$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
