% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_accessor_group_revoke_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest}
\alias{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest}
\title{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest}
\format{
An \code{R6Class} generator object
}
\description{
A request to revoke access of an accessor group. Accessors will be group by submitter and access requirement.

OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRequirementId}}{character [optional]}

\item{\code{submitterId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-new}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-toString}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-isValid}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-print}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-clone}{\code{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$new(
  accessRequirementId = NULL,
  submitterId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{submitterId}}{submitterId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessorGroupRevokeRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
