% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_access_requirement_conversion_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest}
\alias{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest}
\title{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest}
\format{
An \code{R6Class} generator object
}
\description{
A request to convert an ACTAccessRequirement to a ManagedACTAccessRequirement.

OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRequirementId}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{currentVersion}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-new}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-toString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-isValid}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-print}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-clone}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$new(
  accessRequirementId = NULL,
  etag = NULL,
  currentVersion = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{etag}}{etag}

\item{\code{currentVersion}}{currentVersion}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$toJSON(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$isValid(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementConversionRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
