% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_access_requirement_search_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse}
\alias{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse}
\title{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse}
\format{
An \code{R6Class} generator object
}
\description{
Describes the response object for an Access Requirement search

OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{The results of an access requirement search list(\link{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-new}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-toString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-isValid}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-print}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-clone}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$new(
  results = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{The results of an access requirement search}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
