% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_access_requirement_search_result.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult}
\alias{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult}
\title{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult}
\format{
An \code{R6Class} generator object
}
\description{
Result item of an Access Requirement search

OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{type}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{version}}{character [optional]}

\item{\code{relatedProjectIds}}{List of project ids which the AR is direclty or indirectly applied to list(character) [optional]}

\item{\code{reviewerIds}}{List of principal ids that are allowed to review the AR submissions list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-new}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-toString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-isValid}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-print}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-clone}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$new(
  id = NULL,
  type = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  name = NULL,
  version = NULL,
  relatedProjectIds = NULL,
  reviewerIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{type}}{type}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{name}}{name}

\item{\code{version}}{version}

\item{\code{relatedProjectIds}}{List of project ids which the AR is direclty or indirectly applied to}

\item{\code{reviewerIds}}{List of principal ids that are allowed to review the AR submissions}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchResult$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
