% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_access_requirement_search_sort.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort}
\alias{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort}
\title{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort}
\format{
An \code{R6Class} generator object
}
\description{
Identifies both the field and direction for a single sort operation

OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field}}{character [optional]}

\item{\code{direction}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-new}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-toString}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-isValid}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-print}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-clone}{\code{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$new(
  field = NULL,
  direction = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{field}

\item{\code{direction}}{direction}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessAccessRequirementSearchSort$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
