% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_basic_access_requirement_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus}
\alias{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus}
\title{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus}
\format{
An \code{R6Class} generator object
}
\description{
The status of a user meeting a TermsOfUseAccessRequirement.

OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRequirementId}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{isApproved}}{character [optional]}

\item{\code{expiredOn}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-new}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-toString}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-isValid}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-print}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-clone}{\code{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$new(
  concreteType,
  accessRequirementId = NULL,
  isApproved = NULL,
  expiredOn = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{isApproved}}{isApproved}

\item{\code{expiredOn}}{expiredOn}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessBasicAccessRequirementStatus$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
