% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_create_submission_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest}
\alias{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest}
\title{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest}
\format{
An \code{R6Class} generator object
}
\description{
A request to create a Submission.

OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{requestId}}{character [optional]}

\item{\code{requestEtag}}{character [optional]}

\item{\code{subjectId}}{character [optional]}

\item{\code{subjectType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-new}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-toString}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-isValid}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-print}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-clone}{\code{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$new(
  requestId = NULL,
  requestEtag = NULL,
  subjectId = NULL,
  subjectType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requestId}}{requestId}

\item{\code{requestEtag}}{requestEtag}

\item{\code{subjectId}}{subjectId}

\item{\code{subjectType}}{subjectType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessCreateSubmissionRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
