% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_data_access_submission_event.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent}
\alias{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent}
\title{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent}
\format{
An \code{R6Class} generator object
}
\description{
An envent that is fired after a user creates a data access submission

OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{objectId}}{character [optional]}

\item{\code{objectVersion}}{integer [optional]}

\item{\code{objectType}}{character [optional]}

\item{\code{timestamp}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-new}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-toString}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-isValid}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-print}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-clone}{\code{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$new(
  concreteType,
  objectId = NULL,
  objectVersion = NULL,
  objectType = NULL,
  timestamp = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{objectId}}{objectId}

\item{\code{objectVersion}}{objectVersion}

\item{\code{objectType}}{objectType}

\item{\code{timestamp}}{timestamp}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
