% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_managed_act_access_requirement_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus}
\alias{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus}
\title{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus}
\format{
An \code{R6Class} generator object
}
\description{
The status of a user meeting an ACTAccessRequirement.

OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRequirementId}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{isApproved}}{character [optional]}

\item{\code{expiredOn}}{character [optional]}

\item{\code{currentSubmissionStatus}}{\link{OrgSagebionetworksRepoModelDataaccessSubmissionStatus} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-new}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-toString}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-isValid}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-print}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-clone}{\code{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$new(
  concreteType,
  accessRequirementId = NULL,
  isApproved = NULL,
  expiredOn = NULL,
  currentSubmissionStatus = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{isApproved}}{isApproved}

\item{\code{expiredOn}}{expiredOn}

\item{\code{currentSubmissionStatus}}{currentSubmissionStatus}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$isValid(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessManagedACTAccessRequirementStatus$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
