% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_open_submission.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessOpenSubmission}
\alias{OrgSagebionetworksRepoModelDataaccessOpenSubmission}
\title{OrgSagebionetworksRepoModelDataaccessOpenSubmission}
\format{
An \code{R6Class} generator object
}
\description{
This object stores information about the number of submitted Submissions for a single ACTAccessRequirement.

OrgSagebionetworksRepoModelDataaccessOpenSubmission Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessOpenSubmission
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numberOfSubmittedSubmission}}{integer [optional]}

\item{\code{accessRequirementId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-new}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-toString}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-isValid}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-print}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-clone}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmission$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessOpenSubmission class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$new(
  numberOfSubmittedSubmission = NULL,
  accessRequirementId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfSubmittedSubmission}}{numberOfSubmittedSubmission}

\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessOpenSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessOpenSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessOpenSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessOpenSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessOpenSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessOpenSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessOpenSubmission and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessOpenSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmission-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmission$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
