% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_open_submission_page.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage}
\alias{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage}
\title{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage}
\format{
An \code{R6Class} generator object
}
\description{
A page of OpenSubmission.

OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{openSubmissionList}}{list(\link{OrgSagebionetworksRepoModelDataaccessOpenSubmission}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-new}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-toString}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-isValid}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-print}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-clone}{\code{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$new(
  openSubmissionList = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{openSubmissionList}}{openSubmissionList}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessOpenSubmissionPage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
