% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_renewal.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessRenewal}
\alias{OrgSagebionetworksRepoModelDataaccessRenewal}
\title{OrgSagebionetworksRepoModelDataaccessRenewal}
\format{
An \code{R6Class} generator object
}
\description{
A Renewal contains information required by an AccessRequirement and additional information about renewing a request.

OrgSagebionetworksRepoModelDataaccessRenewal Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessRenewal
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{accessRequirementId}}{character [optional]}

\item{\code{researchProjectId}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{ducFileHandleId}}{character [optional]}

\item{\code{irbFileHandleId}}{character [optional]}

\item{\code{attachments}}{The set of file handle ID of attached files to this request. list(character) [optional]}

\item{\code{accessorChanges}}{List of user changes. A user can gain access, renew access or have access revoked. list(\link{OrgSagebionetworksRepoModelDataaccessAccessorChange}) [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{publication}}{character [optional]}

\item{\code{summaryOfUse}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-new}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-toString}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-isValid}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-print}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessRenewal-clone}{\code{OrgSagebionetworksRepoModelDataaccessRenewal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessRenewal class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$new(
  concreteType,
  id = NULL,
  accessRequirementId = NULL,
  researchProjectId = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  ducFileHandleId = NULL,
  irbFileHandleId = NULL,
  attachments = NULL,
  accessorChanges = NULL,
  etag = NULL,
  publication = NULL,
  summaryOfUse = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{id}}{id}

\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{researchProjectId}}{researchProjectId}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{ducFileHandleId}}{ducFileHandleId}

\item{\code{irbFileHandleId}}{irbFileHandleId}

\item{\code{attachments}}{The set of file handle ID of attached files to this request.}

\item{\code{accessorChanges}}{List of user changes. A user can gain access, renew access or have access revoked.}

\item{\code{etag}}{etag}

\item{\code{publication}}{publication}

\item{\code{summaryOfUse}}{summaryOfUse}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessRenewal in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessRenewal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessRenewal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessRenewal in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessRenewal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessRenewal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessRenewal and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessRenewal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessRenewal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessRenewal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessRenewal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
