% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_research_project.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessResearchProject}
\alias{OrgSagebionetworksRepoModelDataaccessResearchProject}
\title{OrgSagebionetworksRepoModelDataaccessResearchProject}
\format{
An \code{R6Class} generator object
}
\description{
A research project describes a project at an institution that used a controlled data set for the purposes that are stated in the Intended Data Use Statement.

OrgSagebionetworksRepoModelDataaccessResearchProject Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessResearchProject
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{accessRequirementId}}{character [optional]}

\item{\code{institution}}{character [optional]}

\item{\code{projectLead}}{character [optional]}

\item{\code{intendedDataUseStatement}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{etag}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-new}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-toString}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-isValid}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-print}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessResearchProject-clone}{\code{OrgSagebionetworksRepoModelDataaccessResearchProject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessResearchProject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$new(
  id = NULL,
  accessRequirementId = NULL,
  institution = NULL,
  projectLead = NULL,
  intendedDataUseStatement = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  etag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{institution}}{institution}

\item{\code{projectLead}}{projectLead}

\item{\code{intendedDataUseStatement}}{intendedDataUseStatement}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{etag}}{etag}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessResearchProject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessResearchProject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessResearchProject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessResearchProject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessResearchProject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessResearchProject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessResearchProject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessResearchProject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessResearchProject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessResearchProject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessResearchProject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
