% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_submission.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessSubmission}
\alias{OrgSagebionetworksRepoModelDataaccessSubmission}
\title{OrgSagebionetworksRepoModelDataaccessSubmission}
\format{
An \code{R6Class} generator object
}
\description{
A submission to request access to controlled data.

OrgSagebionetworksRepoModelDataaccessSubmission Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessSubmission
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{accessRequirementId}}{character [optional]}

\item{\code{accessRequirementVersion}}{integer [optional]}

\item{\code{requestId}}{character [optional]}

\item{\code{ducFileHandleId}}{character [optional]}

\item{\code{irbFileHandleId}}{character [optional]}

\item{\code{attachments}}{The set of file handle ID of attached files to this request. list(character) [optional]}

\item{\code{accessorChanges}}{List of user changes. A user can gain access, renew access or have access revoked. list(\link{OrgSagebionetworksRepoModelDataaccessAccessorChange}) [optional]}

\item{\code{researchProjectSnapshot}}{\link{OrgSagebionetworksRepoModelDataaccessResearchProject} [optional]}

\item{\code{isRenewalSubmission}}{character [optional]}

\item{\code{publication}}{character [optional]}

\item{\code{summaryOfUse}}{character [optional]}

\item{\code{submittedOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{submittedBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{state}}{character [optional]}

\item{\code{rejectedReason}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{subjectId}}{character [optional]}

\item{\code{subjectType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-new}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-toString}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-isValid}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-print}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmission-clone}{\code{OrgSagebionetworksRepoModelDataaccessSubmission$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessSubmission class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$new(
  id = NULL,
  accessRequirementId = NULL,
  accessRequirementVersion = NULL,
  requestId = NULL,
  ducFileHandleId = NULL,
  irbFileHandleId = NULL,
  attachments = NULL,
  accessorChanges = NULL,
  researchProjectSnapshot = NULL,
  isRenewalSubmission = NULL,
  publication = NULL,
  summaryOfUse = NULL,
  submittedOn = NULL,
  modifiedOn = NULL,
  submittedBy = NULL,
  modifiedBy = NULL,
  state = NULL,
  rejectedReason = NULL,
  etag = NULL,
  subjectId = NULL,
  subjectType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{accessRequirementVersion}}{accessRequirementVersion}

\item{\code{requestId}}{requestId}

\item{\code{ducFileHandleId}}{ducFileHandleId}

\item{\code{irbFileHandleId}}{irbFileHandleId}

\item{\code{attachments}}{The set of file handle ID of attached files to this request.}

\item{\code{accessorChanges}}{List of user changes. A user can gain access, renew access or have access revoked.}

\item{\code{researchProjectSnapshot}}{researchProjectSnapshot}

\item{\code{isRenewalSubmission}}{isRenewalSubmission}

\item{\code{publication}}{publication}

\item{\code{summaryOfUse}}{summaryOfUse}

\item{\code{submittedOn}}{submittedOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{submittedBy}}{submittedBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{state}}{state}

\item{\code{rejectedReason}}{rejectedReason}

\item{\code{etag}}{etag}

\item{\code{subjectId}}{subjectId}

\item{\code{subjectType}}{subjectType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessSubmission and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmission-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmission-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmission$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
