% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_submission_info.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessSubmissionInfo}
\alias{OrgSagebionetworksRepoModelDataaccessSubmissionInfo}
\title{OrgSagebionetworksRepoModelDataaccessSubmissionInfo}
\format{
An \code{R6Class} generator object
}
\description{
Information about a request to access controlled data and its related Research Project

OrgSagebionetworksRepoModelDataaccessSubmissionInfo Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessSubmissionInfo
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{institution}}{character [optional]}

\item{\code{projectLead}}{character [optional]}

\item{\code{intendedDataUseStatement}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{submittedBy}}{character [optional]}

\item{\code{accessorChanges}}{List of user changes. A user can gain access, renew access or have access revoked. list(\link{OrgSagebionetworksRepoModelDataaccessAccessorChange}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-new}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-toString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-isValid}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-print}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-clone}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessSubmissionInfo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$new(
  institution = NULL,
  projectLead = NULL,
  intendedDataUseStatement = NULL,
  modifiedOn = NULL,
  submittedBy = NULL,
  accessorChanges = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{institution}}{institution}

\item{\code{projectLead}}{projectLead}

\item{\code{intendedDataUseStatement}}{intendedDataUseStatement}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{submittedBy}}{submittedBy}

\item{\code{accessorChanges}}{List of user changes. A user can gain access, renew access or have access revoked.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmissionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmissionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmissionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmissionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessSubmissionInfo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessSubmissionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
