% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_submission_page_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest}
\alias{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest}
\title{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest}
\format{
An \code{R6Class} generator object
}
\description{
A request to retrieve a page of a Submission.

OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRequirementId}}{character [optional]}

\item{\code{nextPageToken}}{character [optional]}

\item{\code{filterBy}}{character [optional]}

\item{\code{accessorId}}{character [optional]}

\item{\code{orderBy}}{character [optional]}

\item{\code{isAscending}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-new}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-toString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-isValid}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-print}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-clone}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$new(
  accessRequirementId = NULL,
  nextPageToken = NULL,
  filterBy = NULL,
  accessorId = NULL,
  orderBy = NULL,
  isAscending = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{filterBy}}{filterBy}

\item{\code{accessorId}}{accessorId}

\item{\code{orderBy}}{orderBy}

\item{\code{isAscending}}{isAscending}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionPageRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
