% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_submission_search_result.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult}
\alias{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult}
\title{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{accessRequirementId}}{character [optional]}

\item{\code{accessRequirementVersion}}{character [optional]}

\item{\code{accessRequirementName}}{character [optional]}

\item{\code{accessRequirementReviewerIds}}{The list of principal ids that are allowed to review the submission list(character) [optional]}

\item{\code{submitterId}}{character [optional]}

\item{\code{accessorChanges}}{List of user changes. A user can gain access, renew access or have access revoked. list(\link{OrgSagebionetworksRepoModelDataaccessAccessorChange}) [optional]}

\item{\code{state}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-new}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-toString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-isValid}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-print}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-clone}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$new(
  id = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  accessRequirementId = NULL,
  accessRequirementVersion = NULL,
  accessRequirementName = NULL,
  accessRequirementReviewerIds = NULL,
  submitterId = NULL,
  accessorChanges = NULL,
  state = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{accessRequirementVersion}}{accessRequirementVersion}

\item{\code{accessRequirementName}}{accessRequirementName}

\item{\code{accessRequirementReviewerIds}}{The list of principal ids that are allowed to review the submission}

\item{\code{submitterId}}{submitterId}

\item{\code{accessorChanges}}{List of user changes. A user can gain access, renew access or have access revoked.}

\item{\code{state}}{state}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionSearchResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
