% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataaccess_submission_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelDataaccessSubmissionStatus}
\alias{OrgSagebionetworksRepoModelDataaccessSubmissionStatus}
\title{OrgSagebionetworksRepoModelDataaccessSubmissionStatus}
\format{
An \code{R6Class} generator object
}
\description{
The status of a Submission.

OrgSagebionetworksRepoModelDataaccessSubmissionStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelDataaccessSubmissionStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{submissionId}}{character [optional]}

\item{\code{submittedBy}}{character [optional]}

\item{\code{rejectedReason}}{character [optional]}

\item{\code{state}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-new}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-toJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-fromJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-toJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-validateJSON}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-toString}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-isValid}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-print}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-clone}{\code{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDataaccessSubmissionStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$new(
  submissionId = NULL,
  submittedBy = NULL,
  rejectedReason = NULL,
  state = NULL,
  modifiedOn = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submissionId}}{submissionId}

\item{\code{submittedBy}}{submittedBy}

\item{\code{rejectedReason}}{rejectedReason}

\item{\code{state}}{state}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmissionStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDataaccessSubmissionStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDataaccessSubmissionStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDataaccessSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDataaccessSubmissionStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDataaccessSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDataaccessSubmissionStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDataaccessSubmissionStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
