% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_dataset_tracking_data.R
\docType{class}
\name{OrgSagebionetworksRepoModelDatasetTrackingData}
\alias{OrgSagebionetworksRepoModelDatasetTrackingData}
\title{OrgSagebionetworksRepoModelDatasetTrackingData}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for DatasetTrackingData POJO

OrgSagebionetworksRepoModelDatasetTrackingData Class
}
\details{
Create a new OrgSagebionetworksRepoModelDatasetTrackingData
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{acquisitionTrackingData}}{\link{OrgSagebionetworksRepoModelAcquisitionTrackingData} [optional]}

\item{\code{curationTrackingData}}{\link{OrgSagebionetworksRepoModelCurationTrackingData} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-new}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-toJSON}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-fromJSON}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-toJSONString}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-fromJSONString}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-validateJSON}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-toString}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-isValid}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-getInvalidFields}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-print}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDatasetTrackingData-clone}{\code{OrgSagebionetworksRepoModelDatasetTrackingData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDatasetTrackingData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$new(
  acquisitionTrackingData = NULL,
  curationTrackingData = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{acquisitionTrackingData}}{acquisitionTrackingData}

\item{\code{curationTrackingData}}{curationTrackingData}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDatasetTrackingData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDatasetTrackingData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDatasetTrackingData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDatasetTrackingData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDatasetTrackingData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDatasetTrackingData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDatasetTrackingData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDatasetTrackingData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDatasetTrackingData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDatasetTrackingData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDatasetTrackingData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
