% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_create_discussion_reply.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply}
\alias{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply}
\title{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply}
\format{
An \code{R6Class} generator object
}
\description{
The model object represents a single Reply creation.

OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply Class
}
\details{
Create a new OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{threadId}}{character [optional]}

\item{\code{messageMarkdown}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-new}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-validateJSON}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-toString}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-isValid}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-getInvalidFields}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-print}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-clone}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$new(
  threadId = NULL,
  messageMarkdown = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threadId}}{threadId}

\item{\code{messageMarkdown}}{messageMarkdown}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionReply$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
