% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_create_discussion_thread.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread}
\alias{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread}
\title{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread}
\format{
An \code{R6Class} generator object
}
\description{
The model object represents a single Thread creation.

OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread Class
}
\details{
Create a new OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{forumId}}{character [optional]}

\item{\code{title}}{character [optional]}

\item{\code{messageMarkdown}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-new}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-validateJSON}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-toString}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-isValid}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-getInvalidFields}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-print}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-clone}{\code{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$new(
  forumId = NULL,
  title = NULL,
  messageMarkdown = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forumId}}{forumId}

\item{\code{title}}{title}

\item{\code{messageMarkdown}}{messageMarkdown}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionCreateDiscussionThread$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
