% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_discussion_reply_bundle.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle}
\alias{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle}
\title{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle}
\format{
An \code{R6Class} generator object
}
\description{
The Reply model object represents a single reply in a thread.

OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle Class
}
\details{
Create a new OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{threadId}}{character [optional]}

\item{\code{forumId}}{character [optional]}

\item{\code{projectId}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{messageKey}}{character [optional]}

\item{\code{isEdited}}{character [optional]}

\item{\code{isDeleted}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-new}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-validateJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-toString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-isValid}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-getInvalidFields}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-print}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-clone}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$new(
  id = NULL,
  threadId = NULL,
  forumId = NULL,
  projectId = NULL,
  createdOn = NULL,
  createdBy = NULL,
  modifiedOn = NULL,
  etag = NULL,
  messageKey = NULL,
  isEdited = NULL,
  isDeleted = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{threadId}}{threadId}

\item{\code{forumId}}{forumId}

\item{\code{projectId}}{projectId}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{etag}}{etag}

\item{\code{messageKey}}{messageKey}

\item{\code{isEdited}}{isEdited}

\item{\code{isDeleted}}{isDeleted}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionReplyBundle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
