% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_discussion_search_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest}
\alias{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest}
\title{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest}
\format{
An \code{R6Class} generator object
}
\description{
A request to perform a full text search in a discussion forum

OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{searchString}}{character [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-new}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-toString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-isValid}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-print}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-clone}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$new(
  searchString = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchString}}{searchString}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
