% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_discussion_search_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse}
\alias{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse}
\title{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse}
\format{
An \code{R6Class} generator object
}
\description{
The response to a <a href=\"${org.sagebionetworks.repo.model.discussion.DiscussionSearchRequest}\">DiscussionSearchRequest</a>.

OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{matches}}{A single page of results matching the original <a href=\"${org.sagebionetworks.repo.model.discussion.DiscussionSearchRequest}\">DiscussionSearchRequest</a>. list(\link{OrgSagebionetworksRepoModelDiscussionMatch}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-new}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-validateJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-toString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-isValid}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-print}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-clone}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$new(
  matches = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matches}}{A single page of results matching the original <a href=\"${org.sagebionetworks.repo.model.discussion.DiscussionSearchRequest}\">DiscussionSearchRequest</a>.}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionSearchResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
