% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_discussion_thread_bundle.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle}
\alias{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle}
\title{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle}
\format{
An \code{R6Class} generator object
}
\description{
The Thread model object represents a single Thread.

OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle Class
}
\details{
Create a new OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{forumId}}{character [optional]}

\item{\code{projectId}}{character [optional]}

\item{\code{title}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{messageKey}}{character [optional]}

\item{\code{numberOfViews}}{integer [optional]}

\item{\code{numberOfReplies}}{integer [optional]}

\item{\code{lastActivity}}{character [optional]}

\item{\code{activeAuthors}}{The list of userId whose most active on this Thread list(character) [optional]}

\item{\code{isEdited}}{character [optional]}

\item{\code{isDeleted}}{character [optional]}

\item{\code{isPinned}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-new}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-validateJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-toString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-isValid}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-getInvalidFields}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-print}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-clone}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$new(
  id = NULL,
  forumId = NULL,
  projectId = NULL,
  title = NULL,
  createdOn = NULL,
  createdBy = NULL,
  modifiedOn = NULL,
  etag = NULL,
  messageKey = NULL,
  numberOfViews = NULL,
  numberOfReplies = NULL,
  lastActivity = NULL,
  activeAuthors = NULL,
  isEdited = NULL,
  isDeleted = NULL,
  isPinned = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{forumId}}{forumId}

\item{\code{projectId}}{projectId}

\item{\code{title}}{title}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{etag}}{etag}

\item{\code{messageKey}}{messageKey}

\item{\code{numberOfViews}}{numberOfViews}

\item{\code{numberOfReplies}}{numberOfReplies}

\item{\code{lastActivity}}{lastActivity}

\item{\code{activeAuthors}}{The list of userId whose most active on this Thread}

\item{\code{isEdited}}{isEdited}

\item{\code{isDeleted}}{isDeleted}

\item{\code{isPinned}}{isPinned}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadBundle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
