% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_discussion_thread_order.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder}
\alias{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder}
\title{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-new}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-clone}{\code{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionDiscussionThreadOrder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
