% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_entity_thread_count.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionEntityThreadCount}
\alias{OrgSagebionetworksRepoModelDiscussionEntityThreadCount}
\title{OrgSagebionetworksRepoModelDiscussionEntityThreadCount}
\format{
An \code{R6Class} generator object
}
\description{
The model object represents the number of threads that mentioned a particular entity.

OrgSagebionetworksRepoModelDiscussionEntityThreadCount Class
}
\details{
Create a new OrgSagebionetworksRepoModelDiscussionEntityThreadCount
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entityId}}{character [optional]}

\item{\code{count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-new}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-validateJSON}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-toString}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-isValid}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-getInvalidFields}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-print}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-clone}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionEntityThreadCount class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$new(
  entityId = NULL,
  count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entityId}}{entityId}

\item{\code{count}}{count}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionEntityThreadCount in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionEntityThreadCount
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionEntityThreadCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionEntityThreadCount in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionEntityThreadCount
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionEntityThreadCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDiscussionEntityThreadCount and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDiscussionEntityThreadCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCount-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
