% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_entity_thread_counts.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts}
\alias{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts}
\title{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts}
\format{
An \code{R6Class} generator object
}
\description{
The model object represents a list of EntityThreadCounts

OrgSagebionetworksRepoModelDiscussionEntityThreadCounts Class
}
\details{
Create a new OrgSagebionetworksRepoModelDiscussionEntityThreadCounts
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{list}}{list(\link{OrgSagebionetworksRepoModelDiscussionEntityThreadCount}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-new}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-validateJSON}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-toString}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-isValid}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-getInvalidFields}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-print}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-clone}{\code{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionEntityThreadCounts class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$new(list = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list}}{list}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionEntityThreadCounts in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionEntityThreadCounts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionEntityThreadCounts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionEntityThreadCounts in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionEntityThreadCounts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionEntityThreadCounts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDiscussionEntityThreadCounts and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDiscussionEntityThreadCounts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionEntityThreadCounts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionEntityThreadCounts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
