% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_discussion_update_thread_title.R
\docType{class}
\name{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle}
\alias{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle}
\title{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle Class

OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle Class
}
\details{
Create a new OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-new}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-toJSON}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-fromJSON}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-toJSONString}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-fromJSONString}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-validateJSON}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-toString}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-isValid}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-getInvalidFields}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-print}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-clone}{\code{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$new(title = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDiscussionUpdateThreadTitle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
