% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_docker_docker_authorization_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken}
\alias{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken}
\title{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken}
\format{
An \code{R6Class} generator object
}
\description{
Authorization Token for Docker Registry Access

OrgSagebionetworksRepoModelDockerDockerAuthorizationToken Class
}
\details{
Create a new OrgSagebionetworksRepoModelDockerDockerAuthorizationToken
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{token}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-new}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-toJSON}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-fromJSON}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-toJSONString}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-fromJSONString}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-validateJSON}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-toString}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-isValid}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-getInvalidFields}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-print}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-clone}{\code{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDockerDockerAuthorizationToken class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$new(
  token = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{token}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerDockerAuthorizationToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerDockerAuthorizationToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerDockerAuthorizationToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerDockerAuthorizationToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerDockerAuthorizationToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerDockerAuthorizationToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDockerDockerAuthorizationToken and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDockerDockerAuthorizationToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerAuthorizationToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerAuthorizationToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
