% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_docker_docker_registry_event.R
\docType{class}
\name{OrgSagebionetworksRepoModelDockerDockerRegistryEvent}
\alias{OrgSagebionetworksRepoModelDockerDockerRegistryEvent}
\title{OrgSagebionetworksRepoModelDockerDockerRegistryEvent}
\format{
An \code{R6Class} generator object
}
\description{
https://godoc.org/github.com/docker/distribution/notifications#Event

OrgSagebionetworksRepoModelDockerDockerRegistryEvent Class
}
\details{
Create a new OrgSagebionetworksRepoModelDockerDockerRegistryEvent
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{timestamp}}{character [optional]}

\item{\code{action}}{character [optional]}

\item{\code{target}}{\link{OrgSagebionetworksRepoModelDockerRegistryEventTarget} [optional]}

\item{\code{request}}{\link{OrgSagebionetworksRepoModelDockerRegistryEventRequest} [optional]}

\item{\code{actor}}{\link{OrgSagebionetworksRepoModelDockerRegistryEventActor} [optional]}

\item{\code{source}}{\link{OrgSagebionetworksRepoModelDockerRegistryEventSource} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-new}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-toJSON}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-fromJSON}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-toJSONString}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-fromJSONString}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-validateJSON}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-toString}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-isValid}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-getInvalidFields}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-print}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-clone}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDockerDockerRegistryEvent class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$new(
  id = NULL,
  timestamp = NULL,
  action = NULL,
  target = NULL,
  request = NULL,
  actor = NULL,
  source = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{timestamp}}{timestamp}

\item{\code{action}}{action}

\item{\code{target}}{target}

\item{\code{request}}{request}

\item{\code{actor}}{actor}

\item{\code{source}}{source}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerDockerRegistryEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerDockerRegistryEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerDockerRegistryEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerDockerRegistryEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerDockerRegistryEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerDockerRegistryEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDockerDockerRegistryEvent and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDockerDockerRegistryEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
