% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_docker_docker_registry_event_list.R
\docType{class}
\name{OrgSagebionetworksRepoModelDockerDockerRegistryEventList}
\alias{OrgSagebionetworksRepoModelDockerDockerRegistryEventList}
\title{OrgSagebionetworksRepoModelDockerDockerRegistryEventList}
\format{
An \code{R6Class} generator object
}
\description{
List of Registry Events

OrgSagebionetworksRepoModelDockerDockerRegistryEventList Class
}
\details{
Create a new OrgSagebionetworksRepoModelDockerDockerRegistryEventList
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{events}}{A list of events that occurred in a docker registry list(\link{OrgSagebionetworksRepoModelDockerDockerRegistryEvent}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-new}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-toJSON}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-fromJSON}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-toJSONString}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-fromJSONString}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-validateJSON}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-toString}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-isValid}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-getInvalidFields}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-print}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-clone}{\code{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDockerDockerRegistryEventList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$new(
  events = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{events}}{A list of events that occurred in a docker registry}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerDockerRegistryEventList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerDockerRegistryEventList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerDockerRegistryEventList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerDockerRegistryEventList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerDockerRegistryEventList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerDockerRegistryEventList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDockerDockerRegistryEventList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDockerDockerRegistryEventList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRegistryEventList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRegistryEventList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
