% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_docker_docker_repository.R
\docType{class}
\name{OrgSagebionetworksRepoModelDockerDockerRepository}
\alias{OrgSagebionetworksRepoModelDockerDockerRepository}
\title{OrgSagebionetworksRepoModelDockerDockerRepository}
\format{
An \code{R6Class} generator object
}
\description{
A Docker repository is a lightweight virtual machine image.

OrgSagebionetworksRepoModelDockerDockerRepository Class
}
\details{
Create a new OrgSagebionetworksRepoModelDockerDockerRepository
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{parentId}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{repositoryName}}{character [optional]}

\item{\code{isManaged}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-new}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-toJSON}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-fromJSON}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-toJSONString}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-fromJSONString}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-validateJSON}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-toString}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-isValid}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-getInvalidFields}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-print}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerDockerRepository-clone}{\code{OrgSagebionetworksRepoModelDockerDockerRepository$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDockerDockerRepository class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$new(
  concreteType,
  name = NULL,
  description = NULL,
  id = NULL,
  etag = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  parentId = NULL,
  repositoryName = NULL,
  isManaged = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{parentId}}{parentId}

\item{\code{repositoryName}}{repositoryName}

\item{\code{isManaged}}{isManaged}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerDockerRepository in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerDockerRepository
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerDockerRepository
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerDockerRepository in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerDockerRepository
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerDockerRepository
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDockerDockerRepository and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDockerDockerRepository
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerDockerRepository-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerDockerRepository-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerDockerRepository$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
