% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_docker_registry_event_actor.R
\docType{class}
\name{OrgSagebionetworksRepoModelDockerRegistryEventActor}
\alias{OrgSagebionetworksRepoModelDockerRegistryEventActor}
\title{OrgSagebionetworksRepoModelDockerRegistryEventActor}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDockerRegistryEventActor Class

OrgSagebionetworksRepoModelDockerRegistryEventActor Class
}
\details{
Create a new OrgSagebionetworksRepoModelDockerRegistryEventActor
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-new}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-toJSON}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-fromJSON}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-toJSONString}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-fromJSONString}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-validateJSON}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-toString}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-isValid}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-getInvalidFields}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-print}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventActor-clone}{\code{OrgSagebionetworksRepoModelDockerRegistryEventActor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDockerRegistryEventActor class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$new(name = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerRegistryEventActor in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerRegistryEventActor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerRegistryEventActor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerRegistryEventActor in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerRegistryEventActor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerRegistryEventActor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDockerRegistryEventActor and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDockerRegistryEventActor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventActor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventActor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventActor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
