% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_docker_registry_event_source.R
\docType{class}
\name{OrgSagebionetworksRepoModelDockerRegistryEventSource}
\alias{OrgSagebionetworksRepoModelDockerRegistryEventSource}
\title{OrgSagebionetworksRepoModelDockerRegistryEventSource}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDockerRegistryEventSource Class

OrgSagebionetworksRepoModelDockerRegistryEventSource Class
}
\details{
Create a new OrgSagebionetworksRepoModelDockerRegistryEventSource
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{addr}}{character [optional]}

\item{\code{instanceID}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-new}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-toJSON}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-fromJSON}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-toJSONString}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-fromJSONString}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-validateJSON}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-toString}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-isValid}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-getInvalidFields}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-print}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventSource-clone}{\code{OrgSagebionetworksRepoModelDockerRegistryEventSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDockerRegistryEventSource class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$new(
  addr = NULL,
  instanceID = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{addr}}{addr}

\item{\code{instanceID}}{instanceID}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerRegistryEventSource in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerRegistryEventSource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerRegistryEventSource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerRegistryEventSource in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerRegistryEventSource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerRegistryEventSource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDockerRegistryEventSource and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDockerRegistryEventSource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
