% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_docker_registry_event_target.R
\docType{class}
\name{OrgSagebionetworksRepoModelDockerRegistryEventTarget}
\alias{OrgSagebionetworksRepoModelDockerRegistryEventTarget}
\title{OrgSagebionetworksRepoModelDockerRegistryEventTarget}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDockerRegistryEventTarget Class

OrgSagebionetworksRepoModelDockerRegistryEventTarget Class
}
\details{
Create a new OrgSagebionetworksRepoModelDockerRegistryEventTarget
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mediaType}}{character [optional]}

\item{\code{size}}{integer [optional]}

\item{\code{digest}}{character [optional]}

\item{\code{length}}{integer [optional]}

\item{\code{repository}}{character [optional]}

\item{\code{tag}}{character [optional]}

\item{\code{url}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-new}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-toJSON}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-fromJSON}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-toJSONString}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-fromJSONString}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-validateJSON}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-toString}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-isValid}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-getInvalidFields}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-print}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-clone}{\code{OrgSagebionetworksRepoModelDockerRegistryEventTarget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDockerRegistryEventTarget class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$new(
  mediaType = NULL,
  size = NULL,
  digest = NULL,
  length = NULL,
  repository = NULL,
  tag = NULL,
  url = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mediaType}}{mediaType}

\item{\code{size}}{size}

\item{\code{digest}}{digest}

\item{\code{length}}{length}

\item{\code{repository}}{repository}

\item{\code{tag}}{tag}

\item{\code{url}}{url}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerRegistryEventTarget in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerRegistryEventTarget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerRegistryEventTarget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDockerRegistryEventTarget in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDockerRegistryEventTarget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDockerRegistryEventTarget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDockerRegistryEventTarget and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDockerRegistryEventTarget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDockerRegistryEventTarget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDockerRegistryEventTarget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
