% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_doi_doi.R
\docType{class}
\name{OrgSagebionetworksRepoModelDoiDoi}
\alias{OrgSagebionetworksRepoModelDoiDoi}
\title{OrgSagebionetworksRepoModelDoiDoi}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for DOI.

OrgSagebionetworksRepoModelDoiDoi Class
}
\details{
Create a new OrgSagebionetworksRepoModelDoiDoi
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character}

\item{\code{etag}}{character [optional]}

\item{\code{doiStatus}}{character [optional]}

\item{\code{objectId}}{character}

\item{\code{objectType}}{character [optional]}

\item{\code{objectVersion}}{integer [optional]}

\item{\code{createdBy}}{character}

\item{\code{createdOn}}{character [optional]}

\item{\code{updatedOn}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-new}{\code{OrgSagebionetworksRepoModelDoiDoi$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-toJSON}{\code{OrgSagebionetworksRepoModelDoiDoi$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-fromJSON}{\code{OrgSagebionetworksRepoModelDoiDoi$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-toJSONString}{\code{OrgSagebionetworksRepoModelDoiDoi$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-fromJSONString}{\code{OrgSagebionetworksRepoModelDoiDoi$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-validateJSON}{\code{OrgSagebionetworksRepoModelDoiDoi$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-toString}{\code{OrgSagebionetworksRepoModelDoiDoi$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-isValid}{\code{OrgSagebionetworksRepoModelDoiDoi$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-getInvalidFields}{\code{OrgSagebionetworksRepoModelDoiDoi$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-print}{\code{OrgSagebionetworksRepoModelDoiDoi$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiDoi-clone}{\code{OrgSagebionetworksRepoModelDoiDoi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDoiDoi class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$new(
  id,
  objectId,
  createdBy,
  etag = NULL,
  doiStatus = NULL,
  objectType = NULL,
  objectVersion = NULL,
  createdOn = NULL,
  updatedOn = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{objectId}}{objectId}

\item{\code{createdBy}}{createdBy}

\item{\code{etag}}{etag}

\item{\code{doiStatus}}{doiStatus}

\item{\code{objectType}}{objectType}

\item{\code{objectVersion}}{objectVersion}

\item{\code{createdOn}}{createdOn}

\item{\code{updatedOn}}{updatedOn}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiDoi in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiDoi
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiDoi
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiDoi in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiDoi
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiDoi
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDoiDoi and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDoiDoi
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiDoi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiDoi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiDoi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
