% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_doi_v2_datacite_metadata.R
\docType{class}
\name{OrgSagebionetworksRepoModelDoiV2DataciteMetadata}
\alias{OrgSagebionetworksRepoModelDoiV2DataciteMetadata}
\title{OrgSagebionetworksRepoModelDoiV2DataciteMetadata}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDoiV2DataciteMetadata Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelDoiV2DataciteMetadata.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-new}{\code{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-fromJSONString}{\code{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-fromJSON}{\code{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-toJSONString}{\code{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-toJSON}{\code{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-validateJSON}{\code{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-toString}{\code{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-print}{\code{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-clone}{\code{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDoiV2DataciteMetadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelDoiV2Doi"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2DataciteMetadata.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelDoiV2DataciteMetadata.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2DataciteMetadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelDoiV2DataciteMetadata.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelDoiV2DataciteMetadata to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelDoiV2DataciteMetadata.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelDoiV2DataciteMetadata to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelDoiV2DataciteMetadata.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelDoiV2DataciteMetadata and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DataciteMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DataciteMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
