% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_doi_v2_doi.R
\docType{class}
\name{OrgSagebionetworksRepoModelDoiV2Doi}
\alias{OrgSagebionetworksRepoModelDoiV2Doi}
\title{OrgSagebionetworksRepoModelDoiV2Doi}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for fields associated with a DOI and its metadata. This object contains fields from both <a href=\"${org.sagebionetworks.repo.model.doi.v2.DoiAssociation}\">DoiAssociation</a> and <a href=\"${org.sagebionetworks.repo.model.doi.v2.DataciteMetadata}\">DataciteMetadata</a>.<br>To mint a DOI, the following fields are required: <ul> <li>Information that uniquely identifies an object in Synapse <ul><li> The object ID <i>(e.g. \"syn12345\")</i></li><li>The object type <i>(Note: as of this writing, only ENTITY is supported)</i></li><li>The version number of the object <i>(Optional. A \"null\" version will mint a DOI that will point to the most recent version of the object, even if the object changes)</i></li><li>An eTag <i>(Necessary when updating an existing DOI)</i></li></ul></li> <li>Additional supplementary information to mint a DOI <ul><li>Creator(s)</li><li>Title(s)</li><li>The general resource type</li><li>A publication year (no later than one year after the current year)</li></ul></li> </ul><br>For more information, see the <a href=\"${org.sagebionetworks.repo.model.doi.v2.DoiAssociation}\">DoiAssociation</a> and <a href=\"${org.sagebionetworks.repo.model.doi.v2.DataciteMetadata}\">DataciteMetadata</a> objects.

OrgSagebionetworksRepoModelDoiV2Doi Class
}
\details{
Create a new OrgSagebionetworksRepoModelDoiV2Doi
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{creators}}{Required. The main researchers involved in producing the data, or the authors of the publication, in priority order. list(\link{OrgSagebionetworksRepoModelDoiV2DoiCreator}) [optional]}

\item{\code{titles}}{Required. A name or title by which a resource is known. list(\link{OrgSagebionetworksRepoModelDoiV2DoiTitle}) [optional]}

\item{\code{publicationYear}}{integer [optional]}

\item{\code{resourceType}}{\link{OrgSagebionetworksRepoModelDoiV2DoiResourceType} [optional]}

\item{\code{associationId}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{doiUri}}{character [optional]}

\item{\code{doiUrl}}{character [optional]}

\item{\code{objectId}}{character [optional]}

\item{\code{objectType}}{character [optional]}

\item{\code{objectVersion}}{integer [optional]}

\item{\code{associatedBy}}{character [optional]}

\item{\code{associatedOn}}{character [optional]}

\item{\code{updatedBy}}{character [optional]}

\item{\code{updatedOn}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-new}{\code{OrgSagebionetworksRepoModelDoiV2Doi$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-toJSON}{\code{OrgSagebionetworksRepoModelDoiV2Doi$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-fromJSON}{\code{OrgSagebionetworksRepoModelDoiV2Doi$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-toJSONString}{\code{OrgSagebionetworksRepoModelDoiV2Doi$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-fromJSONString}{\code{OrgSagebionetworksRepoModelDoiV2Doi$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-validateJSON}{\code{OrgSagebionetworksRepoModelDoiV2Doi$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-toString}{\code{OrgSagebionetworksRepoModelDoiV2Doi$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-isValid}{\code{OrgSagebionetworksRepoModelDoiV2Doi$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-getInvalidFields}{\code{OrgSagebionetworksRepoModelDoiV2Doi$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-print}{\code{OrgSagebionetworksRepoModelDoiV2Doi$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2Doi-clone}{\code{OrgSagebionetworksRepoModelDoiV2Doi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDoiV2Doi class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$new(
  creators = NULL,
  titles = NULL,
  publicationYear = NULL,
  resourceType = NULL,
  associationId = NULL,
  etag = NULL,
  doiUri = NULL,
  doiUrl = NULL,
  objectId = NULL,
  objectType = NULL,
  objectVersion = NULL,
  associatedBy = NULL,
  associatedOn = NULL,
  updatedBy = NULL,
  updatedOn = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{creators}}{Required. The main researchers involved in producing the data, or the authors of the publication, in priority order.}

\item{\code{titles}}{Required. A name or title by which a resource is known.}

\item{\code{publicationYear}}{publicationYear}

\item{\code{resourceType}}{resourceType}

\item{\code{associationId}}{associationId}

\item{\code{etag}}{etag}

\item{\code{doiUri}}{doiUri}

\item{\code{doiUrl}}{doiUrl}

\item{\code{objectId}}{objectId}

\item{\code{objectType}}{objectType}

\item{\code{objectVersion}}{objectVersion}

\item{\code{associatedBy}}{associatedBy}

\item{\code{associatedOn}}{associatedOn}

\item{\code{updatedBy}}{updatedBy}

\item{\code{updatedOn}}{updatedOn}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2Doi in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2Doi
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiV2Doi
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2Doi in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2Doi
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiV2Doi
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDoiV2Doi and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDoiV2Doi
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2Doi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2Doi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2Doi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
