% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_doi_v2_doi_association.R
\docType{class}
\name{OrgSagebionetworksRepoModelDoiV2DoiAssociation}
\alias{OrgSagebionetworksRepoModelDoiV2DoiAssociation}
\title{OrgSagebionetworksRepoModelDoiV2DoiAssociation}
\format{
An \code{R6Class} generator object
}
\description{
All fields that associate a Synapse object with a DOI. The objectId and objectType are required to create or mint a DOI in all circumstances.

OrgSagebionetworksRepoModelDoiV2DoiAssociation Class
}
\details{
Create a new OrgSagebionetworksRepoModelDoiV2DoiAssociation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{associationId}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{doiUri}}{character [optional]}

\item{\code{doiUrl}}{character [optional]}

\item{\code{objectId}}{character [optional]}

\item{\code{objectType}}{character [optional]}

\item{\code{objectVersion}}{integer [optional]}

\item{\code{associatedBy}}{character [optional]}

\item{\code{associatedOn}}{character [optional]}

\item{\code{updatedBy}}{character [optional]}

\item{\code{updatedOn}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-new}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-toJSON}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-fromJSON}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-toJSONString}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-fromJSONString}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-validateJSON}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-toString}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-isValid}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-getInvalidFields}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-print}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-clone}{\code{OrgSagebionetworksRepoModelDoiV2DoiAssociation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDoiV2DoiAssociation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$new(
  associationId = NULL,
  etag = NULL,
  doiUri = NULL,
  doiUrl = NULL,
  objectId = NULL,
  objectType = NULL,
  objectVersion = NULL,
  associatedBy = NULL,
  associatedOn = NULL,
  updatedBy = NULL,
  updatedOn = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{associationId}}{associationId}

\item{\code{etag}}{etag}

\item{\code{doiUri}}{doiUri}

\item{\code{doiUrl}}{doiUrl}

\item{\code{objectId}}{objectId}

\item{\code{objectType}}{objectType}

\item{\code{objectVersion}}{objectVersion}

\item{\code{associatedBy}}{associatedBy}

\item{\code{associatedOn}}{associatedOn}

\item{\code{updatedBy}}{updatedBy}

\item{\code{updatedOn}}{updatedOn}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2DoiAssociation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2DoiAssociation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiV2DoiAssociation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2DoiAssociation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2DoiAssociation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiV2DoiAssociation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDoiV2DoiAssociation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDoiV2DoiAssociation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiAssociation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiAssociation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
