% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_doi_v2_doi_resource_type_general.R
\docType{class}
\name{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral}
\alias{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral}
\title{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-new}{\code{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-toJSON}{\code{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-fromJSON}{\code{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-toJSONString}{\code{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-fromJSONString}{\code{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-clone}{\code{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2DoiResourceTypeGeneral$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
