% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_doi_v2_name_identifier_scheme.R
\docType{class}
\name{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme}
\alias{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme}
\title{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-new}{\code{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-toJSON}{\code{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-fromJSON}{\code{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-toJSONString}{\code{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-fromJSONString}{\code{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-clone}{\code{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDoiV2NameIdentifierScheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
