% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_add_batch_of_files_to_download_list_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest}
\alias{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest}
\title{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to add a single batch of files to a user's download list.

OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{batchToAdd}}{The batch of files to add to the user's download list. Note: There is a limit of 1000 files per batch. list(\link{OrgSagebionetworksRepoModelDownloadDownloadListItem}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-new}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-toJSON}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-toString}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-isValid}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-print}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-clone}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$new(
  batchToAdd = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{batchToAdd}}{The batch of files to add to the user's download list. Note: There is a limit of 1000 files per batch.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$toJSON(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$isValid(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
