% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_add_batch_of_files_to_download_list_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse}
\alias{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse}
\title{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse}
\format{
An \code{R6Class} generator object
}
\description{
Response to add a single batch of files to a user's download list.

OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numberOfFilesAdded}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-new}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-toJSON}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-toString}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-isValid}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-print}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-clone}{\code{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$new(
  numberOfFilesAdded = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfFilesAdded}}{numberOfFilesAdded}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$toJSON(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$isValid(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$print(
  
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddBatchOfFilesToDownloadListResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
