% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_add_to_download_list_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest}
\alias{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest}
\title{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest}
\format{
An \code{R6Class} generator object
}
\description{
Start an asynchronous job to add files from the given view query or folder to the user's download list,

OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{query}}{\link{OrgSagebionetworksRepoModelTableQuery} [optional]}

\item{\code{parentId}}{character [optional]}

\item{\code{useVersionNumber}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-new}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-toJSON}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-toString}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-isValid}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-print}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-clone}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$new(
  concreteType,
  query = NULL,
  parentId = NULL,
  useVersionNumber = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{query}}{query}

\item{\code{parentId}}{parentId}

\item{\code{useVersionNumber}}{useVersionNumber}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
