% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_add_to_download_list_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse}
\alias{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse}
\title{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse}
\format{
An \code{R6Class} generator object
}
\description{
The results of a job to add the files from a query result or folder to the user's download list.

OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{numberOfFilesAdded}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-new}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-toJSON}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-toString}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-isValid}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-print}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-clone}{\code{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$new(
  concreteType,
  numberOfFilesAdded = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{numberOfFilesAdded}}{numberOfFilesAdded}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAddToDownloadListResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
