% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_available_files_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest}
\alias{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest}
\title{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to get a single page of files from the user's download list for files that are currently available for download.

OrgSagebionetworksRepoModelDownloadAvailableFilesRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadAvailableFilesRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{nextPageToken}}{character [optional]}

\item{\code{sort}}{Optional. Defines how the query results should be sorted. If excluded a default sort will be used. list(\link{OrgSagebionetworksRepoModelDownloadSort}) [optional]}

\item{\code{filter}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-new}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-toJSON}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-toString}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-isValid}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-print}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-clone}{\code{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadAvailableFilesRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$new(
  concreteType,
  nextPageToken = NULL,
  sort = NULL,
  filter = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{sort}}{Optional. Defines how the query results should be sorted. If excluded a default sort will be used.}

\item{\code{filter}}{filter}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAvailableFilesRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAvailableFilesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAvailableFilesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadAvailableFilesRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadAvailableFilesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadAvailableFilesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadAvailableFilesRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadAvailableFilesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadAvailableFilesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadAvailableFilesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
