% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_download_list_item_result.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadDownloadListItemResult}
\alias{OrgSagebionetworksRepoModelDownloadDownloadListItemResult}
\title{OrgSagebionetworksRepoModelDownloadDownloadListItemResult}
\format{
An \code{R6Class} generator object
}
\description{
The results of a single item on a user's download list.

OrgSagebionetworksRepoModelDownloadDownloadListItemResult Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadDownloadListItemResult
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fileName}}{character [optional]}

\item{\code{addedOn}}{character [optional]}

\item{\code{projectId}}{character [optional]}

\item{\code{projectName}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{fileSizeBytes}}{integer [optional]}

\item{\code{isEligibleForPackaging}}{character [optional]}

\item{\code{fileHandleId}}{character [optional]}

\item{\code{fileEntityId}}{character [optional]}

\item{\code{versionNumber}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-new}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-toJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-toString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-isValid}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-print}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-clone}{\code{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadDownloadListItemResult class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$new(
  fileName = NULL,
  addedOn = NULL,
  projectId = NULL,
  projectName = NULL,
  createdBy = NULL,
  createdOn = NULL,
  fileSizeBytes = NULL,
  isEligibleForPackaging = NULL,
  fileHandleId = NULL,
  fileEntityId = NULL,
  versionNumber = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{fileName}

\item{\code{addedOn}}{addedOn}

\item{\code{projectId}}{projectId}

\item{\code{projectName}}{projectName}

\item{\code{createdBy}}{createdBy}

\item{\code{createdOn}}{createdOn}

\item{\code{fileSizeBytes}}{fileSizeBytes}

\item{\code{isEligibleForPackaging}}{isEligibleForPackaging}

\item{\code{fileHandleId}}{fileHandleId}

\item{\code{fileEntityId}}{fileEntityId}

\item{\code{versionNumber}}{versionNumber}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListItemResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListItemResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListItemResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListItemResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListItemResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListItemResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadDownloadListItemResult and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadDownloadListItemResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListItemResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListItemResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
