% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_download_list_manifest_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest}
\alias{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest}
\title{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest}
\format{
An \code{R6Class} generator object
}
\description{
Start an asynchronous job to generate a metadata manifest CSV of the available files on the use’s download list.

OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{csvTableDescriptor}}{\link{OrgSagebionetworksRepoModelTableCsvTableDescriptor} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-new}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-toJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-toString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-isValid}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-print}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-clone}{\code{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$new(
  concreteType,
  csvTableDescriptor = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{csvTableDescriptor}}{csvTableDescriptor}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListManifestRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
