% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_download_list_package_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest}
\alias{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest}
\title{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to package files from a user's download list.

OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{zipFileName}}{character [optional]}

\item{\code{includeManifest}}{character [optional]}

\item{\code{csvTableDescriptor}}{\link{OrgSagebionetworksRepoModelTableCsvTableDescriptor} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-new}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-toJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-toString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-isValid}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-print}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-clone}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$new(
  concreteType,
  zipFileName = NULL,
  includeManifest = NULL,
  csvTableDescriptor = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{zipFileName}}{zipFileName}

\item{\code{includeManifest}}{includeManifest}

\item{\code{csvTableDescriptor}}{csvTableDescriptor}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
