% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_download_list_package_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse}
\alias{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse}
\title{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse}
\format{
An \code{R6Class} generator object
}
\description{
Response of a job to package files from the user's download list.'

OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{resultFileHandleId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-new}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-toJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-toString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-isValid}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-print}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-clone}{\code{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$new(
  concreteType,
  resultFileHandleId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{resultFileHandleId}}{resultFileHandleId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListPackageResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
