% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_download_list_query_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse}
\alias{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse}
\title{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse}
\format{
An \code{R6Class} generator object
}
\description{
The results of an asynchronous job to query a user's download list.

OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{responseDetails}}{\link{OrgSagebionetworksRepoModelDownloadQueryResponseDetails} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-new}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-toJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-toString}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-isValid}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-print}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-clone}{\code{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$new(
  concreteType,
  responseDetails = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{responseDetails}}{responseDetails}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadDownloadListQueryResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
