% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_enable_two_fa.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadEnableTwoFa}
\alias{OrgSagebionetworksRepoModelDownloadEnableTwoFa}
\title{OrgSagebionetworksRepoModelDownloadEnableTwoFa}
\format{
An \code{R6Class} generator object
}
\description{
In order to download a file the user will need to enable two factor authentication.

OrgSagebionetworksRepoModelDownloadEnableTwoFa Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadEnableTwoFa
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{accessRequirementId}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-new}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-toJSON}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-toString}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-isValid}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-print}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-clone}{\code{OrgSagebionetworksRepoModelDownloadEnableTwoFa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadEnableTwoFa class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$new(
  concreteType,
  accessRequirementId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{accessRequirementId}}{accessRequirementId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadEnableTwoFa in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadEnableTwoFa
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadEnableTwoFa
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadEnableTwoFa in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadEnableTwoFa
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadEnableTwoFa
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadEnableTwoFa and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadEnableTwoFa
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadEnableTwoFa-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadEnableTwoFa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
