% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_files_statistics_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest}
\alias{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest}
\title{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to get statistics about the files currently on the user's download list.

OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-new}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-toJSON}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-toString}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-isValid}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-print}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-clone}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$new(
  concreteType,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
