% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_files_statistics_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse}
\alias{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse}
\title{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse}
\format{
An \code{R6Class} generator object
}
\description{
The response of a FilesStatisticsRequest.

OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{totalNumberOfFiles}}{integer [optional]}

\item{\code{numberOfFilesAvailableForDownload}}{integer [optional]}

\item{\code{numberOfFilesAvailableForDownloadAndEligibleForPackaging}}{integer [optional]}

\item{\code{numberOfFilesRequiringAction}}{integer [optional]}

\item{\code{sumOfFileSizesAvailableForDownload}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-new}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-toJSON}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-toString}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-isValid}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-print}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-clone}{\code{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$new(
  concreteType,
  totalNumberOfFiles = NULL,
  numberOfFilesAvailableForDownload = NULL,
  numberOfFilesAvailableForDownloadAndEligibleForPackaging = NULL,
  numberOfFilesRequiringAction = NULL,
  sumOfFileSizesAvailableForDownload = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{totalNumberOfFiles}}{totalNumberOfFiles}

\item{\code{numberOfFilesAvailableForDownload}}{numberOfFilesAvailableForDownload}

\item{\code{numberOfFilesAvailableForDownloadAndEligibleForPackaging}}{numberOfFilesAvailableForDownloadAndEligibleForPackaging}

\item{\code{numberOfFilesRequiringAction}}{numberOfFilesRequiringAction}

\item{\code{sumOfFileSizesAvailableForDownload}}{sumOfFileSizesAvailableForDownload}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
