% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_query_request_details.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadQueryRequestDetails}
\alias{OrgSagebionetworksRepoModelDownloadQueryRequestDetails}
\title{OrgSagebionetworksRepoModelDownloadQueryRequestDetails}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDownloadQueryRequestDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelDownloadQueryRequestDetails.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-new}{\code{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-toJSON}{\code{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-toString}{\code{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-print}{\code{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-clone}{\code{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadQueryRequestDetails.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelDownloadActionRequiredRequest", "OrgSagebionetworksRepoModelDownloadAvailableFilesRequest", "OrgSagebionetworksRepoModelDownloadFilesStatisticsRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadQueryRequestDetails.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelDownloadQueryRequestDetails.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadQueryRequestDetails.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelDownloadQueryRequestDetails.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelDownloadQueryRequestDetails to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelDownloadQueryRequestDetails.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelDownloadQueryRequestDetails to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelDownloadQueryRequestDetails.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelDownloadQueryRequestDetails and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryRequestDetails-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryRequestDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
