% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_query_response_details.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadQueryResponseDetails}
\alias{OrgSagebionetworksRepoModelDownloadQueryResponseDetails}
\title{OrgSagebionetworksRepoModelDownloadQueryResponseDetails}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDownloadQueryResponseDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelDownloadQueryResponseDetails.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-new}{\code{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-toJSON}{\code{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-toString}{\code{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-print}{\code{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-clone}{\code{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadQueryResponseDetails.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelDownloadActionRequiredResponse", "OrgSagebionetworksRepoModelDownloadAvailableFilesResponse", "OrgSagebionetworksRepoModelDownloadFilesStatisticsResponse"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadQueryResponseDetails.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelDownloadQueryResponseDetails.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadQueryResponseDetails.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelDownloadQueryResponseDetails.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelDownloadQueryResponseDetails to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelDownloadQueryResponseDetails.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelDownloadQueryResponseDetails to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelDownloadQueryResponseDetails.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelDownloadQueryResponseDetails and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadQueryResponseDetails-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadQueryResponseDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
