% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_download_remove_batch_of_files_from_download_list_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse}
\alias{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse}
\title{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse}
\format{
An \code{R6Class} generator object
}
\description{
Response to remove a single batch of files from the user's download list.

OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numberOfFilesRemoved}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-new}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-toJSON}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-fromJSON}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-toJSONString}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-validateJSON}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-toString}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-isValid}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-print}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-clone}{\code{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$new(
  numberOfFilesRemoved = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfFilesRemoved}}{numberOfFilesRemoved}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$toJSON(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$isValid(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$print(
  
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDownloadRemoveBatchOfFilesFromDownloadListResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
