% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_drs_access_method.R
\docType{class}
\name{OrgSagebionetworksRepoModelDrsAccessMethod}
\alias{OrgSagebionetworksRepoModelDrsAccessMethod}
\title{OrgSagebionetworksRepoModelDrsAccessMethod}
\format{
An \code{R6Class} generator object
}
\description{
The list of access methods that can be used to fetch the DrsObject. Required for single blobs and bundles have no access method.

OrgSagebionetworksRepoModelDrsAccessMethod Class
}
\details{
Create a new OrgSagebionetworksRepoModelDrsAccessMethod
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{character [optional]}

\item{\code{access_id}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-new}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-toJSON}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-fromJSON}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-toJSONString}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-fromJSONString}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-validateJSON}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-toString}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-isValid}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-getInvalidFields}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-print}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethod-clone}{\code{OrgSagebionetworksRepoModelDrsAccessMethod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDrsAccessMethod class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$new(
  type = NULL,
  access_id = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{type}

\item{\code{access_id}}{access_id}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsAccessMethod in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsAccessMethod
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsAccessMethod
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsAccessMethod in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsAccessMethod
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsAccessMethod
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDrsAccessMethod and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDrsAccessMethod
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethod-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethod-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
