% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_drs_access_method_type.R
\docType{class}
\name{OrgSagebionetworksRepoModelDrsAccessMethodType}
\alias{OrgSagebionetworksRepoModelDrsAccessMethodType}
\title{OrgSagebionetworksRepoModelDrsAccessMethodType}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelDrsAccessMethodType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethodType-new}{\code{OrgSagebionetworksRepoModelDrsAccessMethodType$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethodType-toJSON}{\code{OrgSagebionetworksRepoModelDrsAccessMethodType$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethodType-fromJSON}{\code{OrgSagebionetworksRepoModelDrsAccessMethodType$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethodType-toJSONString}{\code{OrgSagebionetworksRepoModelDrsAccessMethodType$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethodType-fromJSONString}{\code{OrgSagebionetworksRepoModelDrsAccessMethodType$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessMethodType-clone}{\code{OrgSagebionetworksRepoModelDrsAccessMethodType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethodType-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethodType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDrsAccessMethodType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethodType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethodType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethodType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethodType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsAccessMethodType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethodType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethodType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsAccessMethodType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethodType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsAccessMethodType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethodType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethodType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethodType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsAccessMethodType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethodType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethodType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsAccessMethodType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethodType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsAccessMethodType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessMethodType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessMethodType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessMethodType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
