% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_drs_access_url.R
\docType{class}
\name{OrgSagebionetworksRepoModelDrsAccessUrl}
\alias{OrgSagebionetworksRepoModelDrsAccessUrl}
\title{OrgSagebionetworksRepoModelDrsAccessUrl}
\format{
An \code{R6Class} generator object
}
\description{
A fully resolvable presigned URL that can be used to download the object bytes without any authentication.As presigned url has tokens included, which expires with time

OrgSagebionetworksRepoModelDrsAccessUrl Class
}
\details{
Create a new OrgSagebionetworksRepoModelDrsAccessUrl
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-new}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-toJSON}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-fromJSON}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-toJSONString}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-fromJSONString}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-validateJSON}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-toString}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-isValid}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-getInvalidFields}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-print}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsAccessUrl-clone}{\code{OrgSagebionetworksRepoModelDrsAccessUrl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDrsAccessUrl class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$new(url = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{url}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsAccessUrl in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsAccessUrl
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsAccessUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsAccessUrl in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsAccessUrl
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsAccessUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDrsAccessUrl and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDrsAccessUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsAccessUrl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsAccessUrl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsAccessUrl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
