% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_drs_organization_information.R
\docType{class}
\name{OrgSagebionetworksRepoModelDrsOrganizationInformation}
\alias{OrgSagebionetworksRepoModelDrsOrganizationInformation}
\title{OrgSagebionetworksRepoModelDrsOrganizationInformation}
\format{
An \code{R6Class} generator object
}
\description{
The information of the organization, which provides DRS API implementation.

OrgSagebionetworksRepoModelDrsOrganizationInformation Class
}
\details{
Create a new OrgSagebionetworksRepoModelDrsOrganizationInformation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{url}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-new}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-toJSON}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-fromJSON}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-toJSONString}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-fromJSONString}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-validateJSON}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-toString}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-isValid}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-getInvalidFields}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-print}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsOrganizationInformation-clone}{\code{OrgSagebionetworksRepoModelDrsOrganizationInformation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDrsOrganizationInformation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$new(
  name = NULL,
  url = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{url}}{url}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsOrganizationInformation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsOrganizationInformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsOrganizationInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsOrganizationInformation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsOrganizationInformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsOrganizationInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDrsOrganizationInformation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDrsOrganizationInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsOrganizationInformation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsOrganizationInformation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsOrganizationInformation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
