% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_drs_service_information.R
\docType{class}
\name{OrgSagebionetworksRepoModelDrsServiceInformation}
\alias{OrgSagebionetworksRepoModelDrsServiceInformation}
\title{OrgSagebionetworksRepoModelDrsServiceInformation}
\format{
An \code{R6Class} generator object
}
\description{
Drs service information

OrgSagebionetworksRepoModelDrsServiceInformation Class
}
\details{
Create a new OrgSagebionetworksRepoModelDrsServiceInformation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{type}}{\link{OrgSagebionetworksRepoModelDrsPackageInformation} [optional]}

\item{\code{description}}{character [optional]}

\item{\code{organization}}{\link{OrgSagebionetworksRepoModelDrsOrganizationInformation} [optional]}

\item{\code{contactUrl}}{character [optional]}

\item{\code{documentationUrl}}{character [optional]}

\item{\code{createdAt}}{character [optional]}

\item{\code{updatedAt}}{character [optional]}

\item{\code{environment}}{character [optional]}

\item{\code{version}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-new}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$new()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-toJSON}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-fromJSON}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-toJSONString}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-fromJSONString}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-validateJSON}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-toString}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-isValid}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-getInvalidFields}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-print}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$print()}}
\item \href{#method-OrgSagebionetworksRepoModelDrsServiceInformation-clone}{\code{OrgSagebionetworksRepoModelDrsServiceInformation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelDrsServiceInformation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$new(
  id = NULL,
  name = NULL,
  type = NULL,
  description = NULL,
  organization = NULL,
  contactUrl = NULL,
  documentationUrl = NULL,
  createdAt = NULL,
  updatedAt = NULL,
  environment = NULL,
  version = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{type}}{type}

\item{\code{description}}{description}

\item{\code{organization}}{organization}

\item{\code{contactUrl}}{contactUrl}

\item{\code{documentationUrl}}{documentationUrl}

\item{\code{createdAt}}{createdAt}

\item{\code{updatedAt}}{updatedAt}

\item{\code{environment}}{environment}

\item{\code{version}}{version}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsServiceInformation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsServiceInformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsServiceInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelDrsServiceInformation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelDrsServiceInformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelDrsServiceInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelDrsServiceInformation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelDrsServiceInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelDrsServiceInformation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelDrsServiceInformation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelDrsServiceInformation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
