% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_entitybundle_v2_entity_bundle.R
\docType{class}
\name{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle}
\alias{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle}
\title{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle}
\format{
An \code{R6Class} generator object
}
\description{
Bundle to transport an Entity and related data objects

OrgSagebionetworksRepoModelEntitybundleV2EntityBundle Class
}
\details{
Create a new OrgSagebionetworksRepoModelEntitybundleV2EntityBundle
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entity}}{\link{OrgSagebionetworksRepoModelEntity} [optional]}

\item{\code{entityType}}{character [optional]}

\item{\code{annotations}}{\link{OrgSagebionetworksRepoModelAnnotationV2Annotations} [optional]}

\item{\code{permissions}}{\link{OrgSagebionetworksRepoModelAuthUserEntityPermissions} [optional]}

\item{\code{path}}{\link{OrgSagebionetworksRepoModelEntityPath} [optional]}

\item{\code{hasChildren}}{character [optional]}

\item{\code{accessControlList}}{\link{OrgSagebionetworksRepoModelAccessControlList} [optional]}

\item{\code{fileHandles}}{FileHandles associated with this Entity list(\link{OrgSagebionetworksRepoModelFileFileHandle}) [optional]}

\item{\code{tableBundle}}{\link{OrgSagebionetworksRepoModelTableTableBundle} [optional]}

\item{\code{rootWikiId}}{character [optional]}

\item{\code{benefactorAcl}}{\link{OrgSagebionetworksRepoModelAccessControlList} [optional]}

\item{\code{doiAssociation}}{\link{OrgSagebionetworksRepoModelDoiV2DoiAssociation} [optional]}

\item{\code{fileName}}{character [optional]}

\item{\code{threadCount}}{integer [optional]}

\item{\code{restrictionInformation}}{\link{OrgSagebionetworksRepoModelRestrictionInformationResponse} [optional]}

\item{\code{activity}}{\link{OrgSagebionetworksRepoModelProvenanceActivity} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-new}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$new()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-toJSON}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-fromJSON}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-toJSONString}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-fromJSONString}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-validateJSON}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-toString}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-isValid}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-getInvalidFields}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-print}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$print()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-clone}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelEntitybundleV2EntityBundle class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$new(
  entity = NULL,
  entityType = NULL,
  annotations = NULL,
  permissions = NULL,
  path = NULL,
  hasChildren = NULL,
  accessControlList = NULL,
  fileHandles = NULL,
  tableBundle = NULL,
  rootWikiId = NULL,
  benefactorAcl = NULL,
  doiAssociation = NULL,
  fileName = NULL,
  threadCount = NULL,
  restrictionInformation = NULL,
  activity = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{entity}

\item{\code{entityType}}{entityType}

\item{\code{annotations}}{annotations}

\item{\code{permissions}}{permissions}

\item{\code{path}}{path}

\item{\code{hasChildren}}{hasChildren}

\item{\code{accessControlList}}{accessControlList}

\item{\code{fileHandles}}{FileHandles associated with this Entity}

\item{\code{tableBundle}}{tableBundle}

\item{\code{rootWikiId}}{rootWikiId}

\item{\code{benefactorAcl}}{benefactorAcl}

\item{\code{doiAssociation}}{doiAssociation}

\item{\code{fileName}}{fileName}

\item{\code{threadCount}}{threadCount}

\item{\code{restrictionInformation}}{restrictionInformation}

\item{\code{activity}}{activity}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntitybundleV2EntityBundle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntitybundleV2EntityBundle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelEntitybundleV2EntityBundle and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelEntitybundleV2EntityBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
