% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_entitybundle_v2_entity_bundle_create.R
\docType{class}
\name{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate}
\alias{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate}
\title{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate}
\format{
An \code{R6Class} generator object
}
\description{
Used to create Entity and associated metadata in a single transaction

OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate Class
}
\details{
Create a new OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entity}}{\link{OrgSagebionetworksRepoModelEntity} [optional]}

\item{\code{annotations}}{\link{OrgSagebionetworksRepoModelAnnotationV2Annotations} [optional]}

\item{\code{accessControlList}}{\link{OrgSagebionetworksRepoModelAccessControlList} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-new}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$new()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-toJSON}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-fromJSON}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-toJSONString}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-fromJSONString}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-validateJSON}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-toString}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-isValid}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-getInvalidFields}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-print}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$print()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-clone}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$new(
  entity = NULL,
  annotations = NULL,
  accessControlList = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{entity}

\item{\code{annotations}}{annotations}

\item{\code{accessControlList}}{accessControlList}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
