% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_entitybundle_v2_entity_bundle_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest}
\alias{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest}
\title{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest}
\format{
An \code{R6Class} generator object
}
\description{
Specifies what fields to include in an EntityBundle

OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{includeEntity}}{character [optional]}

\item{\code{includeAnnotations}}{character [optional]}

\item{\code{includePermissions}}{character [optional]}

\item{\code{includeEntityPath}}{character [optional]}

\item{\code{includeHasChildren}}{character [optional]}

\item{\code{includeAccessControlList}}{character [optional]}

\item{\code{includeFileHandles}}{character [optional]}

\item{\code{includeTableBundle}}{character [optional]}

\item{\code{includeRootWikiId}}{character [optional]}

\item{\code{includeBenefactorACL}}{character [optional]}

\item{\code{includeDOIAssociation}}{character [optional]}

\item{\code{includeFileName}}{character [optional]}

\item{\code{includeThreadCount}}{character [optional]}

\item{\code{includeRestrictionInformation}}{character [optional]}

\item{\code{includeActivity}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-new}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-toJSON}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-fromJSON}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-toJSONString}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-validateJSON}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-toString}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-isValid}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-print}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-clone}{\code{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$new(
  includeEntity = NULL,
  includeAnnotations = NULL,
  includePermissions = NULL,
  includeEntityPath = NULL,
  includeHasChildren = NULL,
  includeAccessControlList = NULL,
  includeFileHandles = NULL,
  includeTableBundle = NULL,
  includeRootWikiId = NULL,
  includeBenefactorACL = NULL,
  includeDOIAssociation = NULL,
  includeFileName = NULL,
  includeThreadCount = NULL,
  includeRestrictionInformation = NULL,
  includeActivity = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeEntity}}{includeEntity}

\item{\code{includeAnnotations}}{includeAnnotations}

\item{\code{includePermissions}}{includePermissions}

\item{\code{includeEntityPath}}{includeEntityPath}

\item{\code{includeHasChildren}}{includeHasChildren}

\item{\code{includeAccessControlList}}{includeAccessControlList}

\item{\code{includeFileHandles}}{includeFileHandles}

\item{\code{includeTableBundle}}{includeTableBundle}

\item{\code{includeRootWikiId}}{includeRootWikiId}

\item{\code{includeBenefactorACL}}{includeBenefactorACL}

\item{\code{includeDOIAssociation}}{includeDOIAssociation}

\item{\code{includeFileName}}{includeFileName}

\item{\code{includeThreadCount}}{includeThreadCount}

\item{\code{includeRestrictionInformation}}{includeRestrictionInformation}

\item{\code{includeActivity}}{includeActivity}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntitybundleV2EntityBundleRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
