% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_entity_children_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelEntityChildrenRequest}
\alias{OrgSagebionetworksRepoModelEntityChildrenRequest}
\title{OrgSagebionetworksRepoModelEntityChildrenRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request for EntityHeaders of the children of a given parent Entity

OrgSagebionetworksRepoModelEntityChildrenRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelEntityChildrenRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parentId}}{character [optional]}

\item{\code{nextPageToken}}{character [optional]}

\item{\code{includeTypes}}{The types of children to be include. Must include at least one type. list(character) [optional]}

\item{\code{sortBy}}{character [optional]}

\item{\code{sortDirection}}{character [optional]}

\item{\code{includeTotalChildCount}}{character [optional]}

\item{\code{includeSumFileSizes}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-new}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-toJSON}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-fromJSON}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-toJSONString}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-validateJSON}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-toString}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-isValid}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-print}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityChildrenRequest-clone}{\code{OrgSagebionetworksRepoModelEntityChildrenRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelEntityChildrenRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$new(
  parentId = NULL,
  nextPageToken = NULL,
  includeTypes = NULL,
  sortBy = NULL,
  sortDirection = NULL,
  includeTotalChildCount = NULL,
  includeSumFileSizes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parentId}}{parentId}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{includeTypes}}{The types of children to be include. Must include at least one type.}

\item{\code{sortBy}}{sortBy}

\item{\code{sortDirection}}{sortDirection}

\item{\code{includeTotalChildCount}}{includeTotalChildCount}

\item{\code{includeSumFileSizes}}{includeSumFileSizes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntityChildrenRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntityChildrenRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntityChildrenRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntityChildrenRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntityChildrenRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntityChildrenRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelEntityChildrenRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelEntityChildrenRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityChildrenRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityChildrenRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityChildrenRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
