% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_entity_file_handle_update_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest}
\alias{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest}
\title{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest}
\format{
An \code{R6Class} generator object
}
\description{
Reuqest that can be used to update the file handle of an <a href=\"org.sagebionetworks.repo.model.FileEntity>FileEntity</a>\". The MD5 of new file handle should match the MD5 of the current file handle.

OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{oldFileHandleId}}{character [optional]}

\item{\code{newFileHandleId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-new}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-toJSON}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-fromJSON}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-toJSONString}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-validateJSON}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-toString}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-isValid}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-print}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-clone}{\code{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$new(
  oldFileHandleId = NULL,
  newFileHandleId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{oldFileHandleId}}{oldFileHandleId}

\item{\code{newFileHandleId}}{newFileHandleId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityFileHandleUpdateRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
