% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_entity_header.R
\docType{class}
\name{OrgSagebionetworksRepoModelEntityHeader}
\alias{OrgSagebionetworksRepoModelEntityHeader}
\title{OrgSagebionetworksRepoModelEntityHeader}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for EntityHeader POJO

OrgSagebionetworksRepoModelEntityHeader Class
}
\details{
Create a new OrgSagebionetworksRepoModelEntityHeader
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{id}}{character [optional]}

\item{\code{type}}{character [optional]}

\item{\code{versionNumber}}{integer [optional]}

\item{\code{versionLabel}}{character [optional]}

\item{\code{isLatestVersion}}{character [optional]}

\item{\code{benefactorId}}{integer [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-new}{\code{OrgSagebionetworksRepoModelEntityHeader$new()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-toJSON}{\code{OrgSagebionetworksRepoModelEntityHeader$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-fromJSON}{\code{OrgSagebionetworksRepoModelEntityHeader$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-toJSONString}{\code{OrgSagebionetworksRepoModelEntityHeader$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-fromJSONString}{\code{OrgSagebionetworksRepoModelEntityHeader$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-validateJSON}{\code{OrgSagebionetworksRepoModelEntityHeader$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-toString}{\code{OrgSagebionetworksRepoModelEntityHeader$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-isValid}{\code{OrgSagebionetworksRepoModelEntityHeader$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-getInvalidFields}{\code{OrgSagebionetworksRepoModelEntityHeader$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-print}{\code{OrgSagebionetworksRepoModelEntityHeader$print()}}
\item \href{#method-OrgSagebionetworksRepoModelEntityHeader-clone}{\code{OrgSagebionetworksRepoModelEntityHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelEntityHeader class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$new(
  name = NULL,
  id = NULL,
  type = NULL,
  versionNumber = NULL,
  versionLabel = NULL,
  isLatestVersion = NULL,
  benefactorId = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{id}}{id}

\item{\code{type}}{type}

\item{\code{versionNumber}}{versionNumber}

\item{\code{versionLabel}}{versionLabel}

\item{\code{isLatestVersion}}{isLatestVersion}

\item{\code{benefactorId}}{benefactorId}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntityHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntityHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntityHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEntityHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEntityHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEntityHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelEntityHeader and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelEntityHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEntityHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEntityHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEntityHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
