% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_environment_descriptor.R
\docType{class}
\name{OrgSagebionetworksRepoModelEnvironmentDescriptor}
\alias{OrgSagebionetworksRepoModelEnvironmentDescriptor}
\title{OrgSagebionetworksRepoModelEnvironmentDescriptor}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for environment descriptor POJO

OrgSagebionetworksRepoModelEnvironmentDescriptor Class
}
\details{
Create a new OrgSagebionetworksRepoModelEnvironmentDescriptor
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{quantifier}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-new}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$new()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-toJSON}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-fromJSON}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-toJSONString}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-fromJSONString}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-validateJSON}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-toString}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-isValid}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-getInvalidFields}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-print}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$print()}}
\item \href{#method-OrgSagebionetworksRepoModelEnvironmentDescriptor-clone}{\code{OrgSagebionetworksRepoModelEnvironmentDescriptor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelEnvironmentDescriptor class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$new(
  type = NULL,
  name = NULL,
  quantifier = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{type}

\item{\code{name}}{name}

\item{\code{quantifier}}{quantifier}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEnvironmentDescriptor in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEnvironmentDescriptor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEnvironmentDescriptor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelEnvironmentDescriptor in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelEnvironmentDescriptor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelEnvironmentDescriptor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelEnvironmentDescriptor and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelEnvironmentDescriptor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelEnvironmentDescriptor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelEnvironmentDescriptor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelEnvironmentDescriptor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
