% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_error_response_code.R
\docType{class}
\name{OrgSagebionetworksRepoModelErrorResponseCode}
\alias{OrgSagebionetworksRepoModelErrorResponseCode}
\title{OrgSagebionetworksRepoModelErrorResponseCode}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelErrorResponseCode Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelErrorResponseCode-new}{\code{OrgSagebionetworksRepoModelErrorResponseCode$new()}}
\item \href{#method-OrgSagebionetworksRepoModelErrorResponseCode-toJSON}{\code{OrgSagebionetworksRepoModelErrorResponseCode$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelErrorResponseCode-fromJSON}{\code{OrgSagebionetworksRepoModelErrorResponseCode$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelErrorResponseCode-toJSONString}{\code{OrgSagebionetworksRepoModelErrorResponseCode$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelErrorResponseCode-fromJSONString}{\code{OrgSagebionetworksRepoModelErrorResponseCode$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelErrorResponseCode-clone}{\code{OrgSagebionetworksRepoModelErrorResponseCode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelErrorResponseCode-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelErrorResponseCode-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelErrorResponseCode class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelErrorResponseCode$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelErrorResponseCode-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelErrorResponseCode-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelErrorResponseCode$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelErrorResponseCode in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelErrorResponseCode-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelErrorResponseCode-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelErrorResponseCode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelErrorResponseCode$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelErrorResponseCode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelErrorResponseCode-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelErrorResponseCode-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelErrorResponseCode$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelErrorResponseCode in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelErrorResponseCode-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelErrorResponseCode-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelErrorResponseCode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelErrorResponseCode$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelErrorResponseCode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelErrorResponseCode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelErrorResponseCode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelErrorResponseCode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
