% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_example_entity.R
\docType{class}
\name{OrgSagebionetworksRepoModelExampleEntity}
\alias{OrgSagebionetworksRepoModelExampleEntity}
\title{OrgSagebionetworksRepoModelExampleEntity}
\format{
An \code{R6Class} generator object
}
\description{
This object is deprecated and will be removed in future versions of Synapse.

OrgSagebionetworksRepoModelExampleEntity Class
}
\details{
Create a new OrgSagebionetworksRepoModelExampleEntity
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{parentId}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{singleString}}{character [optional]}

\item{\code{stringList}}{This is an example of a list of strings list(character) [optional]}

\item{\code{singleDate}}{character [optional]}

\item{\code{dateList}}{This is an example of a list of dates list(character) [optional]}

\item{\code{singleDouble}}{numeric [optional]}

\item{\code{doubleList}}{This is an example of a double list list(numeric) [optional]}

\item{\code{singleInteger}}{integer [optional]}

\item{\code{integerList}}{This is an example of an Integer List. list(integer) [optional]}

\item{\code{concept}}{character [optional]}

\item{\code{someEnum}}{character [optional]}

\item{\code{references}}{References list(\link{OrgSagebionetworksRepoModelReference}) [optional]}

\item{\code{environmentDescriptors}}{The list of environment descriptors list(\link{OrgSagebionetworksRepoModelEnvironmentDescriptor}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-new}{\code{OrgSagebionetworksRepoModelExampleEntity$new()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-toJSON}{\code{OrgSagebionetworksRepoModelExampleEntity$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-fromJSON}{\code{OrgSagebionetworksRepoModelExampleEntity$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-toJSONString}{\code{OrgSagebionetworksRepoModelExampleEntity$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-fromJSONString}{\code{OrgSagebionetworksRepoModelExampleEntity$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-validateJSON}{\code{OrgSagebionetworksRepoModelExampleEntity$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-toString}{\code{OrgSagebionetworksRepoModelExampleEntity$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-isValid}{\code{OrgSagebionetworksRepoModelExampleEntity$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-getInvalidFields}{\code{OrgSagebionetworksRepoModelExampleEntity$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-print}{\code{OrgSagebionetworksRepoModelExampleEntity$print()}}
\item \href{#method-OrgSagebionetworksRepoModelExampleEntity-clone}{\code{OrgSagebionetworksRepoModelExampleEntity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelExampleEntity class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$new(
  concreteType,
  name = NULL,
  description = NULL,
  id = NULL,
  etag = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  parentId = NULL,
  singleString = NULL,
  stringList = NULL,
  singleDate = NULL,
  dateList = NULL,
  singleDouble = NULL,
  doubleList = NULL,
  singleInteger = NULL,
  integerList = NULL,
  concept = NULL,
  someEnum = NULL,
  references = NULL,
  environmentDescriptors = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{parentId}}{parentId}

\item{\code{singleString}}{singleString}

\item{\code{stringList}}{This is an example of a list of strings}

\item{\code{singleDate}}{singleDate}

\item{\code{dateList}}{This is an example of a list of dates}

\item{\code{singleDouble}}{singleDouble}

\item{\code{doubleList}}{This is an example of a double list}

\item{\code{singleInteger}}{singleInteger}

\item{\code{integerList}}{This is an example of an Integer List.}

\item{\code{concept}}{concept}

\item{\code{someEnum}}{someEnum}

\item{\code{references}}{References}

\item{\code{environmentDescriptors}}{The list of environment descriptors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelExampleEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelExampleEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelExampleEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelExampleEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelExampleEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelExampleEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelExampleEntity and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelExampleEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelExampleEntity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelExampleEntity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelExampleEntity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
