% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_favorite_sort_by.R
\docType{class}
\name{OrgSagebionetworksRepoModelFavoriteSortBy}
\alias{OrgSagebionetworksRepoModelFavoriteSortBy}
\title{OrgSagebionetworksRepoModelFavoriteSortBy}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelFavoriteSortBy Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortBy-new}{\code{OrgSagebionetworksRepoModelFavoriteSortBy$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortBy-toJSON}{\code{OrgSagebionetworksRepoModelFavoriteSortBy$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortBy-fromJSON}{\code{OrgSagebionetworksRepoModelFavoriteSortBy$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortBy-toJSONString}{\code{OrgSagebionetworksRepoModelFavoriteSortBy$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortBy-fromJSONString}{\code{OrgSagebionetworksRepoModelFavoriteSortBy$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortBy-clone}{\code{OrgSagebionetworksRepoModelFavoriteSortBy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortBy-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortBy-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFavoriteSortBy class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortBy$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortBy-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortBy-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortBy$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFavoriteSortBy in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortBy-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortBy-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFavoriteSortBy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortBy$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFavoriteSortBy
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortBy-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortBy-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortBy$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFavoriteSortBy in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortBy-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortBy-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFavoriteSortBy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortBy$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFavoriteSortBy
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortBy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortBy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortBy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
