% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_favorite_sort_direction.R
\docType{class}
\name{OrgSagebionetworksRepoModelFavoriteSortDirection}
\alias{OrgSagebionetworksRepoModelFavoriteSortDirection}
\title{OrgSagebionetworksRepoModelFavoriteSortDirection}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelFavoriteSortDirection Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortDirection-new}{\code{OrgSagebionetworksRepoModelFavoriteSortDirection$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortDirection-toJSON}{\code{OrgSagebionetworksRepoModelFavoriteSortDirection$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortDirection-fromJSON}{\code{OrgSagebionetworksRepoModelFavoriteSortDirection$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortDirection-toJSONString}{\code{OrgSagebionetworksRepoModelFavoriteSortDirection$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortDirection-fromJSONString}{\code{OrgSagebionetworksRepoModelFavoriteSortDirection$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFavoriteSortDirection-clone}{\code{OrgSagebionetworksRepoModelFavoriteSortDirection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortDirection-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortDirection-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFavoriteSortDirection class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortDirection$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortDirection-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortDirection-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortDirection$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFavoriteSortDirection in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortDirection-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortDirection-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFavoriteSortDirection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortDirection$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFavoriteSortDirection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortDirection-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortDirection-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortDirection$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFavoriteSortDirection in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortDirection-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortDirection-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFavoriteSortDirection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortDirection$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFavoriteSortDirection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFavoriteSortDirection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFavoriteSortDirection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFavoriteSortDirection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
