% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_add_file_to_download_list_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest}
\alias{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest}
\title{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to add either the children of a folder or the files of a view query to the user's download list.

OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{folderId}}{character [optional]}

\item{\code{query}}{\link{OrgSagebionetworksRepoModelTableQuery} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-new}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-toString}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-isValid}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-print}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-clone}{\code{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$new(
  concreteType,
  folderId = NULL,
  query = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{folderId}}{folderId}

\item{\code{query}}{query}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileAddFileToDownloadListRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
