% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_base_key_upload_destination.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination}
\alias{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination}
\title{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelFileBaseKeyUploadDestination Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelFileBaseKeyUploadDestination.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-new}{\code{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-fromJSONString}{\code{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-fromJSON}{\code{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-toJSONString}{\code{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-toJSON}{\code{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-validateJSON}{\code{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-toString}{\code{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-print}{\code{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-clone}{\code{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileBaseKeyUploadDestination.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination", "OrgSagebionetworksRepoModelFileExternalS3UploadDestination", "OrgSagebionetworksRepoModelFileS3UploadDestination"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBaseKeyUploadDestination.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelFileBaseKeyUploadDestination.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBaseKeyUploadDestination.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelFileBaseKeyUploadDestination.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelFileBaseKeyUploadDestination to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelFileBaseKeyUploadDestination.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelFileBaseKeyUploadDestination to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelFileBaseKeyUploadDestination.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelFileBaseKeyUploadDestination and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBaseKeyUploadDestination-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBaseKeyUploadDestination$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
