% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_batch_file_handle_copy_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest}
\alias{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest}
\title{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest}
\format{
An \code{R6Class} generator object
}
\description{
Batch Request for copying FileHandles.

OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{copyRequests}}{Defines the copy requests. list(\link{OrgSagebionetworksRepoModelFileFileHandleCopyRequest}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-new}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-toString}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-isValid}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-print}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-clone}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$new(
  copyRequests = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{copyRequests}}{Defines the copy requests.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
