% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_batch_file_handle_copy_result.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult}
\alias{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult}
\title{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult}
\format{
An \code{R6Class} generator object
}
\description{
Batch Result for copying FileHandles.

OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{copyResults}}{The results of each individual requests. list(\link{OrgSagebionetworksRepoModelFileFileHandleCopyResult}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-new}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-toJSON}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-fromJSON}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-toJSONString}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-fromJSONString}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-validateJSON}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-toString}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-isValid}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-print}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-clone}{\code{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$new(
  copyResults = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{copyResults}}{The results of each individual requests.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileHandleCopyResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
