% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_batch_file_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileBatchFileRequest}
\alias{OrgSagebionetworksRepoModelFileBatchFileRequest}
\title{OrgSagebionetworksRepoModelFileBatchFileRequest}
\format{
An \code{R6Class} generator object
}
\description{
Batch Request for file pre-signed-URLs and/or FileHandles.

OrgSagebionetworksRepoModelFileBatchFileRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileBatchFileRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{requestedFiles}}{Defines the files to get. list(\link{OrgSagebionetworksRepoModelFileFileHandleAssociation}) [optional]}

\item{\code{includePreSignedURLs}}{character [optional]}

\item{\code{includeFileHandles}}{character [optional]}

\item{\code{includePreviewPreSignedURLs}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-new}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-toString}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-isValid}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-print}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchFileRequest-clone}{\code{OrgSagebionetworksRepoModelFileBatchFileRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileBatchFileRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$new(
  requestedFiles = NULL,
  includePreSignedURLs = NULL,
  includeFileHandles = NULL,
  includePreviewPreSignedURLs = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{requestedFiles}}{Defines the files to get.}

\item{\code{includePreSignedURLs}}{includePreSignedURLs}

\item{\code{includeFileHandles}}{includeFileHandles}

\item{\code{includePreviewPreSignedURLs}}{includePreviewPreSignedURLs}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchFileRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchFileRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchFileRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchFileRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchFileRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchFileRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileBatchFileRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileBatchFileRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchFileRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchFileRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchFileRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
