% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_batch_presigned_upload_url_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest}
\alias{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest}
\title{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request a batch of multi-part upload or copy pre-signed URLs.

OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uploadId}}{character [optional]}

\item{\code{partNumbers}}{The part numbers to get pre-signed URLs for. list(integer) [optional]}

\item{\code{contentType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-new}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-toString}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-isValid}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-print}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-clone}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$new(
  uploadId = NULL,
  partNumbers = NULL,
  contentType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uploadId}}{uploadId}

\item{\code{partNumbers}}{The part numbers to get pre-signed URLs for.}

\item{\code{contentType}}{contentType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
