% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_batch_presigned_upload_url_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse}
\alias{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse}
\title{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse}
\format{
An \code{R6Class} generator object
}
\description{
Batch response of part pre-signed URLs.

OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{partPresignedUrls}}{List of part pre-signed URLs. list(\link{OrgSagebionetworksRepoModelFilePartPresignedUrl}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-new}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-toJSON}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-fromJSON}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-toJSONString}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-validateJSON}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-toString}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-isValid}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-print}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-clone}{\code{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$new(
  partPresignedUrls = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{partPresignedUrls}}{List of part pre-signed URLs.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBatchPresignedUploadUrlResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
