% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_bulk_file_download_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest}
\alias{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest}
\title{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request for a bulk file download.

OrgSagebionetworksRepoModelFileBulkFileDownloadRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileBulkFileDownloadRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{requestedFiles}}{The files to be included in the download. list(\link{OrgSagebionetworksRepoModelFileFileHandleAssociation}) [optional]}

\item{\code{zipFileName}}{character [optional]}

\item{\code{zipFileFormat}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-new}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-toString}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-isValid}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-print}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-clone}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileBulkFileDownloadRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$new(
  concreteType,
  requestedFiles = NULL,
  zipFileName = NULL,
  zipFileFormat = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{requestedFiles}}{The files to be included in the download.}

\item{\code{zipFileName}}{zipFileName}

\item{\code{zipFileFormat}}{zipFileFormat}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBulkFileDownloadRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBulkFileDownloadRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBulkFileDownloadRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBulkFileDownloadRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBulkFileDownloadRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBulkFileDownloadRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileBulkFileDownloadRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileBulkFileDownloadRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
