% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_bulk_file_download_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse}
\alias{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse}
\title{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse}
\format{
An \code{R6Class} generator object
}
\description{
Response for a bulk file download.

OrgSagebionetworksRepoModelFileBulkFileDownloadResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileBulkFileDownloadResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{resultZipFileHandleId}}{character [optional]}

\item{\code{fileSummary}}{The summary of each requested file. list(\link{OrgSagebionetworksRepoModelFileFileDownloadSummary}) [optional]}

\item{\code{userId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-new}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-toJSON}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-fromJSON}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-toJSONString}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-validateJSON}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-toString}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-isValid}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-print}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-clone}{\code{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileBulkFileDownloadResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$new(
  concreteType,
  resultZipFileHandleId = NULL,
  fileSummary = NULL,
  userId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{resultZipFileHandleId}}{resultZipFileHandleId}

\item{\code{fileSummary}}{The summary of each requested file.}

\item{\code{userId}}{userId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBulkFileDownloadResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBulkFileDownloadResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBulkFileDownloadResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileBulkFileDownloadResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileBulkFileDownloadResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileBulkFileDownloadResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileBulkFileDownloadResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileBulkFileDownloadResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileBulkFileDownloadResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileBulkFileDownloadResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
