% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_cloud_provider_file_handle_interface.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface}
\alias{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface}
\title{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-new}{\code{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-fromJSONString}{\code{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-fromJSON}{\code{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-toJSONString}{\code{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-toJSON}{\code{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-validateJSON}{\code{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-toString}{\code{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-print}{\code{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-clone}{\code{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$new(
  instance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelFileGoogleCloudFileHandle", "OrgSagebionetworksRepoModelFileS3FileHandle"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileCloudProviderFileHandleInterface$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
