% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_download_order.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileDownloadOrder}
\alias{OrgSagebionetworksRepoModelFileDownloadOrder}
\title{OrgSagebionetworksRepoModelFileDownloadOrder}
\format{
An \code{R6Class} generator object
}
\description{
The files of a download order.

OrgSagebionetworksRepoModelFileDownloadOrder Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileDownloadOrder
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{files}}{The list of files included in this order. list(\link{OrgSagebionetworksRepoModelFileFileHandleAssociation}) [optional]}

\item{\code{orderId}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{zipFileName}}{character [optional]}

\item{\code{totalSizeBytes}}{integer [optional]}

\item{\code{totalNumberOfFiles}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-new}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-toJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-fromJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-toJSONString}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-fromJSONString}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-validateJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-toString}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-isValid}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-print}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrder-clone}{\code{OrgSagebionetworksRepoModelFileDownloadOrder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileDownloadOrder class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$new(
  files = NULL,
  orderId = NULL,
  createdBy = NULL,
  createdOn = NULL,
  zipFileName = NULL,
  totalSizeBytes = NULL,
  totalNumberOfFiles = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{The list of files included in this order.}

\item{\code{orderId}}{orderId}

\item{\code{createdBy}}{createdBy}

\item{\code{createdOn}}{createdOn}

\item{\code{zipFileName}}{zipFileName}

\item{\code{totalSizeBytes}}{totalSizeBytes}

\item{\code{totalNumberOfFiles}}{totalNumberOfFiles}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileDownloadOrder in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileDownloadOrder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileDownloadOrder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileDownloadOrder in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileDownloadOrder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileDownloadOrder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileDownloadOrder and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileDownloadOrder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
