% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_download_order_summary.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileDownloadOrderSummary}
\alias{OrgSagebionetworksRepoModelFileDownloadOrderSummary}
\title{OrgSagebionetworksRepoModelFileDownloadOrderSummary}
\format{
An \code{R6Class} generator object
}
\description{
Summary of files included in a download order.

OrgSagebionetworksRepoModelFileDownloadOrderSummary Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileDownloadOrderSummary
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{orderId}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{zipFileName}}{character [optional]}

\item{\code{totalSizeBytes}}{integer [optional]}

\item{\code{totalNumberOfFiles}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-new}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-toJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-fromJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-toJSONString}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-fromJSONString}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-validateJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-toString}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-isValid}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-print}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-clone}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileDownloadOrderSummary class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$new(
  orderId = NULL,
  createdBy = NULL,
  createdOn = NULL,
  zipFileName = NULL,
  totalSizeBytes = NULL,
  totalNumberOfFiles = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orderId}}{orderId}

\item{\code{createdBy}}{createdBy}

\item{\code{createdOn}}{createdOn}

\item{\code{zipFileName}}{zipFileName}

\item{\code{totalSizeBytes}}{totalSizeBytes}

\item{\code{totalNumberOfFiles}}{totalNumberOfFiles}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileDownloadOrderSummary in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileDownloadOrderSummary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileDownloadOrderSummary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileDownloadOrderSummary in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileDownloadOrderSummary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileDownloadOrderSummary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileDownloadOrderSummary and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileDownloadOrderSummary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
