% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_download_order_summary_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest}
\alias{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest}
\title{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request for a single page of DownloadOrderSummary objects.

OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-new}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-toString}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-isValid}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-print}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-clone}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$new(
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
