% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_download_order_summary_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse}
\alias{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse}
\title{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse}
\format{
An \code{R6Class} generator object
}
\description{
Response for a DownloadOrderSummaryRequest.

OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{page}}{A single page of DownloadOrderSummary objects. list(\link{OrgSagebionetworksRepoModelFileDownloadOrderSummary}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-new}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-toJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-fromJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-toJSONString}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-validateJSON}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-toString}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-isValid}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-print}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-clone}{\code{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$new(
  page = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{page}}{A single page of DownloadOrderSummary objects.}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileDownloadOrderSummaryResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
