% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_external_file_handle_interface.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileExternalFileHandleInterface}
\alias{OrgSagebionetworksRepoModelFileExternalFileHandleInterface}
\title{OrgSagebionetworksRepoModelFileExternalFileHandleInterface}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelFileExternalFileHandleInterface Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelFileExternalFileHandleInterface.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-new}{\code{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-fromJSONString}{\code{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-fromJSON}{\code{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-toJSONString}{\code{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-toJSON}{\code{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-validateJSON}{\code{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-toString}{\code{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-print}{\code{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-clone}{\code{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileExternalFileHandleInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelFileExternalFileHandle", "OrgSagebionetworksRepoModelFileExternalObjectStoreFileHandle", "OrgSagebionetworksRepoModelFileProxyFileHandle"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileExternalFileHandleInterface.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelFileExternalFileHandleInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileExternalFileHandleInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelFileExternalFileHandleInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelFileExternalFileHandleInterface to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelFileExternalFileHandleInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelFileExternalFileHandleInterface to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelFileExternalFileHandleInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelFileExternalFileHandleInterface and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalFileHandleInterface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalFileHandleInterface$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
