% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_external_google_cloud_upload_destination.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination}
\alias{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination}
\title{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination}
\format{
An \code{R6Class} generator object
}
\description{
This upload destination contains information to start an upload to an external Google Cloud Storage bucket connected with synapse. The destination is mapped from an <a href=\"${org.sagebionetworks.repo.model.project.ExternalGoogleCloudStorageLocationSetting}\">ExternalGoogleCloudStorageLocationSetting</a>.

OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{uploadType}}{character [optional]}

\item{\code{banner}}{character [optional]}

\item{\code{baseKey}}{character [optional]}

\item{\code{bucket}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-new}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-toJSON}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-fromJSON}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-toJSONString}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-fromJSONString}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-validateJSON}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-toString}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-isValid}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-print}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-clone}{\code{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$new(
  concreteType,
  storageLocationId = NULL,
  uploadType = NULL,
  banner = NULL,
  baseKey = NULL,
  bucket = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{uploadType}}{uploadType}

\item{\code{banner}}{banner}

\item{\code{baseKey}}{baseKey}

\item{\code{bucket}}{bucket}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalGoogleCloudUploadDestination$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
