% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_external_object_store_upload_destination.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination}
\alias{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination}
\title{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination}
\format{
An \code{R6Class} generator object
}
\description{
The upload destination contains information necessary to start an upload to an S3 bucket not managed by Synapse. This destination is mapped from an <a href=\"${org.sagebionetworks.repo.model.project.ExternalObjectStorageLocationSetting}\">ExternalObjectStorageLocationSetting</a>.

OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{uploadType}}{character [optional]}

\item{\code{banner}}{character [optional]}

\item{\code{endpointUrl}}{character [optional]}

\item{\code{bucket}}{character [optional]}

\item{\code{keyPrefixUUID}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-new}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-toJSON}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-fromJSON}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-toJSONString}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-fromJSONString}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-validateJSON}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-toString}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-isValid}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-print}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-clone}{\code{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$new(
  concreteType,
  storageLocationId = NULL,
  uploadType = NULL,
  banner = NULL,
  endpointUrl = NULL,
  bucket = NULL,
  keyPrefixUUID = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{uploadType}}{uploadType}

\item{\code{banner}}{banner}

\item{\code{endpointUrl}}{endpointUrl}

\item{\code{bucket}}{bucket}

\item{\code{keyPrefixUUID}}{keyPrefixUUID}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileExternalObjectStoreUploadDestination$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
