% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_file_download_summary.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileFileDownloadSummary}
\alias{OrgSagebionetworksRepoModelFileFileDownloadSummary}
\title{OrgSagebionetworksRepoModelFileFileDownloadSummary}
\format{
An \code{R6Class} generator object
}
\description{
Summary of a requested file download.

OrgSagebionetworksRepoModelFileFileDownloadSummary Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileFileDownloadSummary
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fileHandleId}}{character [optional]}

\item{\code{associateObjectId}}{character [optional]}

\item{\code{associateObjectType}}{character [optional]}

\item{\code{zipEntryName}}{character [optional]}

\item{\code{status}}{character [optional]}

\item{\code{failureMessage}}{character [optional]}

\item{\code{failureCode}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-new}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-toJSON}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-fromJSON}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-toJSONString}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-fromJSONString}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-validateJSON}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-toString}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-isValid}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-print}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileDownloadSummary-clone}{\code{OrgSagebionetworksRepoModelFileFileDownloadSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileFileDownloadSummary class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$new(
  fileHandleId = NULL,
  associateObjectId = NULL,
  associateObjectType = NULL,
  zipEntryName = NULL,
  status = NULL,
  failureMessage = NULL,
  failureCode = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileHandleId}}{fileHandleId}

\item{\code{associateObjectId}}{associateObjectId}

\item{\code{associateObjectType}}{associateObjectType}

\item{\code{zipEntryName}}{zipEntryName}

\item{\code{status}}{status}

\item{\code{failureMessage}}{failureMessage}

\item{\code{failureCode}}{failureCode}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileFileDownloadSummary in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileFileDownloadSummary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileFileDownloadSummary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileFileDownloadSummary in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileFileDownloadSummary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileFileDownloadSummary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileFileDownloadSummary and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileFileDownloadSummary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileDownloadSummary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileDownloadSummary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileDownloadSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
