% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_file_handle_archival_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest}
\alias{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest}
\title{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request used to tag unlinked file handles for archival. Only administrator can perform this kind of job.

OrgSagebionetworksRepoModelFileFileHandleArchivalRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileFileHandleArchivalRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{limit}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-new}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-toString}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-isValid}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-print}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-clone}{\code{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileFileHandleArchivalRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$new(
  concreteType,
  limit = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{limit}}{limit}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileFileHandleArchivalRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileFileHandleArchivalRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileFileHandleArchivalRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileFileHandleArchivalRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileFileHandleArchivalRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileFileHandleArchivalRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileFileHandleArchivalRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileFileHandleArchivalRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleArchivalRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleArchivalRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
