% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_file_handle_association_list.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileFileHandleAssociationList}
\alias{OrgSagebionetworksRepoModelFileFileHandleAssociationList}
\title{OrgSagebionetworksRepoModelFileFileHandleAssociationList}
\format{
An \code{R6Class} generator object
}
\description{
List of FileHandleAssociation.

OrgSagebionetworksRepoModelFileFileHandleAssociationList Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileFileHandleAssociationList
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{list}}{list(\link{OrgSagebionetworksRepoModelFileFileHandleAssociation}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-new}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-toJSON}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-fromJSON}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-toJSONString}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-fromJSONString}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-validateJSON}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-toString}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-isValid}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-print}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-clone}{\code{OrgSagebionetworksRepoModelFileFileHandleAssociationList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileFileHandleAssociationList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$new(list = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list}}{list}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileFileHandleAssociationList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileFileHandleAssociationList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileFileHandleAssociationList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileFileHandleAssociationList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileFileHandleAssociationList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileFileHandleAssociationList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileFileHandleAssociationList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileFileHandleAssociationList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleAssociationList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleAssociationList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
