% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_file_handle_restore_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest}
\alias{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest}
\title{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request used to restore a batch of file handles that might be UNLINKED or ARCHIVED. Note that only the owner of the file handle can initiate its restore.

OrgSagebionetworksRepoModelFileFileHandleRestoreRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileFileHandleRestoreRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{fileHandleIds}}{The list of file handle ids to restore. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-new}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-toString}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-isValid}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-print}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-clone}{\code{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileFileHandleRestoreRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$new(
  concreteType,
  fileHandleIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{fileHandleIds}}{The list of file handle ids to restore.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileFileHandleRestoreRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileFileHandleRestoreRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileFileHandleRestoreRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileFileHandleRestoreRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileFileHandleRestoreRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileFileHandleRestoreRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileFileHandleRestoreRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileFileHandleRestoreRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileFileHandleRestoreRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileFileHandleRestoreRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
