% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_google_cloud_file_handle.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle}
\alias{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle}
\title{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle}
\format{
An \code{R6Class} generator object
}
\description{
A GoogleCloudFileHandle represents a file stored in Google Cloud Storage.

OrgSagebionetworksRepoModelFileGoogleCloudFileHandle Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileGoogleCloudFileHandle
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{contentType}}{character [optional]}

\item{\code{contentMd5}}{character [optional]}

\item{\code{fileName}}{character [optional]}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{contentSize}}{integer [optional]}

\item{\code{status}}{character [optional]}

\item{\code{bucketName}}{character [optional]}

\item{\code{key}}{character [optional]}

\item{\code{previewId}}{character [optional]}

\item{\code{isPreview}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-new}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-toJSON}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-fromJSON}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-toJSONString}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-fromJSONString}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-validateJSON}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-toString}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-isValid}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-print}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-clone}{\code{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileGoogleCloudFileHandle class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$new(
  concreteType,
  id = NULL,
  etag = NULL,
  createdBy = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  contentType = NULL,
  contentMd5 = NULL,
  fileName = NULL,
  storageLocationId = NULL,
  contentSize = NULL,
  status = NULL,
  bucketName = NULL,
  key = NULL,
  previewId = NULL,
  isPreview = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{createdBy}}{createdBy}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{contentType}}{contentType}

\item{\code{contentMd5}}{contentMd5}

\item{\code{fileName}}{fileName}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{contentSize}}{contentSize}

\item{\code{status}}{status}

\item{\code{bucketName}}{bucketName}

\item{\code{key}}{key}

\item{\code{previewId}}{previewId}

\item{\code{isPreview}}{isPreview}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileGoogleCloudFileHandle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileGoogleCloudFileHandle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileGoogleCloudFileHandle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileGoogleCloudFileHandle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileGoogleCloudFileHandle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileGoogleCloudFileHandle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileGoogleCloudFileHandle and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileGoogleCloudFileHandle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileGoogleCloudFileHandle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileGoogleCloudFileHandle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
