% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_multipart_upload_copy_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest}
\alias{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest}
\title{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest}
\format{
An \code{R6Class} generator object
}
\description{
A multi-part upload that performs a copy of an existing file handle without data transfer from the client. Currently supports only copy from and to S3 buckets that live in the same region.

OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{partSizeBytes}}{integer [optional]}

\item{\code{fileName}}{character [optional]}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{generatePreview}}{character [optional]}

\item{\code{sourceFileHandleAssociation}}{\link{OrgSagebionetworksRepoModelFileFileHandleAssociation} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-new}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-toString}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-isValid}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-print}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-clone}{\code{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$new(
  concreteType,
  partSizeBytes = NULL,
  fileName = NULL,
  storageLocationId = NULL,
  generatePreview = NULL,
  sourceFileHandleAssociation = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{partSizeBytes}}{partSizeBytes}

\item{\code{fileName}}{fileName}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{generatePreview}}{generatePreview}

\item{\code{sourceFileHandleAssociation}}{sourceFileHandleAssociation}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadCopyRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
