% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_multipart_upload_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileMultipartUploadRequest}
\alias{OrgSagebionetworksRepoModelFileMultipartUploadRequest}
\title{OrgSagebionetworksRepoModelFileMultipartUploadRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to start or continue a multi-part file upload.

OrgSagebionetworksRepoModelFileMultipartUploadRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileMultipartUploadRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{partSizeBytes}}{integer [optional]}

\item{\code{fileName}}{character [optional]}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{generatePreview}}{character [optional]}

\item{\code{contentMD5Hex}}{character [optional]}

\item{\code{contentType}}{character [optional]}

\item{\code{fileSizeBytes}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-new}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-toString}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-isValid}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-print}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-clone}{\code{OrgSagebionetworksRepoModelFileMultipartUploadRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileMultipartUploadRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$new(
  concreteType,
  partSizeBytes = NULL,
  fileName = NULL,
  storageLocationId = NULL,
  generatePreview = NULL,
  contentMD5Hex = NULL,
  contentType = NULL,
  fileSizeBytes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{partSizeBytes}}{partSizeBytes}

\item{\code{fileName}}{fileName}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{generatePreview}}{generatePreview}

\item{\code{contentMD5Hex}}{contentMD5Hex}

\item{\code{contentType}}{contentType}

\item{\code{fileSizeBytes}}{fileSizeBytes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileMultipartUploadRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileMultipartUploadRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileMultipartUploadRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileMultipartUploadRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileMultipartUploadRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileMultipartUploadRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileMultipartUploadRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileMultipartUploadRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileMultipartUploadRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileMultipartUploadRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
