% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_proxy_file_handle.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileProxyFileHandle}
\alias{OrgSagebionetworksRepoModelFileProxyFileHandle}
\title{OrgSagebionetworksRepoModelFileProxyFileHandle}
\format{
An \code{R6Class} generator object
}
\description{
Reference to a file that accessible via a proxy server.  Each ProxyFileHandle must reference a ProxyStorageLocationSettings that defines how the files are accessed.

OrgSagebionetworksRepoModelFileProxyFileHandle Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileProxyFileHandle
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{contentType}}{character [optional]}

\item{\code{contentMd5}}{character [optional]}

\item{\code{fileName}}{character [optional]}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{contentSize}}{integer [optional]}

\item{\code{status}}{character [optional]}

\item{\code{filePath}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-new}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-toJSON}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-fromJSON}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-toJSONString}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-fromJSONString}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-validateJSON}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-toString}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-isValid}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-print}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileProxyFileHandle-clone}{\code{OrgSagebionetworksRepoModelFileProxyFileHandle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileProxyFileHandle class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$new(
  concreteType,
  id = NULL,
  etag = NULL,
  createdBy = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  contentType = NULL,
  contentMd5 = NULL,
  fileName = NULL,
  storageLocationId = NULL,
  contentSize = NULL,
  status = NULL,
  filePath = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{createdBy}}{createdBy}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{contentType}}{contentType}

\item{\code{contentMd5}}{contentMd5}

\item{\code{fileName}}{fileName}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{contentSize}}{contentSize}

\item{\code{status}}{status}

\item{\code{filePath}}{filePath}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileProxyFileHandle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileProxyFileHandle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileProxyFileHandle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileProxyFileHandle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileProxyFileHandle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileProxyFileHandle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileProxyFileHandle and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileProxyFileHandle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileProxyFileHandle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileProxyFileHandle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileProxyFileHandle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
