% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_remote_file_preview_generation_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest}
\alias{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest}
\title{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request remote preview generation from an S3 file handle into an S3 file handle.

OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{\link{OrgSagebionetworksRepoModelFileS3FileHandle} [optional]}

\item{\code{destination}}{\link{OrgSagebionetworksRepoModelFileS3FileHandle} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-new}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-toJSON}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-fromJSON}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-toJSONString}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-validateJSON}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-toString}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-isValid}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-print}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-clone}{\code{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$new(
  source = NULL,
  destination = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{source}

\item{\code{destination}}{destination}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileRemoteFilePreviewGenerationRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
