% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_s3_upload_destination.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileS3UploadDestination}
\alias{OrgSagebionetworksRepoModelFileS3UploadDestination}
\title{OrgSagebionetworksRepoModelFileS3UploadDestination}
\format{
An \code{R6Class} generator object
}
\description{
The upload destination contains information to start an upload to the default Synapse storage location that exists on Amazon S3. This destination is mapped from an <a href=\"${org.sagebionetworks.repo.model.project.S3StorageLocationSetting}\">S3StorageLocationSetting</a>.

OrgSagebionetworksRepoModelFileS3UploadDestination Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileS3UploadDestination
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{uploadType}}{character [optional]}

\item{\code{banner}}{character [optional]}

\item{\code{baseKey}}{character [optional]}

\item{\code{stsEnabled}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-new}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-toJSON}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-fromJSON}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-toJSONString}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-fromJSONString}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-validateJSON}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-toString}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-isValid}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-print}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileS3UploadDestination-clone}{\code{OrgSagebionetworksRepoModelFileS3UploadDestination$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileS3UploadDestination class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$new(
  concreteType,
  storageLocationId = NULL,
  uploadType = NULL,
  banner = NULL,
  baseKey = NULL,
  stsEnabled = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{uploadType}}{uploadType}

\item{\code{banner}}{banner}

\item{\code{baseKey}}{baseKey}

\item{\code{stsEnabled}}{stsEnabled}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileS3UploadDestination in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileS3UploadDestination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileS3UploadDestination
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileS3UploadDestination in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileS3UploadDestination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileS3UploadDestination
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileS3UploadDestination and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileS3UploadDestination
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileS3UploadDestination-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileS3UploadDestination-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileS3UploadDestination$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
