% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_file_upload_destination_location.R
\docType{class}
\name{OrgSagebionetworksRepoModelFileUploadDestinationLocation}
\alias{OrgSagebionetworksRepoModelFileUploadDestinationLocation}
\title{OrgSagebionetworksRepoModelFileUploadDestinationLocation}
\format{
An \code{R6Class} generator object
}
\description{
An upload destination location contains the id of the StorageLocation where a file can be uploaded.

OrgSagebionetworksRepoModelFileUploadDestinationLocation Class
}
\details{
Create a new OrgSagebionetworksRepoModelFileUploadDestinationLocation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{storageLocationId}}{integer [optional]}

\item{\code{description}}{character [optional]}

\item{\code{uploadType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-new}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-toJSON}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-fromJSON}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-toJSONString}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-fromJSONString}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-validateJSON}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-toString}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-isValid}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-getInvalidFields}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-print}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-clone}{\code{OrgSagebionetworksRepoModelFileUploadDestinationLocation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFileUploadDestinationLocation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$new(
  storageLocationId = NULL,
  description = NULL,
  uploadType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{storageLocationId}}{storageLocationId}

\item{\code{description}}{description}

\item{\code{uploadType}}{uploadType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileUploadDestinationLocation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileUploadDestinationLocation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileUploadDestinationLocation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFileUploadDestinationLocation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFileUploadDestinationLocation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFileUploadDestinationLocation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFileUploadDestinationLocation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFileUploadDestinationLocation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFileUploadDestinationLocation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFileUploadDestinationLocation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
