% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_form_submission_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelFormSubmissionStatus}
\alias{OrgSagebionetworksRepoModelFormSubmissionStatus}
\title{OrgSagebionetworksRepoModelFormSubmissionStatus}
\format{
An \code{R6Class} generator object
}
\description{
The status of a a submitted FormData object.

OrgSagebionetworksRepoModelFormSubmissionStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelFormSubmissionStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{submittedOn}}{character [optional]}

\item{\code{reviewedOn}}{character [optional]}

\item{\code{reviewedBy}}{character [optional]}

\item{\code{state}}{character [optional]}

\item{\code{rejectionMessage}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-new}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-toJSON}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-fromJSON}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-toJSONString}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-validateJSON}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-toString}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-isValid}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-print}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelFormSubmissionStatus-clone}{\code{OrgSagebionetworksRepoModelFormSubmissionStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelFormSubmissionStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$new(
  submittedOn = NULL,
  reviewedOn = NULL,
  reviewedBy = NULL,
  state = NULL,
  rejectionMessage = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submittedOn}}{submittedOn}

\item{\code{reviewedOn}}{reviewedOn}

\item{\code{reviewedBy}}{reviewedBy}

\item{\code{state}}{state}

\item{\code{rejectionMessage}}{rejectionMessage}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFormSubmissionStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFormSubmissionStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFormSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelFormSubmissionStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelFormSubmissionStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelFormSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelFormSubmissionStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelFormSubmissionStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelFormSubmissionStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelFormSubmissionStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelFormSubmissionStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
