% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_has_accessor_requirement.R
\docType{class}
\name{OrgSagebionetworksRepoModelHasAccessorRequirement}
\alias{OrgSagebionetworksRepoModelHasAccessorRequirement}
\title{OrgSagebionetworksRepoModelHasAccessorRequirement}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelHasAccessorRequirement Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelHasAccessorRequirement.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelHasAccessorRequirement-new}{\code{OrgSagebionetworksRepoModelHasAccessorRequirement$new()}}
\item \href{#method-OrgSagebionetworksRepoModelHasAccessorRequirement-fromJSONString}{\code{OrgSagebionetworksRepoModelHasAccessorRequirement$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelHasAccessorRequirement-fromJSON}{\code{OrgSagebionetworksRepoModelHasAccessorRequirement$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelHasAccessorRequirement-toJSONString}{\code{OrgSagebionetworksRepoModelHasAccessorRequirement$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelHasAccessorRequirement-toJSON}{\code{OrgSagebionetworksRepoModelHasAccessorRequirement$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelHasAccessorRequirement-validateJSON}{\code{OrgSagebionetworksRepoModelHasAccessorRequirement$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelHasAccessorRequirement-toString}{\code{OrgSagebionetworksRepoModelHasAccessorRequirement$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelHasAccessorRequirement-print}{\code{OrgSagebionetworksRepoModelHasAccessorRequirement$print()}}
\item \href{#method-OrgSagebionetworksRepoModelHasAccessorRequirement-clone}{\code{OrgSagebionetworksRepoModelHasAccessorRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelHasAccessorRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelHasAccessorRequirement-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelHasAccessorRequirement.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelHasAccessorRequirement$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelManagedACTAccessRequirement", "OrgSagebionetworksRepoModelSelfSignAccessRequirement"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelHasAccessorRequirement-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelHasAccessorRequirement-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelHasAccessorRequirement.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelHasAccessorRequirement$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelHasAccessorRequirement.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelHasAccessorRequirement-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelHasAccessorRequirement-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelHasAccessorRequirement.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelHasAccessorRequirement$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelHasAccessorRequirement.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelHasAccessorRequirement-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelHasAccessorRequirement-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelHasAccessorRequirement to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelHasAccessorRequirement$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelHasAccessorRequirement.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelHasAccessorRequirement-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelHasAccessorRequirement-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelHasAccessorRequirement to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelHasAccessorRequirement$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelHasAccessorRequirement.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelHasAccessorRequirement-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelHasAccessorRequirement-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelHasAccessorRequirement and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelHasAccessorRequirement$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelHasAccessorRequirement-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelHasAccessorRequirement-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelHasAccessorRequirement$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelHasAccessorRequirement-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelHasAccessorRequirement-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelHasAccessorRequirement$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelHasAccessorRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelHasAccessorRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelHasAccessorRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
