% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_invitee_verification_signed_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelInviteeVerificationSignedToken}
\alias{OrgSagebionetworksRepoModelInviteeVerificationSignedToken}
\title{OrgSagebionetworksRepoModelInviteeVerificationSignedToken}
\format{
An \code{R6Class} generator object
}
\description{
Signed token containing the information needed to update the inviteeId of a MembershipInvitation.

OrgSagebionetworksRepoModelInviteeVerificationSignedToken Class
}
\details{
Create a new OrgSagebionetworksRepoModelInviteeVerificationSignedToken
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{hmac}}{character [optional]}

\item{\code{version}}{integer [optional]}

\item{\code{expiresOn}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{inviteeId}}{character [optional]}

\item{\code{membershipInvitationId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-new}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-toJSON}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-fromJSON}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-toJSONString}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-fromJSONString}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-validateJSON}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-toString}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-isValid}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-getInvalidFields}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-print}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-clone}{\code{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelInviteeVerificationSignedToken class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$new(
  concreteType,
  hmac = NULL,
  version = NULL,
  expiresOn = NULL,
  createdOn = NULL,
  inviteeId = NULL,
  membershipInvitationId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{hmac}}{hmac}

\item{\code{version}}{version}

\item{\code{expiresOn}}{expiresOn}

\item{\code{createdOn}}{createdOn}

\item{\code{inviteeId}}{inviteeId}

\item{\code{membershipInvitationId}}{membershipInvitationId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelInviteeVerificationSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelInviteeVerificationSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelInviteeVerificationSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelInviteeVerificationSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelInviteeVerificationSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelInviteeVerificationSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelInviteeVerificationSignedToken and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelInviteeVerificationSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelInviteeVerificationSignedToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelInviteeVerificationSignedToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
