% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_join_team_signed_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelJoinTeamSignedToken}
\alias{OrgSagebionetworksRepoModelJoinTeamSignedToken}
\title{OrgSagebionetworksRepoModelJoinTeamSignedToken}
\format{
An \code{R6Class} generator object
}
\description{
Signed token containing the information needed to join a new user to a team.

OrgSagebionetworksRepoModelJoinTeamSignedToken Class
}
\details{
Create a new OrgSagebionetworksRepoModelJoinTeamSignedToken
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{hmac}}{character [optional]}

\item{\code{version}}{integer [optional]}

\item{\code{expiresOn}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{userId}}{character [optional]}

\item{\code{teamId}}{character [optional]}

\item{\code{memberId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-new}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-toJSON}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-fromJSON}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-toJSONString}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-fromJSONString}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-validateJSON}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-toString}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-isValid}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-getInvalidFields}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-print}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelJoinTeamSignedToken-clone}{\code{OrgSagebionetworksRepoModelJoinTeamSignedToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelJoinTeamSignedToken class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$new(
  concreteType,
  hmac = NULL,
  version = NULL,
  expiresOn = NULL,
  createdOn = NULL,
  userId = NULL,
  teamId = NULL,
  memberId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{hmac}}{hmac}

\item{\code{version}}{version}

\item{\code{expiresOn}}{expiresOn}

\item{\code{createdOn}}{createdOn}

\item{\code{userId}}{userId}

\item{\code{teamId}}{teamId}

\item{\code{memberId}}{memberId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelJoinTeamSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelJoinTeamSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelJoinTeamSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelJoinTeamSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelJoinTeamSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelJoinTeamSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelJoinTeamSignedToken and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelJoinTeamSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelJoinTeamSignedToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelJoinTeamSignedToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelJoinTeamSignedToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
