% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_managed_act_access_requirement.R
\docType{class}
\name{OrgSagebionetworksRepoModelManagedACTAccessRequirement}
\alias{OrgSagebionetworksRepoModelManagedACTAccessRequirement}
\title{OrgSagebionetworksRepoModelManagedACTAccessRequirement}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for in-Synapse 'Access Control Team' controlled Access Requirement, a 'tier 3' Access Requirement. This access requirement allows the ACT managing the detail requirements, and submissions within Synapse.

OrgSagebionetworksRepoModelManagedACTAccessRequirement Class
}
\details{
Create a new OrgSagebionetworksRepoModelManagedACTAccessRequirement
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{versionNumber}}{integer [optional]}

\item{\code{id}}{integer [optional]}

\item{\code{description}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{subjectsDefinedByAnnotations}}{character [optional]}

\item{\code{subjectIds}}{The IDs of the items controlled by this Access Requirement when 'subjectsDefinedByAnnotations=false'. This property is mutually exclusive with 'subjectsDefinedByAnnotations'.  When 'subjectsDefinedByAnnotations=true' then this property must be empty or excluded.  Required when creating or updating and 'subjectsDefinedByAnnotations=false' or 'subjectsDefinedByAnnotations' is excluded. list(\link{OrgSagebionetworksRepoModelRestrictableObjectDescriptor}) [optional]}

\item{\code{accessType}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{isCertifiedUserRequired}}{character [optional]}

\item{\code{isValidatedProfileRequired}}{character [optional]}

\item{\code{isDUCRequired}}{character [optional]}

\item{\code{ducTemplateFileHandleId}}{character [optional]}

\item{\code{isIRBApprovalRequired}}{character [optional]}

\item{\code{areOtherAttachmentsRequired}}{character [optional]}

\item{\code{expirationPeriod}}{integer [optional]}

\item{\code{isIDUPublic}}{character [optional]}

\item{\code{isIDURequired}}{character [optional]}

\item{\code{isTwoFaRequired}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-new}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$new()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-toJSON}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-fromJSON}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-toJSONString}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-fromJSONString}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-validateJSON}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-toString}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-isValid}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-getInvalidFields}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-print}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$print()}}
\item \href{#method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-clone}{\code{OrgSagebionetworksRepoModelManagedACTAccessRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelManagedACTAccessRequirement class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$new(
  concreteType,
  versionNumber = NULL,
  id = NULL,
  description = NULL,
  name = NULL,
  etag = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  subjectsDefinedByAnnotations = NULL,
  subjectIds = NULL,
  accessType = NULL,
  isCertifiedUserRequired = NULL,
  isValidatedProfileRequired = NULL,
  isDUCRequired = NULL,
  ducTemplateFileHandleId = NULL,
  isIRBApprovalRequired = NULL,
  areOtherAttachmentsRequired = NULL,
  expirationPeriod = NULL,
  isIDUPublic = NULL,
  isIDURequired = NULL,
  isTwoFaRequired = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{versionNumber}}{versionNumber}

\item{\code{id}}{id}

\item{\code{description}}{description}

\item{\code{name}}{name}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{subjectsDefinedByAnnotations}}{subjectsDefinedByAnnotations}

\item{\code{subjectIds}}{The IDs of the items controlled by this Access Requirement when 'subjectsDefinedByAnnotations=false'. This property is mutually exclusive with 'subjectsDefinedByAnnotations'.  When 'subjectsDefinedByAnnotations=true' then this property must be empty or excluded.  Required when creating or updating and 'subjectsDefinedByAnnotations=false' or 'subjectsDefinedByAnnotations' is excluded.}

\item{\code{accessType}}{accessType}

\item{\code{isCertifiedUserRequired}}{isCertifiedUserRequired}

\item{\code{isValidatedProfileRequired}}{isValidatedProfileRequired}

\item{\code{isDUCRequired}}{isDUCRequired}

\item{\code{ducTemplateFileHandleId}}{ducTemplateFileHandleId}

\item{\code{isIRBApprovalRequired}}{isIRBApprovalRequired}

\item{\code{areOtherAttachmentsRequired}}{areOtherAttachmentsRequired}

\item{\code{expirationPeriod}}{expirationPeriod}

\item{\code{isIDUPublic}}{isIDUPublic}

\item{\code{isIDURequired}}{isIDURequired}

\item{\code{isTwoFaRequired}}{isTwoFaRequired}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelManagedACTAccessRequirement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelManagedACTAccessRequirement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelManagedACTAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelManagedACTAccessRequirement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelManagedACTAccessRequirement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelManagedACTAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelManagedACTAccessRequirement and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelManagedACTAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelManagedACTAccessRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelManagedACTAccessRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
