% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_membership_invitation.R
\docType{class}
\name{OrgSagebionetworksRepoModelMembershipInvitation}
\alias{OrgSagebionetworksRepoModelMembershipInvitation}
\title{OrgSagebionetworksRepoModelMembershipInvitation}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for MembershipInvitation POJO. \\nNote: certain fields may be omitted when returned if the field value is null.

OrgSagebionetworksRepoModelMembershipInvitation Class
}
\details{
Create a new OrgSagebionetworksRepoModelMembershipInvitation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{teamId}}{character [optional]}

\item{\code{inviteeId}}{character [optional]}

\item{\code{inviteeEmail}}{character [optional]}

\item{\code{message}}{character [optional]}

\item{\code{expiresOn}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-new}{\code{OrgSagebionetworksRepoModelMembershipInvitation$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-toJSON}{\code{OrgSagebionetworksRepoModelMembershipInvitation$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-fromJSON}{\code{OrgSagebionetworksRepoModelMembershipInvitation$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-toJSONString}{\code{OrgSagebionetworksRepoModelMembershipInvitation$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-fromJSONString}{\code{OrgSagebionetworksRepoModelMembershipInvitation$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-validateJSON}{\code{OrgSagebionetworksRepoModelMembershipInvitation$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-toString}{\code{OrgSagebionetworksRepoModelMembershipInvitation$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-isValid}{\code{OrgSagebionetworksRepoModelMembershipInvitation$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-getInvalidFields}{\code{OrgSagebionetworksRepoModelMembershipInvitation$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-print}{\code{OrgSagebionetworksRepoModelMembershipInvitation$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMembershipInvitation-clone}{\code{OrgSagebionetworksRepoModelMembershipInvitation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMembershipInvitation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$new(
  id = NULL,
  teamId = NULL,
  inviteeId = NULL,
  inviteeEmail = NULL,
  message = NULL,
  expiresOn = NULL,
  createdOn = NULL,
  createdBy = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{teamId}}{teamId}

\item{\code{inviteeId}}{inviteeId}

\item{\code{inviteeEmail}}{inviteeEmail}

\item{\code{message}}{message}

\item{\code{expiresOn}}{expiresOn}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMembershipInvitation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMembershipInvitation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMembershipInvitation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMembershipInvitation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMembershipInvitation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMembershipInvitation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMembershipInvitation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMembershipInvitation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMembershipInvitation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMembershipInvitation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMembershipInvitation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
