% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_cloudmailin_attachment.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageCloudmailinAttachment}
\alias{OrgSagebionetworksRepoModelMessageCloudmailinAttachment}
\title{OrgSagebionetworksRepoModelMessageCloudmailinAttachment}
\format{
An \code{R6Class} generator object
}
\description{
Message Attachment

OrgSagebionetworksRepoModelMessageCloudmailinAttachment Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageCloudmailinAttachment
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{character [optional]}

\item{\code{content_id}}{character [optional]}

\item{\code{url}}{character [optional]}

\item{\code{file_name}}{character [optional]}

\item{\code{content_type}}{character [optional]}

\item{\code{size}}{character [optional]}

\item{\code{disposition}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-new}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-toJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-fromJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-toJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-validateJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-toString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-isValid}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-print}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-clone}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageCloudmailinAttachment class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$new(
  content = NULL,
  content_id = NULL,
  url = NULL,
  file_name = NULL,
  content_type = NULL,
  size = NULL,
  disposition = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{content}

\item{\code{content_id}}{content_id}

\item{\code{url}}{url}

\item{\code{file_name}}{file_name}

\item{\code{content_type}}{content_type}

\item{\code{size}}{size}

\item{\code{disposition}}{disposition}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinAttachment in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinAttachment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinAttachment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinAttachment in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinAttachment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinAttachment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageCloudmailinAttachment and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageCloudmailinAttachment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAttachment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAttachment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
