% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_cloudmailin_authorization_check_header.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader}
\alias{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader}
\title{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader}
\format{
An \code{R6Class} generator object
}
\description{
Message Envelope

OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{to}}{character [optional]}

\item{\code{from}}{character [optional]}

\item{\code{size}}{integer [optional]}

\item{\code{helo_domain}}{character [optional]}

\item{\code{remote_ip}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-new}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-toJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-fromJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-toJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-validateJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-toString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-isValid}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-print}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-clone}{\code{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$new(
  to = NULL,
  from = NULL,
  size = NULL,
  helo_domain = NULL,
  remote_ip = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{to}}{to}

\item{\code{from}}{from}

\item{\code{size}}{size}

\item{\code{helo_domain}}{helo_domain}

\item{\code{remote_ip}}{remote_ip}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinAuthorizationCheckHeader$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
