% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_cloudmailin_envelope.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope}
\alias{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope}
\title{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope}
\format{
An \code{R6Class} generator object
}
\description{
Message Envelope

OrgSagebionetworksRepoModelMessageCloudmailinEnvelope Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageCloudmailinEnvelope
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{to}}{character [optional]}

\item{\code{recipients}}{The full list of recipients that the remote server is attempting to send to in this transaction. list(character) [optional]}

\item{\code{from}}{character [optional]}

\item{\code{helo_domain}}{character [optional]}

\item{\code{remote_ip}}{character [optional]}

\item{\code{spf}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-new}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-toJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-fromJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-toJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-validateJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-toString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-isValid}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-print}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-clone}{\code{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageCloudmailinEnvelope class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$new(
  to = NULL,
  recipients = NULL,
  from = NULL,
  helo_domain = NULL,
  remote_ip = NULL,
  spf = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{to}}{to}

\item{\code{recipients}}{The full list of recipients that the remote server is attempting to send to in this transaction.}

\item{\code{from}}{from}

\item{\code{helo_domain}}{helo_domain}

\item{\code{remote_ip}}{remote_ip}

\item{\code{spf}}{spf}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinEnvelope in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinEnvelope
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinEnvelope
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinEnvelope in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinEnvelope
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinEnvelope
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageCloudmailinEnvelope and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageCloudmailinEnvelope
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinEnvelope-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
