% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_cloudmailin_headers.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageCloudmailinHeaders}
\alias{OrgSagebionetworksRepoModelMessageCloudmailinHeaders}
\title{OrgSagebionetworksRepoModelMessageCloudmailinHeaders}
\format{
An \code{R6Class} generator object
}
\description{
The headers from the email message.

OrgSagebionetworksRepoModelMessageCloudmailinHeaders Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageCloudmailinHeaders
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{To}}{character [optional]}

\item{\code{Cc}}{character [optional]}

\item{\code{Bcc}}{character [optional]}

\item{\code{Subject}}{character [optional]}

\item{\code{From}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-new}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-toJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-fromJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-toJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-validateJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-toString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-isValid}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-print}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-clone}{\code{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageCloudmailinHeaders class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$new(
  To = NULL,
  Cc = NULL,
  Bcc = NULL,
  Subject = NULL,
  From = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{To}}{To}

\item{\code{Cc}}{Cc}

\item{\code{Bcc}}{Bcc}

\item{\code{Subject}}{Subject}

\item{\code{From}}{From}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinHeaders in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinHeaders
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinHeaders
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinHeaders in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinHeaders
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinHeaders
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageCloudmailinHeaders and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageCloudmailinHeaders
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinHeaders-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinHeaders$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
