% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_cloudmailin_message.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageCloudmailinMessage}
\alias{OrgSagebionetworksRepoModelMessageCloudmailinMessage}
\title{OrgSagebionetworksRepoModelMessageCloudmailinMessage}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for the CloudMainIn message format

OrgSagebionetworksRepoModelMessageCloudmailinMessage Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageCloudmailinMessage
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{envelope}}{\link{OrgSagebionetworksRepoModelMessageCloudmailinEnvelope} [optional]}

\item{\code{headers}}{\link{OrgSagebionetworksRepoModelMessageCloudmailinHeaders} [optional]}

\item{\code{plain}}{character [optional]}

\item{\code{html}}{character [optional]}

\item{\code{reply_plain}}{character [optional]}

\item{\code{attachments}}{Attachments to the message. list(\link{OrgSagebionetworksRepoModelMessageCloudmailinAttachment}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-new}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-toJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-fromJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-toJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-validateJSON}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-toString}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-isValid}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-print}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-clone}{\code{OrgSagebionetworksRepoModelMessageCloudmailinMessage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageCloudmailinMessage class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$new(
  envelope = NULL,
  headers = NULL,
  plain = NULL,
  html = NULL,
  reply_plain = NULL,
  attachments = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{envelope}}{envelope}

\item{\code{headers}}{headers}

\item{\code{plain}}{plain}

\item{\code{html}}{html}

\item{\code{reply_plain}}{reply_plain}

\item{\code{attachments}}{Attachments to the message.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinMessage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinMessage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinMessage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageCloudmailinMessage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageCloudmailinMessage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageCloudmailinMessage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageCloudmailinMessage and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageCloudmailinMessage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageCloudmailinMessage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageCloudmailinMessage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
