% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_local_stack_change_messsage.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage}
\alias{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage}
\title{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage}
\format{
An \code{R6Class} generator object
}
\description{
Local stack object that can be used in place of a change message (e.g. this won't migrate through stacks)

OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{objectId}}{character [optional]}

\item{\code{objectVersion}}{integer [optional]}

\item{\code{objectType}}{character [optional]}

\item{\code{timestamp}}{character [optional]}

\item{\code{changeNumber}}{integer [optional]}

\item{\code{changeType}}{character [optional]}

\item{\code{userId}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-new}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-toJSON}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-fromJSON}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-toJSONString}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-validateJSON}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-toString}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-isValid}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-print}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-clone}{\code{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$new(
  concreteType,
  objectId = NULL,
  objectVersion = NULL,
  objectType = NULL,
  timestamp = NULL,
  changeNumber = NULL,
  changeType = NULL,
  userId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{objectId}}{objectId}

\item{\code{objectVersion}}{objectVersion}

\item{\code{objectType}}{objectType}

\item{\code{timestamp}}{timestamp}

\item{\code{changeNumber}}{changeNumber}

\item{\code{changeType}}{changeType}

\item{\code{userId}}{userId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
