% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_local_stack_message.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageLocalStackMessage}
\alias{OrgSagebionetworksRepoModelMessageLocalStackMessage}
\title{OrgSagebionetworksRepoModelMessageLocalStackMessage}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelMessageLocalStackMessage Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelMessageLocalStackMessage.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackMessage-new}{\code{OrgSagebionetworksRepoModelMessageLocalStackMessage$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackMessage-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageLocalStackMessage$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackMessage-fromJSON}{\code{OrgSagebionetworksRepoModelMessageLocalStackMessage$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackMessage-toJSONString}{\code{OrgSagebionetworksRepoModelMessageLocalStackMessage$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackMessage-toJSON}{\code{OrgSagebionetworksRepoModelMessageLocalStackMessage$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackMessage-validateJSON}{\code{OrgSagebionetworksRepoModelMessageLocalStackMessage$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackMessage-toString}{\code{OrgSagebionetworksRepoModelMessageLocalStackMessage$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackMessage-print}{\code{OrgSagebionetworksRepoModelMessageLocalStackMessage$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageLocalStackMessage-clone}{\code{OrgSagebionetworksRepoModelMessageLocalStackMessage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackMessage-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackMessage-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageLocalStackMessage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackMessage$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelDataaccessDataAccessSubmissionEvent", "OrgSagebionetworksRepoModelFileFileEvent", "OrgSagebionetworksRepoModelMessageLocalStackChangeMesssage", "OrgSagebionetworksRepoModelTableQueryCacheHitEvent", "OrgSagebionetworksRepoModelTableTableStatusChangeEvent"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackMessage-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackMessage-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageLocalStackMessage.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackMessage$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelMessageLocalStackMessage.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackMessage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackMessage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageLocalStackMessage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackMessage$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelMessageLocalStackMessage.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackMessage-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackMessage-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelMessageLocalStackMessage to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackMessage$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelMessageLocalStackMessage.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackMessage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackMessage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelMessageLocalStackMessage to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackMessage$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelMessageLocalStackMessage.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackMessage-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackMessage-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelMessageLocalStackMessage and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackMessage$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackMessage-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackMessage-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackMessage$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackMessage-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackMessage-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackMessage$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageLocalStackMessage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageLocalStackMessage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageLocalStackMessage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
