% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_message_recipient_set_interface.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface}
\alias{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface}
\title{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-new}{\code{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-fromJSON}{\code{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-toJSONString}{\code{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-toJSON}{\code{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-validateJSON}{\code{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-toString}{\code{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-print}{\code{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-clone}{\code{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$new(
  instance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelMessageMessageRecipientSet", "OrgSagebionetworksRepoModelMessageMessageToUser"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMessageRecipientSetInterface$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
