% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_multipart_attachment.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageMultipartAttachment}
\alias{OrgSagebionetworksRepoModelMessageMultipartAttachment}
\title{OrgSagebionetworksRepoModelMessageMultipartAttachment}
\format{
An \code{R6Class} generator object
}
\description{
Multipart Message Attachment

OrgSagebionetworksRepoModelMessageMultipartAttachment Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageMultipartAttachment
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{character [optional]}

\item{\code{content_id}}{character [optional]}

\item{\code{url}}{character [optional]}

\item{\code{file_name}}{character [optional]}

\item{\code{content_type}}{character [optional]}

\item{\code{size}}{character [optional]}

\item{\code{disposition}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-new}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-toJSON}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-fromJSON}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-toJSONString}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-validateJSON}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-toString}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-isValid}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-print}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartAttachment-clone}{\code{OrgSagebionetworksRepoModelMessageMultipartAttachment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageMultipartAttachment class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$new(
  content = NULL,
  content_id = NULL,
  url = NULL,
  file_name = NULL,
  content_type = NULL,
  size = NULL,
  disposition = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{content}

\item{\code{content_id}}{content_id}

\item{\code{url}}{url}

\item{\code{file_name}}{file_name}

\item{\code{content_type}}{content_type}

\item{\code{size}}{size}

\item{\code{disposition}}{disposition}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageMultipartAttachment in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageMultipartAttachment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageMultipartAttachment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageMultipartAttachment in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageMultipartAttachment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageMultipartAttachment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageMultipartAttachment and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageMultipartAttachment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartAttachment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartAttachment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartAttachment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
