% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_multipart_message_body.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageMultipartMessageBody}
\alias{OrgSagebionetworksRepoModelMessageMultipartMessageBody}
\title{OrgSagebionetworksRepoModelMessageMultipartMessageBody}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for a multipart message body

OrgSagebionetworksRepoModelMessageMultipartMessageBody Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageMultipartMessageBody
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{plain}}{character [optional]}

\item{\code{html}}{character [optional]}

\item{\code{attachments}}{Attachments to the message. list(\link{OrgSagebionetworksRepoModelMessageMultipartAttachment}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-new}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-toJSON}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-fromJSON}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-toJSONString}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-validateJSON}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-toString}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-isValid}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-print}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-clone}{\code{OrgSagebionetworksRepoModelMessageMultipartMessageBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageMultipartMessageBody class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$new(
  plain = NULL,
  html = NULL,
  attachments = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plain}}{plain}

\item{\code{html}}{html}

\item{\code{attachments}}{Attachments to the message.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageMultipartMessageBody in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageMultipartMessageBody
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageMultipartMessageBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageMultipartMessageBody in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageMultipartMessageBody
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageMultipartMessageBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageMultipartMessageBody and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageMultipartMessageBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageMultipartMessageBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageMultipartMessageBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
