% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_notification_settings_signed_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken}
\alias{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken}
\title{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken}
\format{
An \code{R6Class} generator object
}
\description{
Signed token supporting one-click email suppression.

OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{hmac}}{character [optional]}

\item{\code{version}}{integer [optional]}

\item{\code{expiresOn}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{userId}}{character [optional]}

\item{\code{settings}}{\link{OrgSagebionetworksRepoModelMessageSettings} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-new}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-toJSON}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-fromJSON}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-toJSONString}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-validateJSON}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-toString}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-isValid}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-print}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-clone}{\code{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$new(
  concreteType,
  hmac = NULL,
  version = NULL,
  expiresOn = NULL,
  createdOn = NULL,
  userId = NULL,
  settings = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{hmac}}{hmac}

\item{\code{version}}{version}

\item{\code{expiresOn}}{expiresOn}

\item{\code{createdOn}}{createdOn}

\item{\code{userId}}{userId}

\item{\code{settings}}{settings}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
