% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_settings.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageSettings}
\alias{OrgSagebionetworksRepoModelMessageSettings}
\title{OrgSagebionetworksRepoModelMessageSettings}
\format{
An \code{R6Class} generator object
}
\description{
Contains a user's notification settings

OrgSagebionetworksRepoModelMessageSettings Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageSettings
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sendEmailNotifications}}{character [optional]}

\item{\code{markEmailedMessagesAsRead}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-new}{\code{OrgSagebionetworksRepoModelMessageSettings$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-toJSON}{\code{OrgSagebionetworksRepoModelMessageSettings$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-fromJSON}{\code{OrgSagebionetworksRepoModelMessageSettings$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-toJSONString}{\code{OrgSagebionetworksRepoModelMessageSettings$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageSettings$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-validateJSON}{\code{OrgSagebionetworksRepoModelMessageSettings$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-toString}{\code{OrgSagebionetworksRepoModelMessageSettings$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-isValid}{\code{OrgSagebionetworksRepoModelMessageSettings$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageSettings$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-print}{\code{OrgSagebionetworksRepoModelMessageSettings$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageSettings-clone}{\code{OrgSagebionetworksRepoModelMessageSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageSettings class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$new(
  sendEmailNotifications = NULL,
  markEmailedMessagesAsRead = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sendEmailNotifications}}{sendEmailNotifications}

\item{\code{markEmailedMessagesAsRead}}{markEmailedMessagesAsRead}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageSettings in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageSettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageSettings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageSettings in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageSettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageSettings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageSettings and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageSettings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
