% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_message_unsent_message_range.R
\docType{class}
\name{OrgSagebionetworksRepoModelMessageUnsentMessageRange}
\alias{OrgSagebionetworksRepoModelMessageUnsentMessageRange}
\title{OrgSagebionetworksRepoModelMessageUnsentMessageRange}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for a range of change message numbers to recheck

OrgSagebionetworksRepoModelMessageUnsentMessageRange Class
}
\details{
Create a new OrgSagebionetworksRepoModelMessageUnsentMessageRange
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lowerBound}}{integer [optional]}

\item{\code{upperBound}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-new}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-toJSON}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-fromJSON}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-toJSONString}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-fromJSONString}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-validateJSON}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-toString}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-isValid}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-getInvalidFields}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-print}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-clone}{\code{OrgSagebionetworksRepoModelMessageUnsentMessageRange$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMessageUnsentMessageRange class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$new(
  lowerBound = NULL,
  upperBound = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lowerBound}}{lowerBound}

\item{\code{upperBound}}{upperBound}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageUnsentMessageRange in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageUnsentMessageRange
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageUnsentMessageRange
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMessageUnsentMessageRange in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMessageUnsentMessageRange
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMessageUnsentMessageRange
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMessageUnsentMessageRange and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMessageUnsentMessageRange
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMessageUnsentMessageRange-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMessageUnsentMessageRange$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
