% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_admin_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationAdminRequest}
\alias{OrgSagebionetworksRepoModelMigrationAdminRequest}
\title{OrgSagebionetworksRepoModelMigrationAdminRequest}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelMigrationAdminRequest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelMigrationAdminRequest.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationAdminRequest-new}{\code{OrgSagebionetworksRepoModelMigrationAdminRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAdminRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationAdminRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAdminRequest-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationAdminRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAdminRequest-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationAdminRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAdminRequest-toJSON}{\code{OrgSagebionetworksRepoModelMigrationAdminRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAdminRequest-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationAdminRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAdminRequest-toString}{\code{OrgSagebionetworksRepoModelMigrationAdminRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAdminRequest-print}{\code{OrgSagebionetworksRepoModelMigrationAdminRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAdminRequest-clone}{\code{OrgSagebionetworksRepoModelMigrationAdminRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAdminRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAdminRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationAdminRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAdminRequest$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest", "OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest", "OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeCountRequest", "OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeCountsRequest", "OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest", "OrgSagebionetworksRepoModelMigrationBatchChecksumRequest", "OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest", "OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest", "OrgSagebionetworksRepoModelMigrationRestoreTypeRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAdminRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAdminRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationAdminRequest.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAdminRequest$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelMigrationAdminRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAdminRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAdminRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationAdminRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAdminRequest$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelMigrationAdminRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAdminRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAdminRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelMigrationAdminRequest to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAdminRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelMigrationAdminRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAdminRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAdminRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelMigrationAdminRequest to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAdminRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelMigrationAdminRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAdminRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAdminRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelMigrationAdminRequest and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAdminRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAdminRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAdminRequest-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAdminRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAdminRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAdminRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAdminRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAdminRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAdminRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAdminRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
