% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_async_migration_range_checksum_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest}
\alias{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest}
\title{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request checksum for range of given type.

OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{migrationType}}{character [optional]}

\item{\code{type}}{character [optional]}

\item{\code{salt}}{character [optional]}

\item{\code{minId}}{integer [optional]}

\item{\code{maxId}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-new}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-toJSON}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-toString}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-isValid}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-print}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-clone}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$new(
  concreteType,
  migrationType = NULL,
  type = NULL,
  salt = NULL,
  minId = NULL,
  maxId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{migrationType}}{migrationType}

\item{\code{type}}{type}

\item{\code{salt}}{salt}

\item{\code{minId}}{minId}

\item{\code{maxId}}{maxId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$isValid(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
