% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_async_migration_type_checksum_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest}
\alias{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest}
\title{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request checksum for given type.

OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{migrationType}}{character [optional]}

\item{\code{type}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-new}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-toJSON}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-toString}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-isValid}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-print}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-clone}{\code{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$new(
  concreteType,
  migrationType = NULL,
  type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{migrationType}}{migrationType}

\item{\code{type}}{type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
