% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_backup_manifest.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationBackupManifest}
\alias{OrgSagebionetworksRepoModelMigrationBackupManifest}
\title{OrgSagebionetworksRepoModelMigrationBackupManifest}
\format{
An \code{R6Class} generator object
}
\description{
Manifest of all metadata associated with a single backup file

OrgSagebionetworksRepoModelMigrationBackupManifest Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationBackupManifest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{primaryType}}{\link{OrgSagebionetworksRepoModelMigrationTypeData} [optional]}

\item{\code{aliasType}}{character [optional]}

\item{\code{batchSize}}{integer [optional]}

\item{\code{minimumId}}{integer [optional]}

\item{\code{maximumId}}{integer [optional]}

\item{\code{secondaryTypes}}{The secondary types according to the source of the backup. list(\link{OrgSagebionetworksRepoModelMigrationTypeData}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-new}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-toJSON}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-toString}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-isValid}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-print}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupManifest-clone}{\code{OrgSagebionetworksRepoModelMigrationBackupManifest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationBackupManifest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$new(
  primaryType = NULL,
  aliasType = NULL,
  batchSize = NULL,
  minimumId = NULL,
  maximumId = NULL,
  secondaryTypes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{primaryType}}{primaryType}

\item{\code{aliasType}}{aliasType}

\item{\code{batchSize}}{batchSize}

\item{\code{minimumId}}{minimumId}

\item{\code{maximumId}}{maximumId}

\item{\code{secondaryTypes}}{The secondary types according to the source of the backup.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationBackupManifest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationBackupManifest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationBackupManifest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationBackupManifest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationBackupManifest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationBackupManifest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationBackupManifest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationBackupManifest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupManifest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupManifest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupManifest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
