% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_backup_type_range_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest}
\alias{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest}
\title{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to create a migration backup for a given MigrationType using a range of Ids.

OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{migrationType}}{character [optional]}

\item{\code{aliasType}}{character [optional]}

\item{\code{batchSize}}{integer [optional]}

\item{\code{minimumId}}{integer [optional]}

\item{\code{maximumId}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-new}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-toJSON}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-toString}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-isValid}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-print}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-clone}{\code{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$new(
  concreteType,
  migrationType = NULL,
  aliasType = NULL,
  batchSize = NULL,
  minimumId = NULL,
  maximumId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{migrationType}}{migrationType}

\item{\code{aliasType}}{aliasType}

\item{\code{batchSize}}{batchSize}

\item{\code{minimumId}}{minimumId}

\item{\code{maximumId}}{maximumId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
