% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_batch_checksum_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest}
\alias{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest}
\title{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest}
\format{
An \code{R6Class} generator object
}
\description{
Get n number of checksums for the given ID range.  The given ID range will be batched into sub-ranges and the checksum for each batch will be calculated.

OrgSagebionetworksRepoModelMigrationBatchChecksumRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationBatchChecksumRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{migrationType}}{character [optional]}

\item{\code{salt}}{character [optional]}

\item{\code{minimumId}}{integer [optional]}

\item{\code{maximumId}}{integer [optional]}

\item{\code{batchSize}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-new}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-toJSON}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-toString}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-isValid}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-print}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-clone}{\code{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationBatchChecksumRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$new(
  concreteType,
  migrationType = NULL,
  salt = NULL,
  minimumId = NULL,
  maximumId = NULL,
  batchSize = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{migrationType}}{migrationType}

\item{\code{salt}}{salt}

\item{\code{minimumId}}{minimumId}

\item{\code{maximumId}}{maximumId}

\item{\code{batchSize}}{batchSize}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationBatchChecksumRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationBatchChecksumRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationBatchChecksumRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationBatchChecksumRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationBatchChecksumRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationBatchChecksumRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationBatchChecksumRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationBatchChecksumRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationBatchChecksumRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationBatchChecksumRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
