% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_calculate_optimal_range_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest}
\alias{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest}
\title{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to calculate the ID ranges with the optimal number of rows.

OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{migrationType}}{character [optional]}

\item{\code{minimumId}}{integer [optional]}

\item{\code{maximumId}}{integer [optional]}

\item{\code{optimalRowsPerRange}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-new}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-toJSON}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-toString}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-isValid}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-print}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-clone}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$new(
  concreteType,
  migrationType = NULL,
  minimumId = NULL,
  maximumId = NULL,
  optimalRowsPerRange = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{migrationType}}{migrationType}

\item{\code{minimumId}}{minimumId}

\item{\code{maximumId}}{maximumId}

\item{\code{optimalRowsPerRange}}{optimalRowsPerRange}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
