% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_calculate_optimal_range_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse}
\alias{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse}
\title{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse}
\format{
An \code{R6Class} generator object
}
\description{
Results for a backup type job.

OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{migrationType}}{character [optional]}

\item{\code{ranges}}{List of object IDs to delete list(\link{OrgSagebionetworksRepoModelMigrationIdRange}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-new}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-toJSON}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-toString}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-isValid}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-print}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-clone}{\code{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$new(
  concreteType,
  migrationType = NULL,
  ranges = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{migrationType}}{migrationType}

\item{\code{ranges}}{List of object IDs to delete}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
