% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_dataset_backfill_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest}
\alias{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest}
\title{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest}
\format{
An \code{R6Class} generator object
}
\description{
Dataset back filling of checksum, size and count request.

OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-new}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-toJSON}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-toString}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-isValid}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-print}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-clone}{\code{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$new(
  concreteType,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationDatasetBackfillRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
