% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_has_migration_type.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationHasMigrationType}
\alias{OrgSagebionetworksRepoModelMigrationHasMigrationType}
\title{OrgSagebionetworksRepoModelMigrationHasMigrationType}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelMigrationHasMigrationType Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelMigrationHasMigrationType.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationHasMigrationType-new}{\code{OrgSagebionetworksRepoModelMigrationHasMigrationType$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationHasMigrationType-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationHasMigrationType$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationHasMigrationType-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationHasMigrationType$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationHasMigrationType-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationHasMigrationType$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationHasMigrationType-toJSON}{\code{OrgSagebionetworksRepoModelMigrationHasMigrationType$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationHasMigrationType-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationHasMigrationType$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationHasMigrationType-toString}{\code{OrgSagebionetworksRepoModelMigrationHasMigrationType$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationHasMigrationType-print}{\code{OrgSagebionetworksRepoModelMigrationHasMigrationType$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationHasMigrationType-clone}{\code{OrgSagebionetworksRepoModelMigrationHasMigrationType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationHasMigrationType-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationHasMigrationType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationHasMigrationType.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationHasMigrationType$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelMigrationAsyncMigrationRangeChecksumRequest", "OrgSagebionetworksRepoModelMigrationAsyncMigrationTypeChecksumRequest", "OrgSagebionetworksRepoModelMigrationBackupTypeRangeRequest", "OrgSagebionetworksRepoModelMigrationBatchChecksumRequest", "OrgSagebionetworksRepoModelMigrationBatchChecksumResponse", "OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeRequest", "OrgSagebionetworksRepoModelMigrationCalculateOptimalRangeResponse", "OrgSagebionetworksRepoModelMigrationMigrationTypeChecksum", "OrgSagebionetworksRepoModelMigrationRestoreTypeRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationHasMigrationType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationHasMigrationType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationHasMigrationType.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationHasMigrationType$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelMigrationHasMigrationType.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationHasMigrationType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationHasMigrationType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationHasMigrationType.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationHasMigrationType$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelMigrationHasMigrationType.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationHasMigrationType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationHasMigrationType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelMigrationHasMigrationType to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationHasMigrationType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelMigrationHasMigrationType.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationHasMigrationType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationHasMigrationType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelMigrationHasMigrationType to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationHasMigrationType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelMigrationHasMigrationType.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationHasMigrationType-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationHasMigrationType-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelMigrationHasMigrationType and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationHasMigrationType$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationHasMigrationType-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationHasMigrationType-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationHasMigrationType$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationHasMigrationType-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationHasMigrationType-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationHasMigrationType$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationHasMigrationType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationHasMigrationType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationHasMigrationType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
