% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_id_generator_export.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationIdGeneratorExport}
\alias{OrgSagebionetworksRepoModelMigrationIdGeneratorExport}
\title{OrgSagebionetworksRepoModelMigrationIdGeneratorExport}
\format{
An \code{R6Class} generator object
}
\description{
Export script for the ID generator.

OrgSagebionetworksRepoModelMigrationIdGeneratorExport Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationIdGeneratorExport
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{exportScript}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-new}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-toJSON}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-toString}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-isValid}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-print}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-clone}{\code{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationIdGeneratorExport class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$new(
  exportScript = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exportScript}}{exportScript}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationIdGeneratorExport in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationIdGeneratorExport
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationIdGeneratorExport
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationIdGeneratorExport in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationIdGeneratorExport
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationIdGeneratorExport
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationIdGeneratorExport and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationIdGeneratorExport
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationIdGeneratorExport-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationIdGeneratorExport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
