% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_migration_range_checksum.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum}
\alias{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum}
\title{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum}
\format{
An \code{R6Class} generator object
}
\description{
Checksum (aggregate) for record range from minId to maxId of migration type 'type'

OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{type}}{character [optional]}

\item{\code{checksum}}{character [optional]}

\item{\code{minid}}{integer [optional]}

\item{\code{maxid}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-new}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-toJSON}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-toString}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-isValid}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-print}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-clone}{\code{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$new(
  concreteType,
  type = NULL,
  checksum = NULL,
  minid = NULL,
  maxid = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{type}}{type}

\item{\code{checksum}}{checksum}

\item{\code{minid}}{minid}

\item{\code{maxid}}{maxid}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationRangeChecksum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
