% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_migration_type_counts.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts}
\alias{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts}
\title{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts}
\format{
An \code{R6Class} generator object
}
\description{
List of MigrationTypeCount records

OrgSagebionetworksRepoModelMigrationMigrationTypeCounts Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationMigrationTypeCounts
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{list}}{list(\link{OrgSagebionetworksRepoModelMigrationMigrationTypeCount}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-new}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-toJSON}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-toString}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-isValid}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-print}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-clone}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationMigrationTypeCounts class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$new(
  concreteType,
  list = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{list}}{list}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationMigrationTypeCounts in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationMigrationTypeCounts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationMigrationTypeCounts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationMigrationTypeCounts in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationMigrationTypeCounts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationMigrationTypeCounts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationMigrationTypeCounts and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationMigrationTypeCounts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeCounts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeCounts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
