% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_migration_type_names.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationMigrationTypeNames}
\alias{OrgSagebionetworksRepoModelMigrationMigrationTypeNames}
\title{OrgSagebionetworksRepoModelMigrationMigrationTypeNames}
\format{
An \code{R6Class} generator object
}
\description{
List of MigrationType names

OrgSagebionetworksRepoModelMigrationMigrationTypeNames Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationMigrationTypeNames
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{list}}{list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-new}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-toJSON}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-toString}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-isValid}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-print}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-clone}{\code{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationMigrationTypeNames class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$new(list = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list}}{list}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationMigrationTypeNames in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationMigrationTypeNames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationMigrationTypeNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationMigrationTypeNames in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationMigrationTypeNames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationMigrationTypeNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationMigrationTypeNames and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationMigrationTypeNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationMigrationTypeNames-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationMigrationTypeNames$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
