% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_range_checksum.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationRangeChecksum}
\alias{OrgSagebionetworksRepoModelMigrationRangeChecksum}
\title{OrgSagebionetworksRepoModelMigrationRangeChecksum}
\format{
An \code{R6Class} generator object
}
\description{
The calculated checksum for a single batch of rows.

OrgSagebionetworksRepoModelMigrationRangeChecksum Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationRangeChecksum
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{binNumber}}{integer [optional]}

\item{\code{count}}{integer [optional]}

\item{\code{checksum}}{character [optional]}

\item{\code{minimumId}}{integer [optional]}

\item{\code{maximumId}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-new}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-toJSON}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-toString}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-isValid}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-print}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRangeChecksum-clone}{\code{OrgSagebionetworksRepoModelMigrationRangeChecksum$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationRangeChecksum class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$new(
  binNumber = NULL,
  count = NULL,
  checksum = NULL,
  minimumId = NULL,
  maximumId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{binNumber}}{binNumber}

\item{\code{count}}{count}

\item{\code{checksum}}{checksum}

\item{\code{minimumId}}{minimumId}

\item{\code{maximumId}}{maximumId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationRangeChecksum in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationRangeChecksum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationRangeChecksum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationRangeChecksum in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationRangeChecksum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationRangeChecksum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationRangeChecksum and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationRangeChecksum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRangeChecksum-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRangeChecksum-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRangeChecksum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
