% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_migration_restore_type_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest}
\alias{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest}
\title{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to restore a migration backup file for a given MigrationType.

OrgSagebionetworksRepoModelMigrationRestoreTypeRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelMigrationRestoreTypeRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{migrationType}}{character [optional]}

\item{\code{aliasType}}{character [optional]}

\item{\code{backupFileKey}}{character [optional]}

\item{\code{batchSize}}{integer [optional]}

\item{\code{minimumRowId}}{integer [optional]}

\item{\code{maximumRowId}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-new}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-toJSON}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-fromJSON}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-toJSONString}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-validateJSON}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-toString}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-isValid}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-print}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-clone}{\code{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelMigrationRestoreTypeRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$new(
  concreteType,
  migrationType = NULL,
  aliasType = NULL,
  backupFileKey = NULL,
  batchSize = NULL,
  minimumRowId = NULL,
  maximumRowId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{migrationType}}{migrationType}

\item{\code{aliasType}}{aliasType}

\item{\code{backupFileKey}}{backupFileKey}

\item{\code{batchSize}}{batchSize}

\item{\code{minimumRowId}}{minimumRowId}

\item{\code{maximumRowId}}{maximumRowId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationRestoreTypeRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationRestoreTypeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationRestoreTypeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelMigrationRestoreTypeRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelMigrationRestoreTypeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelMigrationRestoreTypeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelMigrationRestoreTypeRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelMigrationRestoreTypeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelMigrationRestoreTypeRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelMigrationRestoreTypeRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
