% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_node.R
\docType{class}
\name{OrgSagebionetworksRepoModelNode}
\alias{OrgSagebionetworksRepoModelNode}
\title{OrgSagebionetworksRepoModelNode}
\format{
An \code{R6Class} generator object
}
\description{
This is the DTO of a node.

OrgSagebionetworksRepoModelNode Class
}
\details{
Create a new OrgSagebionetworksRepoModelNode
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{parentId}}{character [optional]}

\item{\code{createdByPrincipalId}}{integer [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedByPrincipalId}}{integer [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{nodeType}}{character [optional]}

\item{\code{eTag}}{character [optional]}

\item{\code{versionNumber}}{integer [optional]}

\item{\code{versionComment}}{character [optional]}

\item{\code{versionLabel}}{character [optional]}

\item{\code{isLatestVersion}}{character [optional]}

\item{\code{activityId}}{character [optional]}

\item{\code{fileHandleId}}{character [optional]}

\item{\code{columnModelIds}}{list(character) [optional]}

\item{\code{scopeIds}}{For FileVeiws, the list of IDs the define the scope of the view. list(character) [optional]}

\item{\code{items}}{For Datasets and Dataset Collections, the list of entity references the define the view. list(\link{OrgSagebionetworksRepoModelEntityRef}) [optional]}

\item{\code{reference}}{\link{OrgSagebionetworksRepoModelReference} [optional]}

\item{\code{alias}}{character [optional]}

\item{\code{isSearchEnabled}}{character [optional]}

\item{\code{definingSQL}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelNode-new}{\code{OrgSagebionetworksRepoModelNode$new()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-toJSON}{\code{OrgSagebionetworksRepoModelNode$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-fromJSON}{\code{OrgSagebionetworksRepoModelNode$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-toJSONString}{\code{OrgSagebionetworksRepoModelNode$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-fromJSONString}{\code{OrgSagebionetworksRepoModelNode$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-validateJSON}{\code{OrgSagebionetworksRepoModelNode$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-toString}{\code{OrgSagebionetworksRepoModelNode$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-isValid}{\code{OrgSagebionetworksRepoModelNode$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-getInvalidFields}{\code{OrgSagebionetworksRepoModelNode$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-print}{\code{OrgSagebionetworksRepoModelNode$print()}}
\item \href{#method-OrgSagebionetworksRepoModelNode-clone}{\code{OrgSagebionetworksRepoModelNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelNode class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$new(
  id = NULL,
  name = NULL,
  description = NULL,
  parentId = NULL,
  createdByPrincipalId = NULL,
  createdOn = NULL,
  modifiedByPrincipalId = NULL,
  modifiedOn = NULL,
  nodeType = NULL,
  eTag = NULL,
  versionNumber = NULL,
  versionComment = NULL,
  versionLabel = NULL,
  isLatestVersion = NULL,
  activityId = NULL,
  fileHandleId = NULL,
  columnModelIds = NULL,
  scopeIds = NULL,
  items = NULL,
  reference = NULL,
  alias = NULL,
  isSearchEnabled = NULL,
  definingSQL = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{parentId}}{parentId}

\item{\code{createdByPrincipalId}}{createdByPrincipalId}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedByPrincipalId}}{modifiedByPrincipalId}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{nodeType}}{nodeType}

\item{\code{eTag}}{eTag}

\item{\code{versionNumber}}{versionNumber}

\item{\code{versionComment}}{versionComment}

\item{\code{versionLabel}}{versionLabel}

\item{\code{isLatestVersion}}{isLatestVersion}

\item{\code{activityId}}{activityId}

\item{\code{fileHandleId}}{fileHandleId}

\item{\code{columnModelIds}}{columnModelIds}

\item{\code{scopeIds}}{For FileVeiws, the list of IDs the define the scope of the view.}

\item{\code{items}}{For Datasets and Dataset Collections, the list of entity references the define the view.}

\item{\code{reference}}{reference}

\item{\code{alias}}{alias}

\item{\code{isSearchEnabled}}{isSearchEnabled}

\item{\code{definingSQL}}{definingSQL}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelNode in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelNode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelNode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelNode in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelNode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelNode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelNode and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelNode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
