% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_json_web_key.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthJsonWebKey}
\alias{OrgSagebionetworksRepoModelOauthJsonWebKey}
\title{OrgSagebionetworksRepoModelOauthJsonWebKey}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelOauthJsonWebKey Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelOauthJsonWebKey.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKey-new}{\code{OrgSagebionetworksRepoModelOauthJsonWebKey$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKey-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthJsonWebKey$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKey-fromJSON}{\code{OrgSagebionetworksRepoModelOauthJsonWebKey$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKey-toJSONString}{\code{OrgSagebionetworksRepoModelOauthJsonWebKey$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKey-toJSON}{\code{OrgSagebionetworksRepoModelOauthJsonWebKey$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKey-validateJSON}{\code{OrgSagebionetworksRepoModelOauthJsonWebKey$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKey-toString}{\code{OrgSagebionetworksRepoModelOauthJsonWebKey$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKey-print}{\code{OrgSagebionetworksRepoModelOauthJsonWebKey$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKey-clone}{\code{OrgSagebionetworksRepoModelOauthJsonWebKey$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKey-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKey-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthJsonWebKey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKey$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelOauthJsonWebKeyRSA"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKey-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKey-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthJsonWebKey.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKey$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelOauthJsonWebKey.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKey-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKey-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthJsonWebKey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKey$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelOauthJsonWebKey.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKey-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKey-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelOauthJsonWebKey to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKey$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelOauthJsonWebKey.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKey-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKey-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelOauthJsonWebKey to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKey$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelOauthJsonWebKey.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKey-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKey-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelOauthJsonWebKey and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKey$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKey-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKey-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKey$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKey-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKey-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKey$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKey-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKey-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKey$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
