% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_json_web_key_rsa.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA}
\alias{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA}
\title{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA}
\format{
An \code{R6Class} generator object
}
\description{
JSON Web Key for the RSA algorithm

OrgSagebionetworksRepoModelOauthJsonWebKeyRSA Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthJsonWebKeyRSA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{kid}}{character [optional]}

\item{\code{kty}}{character [optional]}

\item{\code{use}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{e}}{character [optional]}

\item{\code{n}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-new}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-toJSON}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-fromJSON}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-toJSONString}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-validateJSON}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-toString}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-isValid}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-print}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-clone}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthJsonWebKeyRSA class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$new(
  concreteType,
  kid = NULL,
  kty = NULL,
  use = NULL,
  e = NULL,
  n = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{kid}}{kid}

\item{\code{kty}}{kty}

\item{\code{use}}{use}

\item{\code{e}}{e}

\item{\code{n}}{n}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthJsonWebKeyRSA in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthJsonWebKeyRSA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthJsonWebKeyRSA
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthJsonWebKeyRSA in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthJsonWebKeyRSA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthJsonWebKeyRSA
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthJsonWebKeyRSA and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthJsonWebKeyRSA
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeyRSA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeyRSA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
