% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_json_web_key_set.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthJsonWebKeySet}
\alias{OrgSagebionetworksRepoModelOauthJsonWebKeySet}
\title{OrgSagebionetworksRepoModelOauthJsonWebKeySet}
\format{
An \code{R6Class} generator object
}
\description{
List of JSON Web Keys

OrgSagebionetworksRepoModelOauthJsonWebKeySet Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthJsonWebKeySet
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{keys}}{list(\link{OrgSagebionetworksRepoModelOauthJsonWebKey}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-new}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-toJSON}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-fromJSON}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-toJSONString}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-validateJSON}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-toString}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-isValid}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-print}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-clone}{\code{OrgSagebionetworksRepoModelOauthJsonWebKeySet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthJsonWebKeySet class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$new(keys = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{keys}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthJsonWebKeySet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthJsonWebKeySet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthJsonWebKeySet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthJsonWebKeySet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthJsonWebKeySet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthJsonWebKeySet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthJsonWebKeySet and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthJsonWebKeySet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthJsonWebKeySet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthJsonWebKeySet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
