% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_o_auth_client.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOAuthClient}
\alias{OrgSagebionetworksRepoModelOauthOAuthClient}
\title{OrgSagebionetworksRepoModelOauthOAuthClient}
\format{
An \code{R6Class} generator object
}
\description{
OAuth 2.0 Client metadata described in <a href=\"https://openid.net/specs/openid-connect-registration-1_0.html#ClientMetadata\">OpenID Connect Core 1.0 Client Metadata</a>

OrgSagebionetworksRepoModelOauthOAuthClient Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOAuthClient
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client_id}}{character [optional]}

\item{\code{client_name}}{character [optional]}

\item{\code{redirect_uris}}{Array of Redirection URI values used by the Client. One of these registered Redirection URI values MUST exactly match the redirect_uri parameter value used in each Authorization Request. If the 'sector_identifier_uri' is provided then the URIs in this list must be in the list of URIs in the JSON document referenced by the sector_identifier_uri. list(character) [optional]}

\item{\code{verified}}{character [optional]}

\item{\code{client_uri}}{character [optional]}

\item{\code{policy_uri}}{character [optional]}

\item{\code{tos_uri}}{character [optional]}

\item{\code{sector_identifier_uri}}{character [optional]}

\item{\code{sector_identifier}}{character [optional]}

\item{\code{userinfo_signed_response_alg}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{etag}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-new}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-toJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-toString}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-isValid}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-print}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClient-clone}{\code{OrgSagebionetworksRepoModelOauthOAuthClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOAuthClient class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$new(
  client_id = NULL,
  client_name = NULL,
  redirect_uris = NULL,
  verified = NULL,
  client_uri = NULL,
  policy_uri = NULL,
  tos_uri = NULL,
  sector_identifier_uri = NULL,
  sector_identifier = NULL,
  userinfo_signed_response_alg = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  etag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_id}}{client_id}

\item{\code{client_name}}{client_name}

\item{\code{redirect_uris}}{Array of Redirection URI values used by the Client. One of these registered Redirection URI values MUST exactly match the redirect_uri parameter value used in each Authorization Request. If the 'sector_identifier_uri' is provided then the URIs in this list must be in the list of URIs in the JSON document referenced by the sector_identifier_uri.}

\item{\code{verified}}{verified}

\item{\code{client_uri}}{client_uri}

\item{\code{policy_uri}}{policy_uri}

\item{\code{tos_uri}}{tos_uri}

\item{\code{sector_identifier_uri}}{sector_identifier_uri}

\item{\code{sector_identifier}}{sector_identifier}

\item{\code{userinfo_signed_response_alg}}{userinfo_signed_response_alg}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{etag}}{etag}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClient in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClient
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClient in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClient
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOAuthClient and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOAuthClient
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
