% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_o_auth_client_authorization_history.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory}
\alias{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory}
\title{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory}
\format{
An \code{R6Class} generator object
}
\description{
An object representing an OAuth 2.0 client that has been authorized to request a user's resources and identity

OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{\link{OrgSagebionetworksRepoModelOauthOAuthClient} [optional]}

\item{\code{authorizedOn}}{character [optional]}

\item{\code{lastUsed}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-new}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-toJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-toString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-isValid}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-print}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-clone}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$new(
  client = NULL,
  authorizedOn = NULL,
  lastUsed = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client}

\item{\code{authorizedOn}}{authorizedOn}

\item{\code{lastUsed}}{lastUsed}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
