% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_o_auth_client_authorization_history_list.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList}
\alias{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList}
\title{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList}
\format{
An \code{R6Class} generator object
}
\description{
Metadata for a paginated list of OAuth 2.0 clients authorized to access a user's resources in Synapse.

OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistory}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-new}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-toJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-toString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-isValid}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-print}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-clone}{\code{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$new(
  results = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{results}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientAuthorizationHistoryList$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
