% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_o_auth_client_id_and_secret.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret}
\alias{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret}
\title{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret}
\format{
An \code{R6Class} generator object
}
\description{
OAuth 2.0 Client ID and secret, generated when a new client is created

OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client_id}}{character [optional]}

\item{\code{client_secret}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-new}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-toJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-toString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-isValid}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-print}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-clone}{\code{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$new(
  client_id = NULL,
  client_secret = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_id}}{client_id}

\item{\code{client_secret}}{client_secret}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientIdAndSecret$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
