% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_o_auth_client_verification_precheck_result.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult}
\alias{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult}
\title{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult}
\format{
An \code{R6Class} generator object
}
\description{
boolean value indicating whether proposed changes to the OAuth client will require reverification

OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{reverificationRequired}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-new}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-toJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-toString}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-isValid}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-print}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-clone}{\code{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$new(
  reverificationRequired = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reverificationRequired}}{reverificationRequired}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthClientVerificationPrecheckResult$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
