% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_o_auth_refresh_token_information.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation}
\alias{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation}
\title{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation}
\format{
An \code{R6Class} generator object
}
\description{
OAuth 2.0 refresh token metadata. Note that the token itself may not be retrieved after it is issued to a client.

OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tokenId}}{character [optional]}

\item{\code{clientId}}{character [optional]}

\item{\code{principalId}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{scopes}}{An array containing each scope that this refresh token grants access to. list(character) [optional]}

\item{\code{claims}}{\link{OrgSagebionetworksRepoModelOauthOIDCClaimsRequest} [optional]}

\item{\code{authorizedOn}}{character [optional]}

\item{\code{lastUsed}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{etag}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-new}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-toJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-toString}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-isValid}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-print}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-clone}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$new(
  tokenId = NULL,
  clientId = NULL,
  principalId = NULL,
  name = NULL,
  scopes = NULL,
  claims = NULL,
  authorizedOn = NULL,
  lastUsed = NULL,
  modifiedOn = NULL,
  etag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tokenId}}{tokenId}

\item{\code{clientId}}{clientId}

\item{\code{principalId}}{principalId}

\item{\code{name}}{name}

\item{\code{scopes}}{An array containing each scope that this refresh token grants access to.}

\item{\code{claims}}{claims}

\item{\code{authorizedOn}}{authorizedOn}

\item{\code{lastUsed}}{lastUsed}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{etag}}{etag}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
