% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_o_auth_refresh_token_information_list.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList}
\alias{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList}
\title{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList}
\format{
An \code{R6Class} generator object
}
\description{
Metadata for a paginated list of refresh token metadata that give OAuth 2.0 clients access to access a user's resources in Synapse.

OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformation}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-new}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-toJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-toString}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-isValid}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-print}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-clone}{\code{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$new(
  results = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{results}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthRefreshTokenInformationList$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
