% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_o_auth_validation_request_interface.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface}
\alias{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface}
\title{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-new}{\code{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-toJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-toString}{\code{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-print}{\code{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-clone}{\code{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$new(
  instance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelOauthOAuthAccountCreationRequest", "OrgSagebionetworksRepoModelOauthOAuthValidationRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOAuthValidationRequestInterface$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
