% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_oidc_authorization_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest}
\alias{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest}
\title{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest}
\format{
An \code{R6Class} generator object
}
\description{
Content of an authorization request

OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{clientId}}{character [optional]}

\item{\code{scope}}{character [optional]}

\item{\code{claims}}{\link{OrgSagebionetworksRepoModelOauthOIDCClaimsRequest} [optional]}

\item{\code{responseType}}{character [optional]}

\item{\code{redirectUri}}{character [optional]}

\item{\code{nonce}}{character [optional]}

\item{\code{userId}}{character [optional]}

\item{\code{authenticatedAt}}{character [optional]}

\item{\code{authorizedAt}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-new}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-toJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-toString}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-isValid}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-print}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-clone}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$new(
  clientId = NULL,
  scope = NULL,
  claims = NULL,
  responseType = NULL,
  redirectUri = NULL,
  nonce = NULL,
  userId = NULL,
  authenticatedAt = NULL,
  authorizedAt = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clientId}}{clientId}

\item{\code{scope}}{scope}

\item{\code{claims}}{claims}

\item{\code{responseType}}{responseType}

\item{\code{redirectUri}}{redirectUri}

\item{\code{nonce}}{nonce}

\item{\code{userId}}{userId}

\item{\code{authenticatedAt}}{authenticatedAt}

\item{\code{authorizedAt}}{authorizedAt}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
