% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_oidc_authorization_request_description.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription}
\alias{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription}
\title{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription}
\format{
An \code{R6Class} generator object
}
\description{
Human readable description / details about an authorization request

OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{clientId}}{character [optional]}

\item{\code{redirect_uri}}{character [optional]}

\item{\code{scope}}{Scopes and claims requested by client list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-new}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-toJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-toString}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-isValid}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-print}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-clone}{\code{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$new(
  clientId = NULL,
  redirect_uri = NULL,
  scope = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clientId}}{clientId}

\item{\code{redirect_uri}}{redirect_uri}

\item{\code{scope}}{Scopes and claims requested by client}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCAuthorizationRequestDescription$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
