% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_oidc_claims_request_details.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails}
\alias{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails}
\title{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails}
\format{
An \code{R6Class} generator object
}
\description{
<a href=\"https://openid.net/specs/openid-connect-core-1_0.html#ClaimsParameter\">OIDC Core 1.0 Claims Request</a> value

OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails Class
}
\details{
Create a new OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{essential}}{character [optional]}

\item{\code{value}}{character [optional]}

\item{\code{values}}{list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-new}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-toJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-validateJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-toString}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-isValid}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-getInvalidFields}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-print}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$print()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-clone}{\code{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$new(
  essential = NULL,
  value = NULL,
  values = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{essential}}{essential}

\item{\code{value}}{value}

\item{\code{values}}{values}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCClaimsRequestDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
