% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_oidc_signing_algorithm.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm}
\alias{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm}
\title{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-new}{\code{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-toJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-clone}{\code{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSigningAlgorithm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
