% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_oauth_oidc_subject_identifier_type.R
\docType{class}
\name{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType}
\alias{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType}
\title{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-new}{\code{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$new()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-toJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-fromJSON}{\code{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-toJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-fromJSONString}{\code{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-clone}{\code{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelOauthOIDCSubjectIdentifierType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
