% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_paginated_team_ids.R
\docType{class}
\name{OrgSagebionetworksRepoModelPaginatedTeamIds}
\alias{OrgSagebionetworksRepoModelPaginatedTeamIds}
\title{OrgSagebionetworksRepoModelPaginatedTeamIds}
\format{
An \code{R6Class} generator object
}
\description{
Container for a page of team IDs.

OrgSagebionetworksRepoModelPaginatedTeamIds Class
}
\details{
Create a new OrgSagebionetworksRepoModelPaginatedTeamIds
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{teamIds}}{The requested team IDs. list(character) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-new}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-toJSON}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-fromJSON}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-toJSONString}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-fromJSONString}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-validateJSON}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-toString}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-isValid}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-getInvalidFields}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-print}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPaginatedTeamIds-clone}{\code{OrgSagebionetworksRepoModelPaginatedTeamIds$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPaginatedTeamIds class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$new(
  teamIds = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{teamIds}}{The requested team IDs.}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPaginatedTeamIds in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPaginatedTeamIds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPaginatedTeamIds
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPaginatedTeamIds in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPaginatedTeamIds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPaginatedTeamIds
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPaginatedTeamIds and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPaginatedTeamIds
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPaginatedTeamIds-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPaginatedTeamIds-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPaginatedTeamIds$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
