% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_performance_performance_record.R
\docType{class}
\name{OrgSagebionetworksRepoModelPerformancePerformanceRecord}
\alias{OrgSagebionetworksRepoModelPerformancePerformanceRecord}
\title{OrgSagebionetworksRepoModelPerformancePerformanceRecord}
\format{
An \code{R6Class} generator object
}
\description{
Data record for call performance.

OrgSagebionetworksRepoModelPerformancePerformanceRecord Class
}
\details{
Create a new OrgSagebionetworksRepoModelPerformancePerformanceRecord
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{method}}{character [optional]}

\item{\code{totalCount}}{integer [optional]}

\item{\code{totalTime}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-new}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-toJSON}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-fromJSON}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-toJSONString}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-fromJSONString}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-validateJSON}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-toString}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-isValid}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-getInvalidFields}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-print}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-clone}{\code{OrgSagebionetworksRepoModelPerformancePerformanceRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPerformancePerformanceRecord class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$new(
  method = NULL,
  totalCount = NULL,
  totalTime = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{method}

\item{\code{totalCount}}{totalCount}

\item{\code{totalTime}}{totalTime}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPerformancePerformanceRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPerformancePerformanceRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPerformancePerformanceRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPerformancePerformanceRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPerformancePerformanceRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPerformancePerformanceRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPerformancePerformanceRecord and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPerformancePerformanceRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPerformancePerformanceRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPerformancePerformanceRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
