% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_post_message_content_access_requirement.R
\docType{class}
\name{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement}
\alias{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement}
\title{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement}
\format{
An \code{R6Class} generator object
}
\description{
This class is deprecated and should not being used.

OrgSagebionetworksRepoModelPostMessageContentAccessRequirement Class
}
\details{
Create a new OrgSagebionetworksRepoModelPostMessageContentAccessRequirement
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{versionNumber}}{integer [optional]}

\item{\code{id}}{integer [optional]}

\item{\code{description}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{subjectsDefinedByAnnotations}}{character [optional]}

\item{\code{subjectIds}}{The IDs of the items controlled by this Access Requirement when 'subjectsDefinedByAnnotations=false'. This property is mutually exclusive with 'subjectsDefinedByAnnotations'.  When 'subjectsDefinedByAnnotations=true' then this property must be empty or excluded.  Required when creating or updating and 'subjectsDefinedByAnnotations=false' or 'subjectsDefinedByAnnotations' is excluded. list(\link{OrgSagebionetworksRepoModelRestrictableObjectDescriptor}) [optional]}

\item{\code{accessType}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{url}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-new}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-toJSON}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-fromJSON}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-toJSONString}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-fromJSONString}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-validateJSON}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-toString}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-isValid}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-getInvalidFields}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-print}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-clone}{\code{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPostMessageContentAccessRequirement class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$new(
  concreteType,
  versionNumber = NULL,
  id = NULL,
  description = NULL,
  name = NULL,
  etag = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  subjectsDefinedByAnnotations = NULL,
  subjectIds = NULL,
  accessType = NULL,
  url = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{versionNumber}}{versionNumber}

\item{\code{id}}{id}

\item{\code{description}}{description}

\item{\code{name}}{name}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{subjectsDefinedByAnnotations}}{subjectsDefinedByAnnotations}

\item{\code{subjectIds}}{The IDs of the items controlled by this Access Requirement when 'subjectsDefinedByAnnotations=false'. This property is mutually exclusive with 'subjectsDefinedByAnnotations'.  When 'subjectsDefinedByAnnotations=true' then this property must be empty or excluded.  Required when creating or updating and 'subjectsDefinedByAnnotations=false' or 'subjectsDefinedByAnnotations' is excluded.}

\item{\code{accessType}}{accessType}

\item{\code{url}}{url}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPostMessageContentAccessRequirement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPostMessageContentAccessRequirement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPostMessageContentAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPostMessageContentAccessRequirement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPostMessageContentAccessRequirement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPostMessageContentAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPostMessageContentAccessRequirement and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPostMessageContentAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPostMessageContentAccessRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPostMessageContentAccessRequirement$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
