% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_account_creation_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalAccountCreationToken}
\alias{OrgSagebionetworksRepoModelPrincipalAccountCreationToken}
\title{OrgSagebionetworksRepoModelPrincipalAccountCreationToken}
\format{
An \code{R6Class} generator object
}
\description{
Information needed to start the process of creating a new account.

OrgSagebionetworksRepoModelPrincipalAccountCreationToken Class
}
\details{
Create a new OrgSagebionetworksRepoModelPrincipalAccountCreationToken
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{emailValidationSignedToken}}{\link{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken} [optional]}

\item{\code{encodedMembershipInvtnSignedToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-new}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-validateJSON}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-toString}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-isValid}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-getInvalidFields}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-print}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-clone}{\code{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalAccountCreationToken class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$new(
  emailValidationSignedToken = NULL,
  encodedMembershipInvtnSignedToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emailValidationSignedToken}}{emailValidationSignedToken}

\item{\code{encodedMembershipInvtnSignedToken}}{encodedMembershipInvtnSignedToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalAccountCreationToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalAccountCreationToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalAccountCreationToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalAccountCreationToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalAccountCreationToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalAccountCreationToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPrincipalAccountCreationToken and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPrincipalAccountCreationToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountCreationToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountCreationToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
