% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_account_setup_info.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo}
\alias{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo}
\title{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo}
\format{
An \code{R6Class} generator object
}
\description{
Information needed to establish a new user account, following email validation.

OrgSagebionetworksRepoModelPrincipalAccountSetupInfo Class
}
\details{
Create a new OrgSagebionetworksRepoModelPrincipalAccountSetupInfo
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{firstName}}{character [optional]}

\item{\code{lastName}}{character [optional]}

\item{\code{username}}{character [optional]}

\item{\code{password}}{character [optional]}

\item{\code{emailValidationSignedToken}}{\link{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-new}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-validateJSON}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-toString}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-isValid}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-getInvalidFields}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-print}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-clone}{\code{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalAccountSetupInfo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$new(
  firstName = NULL,
  lastName = NULL,
  username = NULL,
  password = NULL,
  emailValidationSignedToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{firstName}}{firstName}

\item{\code{lastName}}{lastName}

\item{\code{username}}{username}

\item{\code{password}}{password}

\item{\code{emailValidationSignedToken}}{emailValidationSignedToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalAccountSetupInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalAccountSetupInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalAccountSetupInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalAccountSetupInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalAccountSetupInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalAccountSetupInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPrincipalAccountSetupInfo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPrincipalAccountSetupInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAccountSetupInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAccountSetupInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
