% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_alias_check_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse}
\alias{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse}
\title{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse}
\format{
An \code{R6Class} generator object
}
\description{
Response to an alias check request.

OrgSagebionetworksRepoModelPrincipalAliasCheckResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelPrincipalAliasCheckResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{available}}{character [optional]}

\item{\code{valid}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-new}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-validateJSON}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-toString}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-isValid}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-print}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-clone}{\code{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalAliasCheckResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$new(
  available = NULL,
  valid = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{available}}{available}

\item{\code{valid}}{valid}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalAliasCheckResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalAliasCheckResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalAliasCheckResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalAliasCheckResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalAliasCheckResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalAliasCheckResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPrincipalAliasCheckResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPrincipalAliasCheckResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasCheckResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasCheckResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
