% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_alias_type.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalAliasType}
\alias{OrgSagebionetworksRepoModelPrincipalAliasType}
\title{OrgSagebionetworksRepoModelPrincipalAliasType}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelPrincipalAliasType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasType-new}{\code{OrgSagebionetworksRepoModelPrincipalAliasType$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasType-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalAliasType$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasType-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalAliasType$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasType-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalAliasType$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasType-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalAliasType$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalAliasType-clone}{\code{OrgSagebionetworksRepoModelPrincipalAliasType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasType-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalAliasType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalAliasType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalAliasType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalAliasType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalAliasType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalAliasType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalAliasType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalAliasType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalAliasType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalAliasType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
