% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_email_quarantine_reason.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason}
\alias{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason}
\title{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-new}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-clone}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineReason$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
