% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_email_quarantine_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus}
\alias{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus}
\title{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus}
\format{
An \code{R6Class} generator object
}
\description{
Information about the quarantine status of a notification email

OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{reason}}{character [optional]}

\item{\code{reasonDetails}}{character [optional]}

\item{\code{expiration}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-new}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-validateJSON}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-toString}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-isValid}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-print}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-clone}{\code{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$new(
  reason = NULL,
  reasonDetails = NULL,
  expiration = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reason}}{reason}

\item{\code{reasonDetails}}{reasonDetails}

\item{\code{expiration}}{expiration}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
