% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_email_validation_signed_token.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken}
\alias{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken}
\title{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken}
\format{
An \code{R6Class} generator object
}
\description{
Signed token containing the the information necessary to create a new account or add an email address to an existing account.

OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken Class
}
\details{
Create a new OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{hmac}}{character [optional]}

\item{\code{version}}{integer [optional]}

\item{\code{expiresOn}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{email}}{character [optional]}

\item{\code{userId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-new}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-validateJSON}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-toString}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-isValid}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-getInvalidFields}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-print}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-clone}{\code{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$new(
  concreteType,
  hmac = NULL,
  version = NULL,
  expiresOn = NULL,
  createdOn = NULL,
  email = NULL,
  userId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{hmac}}{hmac}

\item{\code{version}}{version}

\item{\code{expiresOn}}{expiresOn}

\item{\code{createdOn}}{createdOn}

\item{\code{email}}{email}

\item{\code{userId}}{userId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
