% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_notification_email.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalNotificationEmail}
\alias{OrgSagebionetworksRepoModelPrincipalNotificationEmail}
\title{OrgSagebionetworksRepoModelPrincipalNotificationEmail}
\format{
An \code{R6Class} generator object
}
\description{
Information about the notification email of the user, including its quarantined status if the email address is currently in quarantine

OrgSagebionetworksRepoModelPrincipalNotificationEmail Class
}
\details{
Create a new OrgSagebionetworksRepoModelPrincipalNotificationEmail
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character [optional]}

\item{\code{quarantineStatus}}{\link{OrgSagebionetworksRepoModelPrincipalEmailQuarantineStatus} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-new}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-validateJSON}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-toString}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-isValid}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-getInvalidFields}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-print}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-clone}{\code{OrgSagebionetworksRepoModelPrincipalNotificationEmail$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalNotificationEmail class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$new(
  email = NULL,
  quarantineStatus = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{email}

\item{\code{quarantineStatus}}{quarantineStatus}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalNotificationEmail in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalNotificationEmail
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalNotificationEmail
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalNotificationEmail in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalNotificationEmail
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalNotificationEmail
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPrincipalNotificationEmail and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPrincipalNotificationEmail
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalNotificationEmail-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalNotificationEmail$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
