% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_principal_alias.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalPrincipalAlias}
\alias{OrgSagebionetworksRepoModelPrincipalPrincipalAlias}
\title{OrgSagebionetworksRepoModelPrincipalPrincipalAlias}
\format{
An \code{R6Class} generator object
}
\description{
Alias binding of a principal

OrgSagebionetworksRepoModelPrincipalPrincipalAlias Class
}
\details{
Create a new OrgSagebionetworksRepoModelPrincipalPrincipalAlias
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aliasId}}{integer [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{principalId}}{integer [optional]}

\item{\code{alias}}{character [optional]}

\item{\code{type}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-new}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-validateJSON}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-toString}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-isValid}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-getInvalidFields}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-print}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-clone}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalPrincipalAlias class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$new(
  aliasId = NULL,
  etag = NULL,
  principalId = NULL,
  alias = NULL,
  type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aliasId}}{aliasId}

\item{\code{etag}}{etag}

\item{\code{principalId}}{principalId}

\item{\code{alias}}{alias}

\item{\code{type}}{type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalPrincipalAlias in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalPrincipalAlias
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalPrincipalAlias
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalPrincipalAlias in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalPrincipalAlias
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalPrincipalAlias
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPrincipalPrincipalAlias and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPrincipalPrincipalAlias
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAlias-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAlias$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
