% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_principal_alias_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse}
\alias{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse}
\title{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse}
\format{
An \code{R6Class} generator object
}
\description{
The response of a principal alias lookup request

OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{principalId}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-new}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-validateJSON}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-toString}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-isValid}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-print}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-clone}{\code{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$new(
  principalId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{principalId}}{principalId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalPrincipalAliasResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
