% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_principal_user_group_header_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse}
\alias{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse}
\title{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse}
\format{
An \code{R6Class} generator object
}
\description{
A list of UserGroupHeaders.

OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{list}}{The list of UserGroupHeaders. list(\link{OrgSagebionetworksRepoModelUserGroupHeader}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-new}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-toJSON}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-fromJSON}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-toJSONString}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-validateJSON}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-toString}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-isValid}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-print}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-clone}{\code{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$new(
  list = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list}}{The list of UserGroupHeaders.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelPrincipalUserGroupHeaderResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
