% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_project_base_key_storage_location_setting.R
\docType{class}
\name{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting}
\alias{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting}
\title{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-new}{\code{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-fromJSONString}{\code{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-fromJSON}{\code{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-toJSONString}{\code{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-toJSON}{\code{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-validateJSON}{\code{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-toString}{\code{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-print}{\code{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-clone}{\code{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$new(
  instance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting", "OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting", "OrgSagebionetworksRepoModelProjectS3StorageLocationSetting"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBaseKeyStorageLocationSetting$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
