% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_project_bucket_owner_storage_location_setting.R
\docType{class}
\name{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting}
\alias{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting}
\title{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-new}{\code{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-fromJSONString}{\code{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-fromJSON}{\code{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-toJSONString}{\code{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-toJSON}{\code{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-validateJSON}{\code{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-toString}{\code{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-print}{\code{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-clone}{\code{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$new(
  instance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting", "OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$fromJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectBucketOwnerStorageLocationSetting$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
