% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_project_external_google_cloud_storage_location_setting.R
\docType{class}
\name{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting}
\alias{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting}
\title{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting}
\format{
An \code{R6Class} generator object
}
\description{
The external upload destination setting contains information to generate an external Google Cloud Storage upload destination connected with Synapse.  For details on how to create a custom Google Cloud storage location in the web client, see <a href=\"https://help.synapse.org/docs/Custom-Storage-Locations.2048327803.html\">Custom Storage Locations</a>

OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting Class
}
\details{
Create a new OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{uploadType}}{character [optional]}

\item{\code{banner}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{integer [optional]}

\item{\code{bucket}}{character}

\item{\code{baseKey}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-new}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-toJSON}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-fromJSON}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-toJSONString}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-fromJSONString}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-validateJSON}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-toString}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-isValid}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-getInvalidFields}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-print}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-clone}{\code{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$new(
  concreteType,
  bucket,
  storageLocationId = NULL,
  uploadType = NULL,
  banner = NULL,
  description = NULL,
  etag = NULL,
  createdOn = NULL,
  createdBy = NULL,
  baseKey = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{bucket}}{bucket}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{uploadType}}{uploadType}

\item{\code{banner}}{banner}

\item{\code{description}}{description}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{baseKey}}{baseKey}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$toJSON(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$isValid(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$print(
  
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalGoogleCloudStorageLocationSetting$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
