% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_project_external_object_storage_location_setting.R
\docType{class}
\name{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting}
\alias{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting}
\title{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting}
\format{
An \code{R6Class} generator object
}
\description{
The external upload destination setting contains information to generate an upload destination for an S3 object storage NOT managed by Synapse. This is different from ExternalS3StorageLocationSetting, which contains information about an external S3 storage that is managed by Synapse

OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting Class
}
\details{
Create a new OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{uploadType}}{character [optional]}

\item{\code{banner}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{integer [optional]}

\item{\code{bucket}}{character}

\item{\code{endpointUrl}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-new}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-toJSON}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-fromJSON}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-toJSONString}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-fromJSONString}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-validateJSON}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-toString}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-isValid}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-getInvalidFields}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-print}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-clone}{\code{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$new(
  concreteType,
  bucket,
  endpointUrl,
  storageLocationId = NULL,
  uploadType = NULL,
  banner = NULL,
  description = NULL,
  etag = NULL,
  createdOn = NULL,
  createdBy = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{bucket}}{bucket}

\item{\code{endpointUrl}}{endpointUrl}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{uploadType}}{uploadType}

\item{\code{banner}}{banner}

\item{\code{description}}{description}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$toString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$isValid(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalObjectStorageLocationSetting$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
