% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_project_external_s3_storage_location_setting.R
\docType{class}
\name{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting}
\alias{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting}
\title{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting}
\format{
An \code{R6Class} generator object
}
\description{
The external upload destination setting contains information to generate an external S3 upload destination connected with Synapse. For details on how to create a custom S3 storage location in the web client, see <a href=\"https://help.synapse.org/docs/Custom-Storage-Locations.2048327803.html\">Custom Storage Locations</a>

OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting Class
}
\details{
Create a new OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{uploadType}}{character [optional]}

\item{\code{banner}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{integer [optional]}

\item{\code{bucket}}{character}

\item{\code{baseKey}}{character [optional]}

\item{\code{stsEnabled}}{character [optional]}

\item{\code{endpointUrl}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-new}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-toJSON}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-fromJSON}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-toJSONString}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-fromJSONString}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-validateJSON}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-toString}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-isValid}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-getInvalidFields}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-print}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-clone}{\code{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$new(
  concreteType,
  bucket,
  storageLocationId = NULL,
  uploadType = NULL,
  banner = NULL,
  description = NULL,
  etag = NULL,
  createdOn = NULL,
  createdBy = NULL,
  baseKey = NULL,
  stsEnabled = NULL,
  endpointUrl = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{bucket}}{bucket}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{uploadType}}{uploadType}

\item{\code{banner}}{banner}

\item{\code{description}}{description}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{baseKey}}{baseKey}

\item{\code{stsEnabled}}{stsEnabled}

\item{\code{endpointUrl}}{endpointUrl}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectExternalS3StorageLocationSetting$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
