% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_project_proxy_storage_location_settings.R
\docType{class}
\name{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings}
\alias{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings}
\title{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings}
\format{
An \code{R6Class} generator object
}
\description{
A storage location that uses a HTTPS proxy for all upload and download operations.

OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings Class
}
\details{
Create a new OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{uploadType}}{character [optional]}

\item{\code{banner}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{integer [optional]}

\item{\code{proxyUrl}}{character [optional]}

\item{\code{secretKey}}{character [optional]}

\item{\code{benefactorId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-new}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-toJSON}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-fromJSON}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-toJSONString}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-fromJSONString}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-validateJSON}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-toString}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-isValid}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-getInvalidFields}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-print}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-clone}{\code{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$new(
  concreteType,
  storageLocationId = NULL,
  uploadType = NULL,
  banner = NULL,
  description = NULL,
  etag = NULL,
  createdOn = NULL,
  createdBy = NULL,
  proxyUrl = NULL,
  secretKey = NULL,
  benefactorId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{uploadType}}{uploadType}

\item{\code{banner}}{banner}

\item{\code{description}}{description}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{proxyUrl}}{proxyUrl}

\item{\code{secretKey}}{secretKey}

\item{\code{benefactorId}}{benefactorId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectProxyStorageLocationSettings$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
