% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_project_s3_storage_location_setting.R
\docType{class}
\name{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting}
\alias{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting}
\title{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting}
\format{
An \code{R6Class} generator object
}
\description{
This class refers only to the default Synapse storage location(s) that exist on Amazon S3. To create a storage location that refers to a private S3 bucket, see <a href=\"${org.sagebionetworks.repo.model.project.ExternalS3StorageLocationSetting}\">ExternalS3StorageLocationSetting</a>.

OrgSagebionetworksRepoModelProjectS3StorageLocationSetting Class
}
\details{
Create a new OrgSagebionetworksRepoModelProjectS3StorageLocationSetting
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{storageLocationId}}{integer [optional]}

\item{\code{uploadType}}{character [optional]}

\item{\code{banner}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{integer [optional]}

\item{\code{baseKey}}{character [optional]}

\item{\code{stsEnabled}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-new}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-toJSON}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-fromJSON}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-toJSONString}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-fromJSONString}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-validateJSON}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-toString}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-isValid}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-getInvalidFields}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-print}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-clone}{\code{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProjectS3StorageLocationSetting class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$new(
  concreteType,
  storageLocationId = NULL,
  uploadType = NULL,
  banner = NULL,
  description = NULL,
  etag = NULL,
  createdOn = NULL,
  createdBy = NULL,
  baseKey = NULL,
  stsEnabled = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{storageLocationId}}{storageLocationId}

\item{\code{uploadType}}{uploadType}

\item{\code{banner}}{banner}

\item{\code{description}}{description}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{baseKey}}{baseKey}

\item{\code{stsEnabled}}{stsEnabled}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectS3StorageLocationSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectS3StorageLocationSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectS3StorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectS3StorageLocationSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectS3StorageLocationSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectS3StorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelProjectS3StorageLocationSetting and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelProjectS3StorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectS3StorageLocationSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectS3StorageLocationSetting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
