% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_project_upload_destination_list_setting.R
\docType{class}
\name{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting}
\alias{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting}
\title{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting}
\format{
An \code{R6Class} generator object
}
\description{
This setting type contains the list of upload locations for files in a project. Each id in the locations attribute points to an existing <a href=\"${org.sagebionetworks.repo.model.project.StorageLocationSetting}\">StorageLocationSetting</a>. The maximum number of storage locations that can be stored for a project is 10.

OrgSagebionetworksRepoModelProjectUploadDestinationListSetting Class
}
\details{
Create a new OrgSagebionetworksRepoModelProjectUploadDestinationListSetting
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{id}}{character [optional]}

\item{\code{projectId}}{character [optional]}

\item{\code{settingsType}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{locations}}{List of ids pointing to <a href=\"${org.sagebionetworks.repo.model.project.StorageLocationSetting}\">StorageLocationSetting</a>. The first one in the list is the default location; The maximum number of storage locations for a project is limited to 10. list(integer) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-new}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-toJSON}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-fromJSON}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-toJSONString}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-fromJSONString}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-validateJSON}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-toString}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-isValid}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-getInvalidFields}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-print}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-clone}{\code{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProjectUploadDestinationListSetting class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$new(
  concreteType,
  id = NULL,
  projectId = NULL,
  settingsType = NULL,
  etag = NULL,
  locations = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{id}}{id}

\item{\code{projectId}}{projectId}

\item{\code{settingsType}}{settingsType}

\item{\code{etag}}{etag}

\item{\code{locations}}{List of ids pointing to <a href=\"${org.sagebionetworks.repo.model.project.StorageLocationSetting}\">StorageLocationSetting</a>. The first one in the list is the default location; The maximum number of storage locations for a project is limited to 10.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectUploadDestinationListSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectUploadDestinationListSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectUploadDestinationListSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProjectUploadDestinationListSetting in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProjectUploadDestinationListSetting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProjectUploadDestinationListSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelProjectUploadDestinationListSetting and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelProjectUploadDestinationListSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProjectUploadDestinationListSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProjectUploadDestinationListSetting$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
