% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_provenance_used_entity.R
\docType{class}
\name{OrgSagebionetworksRepoModelProvenanceUsedEntity}
\alias{OrgSagebionetworksRepoModelProvenanceUsedEntity}
\title{OrgSagebionetworksRepoModelProvenanceUsedEntity}
\format{
An \code{R6Class} generator object
}
\description{
Interface for JSON schema for UsedEntity POJO

OrgSagebionetworksRepoModelProvenanceUsedEntity Class
}
\details{
Create a new OrgSagebionetworksRepoModelProvenanceUsedEntity
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{wasExecuted}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{reference}}{\link{OrgSagebionetworksRepoModelReference} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-new}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-toJSON}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-fromJSON}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-toJSONString}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-fromJSONString}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-validateJSON}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-toString}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-isValid}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-getInvalidFields}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-print}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedEntity-clone}{\code{OrgSagebionetworksRepoModelProvenanceUsedEntity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProvenanceUsedEntity class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$new(
  concreteType,
  wasExecuted = NULL,
  reference = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{wasExecuted}}{wasExecuted}

\item{\code{reference}}{reference}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProvenanceUsedEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProvenanceUsedEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProvenanceUsedEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProvenanceUsedEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProvenanceUsedEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProvenanceUsedEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelProvenanceUsedEntity and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelProvenanceUsedEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedEntity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedEntity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedEntity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
