% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_provenance_used_url.R
\docType{class}
\name{OrgSagebionetworksRepoModelProvenanceUsedURL}
\alias{OrgSagebionetworksRepoModelProvenanceUsedURL}
\title{OrgSagebionetworksRepoModelProvenanceUsedURL}
\format{
An \code{R6Class} generator object
}
\description{
Interface for JSON schema for UsedEntity POJO

OrgSagebionetworksRepoModelProvenanceUsedURL Class
}
\details{
Create a new OrgSagebionetworksRepoModelProvenanceUsedURL
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{wasExecuted}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{name}}{character [optional]}

\item{\code{url}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-new}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$new()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-toJSON}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-fromJSON}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-toJSONString}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-fromJSONString}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-validateJSON}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-toString}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-isValid}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-getInvalidFields}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-print}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$print()}}
\item \href{#method-OrgSagebionetworksRepoModelProvenanceUsedURL-clone}{\code{OrgSagebionetworksRepoModelProvenanceUsedURL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelProvenanceUsedURL class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$new(
  concreteType,
  wasExecuted = NULL,
  name = NULL,
  url = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{wasExecuted}}{wasExecuted}

\item{\code{name}}{name}

\item{\code{url}}{url}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProvenanceUsedURL in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProvenanceUsedURL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProvenanceUsedURL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelProvenanceUsedURL in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelProvenanceUsedURL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelProvenanceUsedURL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelProvenanceUsedURL and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelProvenanceUsedURL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelProvenanceUsedURL-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelProvenanceUsedURL-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelProvenanceUsedURL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
