% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_query_query_table_results.R
\docType{class}
\name{OrgSagebionetworksRepoModelQueryQueryTableResults}
\alias{OrgSagebionetworksRepoModelQueryQueryTableResults}
\title{OrgSagebionetworksRepoModelQueryQueryTableResults}
\format{
An \code{R6Class} generator object
}
\description{
An object to store the results of a Synapse Query. Results are stored in row-major tabular format.

OrgSagebionetworksRepoModelQueryQueryTableResults Class
}
\details{
Create a new OrgSagebionetworksRepoModelQueryQueryTableResults
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{headers}}{The ordered list of column headers (attribute names). list(character) [optional]}

\item{\code{rows}}{The ordered list of row data. Each row corresponds to one object. list(\link{OrgSagebionetworksRepoModelQueryRow}) [optional]}

\item{\code{totalNumberOfResults}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-new}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$new()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-toJSON}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-fromJSON}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-toJSONString}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-fromJSONString}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-validateJSON}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-toString}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-isValid}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-getInvalidFields}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-print}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$print()}}
\item \href{#method-OrgSagebionetworksRepoModelQueryQueryTableResults-clone}{\code{OrgSagebionetworksRepoModelQueryQueryTableResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelQueryQueryTableResults class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$new(
  headers = NULL,
  rows = NULL,
  totalNumberOfResults = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{headers}}{The ordered list of column headers (attribute names).}

\item{\code{rows}}{The ordered list of row data. Each row corresponds to one object.}

\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQueryQueryTableResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQueryQueryTableResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQueryQueryTableResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQueryQueryTableResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQueryQueryTableResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQueryQueryTableResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelQueryQueryTableResults and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelQueryQueryTableResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQueryQueryTableResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQueryQueryTableResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQueryQueryTableResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
