% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_quiz_multichoice_answer.R
\docType{class}
\name{OrgSagebionetworksRepoModelQuizMultichoiceAnswer}
\alias{OrgSagebionetworksRepoModelQuizMultichoiceAnswer}
\title{OrgSagebionetworksRepoModelQuizMultichoiceAnswer}
\format{
An \code{R6Class} generator object
}
\description{
One of the multiple choices in a multiple choice question

OrgSagebionetworksRepoModelQuizMultichoiceAnswer Class
}
\details{
Create a new OrgSagebionetworksRepoModelQuizMultichoiceAnswer
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{answerIndex}}{integer [optional]}

\item{\code{prompt}}{character [optional]}

\item{\code{isCorrect}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-new}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$new()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-toJSON}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-fromJSON}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-toJSONString}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-fromJSONString}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-validateJSON}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-toString}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-isValid}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-getInvalidFields}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-print}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$print()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-clone}{\code{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelQuizMultichoiceAnswer class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$new(
  answerIndex = NULL,
  prompt = NULL,
  isCorrect = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{answerIndex}}{answerIndex}

\item{\code{prompt}}{prompt}

\item{\code{isCorrect}}{isCorrect}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizMultichoiceAnswer in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizMultichoiceAnswer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizMultichoiceAnswer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizMultichoiceAnswer in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizMultichoiceAnswer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizMultichoiceAnswer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelQuizMultichoiceAnswer and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelQuizMultichoiceAnswer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceAnswer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceAnswer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
