% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_quiz_multichoice_question.R
\docType{class}
\name{OrgSagebionetworksRepoModelQuizMultichoiceQuestion}
\alias{OrgSagebionetworksRepoModelQuizMultichoiceQuestion}
\title{OrgSagebionetworksRepoModelQuizMultichoiceQuestion}
\format{
An \code{R6Class} generator object
}
\description{
A multiple choice question

OrgSagebionetworksRepoModelQuizMultichoiceQuestion Class
}
\details{
Create a new OrgSagebionetworksRepoModelQuizMultichoiceQuestion
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{questionIndex}}{integer [optional]}

\item{\code{prompt}}{character [optional]}

\item{\code{reference}}{\link{OrgSagebionetworksRepoModelDaoWikiPageKey} [optional]}

\item{\code{docLink}}{character [optional]}

\item{\code{helpText}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{exclusive}}{character [optional]}

\item{\code{answers}}{The possible answers for this question list(\link{OrgSagebionetworksRepoModelQuizMultichoiceAnswer}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-new}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$new()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-toJSON}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-fromJSON}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-toJSONString}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-fromJSONString}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-validateJSON}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-toString}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-isValid}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-getInvalidFields}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-print}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$print()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-clone}{\code{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelQuizMultichoiceQuestion class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$new(
  concreteType,
  questionIndex = NULL,
  prompt = NULL,
  reference = NULL,
  docLink = NULL,
  helpText = NULL,
  exclusive = NULL,
  answers = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{questionIndex}}{questionIndex}

\item{\code{prompt}}{prompt}

\item{\code{reference}}{reference}

\item{\code{docLink}}{docLink}

\item{\code{helpText}}{helpText}

\item{\code{exclusive}}{exclusive}

\item{\code{answers}}{The possible answers for this question}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizMultichoiceQuestion in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizMultichoiceQuestion
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizMultichoiceQuestion
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizMultichoiceQuestion in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizMultichoiceQuestion
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizMultichoiceQuestion
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelQuizMultichoiceQuestion and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelQuizMultichoiceQuestion
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizMultichoiceQuestion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizMultichoiceQuestion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
