% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_quiz_quiz_generator.R
\docType{class}
\name{OrgSagebionetworksRepoModelQuizQuizGenerator}
\alias{OrgSagebionetworksRepoModelQuizQuizGenerator}
\title{OrgSagebionetworksRepoModelQuizQuizGenerator}
\format{
An \code{R6Class} generator object
}
\description{
Schema for a quiz and, optionally, its answer key

OrgSagebionetworksRepoModelQuizQuizGenerator Class
}
\details{
Create a new OrgSagebionetworksRepoModelQuizQuizGenerator
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{header}}{character [optional]}

\item{\code{minimumScore}}{integer [optional]}

\item{\code{questions}}{The questions in the Quiz list(\link{OrgSagebionetworksRepoModelQuizQuestionVariety}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-new}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$new()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-toJSON}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-fromJSON}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-toJSONString}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-fromJSONString}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-validateJSON}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-toString}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-isValid}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-getInvalidFields}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-print}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$print()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizGenerator-clone}{\code{OrgSagebionetworksRepoModelQuizQuizGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelQuizQuizGenerator class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$new(
  id = NULL,
  header = NULL,
  minimumScore = NULL,
  questions = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{header}}{header}

\item{\code{minimumScore}}{minimumScore}

\item{\code{questions}}{The questions in the Quiz}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizQuizGenerator in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizQuizGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizQuizGenerator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizQuizGenerator in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizQuizGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizQuizGenerator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelQuizQuizGenerator and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelQuizQuizGenerator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
