% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_quiz_quiz_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelQuizQuizResponse}
\alias{OrgSagebionetworksRepoModelQuizQuizResponse}
\title{OrgSagebionetworksRepoModelQuizQuizResponse}
\format{
An \code{R6Class} generator object
}
\description{
The set of responses to a Quiz

OrgSagebionetworksRepoModelQuizQuizResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelQuizQuizResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{quizId}}{integer [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{questionResponses}}{The list of responses to the questions in the Quiz list(\link{OrgSagebionetworksRepoModelQuizQuestionResponse}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-new}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-toJSON}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-fromJSON}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-toJSONString}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-validateJSON}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-toString}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-isValid}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-print}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizQuizResponse-clone}{\code{OrgSagebionetworksRepoModelQuizQuizResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelQuizQuizResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$new(
  id = NULL,
  quizId = NULL,
  createdBy = NULL,
  createdOn = NULL,
  questionResponses = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{quizId}}{quizId}

\item{\code{createdBy}}{createdBy}

\item{\code{createdOn}}{createdOn}

\item{\code{questionResponses}}{The list of responses to the questions in the Quiz}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizQuizResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizQuizResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizQuizResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizQuizResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizQuizResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizQuizResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelQuizQuizResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelQuizQuizResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizQuizResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizQuizResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizQuizResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
