% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_quiz_response_correctness.R
\docType{class}
\name{OrgSagebionetworksRepoModelQuizResponseCorrectness}
\alias{OrgSagebionetworksRepoModelQuizResponseCorrectness}
\title{OrgSagebionetworksRepoModelQuizResponseCorrectness}
\format{
An \code{R6Class} generator object
}
\description{
A user response and whether it is correct.

OrgSagebionetworksRepoModelQuizResponseCorrectness Class
}
\details{
Create a new OrgSagebionetworksRepoModelQuizResponseCorrectness
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{question}}{\link{OrgSagebionetworksRepoModelQuizQuestion} [optional]}

\item{\code{response}}{\link{OrgSagebionetworksRepoModelQuizQuestionResponse} [optional]}

\item{\code{isCorrect}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-new}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$new()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-toJSON}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-fromJSON}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-toJSONString}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-fromJSONString}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-validateJSON}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-toString}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-isValid}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-getInvalidFields}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-print}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$print()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizResponseCorrectness-clone}{\code{OrgSagebionetworksRepoModelQuizResponseCorrectness$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelQuizResponseCorrectness class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$new(
  question = NULL,
  response = NULL,
  isCorrect = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{question}}{question}

\item{\code{response}}{response}

\item{\code{isCorrect}}{isCorrect}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizResponseCorrectness in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizResponseCorrectness
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizResponseCorrectness
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizResponseCorrectness in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizResponseCorrectness
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizResponseCorrectness
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelQuizResponseCorrectness and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelQuizResponseCorrectness
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizResponseCorrectness-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizResponseCorrectness-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizResponseCorrectness$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
