% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_quiz_text_field_question.R
\docType{class}
\name{OrgSagebionetworksRepoModelQuizTextFieldQuestion}
\alias{OrgSagebionetworksRepoModelQuizTextFieldQuestion}
\title{OrgSagebionetworksRepoModelQuizTextFieldQuestion}
\format{
An \code{R6Class} generator object
}
\description{
A question whose response is an (unconstrained) text string

OrgSagebionetworksRepoModelQuizTextFieldQuestion Class
}
\details{
Create a new OrgSagebionetworksRepoModelQuizTextFieldQuestion
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{questionIndex}}{integer [optional]}

\item{\code{prompt}}{character [optional]}

\item{\code{reference}}{\link{OrgSagebionetworksRepoModelDaoWikiPageKey} [optional]}

\item{\code{docLink}}{character [optional]}

\item{\code{helpText}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{answer}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-new}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$new()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-toJSON}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-fromJSON}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-toJSONString}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-fromJSONString}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-validateJSON}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-toString}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-isValid}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-getInvalidFields}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-print}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$print()}}
\item \href{#method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-clone}{\code{OrgSagebionetworksRepoModelQuizTextFieldQuestion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelQuizTextFieldQuestion class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$new(
  concreteType,
  questionIndex = NULL,
  prompt = NULL,
  reference = NULL,
  docLink = NULL,
  helpText = NULL,
  answer = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{questionIndex}}{questionIndex}

\item{\code{prompt}}{prompt}

\item{\code{reference}}{reference}

\item{\code{docLink}}{docLink}

\item{\code{helpText}}{helpText}

\item{\code{answer}}{answer}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizTextFieldQuestion in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizTextFieldQuestion
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizTextFieldQuestion
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelQuizTextFieldQuestion in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelQuizTextFieldQuestion
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelQuizTextFieldQuestion
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelQuizTextFieldQuestion and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelQuizTextFieldQuestion
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelQuizTextFieldQuestion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelQuizTextFieldQuestion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
