% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_registry_entity_migration.R
\docType{class}
\name{OrgSagebionetworksRepoModelRegistryEntityMigration}
\alias{OrgSagebionetworksRepoModelRegistryEntityMigration}
\title{OrgSagebionetworksRepoModelRegistryEntityMigration}
\format{
An \code{R6Class} generator object
}
\description{
Metadata about migrating entities

OrgSagebionetworksRepoModelRegistryEntityMigration Class
}
\details{
Create a new OrgSagebionetworksRepoModelRegistryEntityMigration
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{toRename}}{The full path of the entity schema. list(\link{OrgSagebionetworksRepoModelRegistryRenameData}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-new}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-toJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-fromJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-toJSONString}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-fromJSONString}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-validateJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-toString}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-isValid}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-getInvalidFields}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-print}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityMigration-clone}{\code{OrgSagebionetworksRepoModelRegistryEntityMigration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRegistryEntityMigration class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$new(toRename = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{toRename}}{The full path of the entity schema.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryEntityMigration in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryEntityMigration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryEntityMigration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryEntityMigration in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryEntityMigration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryEntityMigration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRegistryEntityMigration and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRegistryEntityMigration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityMigration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityMigration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityMigration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
