% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_registry_entity_registry.R
\docType{class}
\name{OrgSagebionetworksRepoModelRegistryEntityRegistry}
\alias{OrgSagebionetworksRepoModelRegistryEntityRegistry}
\title{OrgSagebionetworksRepoModelRegistryEntityRegistry}
\format{
An \code{R6Class} generator object
}
\description{
Metadata about entities

OrgSagebionetworksRepoModelRegistryEntityRegistry Class
}
\details{
Create a new OrgSagebionetworksRepoModelRegistryEntityRegistry
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entityTypes}}{The full path of the entity schema. list(\link{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-new}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-toJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-fromJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-toJSONString}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-fromJSONString}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-validateJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-toString}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-isValid}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-getInvalidFields}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-print}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityRegistry-clone}{\code{OrgSagebionetworksRepoModelRegistryEntityRegistry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRegistryEntityRegistry class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$new(entityTypes = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entityTypes}}{The full path of the entity schema.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryEntityRegistry in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryEntityRegistry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryEntityRegistry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryEntityRegistry in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryEntityRegistry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryEntityRegistry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRegistryEntityRegistry and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRegistryEntityRegistry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityRegistry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityRegistry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityRegistry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
