% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_registry_entity_type_metadata.R
\docType{class}
\name{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata}
\alias{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata}
\title{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata}
\format{
An \code{R6Class} generator object
}
\description{
Metadata about entities

OrgSagebionetworksRepoModelRegistryEntityTypeMetadata Class
}
\details{
Create a new OrgSagebionetworksRepoModelRegistryEntityTypeMetadata
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character}

\item{\code{displayName}}{character [optional]}

\item{\code{className}}{character [optional]}

\item{\code{entityType}}{character [optional]}

\item{\code{defaultParentPath}}{character [optional]}

\item{\code{validParentTypes}}{The full path of the entity schema. list(character) [optional]}

\item{\code{aliases}}{The list of aliases that can be used to look an entity type with a query. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-new}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-toJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-fromJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-toJSONString}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-fromJSONString}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-validateJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-toString}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-isValid}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-getInvalidFields}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-print}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-clone}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRegistryEntityTypeMetadata class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$new(
  name,
  displayName = NULL,
  className = NULL,
  entityType = NULL,
  defaultParentPath = NULL,
  validParentTypes = NULL,
  aliases = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{displayName}}{displayName}

\item{\code{className}}{className}

\item{\code{entityType}}{entityType}

\item{\code{defaultParentPath}}{defaultParentPath}

\item{\code{validParentTypes}}{The full path of the entity schema.}

\item{\code{aliases}}{The list of aliases that can be used to look an entity type with a query.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryEntityTypeMetadata in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryEntityTypeMetadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryEntityTypeMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryEntityTypeMetadata in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryEntityTypeMetadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryEntityTypeMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRegistryEntityTypeMetadata and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRegistryEntityTypeMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
