% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_registry_entity_type_migration_spec.R
\docType{class}
\name{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec}
\alias{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec}
\title{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec}
\format{
An \code{R6Class} generator object
}
\description{
Metadata to describe the migration of an entity type

OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec Class
}
\details{
Create a new OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entityType}}{character [optional]}

\item{\code{fields}}{list(\link{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-new}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-toJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-fromJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-toJSONString}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-fromJSONString}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-validateJSON}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-toString}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-isValid}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-getInvalidFields}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-print}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-clone}{\code{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$new(
  entityType = NULL,
  fields = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entityType}}{entityType}

\item{\code{fields}}{fields}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryEntityTypeMigrationSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
