% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_registry_field_description.R
\docType{class}
\name{OrgSagebionetworksRepoModelRegistryFieldDescription}
\alias{OrgSagebionetworksRepoModelRegistryFieldDescription}
\title{OrgSagebionetworksRepoModelRegistryFieldDescription}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelRegistryFieldDescription Class

OrgSagebionetworksRepoModelRegistryFieldDescription Class
}
\details{
Create a new OrgSagebionetworksRepoModelRegistryFieldDescription
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{type}}{character [optional]}

\item{\code{bucket}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-new}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-toJSON}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-fromJSON}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-toJSONString}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-fromJSONString}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-validateJSON}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-toString}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-isValid}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-getInvalidFields}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-print}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldDescription-clone}{\code{OrgSagebionetworksRepoModelRegistryFieldDescription$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRegistryFieldDescription class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$new(
  name = NULL,
  type = NULL,
  bucket = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{type}}{type}

\item{\code{bucket}}{bucket}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryFieldDescription in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryFieldDescription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryFieldDescription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryFieldDescription in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryFieldDescription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryFieldDescription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRegistryFieldDescription and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRegistryFieldDescription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldDescription-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldDescription-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldDescription$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
