% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_registry_field_migration_spec.R
\docType{class}
\name{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec}
\alias{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec}
\title{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec}
\format{
An \code{R6Class} generator object
}
\description{
Metadata to describe the migration of a generic field

OrgSagebionetworksRepoModelRegistryFieldMigrationSpec Class
}
\details{
Create a new OrgSagebionetworksRepoModelRegistryFieldMigrationSpec
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{\link{OrgSagebionetworksRepoModelRegistryFieldDescription} [optional]}

\item{\code{destination}}{\link{OrgSagebionetworksRepoModelRegistryFieldDescription} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-new}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-toJSON}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-fromJSON}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-toJSONString}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-fromJSONString}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-validateJSON}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-toString}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-isValid}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-getInvalidFields}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-print}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-clone}{\code{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRegistryFieldMigrationSpec class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$new(
  source = NULL,
  destination = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{source}

\item{\code{destination}}{destination}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryFieldMigrationSpec in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryFieldMigrationSpec
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryFieldMigrationSpec
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRegistryFieldMigrationSpec in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRegistryFieldMigrationSpec
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRegistryFieldMigrationSpec
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRegistryFieldMigrationSpec and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRegistryFieldMigrationSpec
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRegistryFieldMigrationSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRegistryFieldMigrationSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
