% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_report_download_storage_report_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest}
\alias{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest}
\title{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest}
\format{
An \code{R6Class} generator object
}
\description{
A Storage Report CSV download request.

OrgSagebionetworksRepoModelReportDownloadStorageReportRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelReportDownloadStorageReportRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{reportType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-new}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-toJSON}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-fromJSON}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-toJSONString}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-validateJSON}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-toString}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-isValid}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-print}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-clone}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelReportDownloadStorageReportRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$new(
  concreteType,
  reportType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{reportType}}{reportType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelReportDownloadStorageReportRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelReportDownloadStorageReportRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelReportDownloadStorageReportRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelReportDownloadStorageReportRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelReportDownloadStorageReportRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelReportDownloadStorageReportRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelReportDownloadStorageReportRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelReportDownloadStorageReportRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
