% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_report_download_storage_report_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse}
\alias{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse}
\title{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse}
\format{
An \code{R6Class} generator object
}
\description{
The response body of an asynchronous storage report CSV download job.

OrgSagebionetworksRepoModelReportDownloadStorageReportResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelReportDownloadStorageReportResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{resultsFileHandleId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-new}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-toJSON}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-fromJSON}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-toJSONString}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-validateJSON}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-toString}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-isValid}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-print}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-clone}{\code{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelReportDownloadStorageReportResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$new(
  concreteType,
  resultsFileHandleId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{resultsFileHandleId}}{resultsFileHandleId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelReportDownloadStorageReportResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelReportDownloadStorageReportResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelReportDownloadStorageReportResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelReportDownloadStorageReportResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelReportDownloadStorageReportResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelReportDownloadStorageReportResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelReportDownloadStorageReportResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelReportDownloadStorageReportResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelReportDownloadStorageReportResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelReportDownloadStorageReportResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
