% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_request_reference_list.R
\docType{class}
\name{OrgSagebionetworksRepoModelRequestReferenceList}
\alias{OrgSagebionetworksRepoModelRequestReferenceList}
\title{OrgSagebionetworksRepoModelRequestReferenceList}
\format{
An \code{R6Class} generator object
}
\description{
An array of Reference objects

OrgSagebionetworksRepoModelRequestReferenceList Class
}
\details{
Create a new OrgSagebionetworksRepoModelRequestReferenceList
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{references}}{The list of references list(\link{OrgSagebionetworksRepoModelReference}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-new}{\code{OrgSagebionetworksRepoModelRequestReferenceList$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-toJSON}{\code{OrgSagebionetworksRepoModelRequestReferenceList$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-fromJSON}{\code{OrgSagebionetworksRepoModelRequestReferenceList$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-toJSONString}{\code{OrgSagebionetworksRepoModelRequestReferenceList$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-fromJSONString}{\code{OrgSagebionetworksRepoModelRequestReferenceList$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-validateJSON}{\code{OrgSagebionetworksRepoModelRequestReferenceList$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-toString}{\code{OrgSagebionetworksRepoModelRequestReferenceList$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-isValid}{\code{OrgSagebionetworksRepoModelRequestReferenceList$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-getInvalidFields}{\code{OrgSagebionetworksRepoModelRequestReferenceList$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-print}{\code{OrgSagebionetworksRepoModelRequestReferenceList$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRequestReferenceList-clone}{\code{OrgSagebionetworksRepoModelRequestReferenceList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRequestReferenceList class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$new(references = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{references}}{The list of references}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRequestReferenceList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRequestReferenceList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRequestReferenceList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRequestReferenceList in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRequestReferenceList
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRequestReferenceList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRequestReferenceList and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRequestReferenceList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRequestReferenceList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRequestReferenceList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRequestReferenceList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
