% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_resource_access.R
\docType{class}
\name{OrgSagebionetworksRepoModelResourceAccess}
\alias{OrgSagebionetworksRepoModelResourceAccess}
\title{OrgSagebionetworksRepoModelResourceAccess}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for a ResourceAccess POJO

OrgSagebionetworksRepoModelResourceAccess Class
}
\details{
Create a new OrgSagebionetworksRepoModelResourceAccess
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{principalId}}{integer [optional]}

\item{\code{accessType}}{A list of types of access granted to the given principal list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-new}{\code{OrgSagebionetworksRepoModelResourceAccess$new()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-toJSON}{\code{OrgSagebionetworksRepoModelResourceAccess$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-fromJSON}{\code{OrgSagebionetworksRepoModelResourceAccess$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-toJSONString}{\code{OrgSagebionetworksRepoModelResourceAccess$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-fromJSONString}{\code{OrgSagebionetworksRepoModelResourceAccess$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-validateJSON}{\code{OrgSagebionetworksRepoModelResourceAccess$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-toString}{\code{OrgSagebionetworksRepoModelResourceAccess$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-isValid}{\code{OrgSagebionetworksRepoModelResourceAccess$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-getInvalidFields}{\code{OrgSagebionetworksRepoModelResourceAccess$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-print}{\code{OrgSagebionetworksRepoModelResourceAccess$print()}}
\item \href{#method-OrgSagebionetworksRepoModelResourceAccess-clone}{\code{OrgSagebionetworksRepoModelResourceAccess$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelResourceAccess class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$new(
  principalId = NULL,
  accessType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{principalId}}{principalId}

\item{\code{accessType}}{A list of types of access granted to the given principal}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelResourceAccess in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelResourceAccess
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelResourceAccess
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelResourceAccess in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelResourceAccess
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelResourceAccess
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelResourceAccess and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelResourceAccess
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelResourceAccess-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelResourceAccess-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelResourceAccess$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
