% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_restrictable_object_descriptor.R
\docType{class}
\name{OrgSagebionetworksRepoModelRestrictableObjectDescriptor}
\alias{OrgSagebionetworksRepoModelRestrictableObjectDescriptor}
\title{OrgSagebionetworksRepoModelRestrictableObjectDescriptor}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelRestrictableObjectDescriptor Class

OrgSagebionetworksRepoModelRestrictableObjectDescriptor Class
}
\details{
Create a new OrgSagebionetworksRepoModelRestrictableObjectDescriptor
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{type}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-new}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-toJSON}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-fromJSON}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-toJSONString}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-fromJSONString}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-validateJSON}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-toString}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-isValid}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-getInvalidFields}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-print}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-clone}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRestrictableObjectDescriptor class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$new(
  id = NULL,
  type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{type}}{type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictableObjectDescriptor in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictableObjectDescriptor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictableObjectDescriptor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictableObjectDescriptor in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictableObjectDescriptor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictableObjectDescriptor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRestrictableObjectDescriptor and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRestrictableObjectDescriptor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
