% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_restrictable_object_descriptor_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse}
\alias{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse}
\title{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse Class

OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subjects}}{a page of subjects list(\link{OrgSagebionetworksRepoModelRestrictableObjectDescriptor}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-new}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-toJSON}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-fromJSON}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-toJSONString}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-validateJSON}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-toString}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-isValid}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-print}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-clone}{\code{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$new(
  subjects = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subjects}}{a page of subjects}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictableObjectDescriptorResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
