% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_restriction_information_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelRestrictionInformationRequest}
\alias{OrgSagebionetworksRepoModelRestrictionInformationRequest}
\title{OrgSagebionetworksRepoModelRestrictionInformationRequest}
\format{
An \code{R6Class} generator object
}
\description{
A request to retrieve the information about restriction level on a restrict-able object.

OrgSagebionetworksRepoModelRestrictionInformationRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelRestrictionInformationRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{restrictableObjectType}}{character [optional]}

\item{\code{objectId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-new}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-toJSON}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-fromJSON}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-toJSONString}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-validateJSON}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-toString}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-isValid}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-print}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationRequest-clone}{\code{OrgSagebionetworksRepoModelRestrictionInformationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRestrictionInformationRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$new(
  restrictableObjectType = NULL,
  objectId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{restrictableObjectType}}{restrictableObjectType}

\item{\code{objectId}}{objectId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictionInformationRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictionInformationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictionInformationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictionInformationRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictionInformationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictionInformationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRestrictionInformationRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRestrictionInformationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
