% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_restriction_information_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelRestrictionInformationResponse}
\alias{OrgSagebionetworksRepoModelRestrictionInformationResponse}
\title{OrgSagebionetworksRepoModelRestrictionInformationResponse}
\format{
An \code{R6Class} generator object
}
\description{
The information about restriction level on a restrict-able object.

OrgSagebionetworksRepoModelRestrictionInformationResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelRestrictionInformationResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{restrictionLevel}}{character [optional]}

\item{\code{hasUnmetAccessRequirement}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-new}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-toJSON}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-fromJSON}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-toJSONString}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-validateJSON}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-toString}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-isValid}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-print}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionInformationResponse-clone}{\code{OrgSagebionetworksRepoModelRestrictionInformationResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRestrictionInformationResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$new(
  restrictionLevel = NULL,
  hasUnmetAccessRequirement = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{restrictionLevel}}{restrictionLevel}

\item{\code{hasUnmetAccessRequirement}}{hasUnmetAccessRequirement}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictionInformationResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictionInformationResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictionInformationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictionInformationResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictionInformationResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictionInformationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRestrictionInformationResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRestrictionInformationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionInformationResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionInformationResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionInformationResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
