% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_restriction_level.R
\docType{class}
\name{OrgSagebionetworksRepoModelRestrictionLevel}
\alias{OrgSagebionetworksRepoModelRestrictionLevel}
\title{OrgSagebionetworksRepoModelRestrictionLevel}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelRestrictionLevel Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRestrictionLevel-new}{\code{OrgSagebionetworksRepoModelRestrictionLevel$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionLevel-toJSON}{\code{OrgSagebionetworksRepoModelRestrictionLevel$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionLevel-fromJSON}{\code{OrgSagebionetworksRepoModelRestrictionLevel$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionLevel-toJSONString}{\code{OrgSagebionetworksRepoModelRestrictionLevel$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionLevel-fromJSONString}{\code{OrgSagebionetworksRepoModelRestrictionLevel$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRestrictionLevel-clone}{\code{OrgSagebionetworksRepoModelRestrictionLevel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionLevel-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionLevel-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRestrictionLevel class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionLevel$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionLevel-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionLevel-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionLevel$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictionLevel in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionLevel-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionLevel-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictionLevel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionLevel$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictionLevel
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionLevel-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionLevel-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionLevel$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRestrictionLevel in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionLevel-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionLevel-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRestrictionLevel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionLevel$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRestrictionLevel
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRestrictionLevel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRestrictionLevel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRestrictionLevel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
