% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_rss_entry.R
\docType{class}
\name{OrgSagebionetworksRepoModelRSSEntry}
\alias{OrgSagebionetworksRepoModelRSSEntry}
\title{OrgSagebionetworksRepoModelRSSEntry}
\format{
An \code{R6Class} generator object
}
\description{
This contains items from an rss feed, and meta data about the rss feed itself

OrgSagebionetworksRepoModelRSSEntry Class
}
\details{
Create a new OrgSagebionetworksRepoModelRSSEntry
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{character [optional]}

\item{\code{date}}{character [optional]}

\item{\code{author}}{character [optional]}

\item{\code{content}}{character [optional]}

\item{\code{link}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-new}{\code{OrgSagebionetworksRepoModelRSSEntry$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-toJSON}{\code{OrgSagebionetworksRepoModelRSSEntry$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-fromJSON}{\code{OrgSagebionetworksRepoModelRSSEntry$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-toJSONString}{\code{OrgSagebionetworksRepoModelRSSEntry$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-fromJSONString}{\code{OrgSagebionetworksRepoModelRSSEntry$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-validateJSON}{\code{OrgSagebionetworksRepoModelRSSEntry$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-toString}{\code{OrgSagebionetworksRepoModelRSSEntry$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-isValid}{\code{OrgSagebionetworksRepoModelRSSEntry$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-getInvalidFields}{\code{OrgSagebionetworksRepoModelRSSEntry$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-print}{\code{OrgSagebionetworksRepoModelRSSEntry$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSEntry-clone}{\code{OrgSagebionetworksRepoModelRSSEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRSSEntry class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$new(
  title = NULL,
  date = NULL,
  author = NULL,
  content = NULL,
  link = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{date}}{date}

\item{\code{author}}{author}

\item{\code{content}}{content}

\item{\code{link}}{link}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRSSEntry in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRSSEntry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRSSEntry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRSSEntry in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRSSEntry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRSSEntry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRSSEntry and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRSSEntry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
