% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_rss_feed.R
\docType{class}
\name{OrgSagebionetworksRepoModelRSSFeed}
\alias{OrgSagebionetworksRepoModelRSSFeed}
\title{OrgSagebionetworksRepoModelRSSFeed}
\format{
An \code{R6Class} generator object
}
\description{
This contains items from an rss feed, and meta data about the rss feed itself

OrgSagebionetworksRepoModelRSSFeed Class
}
\details{
Create a new OrgSagebionetworksRepoModelRSSFeed
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{character [optional]}

\item{\code{uri}}{character [optional]}

\item{\code{author}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{entries}}{The list of rss entries. list(\link{OrgSagebionetworksRepoModelRSSEntry}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-new}{\code{OrgSagebionetworksRepoModelRSSFeed$new()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-toJSON}{\code{OrgSagebionetworksRepoModelRSSFeed$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-fromJSON}{\code{OrgSagebionetworksRepoModelRSSFeed$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-toJSONString}{\code{OrgSagebionetworksRepoModelRSSFeed$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-fromJSONString}{\code{OrgSagebionetworksRepoModelRSSFeed$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-validateJSON}{\code{OrgSagebionetworksRepoModelRSSFeed$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-toString}{\code{OrgSagebionetworksRepoModelRSSFeed$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-isValid}{\code{OrgSagebionetworksRepoModelRSSFeed$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-getInvalidFields}{\code{OrgSagebionetworksRepoModelRSSFeed$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-print}{\code{OrgSagebionetworksRepoModelRSSFeed$print()}}
\item \href{#method-OrgSagebionetworksRepoModelRSSFeed-clone}{\code{OrgSagebionetworksRepoModelRSSFeed$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelRSSFeed class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$new(
  title = NULL,
  uri = NULL,
  author = NULL,
  description = NULL,
  entries = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{uri}}{uri}

\item{\code{author}}{author}

\item{\code{description}}{description}

\item{\code{entries}}{The list of rss entries.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRSSFeed in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRSSFeed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRSSFeed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelRSSFeed in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelRSSFeed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelRSSFeed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelRSSFeed and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelRSSFeed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelRSSFeed-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelRSSFeed-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelRSSFeed$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
