% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_schema_create_schema_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse}
\alias{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse}
\title{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse}
\format{
An \code{R6Class} generator object
}
\description{
The response of a create JsonSchema request.

OrgSagebionetworksRepoModelSchemaCreateSchemaResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelSchemaCreateSchemaResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{newVersionInfo}}{\link{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo} [optional]}

\item{\code{validationSchema}}{\link{OrgSagebionetworksRepoModelSchemaJsonSchema} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-new}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-toJSON}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-fromJSON}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-toJSONString}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-validateJSON}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-toString}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-isValid}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-print}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-clone}{\code{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSchemaCreateSchemaResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$new(
  concreteType,
  newVersionInfo = NULL,
  validationSchema = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{newVersionInfo}}{newVersionInfo}

\item{\code{validationSchema}}{validationSchema}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaCreateSchemaResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaCreateSchemaResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaCreateSchemaResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaCreateSchemaResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaCreateSchemaResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaCreateSchemaResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSchemaCreateSchemaResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSchemaCreateSchemaResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaCreateSchemaResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaCreateSchemaResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
