% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_schema_get_validation_schema_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest}
\alias{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest}
\title{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest}
\format{
An \code{R6Class} generator object
}
\description{
To use a JSON schema for validation, the JSON schema plus all of its dependency schemas must be provided to a JSON schema validator.  The 'validation' schema is simply a JSON schema with all of its dependencies added to the 'definitions' section of the schema, making the schema self-contained.

OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{$id}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-new}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-toJSON}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-fromJSON}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-toJSONString}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-validateJSON}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-toString}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-isValid}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-print}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-clone}{\code{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$new(
  concreteType,
  `$id` = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{$id}}{$id}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaGetValidationSchemaRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
