% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_schema_json_schema_info.R
\docType{class}
\name{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo}
\alias{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo}
\title{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo}
\format{
An \code{R6Class} generator object
}
\description{
Information about a single JSON schema.

OrgSagebionetworksRepoModelSchemaJsonSchemaInfo Class
}
\details{
Create a new OrgSagebionetworksRepoModelSchemaJsonSchemaInfo
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{organizationId}}{character [optional]}

\item{\code{organizationName}}{character [optional]}

\item{\code{schemaId}}{character [optional]}

\item{\code{schemaName}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-new}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-toJSON}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-fromJSON}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-toJSONString}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-fromJSONString}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-validateJSON}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-toString}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-isValid}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-getInvalidFields}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-print}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-clone}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSchemaJsonSchemaInfo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$new(
  organizationId = NULL,
  organizationName = NULL,
  schemaId = NULL,
  schemaName = NULL,
  createdOn = NULL,
  createdBy = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{organizationId}}{organizationId}

\item{\code{organizationName}}{organizationName}

\item{\code{schemaId}}{schemaId}

\item{\code{schemaName}}{schemaName}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaJsonSchemaInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaJsonSchemaInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaJsonSchemaInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaJsonSchemaInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaJsonSchemaInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaJsonSchemaInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSchemaJsonSchemaInfo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSchemaJsonSchemaInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
