% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_schema_json_schema_version_info.R
\docType{class}
\name{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo}
\alias{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo}
\title{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo}
\format{
An \code{R6Class} generator object
}
\description{
Information about a single version of a JSON schema.

OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo Class
}
\details{
Create a new OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{organizationId}}{character [optional]}

\item{\code{organizationName}}{character [optional]}

\item{\code{schemaId}}{character [optional]}

\item{\code{schemaName}}{character [optional]}

\item{\code{versionId}}{character [optional]}

\item{\code{$id}}{character [optional]}

\item{\code{semanticVersion}}{character [optional]}

\item{\code{jsonSHA256Hex}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-new}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-toJSON}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-fromJSON}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-toJSONString}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-fromJSONString}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-validateJSON}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-toString}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-isValid}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-getInvalidFields}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-print}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-clone}{\code{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$new(
  organizationId = NULL,
  organizationName = NULL,
  schemaId = NULL,
  schemaName = NULL,
  versionId = NULL,
  `$id` = NULL,
  semanticVersion = NULL,
  jsonSHA256Hex = NULL,
  createdOn = NULL,
  createdBy = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{organizationId}}{organizationId}

\item{\code{organizationName}}{organizationName}

\item{\code{schemaId}}{schemaId}

\item{\code{schemaName}}{schemaName}

\item{\code{versionId}}{versionId}

\item{\code{$id}}{$id}

\item{\code{semanticVersion}}{semanticVersion}

\item{\code{jsonSHA256Hex}}{jsonSHA256Hex}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
