% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_schema_list_json_schema_info_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest}
\alias{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest}
\title{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to list a single page of JsonSchemaInfo for the given organization name.

OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{organizationName}}{character [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-new}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-toJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-fromJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-toJSONString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-validateJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-toString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-isValid}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-print}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-clone}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$new(
  organizationName = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{organizationName}}{organizationName}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
