% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_schema_list_json_schema_info_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse}
\alias{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse}
\title{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse}
\format{
An \code{R6Class} generator object
}
\description{
A single page of the currently registered Organizations

OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{page}}{list(\link{OrgSagebionetworksRepoModelSchemaJsonSchemaInfo}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-new}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-toJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-fromJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-toJSONString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-validateJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-toString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-isValid}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-print}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-clone}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$new(
  page = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{page}}{page}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaInfoResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
