% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_schema_list_json_schema_version_info_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest}
\alias{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest}
\title{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request to get a single page of schema version metadata for the given organization and schema name.

OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{organizationName}}{character [optional]}

\item{\code{schemaName}}{character [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-new}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-toJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-fromJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-toJSONString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-validateJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-toString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-isValid}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-print}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-clone}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$new(
  organizationName = NULL,
  schemaName = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{organizationName}}{organizationName}

\item{\code{schemaName}}{schemaName}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
