% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_schema_list_json_schema_version_info_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse}
\alias{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse}
\title{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse}
\format{
An \code{R6Class} generator object
}
\description{
A single page of schema version metadata for the provided organization and schema name

OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{page}}{list(\link{OrgSagebionetworksRepoModelSchemaJsonSchemaVersionInfo}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-new}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-toJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-fromJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-toJSONString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-validateJSON}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-toString}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-isValid}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-print}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-clone}{\code{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$new(
  page = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{page}}{page}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaListJsonSchemaVersionInfoResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
