% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_schema_validation_summary_statistics.R
\docType{class}
\name{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics}
\alias{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics}
\title{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics}
\format{
An \code{R6Class} generator object
}
\description{
Summary statistics for the JSON schema validation results for the children of an Entity container (Project or Folder)

OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics Class
}
\details{
Create a new OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{containerId}}{character [optional]}

\item{\code{totalNumberOfChildren}}{integer [optional]}

\item{\code{numberOfValidChildren}}{integer [optional]}

\item{\code{numberOfInvalidChildren}}{integer [optional]}

\item{\code{numberOfUnknownChildren}}{integer [optional]}

\item{\code{generatedOn}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-new}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-toJSON}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-fromJSON}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-toJSONString}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-fromJSONString}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-validateJSON}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-toString}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-isValid}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-getInvalidFields}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-print}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-clone}{\code{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$new(
  containerId = NULL,
  totalNumberOfChildren = NULL,
  numberOfValidChildren = NULL,
  numberOfInvalidChildren = NULL,
  numberOfUnknownChildren = NULL,
  generatedOn = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{containerId}}{containerId}

\item{\code{totalNumberOfChildren}}{totalNumberOfChildren}

\item{\code{numberOfValidChildren}}{numberOfValidChildren}

\item{\code{numberOfInvalidChildren}}{numberOfInvalidChildren}

\item{\code{numberOfUnknownChildren}}{numberOfUnknownChildren}

\item{\code{generatedOn}}{generatedOn}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSchemaValidationSummaryStatistics$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
