% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_search_document_fields.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchDocumentFields}
\alias{OrgSagebionetworksRepoModelSearchDocumentFields}
\title{OrgSagebionetworksRepoModelSearchDocumentFields}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for the fields of a search document.  Note that awesome search does not support camel case so we have to deviate from the usual naming scheme for properties.

OrgSagebionetworksRepoModelSearchDocumentFields Class
}
\details{
Create a new OrgSagebionetworksRepoModelSearchDocumentFields
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{parent_id}}{character [optional]}

\item{\code{node_type}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{created_on}}{integer [optional]}

\item{\code{modified_on}}{integer [optional]}

\item{\code{created_by}}{character [optional]}

\item{\code{modified_by}}{character [optional]}

\item{\code{acl}}{All group names with READ access to this entity list(character) [optional]}

\item{\code{update_acl}}{All group names with UPDATE access to this entity list(character) [optional]}

\item{\code{diagnosis}}{character [optional]}

\item{\code{tissue}}{character [optional]}

\item{\code{consortium}}{character [optional]}

\item{\code{organ}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-new}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-toJSON}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-fromJSON}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-toJSONString}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-validateJSON}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-toString}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-isValid}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-getInvalidFields}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-print}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentFields-clone}{\code{OrgSagebionetworksRepoModelSearchDocumentFields$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchDocumentFields class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$new(
  name = NULL,
  description = NULL,
  parent_id = NULL,
  node_type = NULL,
  etag = NULL,
  created_on = NULL,
  modified_on = NULL,
  created_by = NULL,
  modified_by = NULL,
  acl = NULL,
  update_acl = NULL,
  diagnosis = NULL,
  tissue = NULL,
  consortium = NULL,
  organ = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{parent_id}}{parent_id}

\item{\code{node_type}}{node_type}

\item{\code{etag}}{etag}

\item{\code{created_on}}{created_on}

\item{\code{modified_on}}{modified_on}

\item{\code{created_by}}{created_by}

\item{\code{modified_by}}{modified_by}

\item{\code{acl}}{All group names with READ access to this entity}

\item{\code{update_acl}}{All group names with UPDATE access to this entity}

\item{\code{diagnosis}}{diagnosis}

\item{\code{tissue}}{tissue}

\item{\code{consortium}}{consortium}

\item{\code{organ}}{organ}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchDocumentFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchDocumentFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchDocumentFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchDocumentFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchDocumentFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchDocumentFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSearchDocumentFields and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSearchDocumentFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentFields-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentFields-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentFields$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
