% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_search_document_type_names.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchDocumentTypeNames}
\alias{OrgSagebionetworksRepoModelSearchDocumentTypeNames}
\title{OrgSagebionetworksRepoModelSearchDocumentTypeNames}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelSearchDocumentTypeNames Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-new}{\code{OrgSagebionetworksRepoModelSearchDocumentTypeNames$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-toJSON}{\code{OrgSagebionetworksRepoModelSearchDocumentTypeNames$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-fromJSON}{\code{OrgSagebionetworksRepoModelSearchDocumentTypeNames$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-toJSONString}{\code{OrgSagebionetworksRepoModelSearchDocumentTypeNames$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchDocumentTypeNames$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-clone}{\code{OrgSagebionetworksRepoModelSearchDocumentTypeNames$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchDocumentTypeNames class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentTypeNames$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentTypeNames$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchDocumentTypeNames in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchDocumentTypeNames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentTypeNames$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchDocumentTypeNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentTypeNames$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchDocumentTypeNames in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchDocumentTypeNames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentTypeNames$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchDocumentTypeNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchDocumentTypeNames-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchDocumentTypeNames$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
