% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_search_facet.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchFacet}
\alias{OrgSagebionetworksRepoModelSearchFacet}
\title{OrgSagebionetworksRepoModelSearchFacet}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for a continuous or literal value Facet found in the search results.

OrgSagebionetworksRepoModelSearchFacet Class
}
\details{
Create a new OrgSagebionetworksRepoModelSearchFacet
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{type}}{character [optional]}

\item{\code{min}}{integer [optional]}

\item{\code{max}}{integer [optional]}

\item{\code{constraints}}{The list of constraints for this facet list(\link{OrgSagebionetworksRepoModelSearchFacetConstraint}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-new}{\code{OrgSagebionetworksRepoModelSearchFacet$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-toJSON}{\code{OrgSagebionetworksRepoModelSearchFacet$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-fromJSON}{\code{OrgSagebionetworksRepoModelSearchFacet$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-toJSONString}{\code{OrgSagebionetworksRepoModelSearchFacet$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchFacet$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-validateJSON}{\code{OrgSagebionetworksRepoModelSearchFacet$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-toString}{\code{OrgSagebionetworksRepoModelSearchFacet$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-isValid}{\code{OrgSagebionetworksRepoModelSearchFacet$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-getInvalidFields}{\code{OrgSagebionetworksRepoModelSearchFacet$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-print}{\code{OrgSagebionetworksRepoModelSearchFacet$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchFacet-clone}{\code{OrgSagebionetworksRepoModelSearchFacet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchFacet class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$new(
  name = NULL,
  type = NULL,
  min = NULL,
  max = NULL,
  constraints = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{type}}{type}

\item{\code{min}}{min}

\item{\code{max}}{max}

\item{\code{constraints}}{The list of constraints for this facet}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchFacet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchFacet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchFacet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchFacet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchFacet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchFacet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSearchFacet and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSearchFacet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchFacet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchFacet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchFacet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
