% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_search_hit.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchHit}
\alias{OrgSagebionetworksRepoModelSearchHit}
\title{OrgSagebionetworksRepoModelSearchHit}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for a Hit in the search results.  Each Hit contains a subset of entity data.

OrgSagebionetworksRepoModelSearchHit Class
}
\details{
Create a new OrgSagebionetworksRepoModelSearchHit
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{alias}}{character [optional]}

\item{\code{path}}{\link{OrgSagebionetworksRepoModelEntityPath} [optional]}

\item{\code{description}}{character [optional]}

\item{\code{node_type}}{character [optional]}

\item{\code{created_on}}{integer [optional]}

\item{\code{modified_on}}{integer [optional]}

\item{\code{created_by}}{character [optional]}

\item{\code{modified_by}}{character [optional]}

\item{\code{diagnosis}}{character [optional]}

\item{\code{tissue}}{character [optional]}

\item{\code{consortium}}{character [optional]}

\item{\code{organ}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-new}{\code{OrgSagebionetworksRepoModelSearchHit$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-toJSON}{\code{OrgSagebionetworksRepoModelSearchHit$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-fromJSON}{\code{OrgSagebionetworksRepoModelSearchHit$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-toJSONString}{\code{OrgSagebionetworksRepoModelSearchHit$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchHit$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-validateJSON}{\code{OrgSagebionetworksRepoModelSearchHit$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-toString}{\code{OrgSagebionetworksRepoModelSearchHit$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-isValid}{\code{OrgSagebionetworksRepoModelSearchHit$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-getInvalidFields}{\code{OrgSagebionetworksRepoModelSearchHit$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-print}{\code{OrgSagebionetworksRepoModelSearchHit$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchHit-clone}{\code{OrgSagebionetworksRepoModelSearchHit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchHit class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$new(
  id = NULL,
  etag = NULL,
  name = NULL,
  alias = NULL,
  path = NULL,
  description = NULL,
  node_type = NULL,
  created_on = NULL,
  modified_on = NULL,
  created_by = NULL,
  modified_by = NULL,
  diagnosis = NULL,
  tissue = NULL,
  consortium = NULL,
  organ = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{name}}{name}

\item{\code{alias}}{alias}

\item{\code{path}}{path}

\item{\code{description}}{description}

\item{\code{node_type}}{node_type}

\item{\code{created_on}}{created_on}

\item{\code{modified_on}}{modified_on}

\item{\code{created_by}}{created_by}

\item{\code{modified_by}}{modified_by}

\item{\code{diagnosis}}{diagnosis}

\item{\code{tissue}}{tissue}

\item{\code{consortium}}{consortium}

\item{\code{organ}}{organ}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchHit in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchHit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchHit
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchHit in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchHit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchHit
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSearchHit and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSearchHit
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchHit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchHit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchHit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
