% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_search_query_facet_sort.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchQueryFacetSort}
\alias{OrgSagebionetworksRepoModelSearchQueryFacetSort}
\title{OrgSagebionetworksRepoModelSearchQueryFacetSort}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for a facet field sort key-value pair.

OrgSagebionetworksRepoModelSearchQueryFacetSort Class
}
\details{
Create a new OrgSagebionetworksRepoModelSearchQueryFacetSort
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{facetName}}{character [optional]}

\item{\code{sortType}}{character [optional]}

\item{\code{sumFields}}{Sort the facet values by the sum of the values in two or more fields. This is a list of names of the fields to sum. list(character) [optional]}

\item{\code{maxfield}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-new}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-toJSON}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-fromJSON}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-toJSONString}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-validateJSON}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-toString}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-isValid}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-getInvalidFields}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-print}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSort-clone}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchQueryFacetSort class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$new(
  facetName = NULL,
  sortType = NULL,
  sumFields = NULL,
  maxfield = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{facetName}}{facetName}

\item{\code{sortType}}{sortType}

\item{\code{sumFields}}{Sort the facet values by the sum of the values in two or more fields. This is a list of names of the fields to sum.}

\item{\code{maxfield}}{maxfield}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQueryFacetSort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQueryFacetSort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQueryFacetSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQueryFacetSort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQueryFacetSort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQueryFacetSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSearchQueryFacetSort and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSearchQueryFacetSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
