% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_search_query_facet_sort_options.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions}
\alias{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions}
\title{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelSearchQueryFacetSortOptions Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-new}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-toJSON}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-fromJSON}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-toJSONString}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-clone}{\code{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchQueryFacetSortOptions class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQueryFacetSortOptions in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQueryFacetSortOptions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQueryFacetSortOptions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQueryFacetSortOptions in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQueryFacetSortOptions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQueryFacetSortOptions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryFacetSortOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryFacetSortOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
