% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_search_query_key_value.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchQueryKeyValue}
\alias{OrgSagebionetworksRepoModelSearchQueryKeyValue}
\title{OrgSagebionetworksRepoModelSearchQueryKeyValue}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for a faceted key-value pair.

OrgSagebionetworksRepoModelSearchQueryKeyValue Class
}
\details{
Create a new OrgSagebionetworksRepoModelSearchQueryKeyValue
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{key}}{character [optional]}

\item{\code{value}}{character [optional]}

\item{\code{not}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-new}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-toJSON}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-fromJSON}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-toJSONString}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-validateJSON}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-toString}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-isValid}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-getInvalidFields}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-print}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQueryKeyValue-clone}{\code{OrgSagebionetworksRepoModelSearchQueryKeyValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchQueryKeyValue class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$new(
  key = NULL,
  value = NULL,
  not = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{key}

\item{\code{value}}{value}

\item{\code{not}}{not}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQueryKeyValue in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQueryKeyValue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQueryKeyValue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQueryKeyValue in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQueryKeyValue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQueryKeyValue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSearchQueryKeyValue and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSearchQueryKeyValue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQueryKeyValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQueryKeyValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQueryKeyValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
