% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_search_query_search_facet_option.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption}
\alias{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption}
\title{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption}
\format{
An \code{R6Class} generator object
}
\description{
Specify which field should be returned as facets and the format of the results for each field

OrgSagebionetworksRepoModelSearchQuerySearchFacetOption Class
}
\details{
Create a new OrgSagebionetworksRepoModelSearchQuerySearchFacetOption
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{maxResultCount}}{integer [optional]}

\item{\code{sortType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-new}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-toJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-fromJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-toJSONString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-validateJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-toString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-isValid}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-getInvalidFields}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-print}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-clone}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchQuerySearchFacetOption class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$new(
  name = NULL,
  maxResultCount = NULL,
  sortType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{maxResultCount}}{maxResultCount}

\item{\code{sortType}}{sortType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQuerySearchFacetOption in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQuerySearchFacetOption
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQuerySearchFacetOption
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQuerySearchFacetOption in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQuerySearchFacetOption
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQuerySearchFacetOption
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSearchQuerySearchFacetOption and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSearchQuerySearchFacetOption
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetOption-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
