% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_search_query_search_facet_sort.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort}
\alias{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort}
\title{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelSearchQuerySearchFacetSort Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-new}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-toJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-fromJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-toJSONString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-clone}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchQuerySearchFacetSort class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQuerySearchFacetSort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQuerySearchFacetSort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQuerySearchFacetSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQuerySearchFacetSort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQuerySearchFacetSort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQuerySearchFacetSort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFacetSort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFacetSort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
