% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_search_query_search_field_name.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchQuerySearchFieldName}
\alias{OrgSagebionetworksRepoModelSearchQuerySearchFieldName}
\title{OrgSagebionetworksRepoModelSearchQuerySearchFieldName}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelSearchQuerySearchFieldName Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-new}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-toJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-fromJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-toJSONString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-clone}{\code{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchQuerySearchFieldName class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQuerySearchFieldName in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQuerySearchFieldName
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQuerySearchFieldName
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQuerySearchFieldName in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQuerySearchFieldName
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQuerySearchFieldName
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchFieldName-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchFieldName$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
