% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_search_query_search_query.R
\docType{class}
\name{OrgSagebionetworksRepoModelSearchQuerySearchQuery}
\alias{OrgSagebionetworksRepoModelSearchQuerySearchQuery}
\title{OrgSagebionetworksRepoModelSearchQuerySearchQuery}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for a search query.

OrgSagebionetworksRepoModelSearchQuerySearchQuery Class
}
\details{
Create a new OrgSagebionetworksRepoModelSearchQuerySearchQuery
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{queryTerm}}{The free text search terms. This will search the name of the Entity and its Wiki contents list(character) [optional]}

\item{\code{booleanQuery}}{One or more key-value pairs that define a boolean search. Multiple expressions are joined with a top-level AND. Key is the facet field name, value is the facet value. list(\link{OrgSagebionetworksRepoModelSearchQueryKeyValue}) [optional]}

\item{\code{rangeQuery}}{One or more key-value-range filters that filter values of a key based on the specified range where min<=value<=max. Multiple expressions are joined with a top-level AND. list(\link{OrgSagebionetworksRepoModelSearchQueryKeyRange}) [optional]}

\item{\code{facetOptions}}{Specify which fields should be returned as facets and the format of the results for each field list(\link{OrgSagebionetworksRepoModelSearchQuerySearchFacetOption}) [optional]}

\item{\code{returnFields}}{Specifies the document fields to include in the response. By default, only the document ids of the hits are returned. list(character) [optional]}

\item{\code{start}}{integer [optional]}

\item{\code{size}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-new}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-toJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-fromJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-toJSONString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-fromJSONString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-validateJSON}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-toString}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-isValid}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-getInvalidFields}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-print}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-clone}{\code{OrgSagebionetworksRepoModelSearchQuerySearchQuery$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSearchQuerySearchQuery class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$new(
  queryTerm = NULL,
  booleanQuery = NULL,
  rangeQuery = NULL,
  facetOptions = NULL,
  returnFields = NULL,
  start = NULL,
  size = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queryTerm}}{The free text search terms. This will search the name of the Entity and its Wiki contents}

\item{\code{booleanQuery}}{One or more key-value pairs that define a boolean search. Multiple expressions are joined with a top-level AND. Key is the facet field name, value is the facet value.}

\item{\code{rangeQuery}}{One or more key-value-range filters that filter values of a key based on the specified range where min<=value<=max. Multiple expressions are joined with a top-level AND.}

\item{\code{facetOptions}}{Specify which fields should be returned as facets and the format of the results for each field}

\item{\code{returnFields}}{Specifies the document fields to include in the response. By default, only the document ids of the hits are returned.}

\item{\code{start}}{start}

\item{\code{size}}{size}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQuerySearchQuery in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQuerySearchQuery
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQuerySearchQuery
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSearchQuerySearchQuery in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSearchQuerySearchQuery
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSearchQuerySearchQuery
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSearchQuerySearchQuery and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSearchQuerySearchQuery
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSearchQuerySearchQuery-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSearchQuerySearchQuery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
