% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_self_sign_access_requirement.R
\docType{class}
\name{OrgSagebionetworksRepoModelSelfSignAccessRequirement}
\alias{OrgSagebionetworksRepoModelSelfSignAccessRequirement}
\title{OrgSagebionetworksRepoModelSelfSignAccessRequirement}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for Terms of Use Access Requirement, used for a 'tier 2' Access Requirement

OrgSagebionetworksRepoModelSelfSignAccessRequirement Class
}
\details{
Create a new OrgSagebionetworksRepoModelSelfSignAccessRequirement
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{versionNumber}}{integer [optional]}

\item{\code{id}}{integer [optional]}

\item{\code{description}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{subjectsDefinedByAnnotations}}{character [optional]}

\item{\code{subjectIds}}{The IDs of the items controlled by this Access Requirement when 'subjectsDefinedByAnnotations=false'. This property is mutually exclusive with 'subjectsDefinedByAnnotations'.  When 'subjectsDefinedByAnnotations=true' then this property must be empty or excluded.  Required when creating or updating and 'subjectsDefinedByAnnotations=false' or 'subjectsDefinedByAnnotations' is excluded. list(\link{OrgSagebionetworksRepoModelRestrictableObjectDescriptor}) [optional]}

\item{\code{accessType}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{isCertifiedUserRequired}}{character [optional]}

\item{\code{isValidatedProfileRequired}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-new}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-toJSON}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-fromJSON}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-toJSONString}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-fromJSONString}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-validateJSON}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-toString}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-isValid}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-getInvalidFields}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-print}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-clone}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSelfSignAccessRequirement class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$new(
  concreteType,
  versionNumber = NULL,
  id = NULL,
  description = NULL,
  name = NULL,
  etag = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  subjectsDefinedByAnnotations = NULL,
  subjectIds = NULL,
  accessType = NULL,
  isCertifiedUserRequired = NULL,
  isValidatedProfileRequired = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{versionNumber}}{versionNumber}

\item{\code{id}}{id}

\item{\code{description}}{description}

\item{\code{name}}{name}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{subjectsDefinedByAnnotations}}{subjectsDefinedByAnnotations}

\item{\code{subjectIds}}{The IDs of the items controlled by this Access Requirement when 'subjectsDefinedByAnnotations=false'. This property is mutually exclusive with 'subjectsDefinedByAnnotations'.  When 'subjectsDefinedByAnnotations=true' then this property must be empty or excluded.  Required when creating or updating and 'subjectsDefinedByAnnotations=false' or 'subjectsDefinedByAnnotations' is excluded.}

\item{\code{accessType}}{accessType}

\item{\code{isCertifiedUserRequired}}{isCertifiedUserRequired}

\item{\code{isValidatedProfileRequired}}{isValidatedProfileRequired}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSelfSignAccessRequirement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSelfSignAccessRequirement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSelfSignAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSelfSignAccessRequirement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSelfSignAccessRequirement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSelfSignAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSelfSignAccessRequirement and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSelfSignAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
