% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_self_sign_access_requirement_interface.R
\docType{class}
\name{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface}
\alias{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface}
\title{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-new}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-fromJSONString}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-fromJSON}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-toJSONString}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-toJSON}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-validateJSON}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-toString}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-print}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-clone}{\code{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$new(
  instance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelPostMessageContentAccessRequirement", "OrgSagebionetworksRepoModelSelfSignAccessRequirement", "OrgSagebionetworksRepoModelTermsOfUseAccessRequirement"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSelfSignAccessRequirementInterface$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
