% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_signed_token_interface.R
\docType{class}
\name{OrgSagebionetworksRepoModelSignedTokenInterface}
\alias{OrgSagebionetworksRepoModelSignedTokenInterface}
\title{OrgSagebionetworksRepoModelSignedTokenInterface}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelSignedTokenInterface Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelSignedTokenInterface.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSignedTokenInterface-new}{\code{OrgSagebionetworksRepoModelSignedTokenInterface$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSignedTokenInterface-fromJSONString}{\code{OrgSagebionetworksRepoModelSignedTokenInterface$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSignedTokenInterface-fromJSON}{\code{OrgSagebionetworksRepoModelSignedTokenInterface$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSignedTokenInterface-toJSONString}{\code{OrgSagebionetworksRepoModelSignedTokenInterface$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSignedTokenInterface-toJSON}{\code{OrgSagebionetworksRepoModelSignedTokenInterface$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSignedTokenInterface-validateJSON}{\code{OrgSagebionetworksRepoModelSignedTokenInterface$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSignedTokenInterface-toString}{\code{OrgSagebionetworksRepoModelSignedTokenInterface$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSignedTokenInterface-print}{\code{OrgSagebionetworksRepoModelSignedTokenInterface$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSignedTokenInterface-clone}{\code{OrgSagebionetworksRepoModelSignedTokenInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSignedTokenInterface-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSignedTokenInterface-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSignedTokenInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSignedTokenInterface$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelAuthAuthenticationReceiptToken", "OrgSagebionetworksRepoModelAuthNewUserSignedToken", "OrgSagebionetworksRepoModelAuthPasswordResetSignedToken", "OrgSagebionetworksRepoModelAuthTwoFactorAuthToken", "OrgSagebionetworksRepoModelInviteeVerificationSignedToken", "OrgSagebionetworksRepoModelJoinTeamSignedToken", "OrgSagebionetworksRepoModelMembershipInvtnSignedToken", "OrgSagebionetworksRepoModelMessageNotificationSettingsSignedToken", "OrgSagebionetworksRepoModelPrincipalEmailValidationSignedToken"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSignedTokenInterface-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSignedTokenInterface-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSignedTokenInterface.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSignedTokenInterface$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelSignedTokenInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSignedTokenInterface-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSignedTokenInterface-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSignedTokenInterface.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSignedTokenInterface$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelSignedTokenInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSignedTokenInterface-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSignedTokenInterface-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelSignedTokenInterface to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSignedTokenInterface$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelSignedTokenInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSignedTokenInterface-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSignedTokenInterface-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelSignedTokenInterface to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSignedTokenInterface$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelSignedTokenInterface.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSignedTokenInterface-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSignedTokenInterface-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelSignedTokenInterface and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSignedTokenInterface$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSignedTokenInterface-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSignedTokenInterface-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSignedTokenInterface$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSignedTokenInterface-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSignedTokenInterface-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSignedTokenInterface$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSignedTokenInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSignedTokenInterface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSignedTokenInterface$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
