% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_statistics_files_count_statistics.R
\docType{class}
\name{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics}
\alias{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics}
\title{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics}
\format{
An \code{R6Class} generator object
}
\description{
Describes an aggregate of file actions (e.g. downloads/uploads) and (unique) users count for a specific time frame

OrgSagebionetworksRepoModelStatisticsFilesCountStatistics Class
}
\details{
Create a new OrgSagebionetworksRepoModelStatisticsFilesCountStatistics
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rangeStart}}{character [optional]}

\item{\code{rangeEnd}}{character [optional]}

\item{\code{filesCount}}{integer [optional]}

\item{\code{usersCount}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-new}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$new()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-toJSON}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-fromJSON}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-toJSONString}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-fromJSONString}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-validateJSON}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-toString}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-isValid}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-getInvalidFields}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-print}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$print()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-clone}{\code{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelStatisticsFilesCountStatistics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$new(
  rangeStart = NULL,
  rangeEnd = NULL,
  filesCount = NULL,
  usersCount = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeStart}}{rangeStart}

\item{\code{rangeEnd}}{rangeEnd}

\item{\code{filesCount}}{filesCount}

\item{\code{usersCount}}{usersCount}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatisticsFilesCountStatistics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsFilesCountStatistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatisticsFilesCountStatistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatisticsFilesCountStatistics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsFilesCountStatistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatisticsFilesCountStatistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelStatisticsFilesCountStatistics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelStatisticsFilesCountStatistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsFilesCountStatistics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsFilesCountStatistics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
