% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_statistics_object_statistics_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest}
\alias{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest}
\title{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-new}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-fromJSON}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-toJSONString}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-toJSON}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-validateJSON}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-toString}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-print}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-clone}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$new(
  instance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
