% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_statistics_object_statistics_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse}
\alias{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse}
\title{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-new}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-fromJSON}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-toJSONString}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-toJSON}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-validateJSON}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-toString}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-print}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-clone}{\code{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$new(
  instance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$fromJSONString(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsObjectStatisticsResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
