% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_statistics_project_files_statistics_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest}
\alias{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest}
\title{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request body for retrieving the statistics relative to files for a specific project. Allows to retrieve the monthly count of files downloaded and/or uploaded as well as the (unique) number of users that performed a download/upload for the past 12 months (excluding the current month). The response type for this request is <a href=\"${org.sagebionetworks.repo.model.statistics.ProjectFilesStatisticsResponse}\">ProjectFilesStatisticsResponse</a>

OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{objectId}}{character [optional]}

\item{\code{fileDownloads}}{character [optional]}

\item{\code{fileUploads}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-new}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-toJSON}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-fromJSON}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-toJSONString}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-validateJSON}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-toString}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-isValid}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-print}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-clone}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$new(
  concreteType,
  objectId = NULL,
  fileDownloads = NULL,
  fileUploads = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{objectId}}{objectId}

\item{\code{fileDownloads}}{fileDownloads}

\item{\code{fileUploads}}{fileUploads}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
