% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_statistics_project_files_statistics_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse}
\alias{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse}
\title{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Container for statistics related to project files. Exposes the monthly aggregates of files and (unique) users that downloaded/uploaded for the past 12 months (excluding the current month).

OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objectId}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{fileDownloads}}{\link{OrgSagebionetworksRepoModelStatisticsMonthlyFilesStatistics} [optional]}

\item{\code{fileUploads}}{\link{OrgSagebionetworksRepoModelStatisticsMonthlyFilesStatistics} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-new}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-toJSON}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-fromJSON}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-toJSONString}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-validateJSON}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-toString}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-isValid}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-print}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-clone}{\code{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$new(
  concreteType,
  objectId = NULL,
  fileDownloads = NULL,
  fileUploads = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{objectId}}{objectId}

\item{\code{fileDownloads}}{fileDownloads}

\item{\code{fileUploads}}{fileUploads}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$toJSONString(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatisticsProjectFilesStatisticsResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
