% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_status_history_record.R
\docType{class}
\name{OrgSagebionetworksRepoModelStatusHistoryRecord}
\alias{OrgSagebionetworksRepoModelStatusHistoryRecord}
\title{OrgSagebionetworksRepoModelStatusHistoryRecord}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for StatusHistoryRecord POJO

OrgSagebionetworksRepoModelStatusHistoryRecord Class
}
\details{
Create a new OrgSagebionetworksRepoModelStatusHistoryRecord
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{statusName}}{character [optional]}

\item{\code{date}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-new}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$new()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-toJSON}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-fromJSON}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-toJSONString}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-fromJSONString}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-validateJSON}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-toString}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-isValid}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-getInvalidFields}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-print}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$print()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusHistoryRecord-clone}{\code{OrgSagebionetworksRepoModelStatusHistoryRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelStatusHistoryRecord class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$new(
  statusName = NULL,
  date = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statusName}}{statusName}

\item{\code{date}}{date}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatusHistoryRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatusHistoryRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatusHistoryRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatusHistoryRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatusHistoryRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatusHistoryRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelStatusHistoryRecord and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelStatusHistoryRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusHistoryRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusHistoryRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusHistoryRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
