% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_status_stack_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelStatusStackStatus}
\alias{OrgSagebionetworksRepoModelStatusStackStatus}
\title{OrgSagebionetworksRepoModelStatusStackStatus}
\format{
An \code{R6Class} generator object
}
\description{
The Status of the stack

OrgSagebionetworksRepoModelStatusStackStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelStatusStackStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{status}}{character [optional]}

\item{\code{currentMessage}}{character [optional]}

\item{\code{pendingMaintenanceMessage}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-new}{\code{OrgSagebionetworksRepoModelStatusStackStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-toJSON}{\code{OrgSagebionetworksRepoModelStatusStackStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-fromJSON}{\code{OrgSagebionetworksRepoModelStatusStackStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-toJSONString}{\code{OrgSagebionetworksRepoModelStatusStackStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelStatusStackStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-validateJSON}{\code{OrgSagebionetworksRepoModelStatusStackStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-toString}{\code{OrgSagebionetworksRepoModelStatusStackStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-isValid}{\code{OrgSagebionetworksRepoModelStatusStackStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelStatusStackStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-print}{\code{OrgSagebionetworksRepoModelStatusStackStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelStatusStackStatus-clone}{\code{OrgSagebionetworksRepoModelStatusStackStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelStatusStackStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$new(
  status = NULL,
  currentMessage = NULL,
  pendingMaintenanceMessage = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{status}

\item{\code{currentMessage}}{currentMessage}

\item{\code{pendingMaintenanceMessage}}{pendingMaintenanceMessage}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatusStackStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatusStackStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatusStackStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStatusStackStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStatusStackStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStatusStackStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelStatusStackStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelStatusStackStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStatusStackStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStatusStackStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStatusStackStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
