% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_sts_sts_credentials.R
\docType{class}
\name{OrgSagebionetworksRepoModelStsStsCredentials}
\alias{OrgSagebionetworksRepoModelStsStsCredentials}
\title{OrgSagebionetworksRepoModelStsStsCredentials}
\format{
An \code{R6Class} generator object
}
\description{
Temporary AWS credentials vended by STS (Security Token Service).

OrgSagebionetworksRepoModelStsStsCredentials Class
}
\details{
Create a new OrgSagebionetworksRepoModelStsStsCredentials
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bucket}}{character [optional]}

\item{\code{baseKey}}{character [optional]}

\item{\code{accessKeyId}}{character [optional]}

\item{\code{secretAccessKey}}{character [optional]}

\item{\code{sessionToken}}{character [optional]}

\item{\code{expiration}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-new}{\code{OrgSagebionetworksRepoModelStsStsCredentials$new()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-toJSON}{\code{OrgSagebionetworksRepoModelStsStsCredentials$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-fromJSON}{\code{OrgSagebionetworksRepoModelStsStsCredentials$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-toJSONString}{\code{OrgSagebionetworksRepoModelStsStsCredentials$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-fromJSONString}{\code{OrgSagebionetworksRepoModelStsStsCredentials$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-validateJSON}{\code{OrgSagebionetworksRepoModelStsStsCredentials$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-toString}{\code{OrgSagebionetworksRepoModelStsStsCredentials$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-isValid}{\code{OrgSagebionetworksRepoModelStsStsCredentials$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-getInvalidFields}{\code{OrgSagebionetworksRepoModelStsStsCredentials$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-print}{\code{OrgSagebionetworksRepoModelStsStsCredentials$print()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsCredentials-clone}{\code{OrgSagebionetworksRepoModelStsStsCredentials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelStsStsCredentials class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$new(
  bucket = NULL,
  baseKey = NULL,
  accessKeyId = NULL,
  secretAccessKey = NULL,
  sessionToken = NULL,
  expiration = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bucket}}{bucket}

\item{\code{baseKey}}{baseKey}

\item{\code{accessKeyId}}{accessKeyId}

\item{\code{secretAccessKey}}{secretAccessKey}

\item{\code{sessionToken}}{sessionToken}

\item{\code{expiration}}{expiration}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStsStsCredentials in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStsStsCredentials
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStsStsCredentials
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStsStsCredentials in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStsStsCredentials
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStsStsCredentials
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelStsStsCredentials and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelStsStsCredentials
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsCredentials-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsCredentials-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsCredentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
