% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_sts_sts_permission.R
\docType{class}
\name{OrgSagebionetworksRepoModelStsStsPermission}
\alias{OrgSagebionetworksRepoModelStsStsPermission}
\title{OrgSagebionetworksRepoModelStsStsPermission}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelStsStsPermission Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelStsStsPermission-new}{\code{OrgSagebionetworksRepoModelStsStsPermission$new()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsPermission-toJSON}{\code{OrgSagebionetworksRepoModelStsStsPermission$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsPermission-fromJSON}{\code{OrgSagebionetworksRepoModelStsStsPermission$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsPermission-toJSONString}{\code{OrgSagebionetworksRepoModelStsStsPermission$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsPermission-fromJSONString}{\code{OrgSagebionetworksRepoModelStsStsPermission$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelStsStsPermission-clone}{\code{OrgSagebionetworksRepoModelStsStsPermission$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsPermission-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsPermission-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelStsStsPermission class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsPermission$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsPermission-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsPermission-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsPermission$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStsStsPermission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsPermission-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsPermission-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStsStsPermission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsPermission$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStsStsPermission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsPermission-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsPermission-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsPermission$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelStsStsPermission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsPermission-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsPermission-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelStsStsPermission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsPermission$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelStsStsPermission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelStsStsPermission-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelStsStsPermission-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelStsStsPermission$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
