% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_submission_team_paged_results.R
\docType{class}
\name{OrgSagebionetworksRepoModelSubmissionTeamPagedResults}
\alias{OrgSagebionetworksRepoModelSubmissionTeamPagedResults}
\title{OrgSagebionetworksRepoModelSubmissionTeamPagedResults}
\format{
An \code{R6Class} generator object
}
\description{
A paginated list of SubmissionTeam objects

OrgSagebionetworksRepoModelSubmissionTeamPagedResults Class
}
\details{
Create a new OrgSagebionetworksRepoModelSubmissionTeamPagedResults
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{The list of results for this page list(\link{OrgSagebionetworksRepoModelSubmissionTeam}) [optional]}

\item{\code{totalNumberOfResults}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-new}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-toJSON}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-fromJSON}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-toJSONString}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-fromJSONString}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-validateJSON}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-toString}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-isValid}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-getInvalidFields}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-print}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-clone}{\code{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSubmissionTeamPagedResults class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$new(
  results = NULL,
  totalNumberOfResults = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{The list of results for this page}

\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubmissionTeamPagedResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubmissionTeamPagedResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubmissionTeamPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubmissionTeamPagedResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubmissionTeamPagedResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubmissionTeamPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSubmissionTeamPagedResults and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSubmissionTeamPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubmissionTeamPagedResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubmissionTeamPagedResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
