% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_subscription_subscriber_count.R
\docType{class}
\name{OrgSagebionetworksRepoModelSubscriptionSubscriberCount}
\alias{OrgSagebionetworksRepoModelSubscriptionSubscriberCount}
\title{OrgSagebionetworksRepoModelSubscriptionSubscriberCount}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelSubscriptionSubscriberCount Class

OrgSagebionetworksRepoModelSubscriptionSubscriberCount Class
}
\details{
Create a new OrgSagebionetworksRepoModelSubscriptionSubscriberCount
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-new}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-toJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-fromJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-toJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-fromJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-validateJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-toString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-isValid}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-getInvalidFields}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-print}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-clone}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSubscriptionSubscriberCount class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$new(count = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{count}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriberCount in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionSubscriberCount
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionSubscriberCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriberCount in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionSubscriberCount
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionSubscriberCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSubscriptionSubscriberCount and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSubscriptionSubscriberCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriberCount-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriberCount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
