% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_subscription_subscription.R
\docType{class}
\name{OrgSagebionetworksRepoModelSubscriptionSubscription}
\alias{OrgSagebionetworksRepoModelSubscriptionSubscription}
\title{OrgSagebionetworksRepoModelSubscriptionSubscription}
\format{
An \code{R6Class} generator object
}
\description{
The model object represents a single Subscription.

OrgSagebionetworksRepoModelSubscriptionSubscription Class
}
\details{
Create a new OrgSagebionetworksRepoModelSubscriptionSubscription
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subscriptionId}}{character [optional]}

\item{\code{subscriberId}}{character [optional]}

\item{\code{objectId}}{character [optional]}

\item{\code{objectType}}{character [optional]}

\item{\code{createdOn}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-new}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-toJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-fromJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-toJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-fromJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-validateJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-toString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-isValid}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-getInvalidFields}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-print}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscription-clone}{\code{OrgSagebionetworksRepoModelSubscriptionSubscription$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSubscriptionSubscription class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$new(
  subscriptionId = NULL,
  subscriberId = NULL,
  objectId = NULL,
  objectType = NULL,
  createdOn = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subscriptionId}}{subscriptionId}

\item{\code{subscriberId}}{subscriberId}

\item{\code{objectId}}{objectId}

\item{\code{objectType}}{objectType}

\item{\code{createdOn}}{createdOn}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscription in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionSubscription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionSubscription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscription in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionSubscription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionSubscription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSubscriptionSubscription and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSubscriptionSubscription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscription-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscription-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscription$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
