% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_subscription_subscription_object_type.R
\docType{class}
\name{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType}
\alias{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType}
\title{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-new}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-toJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-fromJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-toJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-fromJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-clone}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
