% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_subscription_subscription_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest}
\alias{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest}
\title{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest}
\format{
An \code{R6Class} generator object
}
\description{
A object that is used to request a list of Subscription

OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objectType}}{character [optional]}

\item{\code{idList}}{The list of objectId list(character) [optional]}

\item{\code{sortByType}}{character [optional]}

\item{\code{sortDirection}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-new}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-toJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-fromJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-toJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-validateJSON}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-toString}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-isValid}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-print}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-clone}{\code{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$new(
  objectType = NULL,
  idList = NULL,
  sortByType = NULL,
  sortDirection = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objectType}}{objectType}

\item{\code{idList}}{The list of objectId}

\item{\code{sortByType}}{sortByType}

\item{\code{sortDirection}}{sortDirection}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
