% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_subscription_topic.R
\docType{class}
\name{OrgSagebionetworksRepoModelSubscriptionTopic}
\alias{OrgSagebionetworksRepoModelSubscriptionTopic}
\title{OrgSagebionetworksRepoModelSubscriptionTopic}
\format{
An \code{R6Class} generator object
}
\description{
The model object represents a topic of subscription.

OrgSagebionetworksRepoModelSubscriptionTopic Class
}
\details{
Create a new OrgSagebionetworksRepoModelSubscriptionTopic
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objectId}}{character [optional]}

\item{\code{objectType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-new}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$new()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-toJSON}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-fromJSON}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-toJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-fromJSONString}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-validateJSON}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-toString}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-isValid}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-getInvalidFields}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-print}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$print()}}
\item \href{#method-OrgSagebionetworksRepoModelSubscriptionTopic-clone}{\code{OrgSagebionetworksRepoModelSubscriptionTopic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelSubscriptionTopic class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$new(
  objectId = NULL,
  objectType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objectId}}{objectId}

\item{\code{objectType}}{objectType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionTopic in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionTopic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionTopic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionTopic in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelSubscriptionTopic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelSubscriptionTopic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelSubscriptionTopic and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelSubscriptionTopic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelSubscriptionTopic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelSubscriptionTopic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelSubscriptionTopic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
