% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_appendable_row_set_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableAppendableRowSetRequest}
\alias{OrgSagebionetworksRepoModelTableAppendableRowSetRequest}
\title{OrgSagebionetworksRepoModelTableAppendableRowSetRequest}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelTableAppendableRowSetRequest Class

OrgSagebionetworksRepoModelTableAppendableRowSetRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableAppendableRowSetRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{entityId}}{character [optional]}

\item{\code{toAppend}}{\link{OrgSagebionetworksRepoModelTableAppendableRowSet} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-new}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-toJSON}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-fromJSON}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-toJSONString}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-validateJSON}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-toString}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-isValid}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-print}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-clone}{\code{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableAppendableRowSetRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$new(
  concreteType,
  entityId = NULL,
  toAppend = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{entityId}}{entityId}

\item{\code{toAppend}}{toAppend}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableAppendableRowSetRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableAppendableRowSetRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableAppendableRowSetRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableAppendableRowSetRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableAppendableRowSetRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableAppendableRowSetRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableAppendableRowSetRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableAppendableRowSetRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableAppendableRowSetRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableAppendableRowSetRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
