% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_column_single_value_filter_operator.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator}
\alias{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator}
\title{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-new}{\code{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-toJSON}{\code{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-fromJSON}{\code{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-toJSONString}{\code{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-fromJSONString}{\code{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-clone}{\code{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableColumnSingleValueFilterOperator$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
