% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_csv_table_descriptor.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableCsvTableDescriptor}
\alias{OrgSagebionetworksRepoModelTableCsvTableDescriptor}
\title{OrgSagebionetworksRepoModelTableCsvTableDescriptor}
\format{
An \code{R6Class} generator object
}
\description{
The description of a csv for upload or download.

OrgSagebionetworksRepoModelTableCsvTableDescriptor Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableCsvTableDescriptor
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{separator}}{character [optional]}

\item{\code{quoteCharacter}}{character [optional]}

\item{\code{escapeCharacter}}{character [optional]}

\item{\code{lineEnd}}{character [optional]}

\item{\code{isFirstLineHeader}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-new}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-toJSON}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-fromJSON}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-toJSONString}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-fromJSONString}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-validateJSON}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-toString}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-isValid}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-print}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-clone}{\code{OrgSagebionetworksRepoModelTableCsvTableDescriptor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableCsvTableDescriptor class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$new(
  separator = NULL,
  quoteCharacter = NULL,
  escapeCharacter = NULL,
  lineEnd = NULL,
  isFirstLineHeader = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{separator}}{separator}

\item{\code{quoteCharacter}}{quoteCharacter}

\item{\code{escapeCharacter}}{escapeCharacter}

\item{\code{lineEnd}}{lineEnd}

\item{\code{isFirstLineHeader}}{isFirstLineHeader}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableCsvTableDescriptor in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableCsvTableDescriptor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableCsvTableDescriptor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableCsvTableDescriptor in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableCsvTableDescriptor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableCsvTableDescriptor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableCsvTableDescriptor and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableCsvTableDescriptor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableCsvTableDescriptor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableCsvTableDescriptor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
