% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_defining_sql_entity_type.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableDefiningSqlEntityType}
\alias{OrgSagebionetworksRepoModelTableDefiningSqlEntityType}
\title{OrgSagebionetworksRepoModelTableDefiningSqlEntityType}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelTableDefiningSqlEntityType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-new}{\code{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-toJSON}{\code{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-fromJSON}{\code{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-toJSONString}{\code{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-fromJSONString}{\code{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-clone}{\code{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableDefiningSqlEntityType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableDefiningSqlEntityType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableDefiningSqlEntityType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableDefiningSqlEntityType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableDefiningSqlEntityType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableDefiningSqlEntityType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableDefiningSqlEntityType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDefiningSqlEntityType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDefiningSqlEntityType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
