% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_download_from_table_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableDownloadFromTableRequest}
\alias{OrgSagebionetworksRepoModelTableDownloadFromTableRequest}
\title{OrgSagebionetworksRepoModelTableDownloadFromTableRequest}
\format{
An \code{R6Class} generator object
}
\description{
A CSV table download request.

OrgSagebionetworksRepoModelTableDownloadFromTableRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableDownloadFromTableRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{entityId}}{character [optional]}

\item{\code{writeHeader}}{character [optional]}

\item{\code{includeRowIdAndRowVersion}}{character [optional]}

\item{\code{csvTableDescriptor}}{\link{OrgSagebionetworksRepoModelTableCsvTableDescriptor} [optional]}

\item{\code{fileName}}{character [optional]}

\item{\code{sql}}{character [optional]}

\item{\code{additionalFilters}}{Appends additional filters to the SQL query. These are applied before facets. Filters within the list have an AND relationship. If a WHERE clause already exists on the SQL query or facets are selected, it will also be ANDed with the query generated by these additional filters. list(\link{OrgSagebionetworksRepoModelTableQueryFilter}) [optional]}

\item{\code{selectedFacets}}{The selected facet filters. list(\link{OrgSagebionetworksRepoModelTableFacetColumnRequest}) [optional]}

\item{\code{includeEntityEtag}}{character [optional]}

\item{\code{selectFileColumn}}{integer [optional]}

\item{\code{offset}}{integer [optional]}

\item{\code{limit}}{integer [optional]}

\item{\code{sort}}{list(\link{OrgSagebionetworksRepoModelTableSortItem}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-new}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-toJSON}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-fromJSON}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-toJSONString}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-validateJSON}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-toString}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-isValid}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-print}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-clone}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableDownloadFromTableRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$new(
  concreteType,
  entityId = NULL,
  writeHeader = NULL,
  includeRowIdAndRowVersion = NULL,
  csvTableDescriptor = NULL,
  fileName = NULL,
  sql = NULL,
  additionalFilters = NULL,
  selectedFacets = NULL,
  includeEntityEtag = NULL,
  selectFileColumn = NULL,
  offset = NULL,
  limit = NULL,
  sort = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{entityId}}{entityId}

\item{\code{writeHeader}}{writeHeader}

\item{\code{includeRowIdAndRowVersion}}{includeRowIdAndRowVersion}

\item{\code{csvTableDescriptor}}{csvTableDescriptor}

\item{\code{fileName}}{fileName}

\item{\code{sql}}{sql}

\item{\code{additionalFilters}}{Appends additional filters to the SQL query. These are applied before facets. Filters within the list have an AND relationship. If a WHERE clause already exists on the SQL query or facets are selected, it will also be ANDed with the query generated by these additional filters.}

\item{\code{selectedFacets}}{The selected facet filters.}

\item{\code{includeEntityEtag}}{includeEntityEtag}

\item{\code{selectFileColumn}}{selectFileColumn}

\item{\code{offset}}{offset}

\item{\code{limit}}{limit}

\item{\code{sort}}{sort}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableDownloadFromTableRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableDownloadFromTableRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableDownloadFromTableRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableDownloadFromTableRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableDownloadFromTableRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableDownloadFromTableRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableDownloadFromTableRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableDownloadFromTableRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
