% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_download_from_table_result.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableDownloadFromTableResult}
\alias{OrgSagebionetworksRepoModelTableDownloadFromTableResult}
\title{OrgSagebionetworksRepoModelTableDownloadFromTableResult}
\format{
An \code{R6Class} generator object
}
\description{
The response body of an asynchronous CSV table download job.

OrgSagebionetworksRepoModelTableDownloadFromTableResult Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableDownloadFromTableResult
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{resultsFileHandleId}}{character [optional]}

\item{\code{tableId}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{headers}}{The list of SelectColumns that describes the rows of this set. list(\link{OrgSagebionetworksRepoModelTableSelectColumn}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-new}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-toJSON}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-fromJSON}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-toJSONString}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-fromJSONString}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-validateJSON}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-toString}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-isValid}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-print}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-clone}{\code{OrgSagebionetworksRepoModelTableDownloadFromTableResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableDownloadFromTableResult class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$new(
  concreteType,
  resultsFileHandleId = NULL,
  tableId = NULL,
  etag = NULL,
  headers = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{resultsFileHandleId}}{resultsFileHandleId}

\item{\code{tableId}}{tableId}

\item{\code{etag}}{etag}

\item{\code{headers}}{The list of SelectColumns that describes the rows of this set.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableDownloadFromTableResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableDownloadFromTableResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableDownloadFromTableResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableDownloadFromTableResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableDownloadFromTableResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableDownloadFromTableResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableDownloadFromTableResult and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableDownloadFromTableResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableDownloadFromTableResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableDownloadFromTableResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
