% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_entity_update_failure_code.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode}
\alias{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode}
\title{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelTableEntityUpdateFailureCode Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-new}{\code{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-toJSON}{\code{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-fromJSON}{\code{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-toJSONString}{\code{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-fromJSONString}{\code{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-clone}{\code{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableEntityUpdateFailureCode class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableEntityUpdateFailureCode in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableEntityUpdateFailureCode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableEntityUpdateFailureCode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableEntityUpdateFailureCode in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableEntityUpdateFailureCode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableEntityUpdateFailureCode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityUpdateFailureCode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityUpdateFailureCode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
