% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_entity_view.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableEntityView}
\alias{OrgSagebionetworksRepoModelTableEntityView}
\title{OrgSagebionetworksRepoModelTableEntityView}
\format{
An \code{R6Class} generator object
}
\description{
A view of Entities within a defined scope.

OrgSagebionetworksRepoModelTableEntityView Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableEntityView
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{parentId}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{versionNumber}}{integer [optional]}

\item{\code{versionLabel}}{character [optional]}

\item{\code{versionComment}}{character [optional]}

\item{\code{isLatestVersion}}{character [optional]}

\item{\code{columnIds}}{The list of ColumnModel IDs that define the schema of the object. list(character) [optional]}

\item{\code{isSearchEnabled}}{character [optional]}

\item{\code{viewTypeMask}}{integer [optional]}

\item{\code{type}}{character [optional]}

\item{\code{scopeIds}}{The list of container ids that define the scope of this view. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-new}{\code{OrgSagebionetworksRepoModelTableEntityView$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-toJSON}{\code{OrgSagebionetworksRepoModelTableEntityView$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-fromJSON}{\code{OrgSagebionetworksRepoModelTableEntityView$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-toJSONString}{\code{OrgSagebionetworksRepoModelTableEntityView$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-fromJSONString}{\code{OrgSagebionetworksRepoModelTableEntityView$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-validateJSON}{\code{OrgSagebionetworksRepoModelTableEntityView$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-toString}{\code{OrgSagebionetworksRepoModelTableEntityView$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-isValid}{\code{OrgSagebionetworksRepoModelTableEntityView$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableEntityView$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-print}{\code{OrgSagebionetworksRepoModelTableEntityView$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableEntityView-clone}{\code{OrgSagebionetworksRepoModelTableEntityView$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableEntityView class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$new(
  concreteType,
  name = NULL,
  description = NULL,
  id = NULL,
  etag = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  parentId = NULL,
  versionNumber = NULL,
  versionLabel = NULL,
  versionComment = NULL,
  isLatestVersion = NULL,
  columnIds = NULL,
  isSearchEnabled = NULL,
  viewTypeMask = NULL,
  type = NULL,
  scopeIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{parentId}}{parentId}

\item{\code{versionNumber}}{versionNumber}

\item{\code{versionLabel}}{versionLabel}

\item{\code{versionComment}}{versionComment}

\item{\code{isLatestVersion}}{isLatestVersion}

\item{\code{columnIds}}{The list of ColumnModel IDs that define the schema of the object.}

\item{\code{isSearchEnabled}}{isSearchEnabled}

\item{\code{viewTypeMask}}{viewTypeMask}

\item{\code{type}}{type}

\item{\code{scopeIds}}{The list of container ids that define the scope of this view.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableEntityView in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableEntityView
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableEntityView
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableEntityView in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableEntityView
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableEntityView
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableEntityView and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableEntityView
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableEntityView-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableEntityView-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableEntityView$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
