% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_facet_column_values_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest}
\alias{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest}
\title{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest}
\format{
An \code{R6Class} generator object
}
\description{
Contains the name and type of the column to be filtered, and the facet values to filter on.

OrgSagebionetworksRepoModelTableFacetColumnValuesRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableFacetColumnValuesRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{columnName}}{character [optional]}

\item{\code{jsonPath}}{character [optional]}

\item{\code{facetValues}}{The set of facet values that were selected list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-new}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-toJSON}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-fromJSON}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-toJSONString}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-validateJSON}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-toString}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-isValid}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-print}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-clone}{\code{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableFacetColumnValuesRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$new(
  concreteType,
  columnName = NULL,
  jsonPath = NULL,
  facetValues = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{columnName}}{columnName}

\item{\code{jsonPath}}{jsonPath}

\item{\code{facetValues}}{The set of facet values that were selected}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableFacetColumnValuesRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableFacetColumnValuesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableFacetColumnValuesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableFacetColumnValuesRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableFacetColumnValuesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableFacetColumnValuesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableFacetColumnValuesRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableFacetColumnValuesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableFacetColumnValuesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableFacetColumnValuesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
