% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_json_sub_column_model.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableJsonSubColumnModel}
\alias{OrgSagebionetworksRepoModelTableJsonSubColumnModel}
\title{OrgSagebionetworksRepoModelTableJsonSubColumnModel}
\format{
An \code{R6Class} generator object
}
\description{
Defines a single sub-column of a column of type JSON.  This can be used to enable the faceting of the sub-columns.

OrgSagebionetworksRepoModelTableJsonSubColumnModel Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableJsonSubColumnModel
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{columnType}}{character [optional]}

\item{\code{facetType}}{character [optional]}

\item{\code{jsonPath}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-new}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-toJSON}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-fromJSON}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-toJSONString}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-fromJSONString}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-validateJSON}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-toString}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-isValid}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-print}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-clone}{\code{OrgSagebionetworksRepoModelTableJsonSubColumnModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableJsonSubColumnModel class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$new(
  name = NULL,
  columnType = NULL,
  facetType = NULL,
  jsonPath = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{columnType}}{columnType}

\item{\code{facetType}}{facetType}

\item{\code{jsonPath}}{jsonPath}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableJsonSubColumnModel in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableJsonSubColumnModel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableJsonSubColumnModel
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableJsonSubColumnModel in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableJsonSubColumnModel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableJsonSubColumnModel
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableJsonSubColumnModel and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableJsonSubColumnModel
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableJsonSubColumnModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableJsonSubColumnModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
