% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_paginated_row_set.R
\docType{class}
\name{OrgSagebionetworksRepoModelTablePaginatedRowSet}
\alias{OrgSagebionetworksRepoModelTablePaginatedRowSet}
\title{OrgSagebionetworksRepoModelTablePaginatedRowSet}
\format{
An \code{R6Class} generator object
}
\description{
Paginated results of a RowSet

OrgSagebionetworksRepoModelTablePaginatedRowSet Class
}
\details{
Create a new OrgSagebionetworksRepoModelTablePaginatedRowSet
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{\link{OrgSagebionetworksRepoModelTableRowSet} [optional]}

\item{\code{totalNumberOfResults}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-new}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-toJSON}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-fromJSON}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-toJSONString}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-fromJSONString}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-validateJSON}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-toString}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-isValid}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-getInvalidFields}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-print}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePaginatedRowSet-clone}{\code{OrgSagebionetworksRepoModelTablePaginatedRowSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTablePaginatedRowSet class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$new(
  results = NULL,
  totalNumberOfResults = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{results}

\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTablePaginatedRowSet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTablePaginatedRowSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTablePaginatedRowSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTablePaginatedRowSet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTablePaginatedRowSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTablePaginatedRowSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTablePaginatedRowSet and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTablePaginatedRowSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePaginatedRowSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePaginatedRowSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePaginatedRowSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
