% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_partial_row_set.R
\docType{class}
\name{OrgSagebionetworksRepoModelTablePartialRowSet}
\alias{OrgSagebionetworksRepoModelTablePartialRowSet}
\title{OrgSagebionetworksRepoModelTablePartialRowSet}
\format{
An \code{R6Class} generator object
}
\description{
Represents a set of row of a TableEntity

OrgSagebionetworksRepoModelTablePartialRowSet Class
}
\details{
Create a new OrgSagebionetworksRepoModelTablePartialRowSet
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{tableId}}{character [optional]}

\item{\code{rows}}{The partial rows of this set. list(\link{OrgSagebionetworksRepoModelTablePartialRow}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-new}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-toJSON}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-fromJSON}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-toJSONString}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-fromJSONString}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-validateJSON}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-toString}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-isValid}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-getInvalidFields}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-print}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTablePartialRowSet-clone}{\code{OrgSagebionetworksRepoModelTablePartialRowSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTablePartialRowSet class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$new(
  concreteType,
  tableId = NULL,
  rows = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{tableId}}{tableId}

\item{\code{rows}}{The partial rows of this set.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTablePartialRowSet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTablePartialRowSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTablePartialRowSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTablePartialRowSet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTablePartialRowSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTablePartialRowSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTablePartialRowSet and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTablePartialRowSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTablePartialRowSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTablePartialRowSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTablePartialRowSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
