% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_query_cache_hit_event.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableQueryCacheHitEvent}
\alias{OrgSagebionetworksRepoModelTableQueryCacheHitEvent}
\title{OrgSagebionetworksRepoModelTableQueryCacheHitEvent}
\format{
An \code{R6Class} generator object
}
\description{
An event fired when a user runs a query that results in a cache hit.

OrgSagebionetworksRepoModelTableQueryCacheHitEvent Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableQueryCacheHitEvent
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{objectId}}{character [optional]}

\item{\code{objectVersion}}{integer [optional]}

\item{\code{objectType}}{character [optional]}

\item{\code{timestamp}}{character [optional]}

\item{\code{queryRequestHash}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-new}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-toJSON}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-fromJSON}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-toJSONString}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-fromJSONString}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-validateJSON}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-toString}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-isValid}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-print}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-clone}{\code{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableQueryCacheHitEvent class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$new(
  concreteType,
  objectId = NULL,
  objectVersion = NULL,
  objectType = NULL,
  timestamp = NULL,
  queryRequestHash = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{objectId}}{objectId}

\item{\code{objectVersion}}{objectVersion}

\item{\code{objectType}}{objectType}

\item{\code{timestamp}}{timestamp}

\item{\code{queryRequestHash}}{queryRequestHash}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableQueryCacheHitEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableQueryCacheHitEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableQueryCacheHitEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableQueryCacheHitEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableQueryCacheHitEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableQueryCacheHitEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableQueryCacheHitEvent and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableQueryCacheHitEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableQueryCacheHitEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableQueryCacheHitEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
