% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_row_reference_set.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableRowReferenceSet}
\alias{OrgSagebionetworksRepoModelTableRowReferenceSet}
\title{OrgSagebionetworksRepoModelTableRowReferenceSet}
\format{
An \code{R6Class} generator object
}
\description{
Represents a set of RowReferences of a TableEntity

OrgSagebionetworksRepoModelTableRowReferenceSet Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableRowReferenceSet
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tableId}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{headers}}{The list of ColumnModels ID that describes the rows of this set. list(\link{OrgSagebionetworksRepoModelTableSelectColumn}) [optional]}

\item{\code{rows}}{Each RowReference of this list refers to a single version of a single row of a TableEntity. list(\link{OrgSagebionetworksRepoModelTableRowReference}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-new}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-toJSON}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-fromJSON}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-toJSONString}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-fromJSONString}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-validateJSON}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-toString}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-isValid}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-print}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSet-clone}{\code{OrgSagebionetworksRepoModelTableRowReferenceSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableRowReferenceSet class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$new(
  tableId = NULL,
  etag = NULL,
  headers = NULL,
  rows = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tableId}}{tableId}

\item{\code{etag}}{etag}

\item{\code{headers}}{The list of ColumnModels ID that describes the rows of this set.}

\item{\code{rows}}{Each RowReference of this list refers to a single version of a single row of a TableEntity.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableRowReferenceSet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableRowReferenceSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableRowReferenceSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableRowReferenceSet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableRowReferenceSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableRowReferenceSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableRowReferenceSet and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableRowReferenceSet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
