% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_row_reference_set_results.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableRowReferenceSetResults}
\alias{OrgSagebionetworksRepoModelTableRowReferenceSetResults}
\title{OrgSagebionetworksRepoModelTableRowReferenceSetResults}
\format{
An \code{R6Class} generator object
}
\description{
Response of an asynchronous that returns RowReferenceSet.

OrgSagebionetworksRepoModelTableRowReferenceSetResults Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableRowReferenceSetResults
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{rowReferenceSet}}{\link{OrgSagebionetworksRepoModelTableRowReferenceSet} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-new}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-toJSON}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-fromJSON}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-toJSONString}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-fromJSONString}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-validateJSON}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-toString}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-isValid}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-print}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-clone}{\code{OrgSagebionetworksRepoModelTableRowReferenceSetResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableRowReferenceSetResults class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$new(
  concreteType,
  rowReferenceSet = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{rowReferenceSet}}{rowReferenceSet}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableRowReferenceSetResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableRowReferenceSetResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableRowReferenceSetResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableRowReferenceSetResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableRowReferenceSetResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableRowReferenceSetResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableRowReferenceSetResults and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableRowReferenceSetResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableRowReferenceSetResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableRowReferenceSetResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
