% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_sparse_change_set_dto.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableSparseChangeSetDto}
\alias{OrgSagebionetworksRepoModelTableSparseChangeSetDto}
\title{OrgSagebionetworksRepoModelTableSparseChangeSetDto}
\format{
An \code{R6Class} generator object
}
\description{
Data Transfer Object (DTO) for a SparseChangeSet

OrgSagebionetworksRepoModelTableSparseChangeSetDto Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableSparseChangeSetDto
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tableId}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{columnIds}}{The column IDs that define the schema of this change set. list(character) [optional]}

\item{\code{rows}}{The partial rows of this set. list(\link{OrgSagebionetworksRepoModelTableSparseRowDto}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-new}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-toJSON}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-fromJSON}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-toJSONString}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-fromJSONString}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-validateJSON}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-toString}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-isValid}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-print}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-clone}{\code{OrgSagebionetworksRepoModelTableSparseChangeSetDto$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableSparseChangeSetDto class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$new(
  tableId = NULL,
  etag = NULL,
  columnIds = NULL,
  rows = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tableId}}{tableId}

\item{\code{etag}}{etag}

\item{\code{columnIds}}{The column IDs that define the schema of this change set.}

\item{\code{rows}}{The partial rows of this set.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableSparseChangeSetDto in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableSparseChangeSetDto
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableSparseChangeSetDto
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableSparseChangeSetDto in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableSparseChangeSetDto
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableSparseChangeSetDto
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableSparseChangeSetDto and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableSparseChangeSetDto
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseChangeSetDto-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseChangeSetDto$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
