% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_sparse_row_dto.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableSparseRowDto}
\alias{OrgSagebionetworksRepoModelTableSparseRowDto}
\title{OrgSagebionetworksRepoModelTableSparseRowDto}
\format{
An \code{R6Class} generator object
}
\description{
Represents a single partial row of a TableEntity

OrgSagebionetworksRepoModelTableSparseRowDto Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableSparseRowDto
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{etag}}{character [optional]}

\item{\code{rowId}}{integer [optional]}

\item{\code{versionNumber}}{integer [optional]}

\item{\code{values}}{The values for columns of this row as a map. The key is the columnId and value is the cell value named list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-new}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-toJSON}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-fromJSON}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-toJSONString}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-fromJSONString}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-validateJSON}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-toString}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-isValid}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-print}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableSparseRowDto-clone}{\code{OrgSagebionetworksRepoModelTableSparseRowDto$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableSparseRowDto class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$new(
  etag = NULL,
  rowId = NULL,
  versionNumber = NULL,
  values = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{etag}}{etag}

\item{\code{rowId}}{rowId}

\item{\code{versionNumber}}{versionNumber}

\item{\code{values}}{The values for columns of this row as a map. The key is the columnId and value is the cell value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableSparseRowDto in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableSparseRowDto
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableSparseRowDto
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableSparseRowDto in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableSparseRowDto
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableSparseRowDto
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableSparseRowDto and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableSparseRowDto
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableSparseRowDto-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableSparseRowDto-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableSparseRowDto$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
