% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_table_file_handle_results.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableTableFileHandleResults}
\alias{OrgSagebionetworksRepoModelTableTableFileHandleResults}
\title{OrgSagebionetworksRepoModelTableTableFileHandleResults}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for TableFileHandleResults.

OrgSagebionetworksRepoModelTableTableFileHandleResults Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableTableFileHandleResults
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tableId}}{character [optional]}

\item{\code{headers}}{The list of ColumnModels ID that describes the rows of this set. list(\link{OrgSagebionetworksRepoModelTableSelectColumn}) [optional]}

\item{\code{rows}}{For each row a list of file handles for each requested column list(\link{OrgSagebionetworksRepoModelFileFileHandleResults}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-new}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-toJSON}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-fromJSON}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-toJSONString}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-fromJSONString}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-validateJSON}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-toString}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-isValid}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-print}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableFileHandleResults-clone}{\code{OrgSagebionetworksRepoModelTableTableFileHandleResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableTableFileHandleResults class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$new(
  tableId = NULL,
  headers = NULL,
  rows = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tableId}}{tableId}

\item{\code{headers}}{The list of ColumnModels ID that describes the rows of this set.}

\item{\code{rows}}{For each row a list of file handles for each requested column}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableFileHandleResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableFileHandleResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableFileHandleResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableFileHandleResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableFileHandleResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableFileHandleResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableTableFileHandleResults and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableTableFileHandleResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableFileHandleResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableFileHandleResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableFileHandleResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
