% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_table_row_change.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableTableRowChange}
\alias{OrgSagebionetworksRepoModelTableTableRowChange}
\title{OrgSagebionetworksRepoModelTableTableRowChange}
\format{
An \code{R6Class} generator object
}
\description{
Metadata about Table row change set.

OrgSagebionetworksRepoModelTableTableRowChange Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableTableRowChange
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{tableId}}{character [optional]}

\item{\code{rowVersion}}{integer [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{bucket}}{character [optional]}

\item{\code{keyNew}}{character [optional]}

\item{\code{rowCount}}{integer [optional]}

\item{\code{changeType}}{character [optional]}

\item{\code{transactionId}}{integer [optional]}

\item{\code{hasFileRefs}}{character [optional]}

\item{\code{isSearchEnabled}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-new}{\code{OrgSagebionetworksRepoModelTableTableRowChange$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-toJSON}{\code{OrgSagebionetworksRepoModelTableTableRowChange$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-fromJSON}{\code{OrgSagebionetworksRepoModelTableTableRowChange$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-toJSONString}{\code{OrgSagebionetworksRepoModelTableTableRowChange$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-fromJSONString}{\code{OrgSagebionetworksRepoModelTableTableRowChange$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-validateJSON}{\code{OrgSagebionetworksRepoModelTableTableRowChange$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-toString}{\code{OrgSagebionetworksRepoModelTableTableRowChange$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-isValid}{\code{OrgSagebionetworksRepoModelTableTableRowChange$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableTableRowChange$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-print}{\code{OrgSagebionetworksRepoModelTableTableRowChange$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableRowChange-clone}{\code{OrgSagebionetworksRepoModelTableTableRowChange$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableTableRowChange class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$new(
  id = NULL,
  tableId = NULL,
  rowVersion = NULL,
  etag = NULL,
  createdOn = NULL,
  createdBy = NULL,
  bucket = NULL,
  keyNew = NULL,
  rowCount = NULL,
  changeType = NULL,
  transactionId = NULL,
  hasFileRefs = NULL,
  isSearchEnabled = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{tableId}}{tableId}

\item{\code{rowVersion}}{rowVersion}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{bucket}}{bucket}

\item{\code{keyNew}}{keyNew}

\item{\code{rowCount}}{rowCount}

\item{\code{changeType}}{changeType}

\item{\code{transactionId}}{transactionId}

\item{\code{hasFileRefs}}{hasFileRefs}

\item{\code{isSearchEnabled}}{isSearchEnabled}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableRowChange in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableRowChange
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableRowChange
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableRowChange in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableRowChange
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableRowChange
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableTableRowChange and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableTableRowChange
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableRowChange-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableRowChange-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableRowChange$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
