% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_table_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableTableStatus}
\alias{OrgSagebionetworksRepoModelTableTableStatus}
\title{OrgSagebionetworksRepoModelTableTableStatus}
\format{
An \code{R6Class} generator object
}
\description{
Object used to track the status of a Table.

OrgSagebionetworksRepoModelTableTableStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableTableStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{state}}{character [optional]}

\item{\code{tableId}}{character [optional]}

\item{\code{version}}{integer [optional]}

\item{\code{startedOn}}{character [optional]}

\item{\code{changedOn}}{character [optional]}

\item{\code{resetToken}}{character [optional]}

\item{\code{lastTableChangeEtag}}{character [optional]}

\item{\code{progressMessage}}{character [optional]}

\item{\code{progressCurrent}}{integer [optional]}

\item{\code{progressTotal}}{integer [optional]}

\item{\code{errorMessage}}{character [optional]}

\item{\code{errorDetails}}{character [optional]}

\item{\code{totalTimeMS}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-new}{\code{OrgSagebionetworksRepoModelTableTableStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-toJSON}{\code{OrgSagebionetworksRepoModelTableTableStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-fromJSON}{\code{OrgSagebionetworksRepoModelTableTableStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-toJSONString}{\code{OrgSagebionetworksRepoModelTableTableStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelTableTableStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-validateJSON}{\code{OrgSagebionetworksRepoModelTableTableStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-toString}{\code{OrgSagebionetworksRepoModelTableTableStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-isValid}{\code{OrgSagebionetworksRepoModelTableTableStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableTableStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-print}{\code{OrgSagebionetworksRepoModelTableTableStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatus-clone}{\code{OrgSagebionetworksRepoModelTableTableStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableTableStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$new(
  state = NULL,
  tableId = NULL,
  version = NULL,
  startedOn = NULL,
  changedOn = NULL,
  resetToken = NULL,
  lastTableChangeEtag = NULL,
  progressMessage = NULL,
  progressCurrent = NULL,
  progressTotal = NULL,
  errorMessage = NULL,
  errorDetails = NULL,
  totalTimeMS = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{state}

\item{\code{tableId}}{tableId}

\item{\code{version}}{version}

\item{\code{startedOn}}{startedOn}

\item{\code{changedOn}}{changedOn}

\item{\code{resetToken}}{resetToken}

\item{\code{lastTableChangeEtag}}{lastTableChangeEtag}

\item{\code{progressMessage}}{progressMessage}

\item{\code{progressCurrent}}{progressCurrent}

\item{\code{progressTotal}}{progressTotal}

\item{\code{errorMessage}}{errorMessage}

\item{\code{errorDetails}}{errorDetails}

\item{\code{totalTimeMS}}{totalTimeMS}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableTableStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableTableStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
