% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_table_status_change_event.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableTableStatusChangeEvent}
\alias{OrgSagebionetworksRepoModelTableTableStatusChangeEvent}
\title{OrgSagebionetworksRepoModelTableTableStatusChangeEvent}
\format{
An \code{R6Class} generator object
}
\description{
Object used to propagate table status changes

OrgSagebionetworksRepoModelTableTableStatusChangeEvent Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableTableStatusChangeEvent
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{objectId}}{character [optional]}

\item{\code{objectVersion}}{integer [optional]}

\item{\code{objectType}}{character [optional]}

\item{\code{timestamp}}{character [optional]}

\item{\code{state}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-new}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-toJSON}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-fromJSON}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-toJSONString}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-fromJSONString}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-validateJSON}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-toString}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-isValid}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-print}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-clone}{\code{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableTableStatusChangeEvent class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$new(
  concreteType,
  objectId = NULL,
  objectVersion = NULL,
  objectType = NULL,
  timestamp = NULL,
  state = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{objectId}}{objectId}

\item{\code{objectVersion}}{objectVersion}

\item{\code{objectType}}{objectType}

\item{\code{timestamp}}{timestamp}

\item{\code{state}}{state}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableStatusChangeEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableStatusChangeEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableStatusChangeEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableStatusChangeEvent in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableStatusChangeEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableStatusChangeEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableTableStatusChangeEvent and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableTableStatusChangeEvent
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableStatusChangeEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableStatusChangeEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
