% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_table_update_transaction_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest}
\alias{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest}
\title{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest}
\format{
An \code{R6Class} generator object
}
\description{
An AsynchronousRequestBody to used make multiple changes to a table or view as a single 'transaction'. All changes will either succeed or fail as a unit.

OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{entityId}}{character [optional]}

\item{\code{changes}}{List of changes that describes schema and/or row changes to a table. For a view the changes are eventually consistent. list(\link{OrgSagebionetworksRepoModelTableTableUpdateRequest}) [optional]}

\item{\code{createSnapshot}}{character [optional]}

\item{\code{snapshotOptions}}{\link{OrgSagebionetworksRepoModelTableSnapshotRequest} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-new}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-toJSON}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-fromJSON}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-toJSONString}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-validateJSON}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-toString}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-isValid}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-print}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-clone}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$new(
  concreteType,
  entityId = NULL,
  changes = NULL,
  createSnapshot = NULL,
  snapshotOptions = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{entityId}}{entityId}

\item{\code{changes}}{List of changes that describes schema and/or row changes to a table. For a view the changes are eventually consistent.}

\item{\code{createSnapshot}}{createSnapshot}

\item{\code{snapshotOptions}}{snapshotOptions}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionRequest$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
