% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_table_update_transaction_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse}
\alias{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse}
\title{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse}
\format{
An \code{R6Class} generator object
}
\description{
An AsynchronousResponseBody returned from a table update transaction.

OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{results}}{List of responses. There will be one response for each request in the transaction. list(\link{OrgSagebionetworksRepoModelTableTableUpdateResponse}) [optional]}

\item{\code{snapshotVersionNumber}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-new}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-toJSON}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-fromJSON}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-toJSONString}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-validateJSON}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-toString}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-isValid}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-print}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-clone}{\code{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$new(
  concreteType,
  results = NULL,
  snapshotVersionNumber = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{results}}{List of responses. There will be one response for each request in the transaction.}

\item{\code{snapshotVersionNumber}}{snapshotVersionNumber}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTableUpdateTransactionResponse$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
