% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_text_matches_query_filter.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter}
\alias{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter}
\title{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter}
\format{
An \code{R6Class} generator object
}
\description{
A filter to apply a TEXT_MATCHES predicate (full text search) to the query using the given searchExpression as value.

OrgSagebionetworksRepoModelTableTextMatchesQueryFilter Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableTextMatchesQueryFilter
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{isDefiningCondition}}{character [optional]}

\item{\code{searchExpression}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-new}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-toJSON}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-fromJSON}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-toJSONString}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-fromJSONString}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-validateJSON}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-toString}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-isValid}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-print}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-clone}{\code{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableTextMatchesQueryFilter class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$new(
  concreteType,
  isDefiningCondition = NULL,
  searchExpression = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{isDefiningCondition}}{isDefiningCondition}

\item{\code{searchExpression}}{searchExpression}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTextMatchesQueryFilter in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTextMatchesQueryFilter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTextMatchesQueryFilter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableTextMatchesQueryFilter in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableTextMatchesQueryFilter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableTextMatchesQueryFilter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableTextMatchesQueryFilter and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableTextMatchesQueryFilter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableTextMatchesQueryFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableTextMatchesQueryFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
