% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_upload_to_table_preview_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest}
\alias{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest}
\title{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest}
\format{
An \code{R6Class} generator object
}
\description{
Request for a preview of an upload to a Table.

OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{uploadFileHandleId}}{character [optional]}

\item{\code{linesToSkip}}{integer [optional]}

\item{\code{csvTableDescriptor}}{\link{OrgSagebionetworksRepoModelTableCsvTableDescriptor} [optional]}

\item{\code{doFullFileScan}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-new}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-toJSON}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-fromJSON}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-toJSONString}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-validateJSON}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-toString}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-isValid}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-print}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-clone}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$new(
  concreteType,
  uploadFileHandleId = NULL,
  linesToSkip = NULL,
  csvTableDescriptor = NULL,
  doFullFileScan = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{uploadFileHandleId}}{uploadFileHandleId}

\item{\code{linesToSkip}}{linesToSkip}

\item{\code{csvTableDescriptor}}{csvTableDescriptor}

\item{\code{doFullFileScan}}{doFullFileScan}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
