% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_upload_to_table_preview_result.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult}
\alias{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult}
\title{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult}
\format{
An \code{R6Class} generator object
}
\description{
Resulting preview of a CSV upload to a Table.

OrgSagebionetworksRepoModelTableUploadToTablePreviewResult Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableUploadToTablePreviewResult
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{suggestedColumns}}{The list of ColumnModels that are compatible with the CSV file.  This is more accurate for a full scan than a partial scan. list(\link{OrgSagebionetworksRepoModelTableColumnModel}) [optional]}

\item{\code{sampleRows}}{A sampling of the first few rows from the CSV as they would be applied with the provided parameters. list(\link{OrgSagebionetworksRepoModelTableRow}) [optional]}

\item{\code{rowsScanned}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-new}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-toJSON}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-fromJSON}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-toJSONString}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-fromJSONString}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-validateJSON}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-toString}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-isValid}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-print}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-clone}{\code{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableUploadToTablePreviewResult class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$new(
  concreteType,
  suggestedColumns = NULL,
  sampleRows = NULL,
  rowsScanned = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{suggestedColumns}}{The list of ColumnModels that are compatible with the CSV file.  This is more accurate for a full scan than a partial scan.}

\item{\code{sampleRows}}{A sampling of the first few rows from the CSV as they would be applied with the provided parameters.}

\item{\code{rowsScanned}}{rowsScanned}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableUploadToTablePreviewResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableUploadToTablePreviewResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableUploadToTablePreviewResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableUploadToTablePreviewResult in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableUploadToTablePreviewResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableUploadToTablePreviewResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableUploadToTablePreviewResult and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableUploadToTablePreviewResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTablePreviewResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTablePreviewResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
