% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_upload_to_table_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableUploadToTableRequest}
\alias{OrgSagebionetworksRepoModelTableUploadToTableRequest}
\title{OrgSagebionetworksRepoModelTableUploadToTableRequest}
\format{
An \code{R6Class} generator object
}
\description{
A CSV table upload request.

OrgSagebionetworksRepoModelTableUploadToTableRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableUploadToTableRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{entityId}}{character [optional]}

\item{\code{tableId}}{character [optional]}

\item{\code{uploadFileHandleId}}{character [optional]}

\item{\code{updateEtag}}{character [optional]}

\item{\code{linesToSkip}}{integer [optional]}

\item{\code{csvTableDescriptor}}{\link{OrgSagebionetworksRepoModelTableCsvTableDescriptor} [optional]}

\item{\code{columnIds}}{Deprecated. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-new}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-toJSON}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-fromJSON}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-toJSONString}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-validateJSON}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-toString}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-isValid}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-print}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableUploadToTableRequest-clone}{\code{OrgSagebionetworksRepoModelTableUploadToTableRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableUploadToTableRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$new(
  concreteType,
  entityId = NULL,
  tableId = NULL,
  uploadFileHandleId = NULL,
  updateEtag = NULL,
  linesToSkip = NULL,
  csvTableDescriptor = NULL,
  columnIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{entityId}}{entityId}

\item{\code{tableId}}{tableId}

\item{\code{uploadFileHandleId}}{uploadFileHandleId}

\item{\code{updateEtag}}{updateEtag}

\item{\code{linesToSkip}}{linesToSkip}

\item{\code{csvTableDescriptor}}{csvTableDescriptor}

\item{\code{columnIds}}{Deprecated.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableUploadToTableRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableUploadToTableRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableUploadToTableRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableUploadToTableRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableUploadToTableRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableUploadToTableRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableUploadToTableRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableUploadToTableRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableUploadToTableRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableUploadToTableRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableUploadToTableRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
