% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_validate_defining_sql_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest}
\alias{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest}
\title{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest}
\format{
An \code{R6Class} generator object
}
\description{
The request body for validateDefiningSql

OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{definingSql}}{character [optional]}

\item{\code{entityType}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-new}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-toJSON}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-fromJSON}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-toJSONString}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-validateJSON}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-toString}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-isValid}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-print}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-clone}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$new(
  definingSql = NULL,
  entityType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{definingSql}}{definingSql}

\item{\code{entityType}}{entityType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
