% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_validate_defining_sql_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse}
\alias{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse}
\title{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse}
\format{
An \code{R6Class} generator object
}
\description{
The response body from validateDefiningSql

OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isValid}}{character [optional]}

\item{\code{invalidReason}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-new}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-toJSON}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-fromJSON}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-toJSONString}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-validateJSON}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-toString}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-print}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-clone}{\code{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$new(
  isValid = NULL,
  invalidReason = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{isValid}

\item{\code{invalidReason}}{invalidReason}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return true if the values in all fields are valid.


Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableValidateDefiningSqlResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
