% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_view_column_model_request.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableViewColumnModelRequest}
\alias{OrgSagebionetworksRepoModelTableViewColumnModelRequest}
\title{OrgSagebionetworksRepoModelTableViewColumnModelRequest}
\format{
An \code{R6Class} generator object
}
\description{
An asynchronous request to compute the column models from the annotations in a given scope, note that the result won't include the default column model.

OrgSagebionetworksRepoModelTableViewColumnModelRequest Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableViewColumnModelRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{viewScope}}{\link{OrgSagebionetworksRepoModelTableViewScope} [optional]}

\item{\code{includeDerivedAnnotations}}{character [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-new}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-toJSON}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-fromJSON}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-toJSONString}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-fromJSONString}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-validateJSON}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-toString}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-isValid}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-print}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-clone}{\code{OrgSagebionetworksRepoModelTableViewColumnModelRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableViewColumnModelRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$new(
  concreteType,
  viewScope = NULL,
  includeDerivedAnnotations = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{viewScope}}{viewScope}

\item{\code{includeDerivedAnnotations}}{includeDerivedAnnotations}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableViewColumnModelRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableViewColumnModelRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableViewColumnModelRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableViewColumnModelRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableViewColumnModelRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableViewColumnModelRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableViewColumnModelRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableViewColumnModelRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
