% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_table_view_column_model_response.R
\docType{class}
\name{OrgSagebionetworksRepoModelTableViewColumnModelResponse}
\alias{OrgSagebionetworksRepoModelTableViewColumnModelResponse}
\title{OrgSagebionetworksRepoModelTableViewColumnModelResponse}
\format{
An \code{R6Class} generator object
}
\description{
Response for the <a href=\"${org.sagebionetworks.repo.model.table.ViewColumnModelRequest}\">ViewColumnModelRequest</a> containing a page of suggested column models

OrgSagebionetworksRepoModelTableViewColumnModelResponse Class
}
\details{
Create a new OrgSagebionetworksRepoModelTableViewColumnModelResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{concreteType}}{character}

\item{\code{results}}{A page of <a href=\"${org.sagebionetworks.repo.model.table.ColumnModel}\">ColumnModels</a> list(\link{OrgSagebionetworksRepoModelTableColumnModel}) [optional]}

\item{\code{nextPageToken}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-new}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-toJSON}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-fromJSON}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-toJSONString}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-fromJSONString}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-validateJSON}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-toString}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-isValid}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-getInvalidFields}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-print}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-clone}{\code{OrgSagebionetworksRepoModelTableViewColumnModelResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTableViewColumnModelResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$new(
  concreteType,
  results = NULL,
  nextPageToken = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{results}}{A page of <a href=\"${org.sagebionetworks.repo.model.table.ColumnModel}\">ColumnModels</a>}

\item{\code{nextPageToken}}{nextPageToken}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableViewColumnModelResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableViewColumnModelResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableViewColumnModelResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTableViewColumnModelResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTableViewColumnModelResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTableViewColumnModelResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTableViewColumnModelResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTableViewColumnModelResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTableViewColumnModelResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTableViewColumnModelResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
