% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_team_membership_status.R
\docType{class}
\name{OrgSagebionetworksRepoModelTeamMembershipStatus}
\alias{OrgSagebionetworksRepoModelTeamMembershipStatus}
\title{OrgSagebionetworksRepoModelTeamMembershipStatus}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for the possible status of a User with respect to Team membership.

OrgSagebionetworksRepoModelTeamMembershipStatus Class
}
\details{
Create a new OrgSagebionetworksRepoModelTeamMembershipStatus
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{teamId}}{character [optional]}

\item{\code{userId}}{character [optional]}

\item{\code{isMember}}{character [optional]}

\item{\code{hasOpenInvitation}}{character [optional]}

\item{\code{hasOpenRequest}}{character [optional]}

\item{\code{canJoin}}{character [optional]}

\item{\code{membershipApprovalRequired}}{character [optional]}

\item{\code{hasUnmetAccessRequirement}}{character [optional]}

\item{\code{canSendEmail}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-new}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-toJSON}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-fromJSON}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-toJSONString}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-fromJSONString}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-validateJSON}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-toString}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-isValid}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-getInvalidFields}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-print}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMembershipStatus-clone}{\code{OrgSagebionetworksRepoModelTeamMembershipStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTeamMembershipStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$new(
  teamId = NULL,
  userId = NULL,
  isMember = NULL,
  hasOpenInvitation = NULL,
  hasOpenRequest = NULL,
  canJoin = NULL,
  membershipApprovalRequired = NULL,
  hasUnmetAccessRequirement = NULL,
  canSendEmail = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{teamId}}{teamId}

\item{\code{userId}}{userId}

\item{\code{isMember}}{isMember}

\item{\code{hasOpenInvitation}}{hasOpenInvitation}

\item{\code{hasOpenRequest}}{hasOpenRequest}

\item{\code{canJoin}}{canJoin}

\item{\code{membershipApprovalRequired}}{membershipApprovalRequired}

\item{\code{hasUnmetAccessRequirement}}{hasUnmetAccessRequirement}

\item{\code{canSendEmail}}{canSendEmail}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTeamMembershipStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTeamMembershipStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTeamMembershipStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTeamMembershipStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTeamMembershipStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTeamMembershipStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTeamMembershipStatus and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTeamMembershipStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMembershipStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMembershipStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMembershipStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
