% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_team_member_type_filter_options.R
\docType{class}
\name{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions}
\alias{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions}
\title{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-new}{\code{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-toJSON}{\code{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-fromJSON}{\code{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-toJSONString}{\code{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-fromJSONString}{\code{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-clone}{\code{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
