% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_terms_of_use_access_requirement.R
\docType{class}
\name{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement}
\alias{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement}
\title{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for Terms of Use Access Requirement, used for a 'tier 2' Access Requirement

OrgSagebionetworksRepoModelTermsOfUseAccessRequirement Class
}
\details{
Create a new OrgSagebionetworksRepoModelTermsOfUseAccessRequirement
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{versionNumber}}{integer [optional]}

\item{\code{id}}{integer [optional]}

\item{\code{description}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{subjectsDefinedByAnnotations}}{character [optional]}

\item{\code{subjectIds}}{The IDs of the items controlled by this Access Requirement when 'subjectsDefinedByAnnotations=false'. This property is mutually exclusive with 'subjectsDefinedByAnnotations'.  When 'subjectsDefinedByAnnotations=true' then this property must be empty or excluded.  Required when creating or updating and 'subjectsDefinedByAnnotations=false' or 'subjectsDefinedByAnnotations' is excluded. list(\link{OrgSagebionetworksRepoModelRestrictableObjectDescriptor}) [optional]}

\item{\code{accessType}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{termsOfUse}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-new}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$new()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-toJSON}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-fromJSON}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-toJSONString}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-fromJSONString}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-validateJSON}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-toString}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-isValid}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-getInvalidFields}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-print}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$print()}}
\item \href{#method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-clone}{\code{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelTermsOfUseAccessRequirement class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$new(
  concreteType,
  versionNumber = NULL,
  id = NULL,
  description = NULL,
  name = NULL,
  etag = NULL,
  createdOn = NULL,
  modifiedOn = NULL,
  createdBy = NULL,
  modifiedBy = NULL,
  subjectsDefinedByAnnotations = NULL,
  subjectIds = NULL,
  accessType = NULL,
  termsOfUse = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{versionNumber}}{versionNumber}

\item{\code{id}}{id}

\item{\code{description}}{description}

\item{\code{name}}{name}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{subjectsDefinedByAnnotations}}{subjectsDefinedByAnnotations}

\item{\code{subjectIds}}{The IDs of the items controlled by this Access Requirement when 'subjectsDefinedByAnnotations=false'. This property is mutually exclusive with 'subjectsDefinedByAnnotations'.  When 'subjectsDefinedByAnnotations=true' then this property must be empty or excluded.  Required when creating or updating and 'subjectsDefinedByAnnotations=false' or 'subjectsDefinedByAnnotations' is excluded.}

\item{\code{accessType}}{accessType}

\item{\code{termsOfUse}}{termsOfUse}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTermsOfUseAccessRequirement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTermsOfUseAccessRequirement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTermsOfUseAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelTermsOfUseAccessRequirement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelTermsOfUseAccessRequirement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelTermsOfUseAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelTermsOfUseAccessRequirement and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelTermsOfUseAccessRequirement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelTermsOfUseAccessRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelTermsOfUseAccessRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
