% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_user_group.R
\docType{class}
\name{OrgSagebionetworksRepoModelUserGroup}
\alias{OrgSagebionetworksRepoModelUserGroup}
\title{OrgSagebionetworksRepoModelUserGroup}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for UserGroup POJO

OrgSagebionetworksRepoModelUserGroup Class
}
\details{
Create a new OrgSagebionetworksRepoModelUserGroup
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{isIndividual}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{creationDate}}{character [optional]}

\item{\code{uri}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-new}{\code{OrgSagebionetworksRepoModelUserGroup$new()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-toJSON}{\code{OrgSagebionetworksRepoModelUserGroup$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-fromJSON}{\code{OrgSagebionetworksRepoModelUserGroup$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-toJSONString}{\code{OrgSagebionetworksRepoModelUserGroup$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-fromJSONString}{\code{OrgSagebionetworksRepoModelUserGroup$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-validateJSON}{\code{OrgSagebionetworksRepoModelUserGroup$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-toString}{\code{OrgSagebionetworksRepoModelUserGroup$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-isValid}{\code{OrgSagebionetworksRepoModelUserGroup$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-getInvalidFields}{\code{OrgSagebionetworksRepoModelUserGroup$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-print}{\code{OrgSagebionetworksRepoModelUserGroup$print()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroup-clone}{\code{OrgSagebionetworksRepoModelUserGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelUserGroup class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$new(
  id = NULL,
  isIndividual = NULL,
  etag = NULL,
  creationDate = NULL,
  uri = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{isIndividual}}{isIndividual}

\item{\code{etag}}{etag}

\item{\code{creationDate}}{creationDate}

\item{\code{uri}}{uri}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserGroup in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserGroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserGroup
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserGroup in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserGroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserGroup
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelUserGroup and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelUserGroup
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
