% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_user_group_header.R
\docType{class}
\name{OrgSagebionetworksRepoModelUserGroupHeader}
\alias{OrgSagebionetworksRepoModelUserGroupHeader}
\title{OrgSagebionetworksRepoModelUserGroupHeader}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for UserHeader POJO

OrgSagebionetworksRepoModelUserGroupHeader Class
}
\details{
Create a new OrgSagebionetworksRepoModelUserGroupHeader
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ownerId}}{character [optional]}

\item{\code{firstName}}{character [optional]}

\item{\code{lastName}}{character [optional]}

\item{\code{userName}}{character [optional]}

\item{\code{email}}{character [optional]}

\item{\code{displayName}}{character [optional]}

\item{\code{isIndividual}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-new}{\code{OrgSagebionetworksRepoModelUserGroupHeader$new()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-toJSON}{\code{OrgSagebionetworksRepoModelUserGroupHeader$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-fromJSON}{\code{OrgSagebionetworksRepoModelUserGroupHeader$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-toJSONString}{\code{OrgSagebionetworksRepoModelUserGroupHeader$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-fromJSONString}{\code{OrgSagebionetworksRepoModelUserGroupHeader$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-validateJSON}{\code{OrgSagebionetworksRepoModelUserGroupHeader$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-toString}{\code{OrgSagebionetworksRepoModelUserGroupHeader$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-isValid}{\code{OrgSagebionetworksRepoModelUserGroupHeader$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-getInvalidFields}{\code{OrgSagebionetworksRepoModelUserGroupHeader$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-print}{\code{OrgSagebionetworksRepoModelUserGroupHeader$print()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeader-clone}{\code{OrgSagebionetworksRepoModelUserGroupHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelUserGroupHeader class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$new(
  ownerId = NULL,
  firstName = NULL,
  lastName = NULL,
  userName = NULL,
  email = NULL,
  displayName = NULL,
  isIndividual = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ownerId}}{ownerId}

\item{\code{firstName}}{firstName}

\item{\code{lastName}}{lastName}

\item{\code{userName}}{userName}

\item{\code{email}}{email}

\item{\code{displayName}}{displayName}

\item{\code{isIndividual}}{isIndividual}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserGroupHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserGroupHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserGroupHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserGroupHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserGroupHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserGroupHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelUserGroupHeader and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelUserGroupHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
