% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_user_group_header_response_page.R
\docType{class}
\name{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage}
\alias{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage}
\title{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage}
\format{
An \code{R6Class} generator object
}
\description{
A single page of a users/groups info query response.

OrgSagebionetworksRepoModelUserGroupHeaderResponsePage Class
}
\details{
Create a new OrgSagebionetworksRepoModelUserGroupHeaderResponsePage
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{children}}{The list of children that match the requested concept. list(\link{OrgSagebionetworksRepoModelUserGroupHeader}) [optional]}

\item{\code{prefixFilter}}{character [optional]}

\item{\code{totalNumberOfResults}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-new}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$new()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-toJSON}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-fromJSON}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-toJSONString}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-fromJSONString}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-validateJSON}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-toString}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-isValid}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-getInvalidFields}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-print}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$print()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-clone}{\code{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelUserGroupHeaderResponsePage class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$new(
  children = NULL,
  prefixFilter = NULL,
  totalNumberOfResults = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{children}}{The list of children that match the requested concept.}

\item{\code{prefixFilter}}{prefixFilter}

\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserGroupHeaderResponsePage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserGroupHeaderResponsePage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserGroupHeaderResponsePage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserGroupHeaderResponsePage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserGroupHeaderResponsePage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserGroupHeaderResponsePage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelUserGroupHeaderResponsePage and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelUserGroupHeaderResponsePage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupHeaderResponsePage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupHeaderResponsePage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
