% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_user_group_int.R
\docType{class}
\name{OrgSagebionetworksRepoModelUserGroupInt}
\alias{OrgSagebionetworksRepoModelUserGroupInt}
\title{OrgSagebionetworksRepoModelUserGroupInt}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelUserGroupInt Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelUserGroupInt.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelUserGroupInt-new}{\code{OrgSagebionetworksRepoModelUserGroupInt$new()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupInt-fromJSONString}{\code{OrgSagebionetworksRepoModelUserGroupInt$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupInt-fromJSON}{\code{OrgSagebionetworksRepoModelUserGroupInt$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupInt-toJSONString}{\code{OrgSagebionetworksRepoModelUserGroupInt$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupInt-toJSON}{\code{OrgSagebionetworksRepoModelUserGroupInt$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupInt-validateJSON}{\code{OrgSagebionetworksRepoModelUserGroupInt$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupInt-toString}{\code{OrgSagebionetworksRepoModelUserGroupInt$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupInt-print}{\code{OrgSagebionetworksRepoModelUserGroupInt$print()}}
\item \href{#method-OrgSagebionetworksRepoModelUserGroupInt-clone}{\code{OrgSagebionetworksRepoModelUserGroupInt$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupInt-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupInt-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelUserGroupInt.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupInt$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelUserGroup"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupInt-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupInt-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserGroupInt.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupInt$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelUserGroupInt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupInt-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupInt-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserGroupInt.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupInt$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelUserGroupInt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupInt-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupInt-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelUserGroupInt to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupInt$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelUserGroupInt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupInt-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupInt-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelUserGroupInt to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupInt$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelUserGroupInt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupInt-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupInt-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelUserGroupInt and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupInt$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupInt-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupInt-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupInt$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupInt-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupInt-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupInt$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserGroupInt-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserGroupInt-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserGroupInt$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
