% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_user_preference_boolean.R
\docType{class}
\name{OrgSagebionetworksRepoModelUserPreferenceBoolean}
\alias{OrgSagebionetworksRepoModelUserPreferenceBoolean}
\title{OrgSagebionetworksRepoModelUserPreferenceBoolean}
\format{
An \code{R6Class} generator object
}
\description{
A user preference of type boolean

OrgSagebionetworksRepoModelUserPreferenceBoolean Class
}
\details{
Create a new OrgSagebionetworksRepoModelUserPreferenceBoolean
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{concreteType}}{character}

\item{\code{value}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-new}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$new()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-toJSON}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-fromJSON}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-toJSONString}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-fromJSONString}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-validateJSON}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-toString}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-isValid}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-getInvalidFields}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-print}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$print()}}
\item \href{#method-OrgSagebionetworksRepoModelUserPreferenceBoolean-clone}{\code{OrgSagebionetworksRepoModelUserPreferenceBoolean$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelUserPreferenceBoolean class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$new(
  concreteType,
  name = NULL,
  value = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concreteType}}{concreteType}

\item{\code{name}}{name}

\item{\code{value}}{value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserPreferenceBoolean in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserPreferenceBoolean
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserPreferenceBoolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserPreferenceBoolean in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserPreferenceBoolean
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserPreferenceBoolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelUserPreferenceBoolean and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelUserPreferenceBoolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserPreferenceBoolean-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserPreferenceBoolean-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserPreferenceBoolean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
