% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_user_profile.R
\docType{class}
\name{OrgSagebionetworksRepoModelUserProfile}
\alias{OrgSagebionetworksRepoModelUserProfile}
\title{OrgSagebionetworksRepoModelUserProfile}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for UserProfile POJO

OrgSagebionetworksRepoModelUserProfile Class
}
\details{
Create a new OrgSagebionetworksRepoModelUserProfile
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ownerId}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{firstName}}{character [optional]}

\item{\code{lastName}}{character [optional]}

\item{\code{email}}{character [optional]}

\item{\code{emails}}{The list of user email addresses registered to this user. list(character) [optional]}

\item{\code{openIds}}{The list of OpenIds bound to this user's account. list(character) [optional]}

\item{\code{userName}}{character [optional]}

\item{\code{displayName}}{character [optional]}

\item{\code{rStudioUrl}}{character [optional]}

\item{\code{summary}}{character [optional]}

\item{\code{position}}{character [optional]}

\item{\code{location}}{character [optional]}

\item{\code{industry}}{character [optional]}

\item{\code{company}}{character [optional]}

\item{\code{profilePicureFileHandleId}}{character [optional]}

\item{\code{url}}{character [optional]}

\item{\code{teamName}}{character [optional]}

\item{\code{notificationSettings}}{\link{OrgSagebionetworksRepoModelMessageSettings} [optional]}

\item{\code{preferences}}{list(\link{OrgSagebionetworksRepoModelUserPreference}) [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{twoFactorAuthEnabled}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-new}{\code{OrgSagebionetworksRepoModelUserProfile$new()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-toJSON}{\code{OrgSagebionetworksRepoModelUserProfile$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-fromJSON}{\code{OrgSagebionetworksRepoModelUserProfile$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-toJSONString}{\code{OrgSagebionetworksRepoModelUserProfile$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-fromJSONString}{\code{OrgSagebionetworksRepoModelUserProfile$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-validateJSON}{\code{OrgSagebionetworksRepoModelUserProfile$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-toString}{\code{OrgSagebionetworksRepoModelUserProfile$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-isValid}{\code{OrgSagebionetworksRepoModelUserProfile$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-getInvalidFields}{\code{OrgSagebionetworksRepoModelUserProfile$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-print}{\code{OrgSagebionetworksRepoModelUserProfile$print()}}
\item \href{#method-OrgSagebionetworksRepoModelUserProfile-clone}{\code{OrgSagebionetworksRepoModelUserProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelUserProfile class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$new(
  ownerId = NULL,
  etag = NULL,
  firstName = NULL,
  lastName = NULL,
  email = NULL,
  emails = NULL,
  openIds = NULL,
  userName = NULL,
  displayName = NULL,
  rStudioUrl = NULL,
  summary = NULL,
  position = NULL,
  location = NULL,
  industry = NULL,
  company = NULL,
  profilePicureFileHandleId = NULL,
  url = NULL,
  teamName = NULL,
  notificationSettings = NULL,
  preferences = NULL,
  createdOn = NULL,
  twoFactorAuthEnabled = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ownerId}}{ownerId}

\item{\code{etag}}{etag}

\item{\code{firstName}}{firstName}

\item{\code{lastName}}{lastName}

\item{\code{email}}{email}

\item{\code{emails}}{The list of user email addresses registered to this user.}

\item{\code{openIds}}{The list of OpenIds bound to this user's account.}

\item{\code{userName}}{userName}

\item{\code{displayName}}{displayName}

\item{\code{rStudioUrl}}{rStudioUrl}

\item{\code{summary}}{summary}

\item{\code{position}}{position}

\item{\code{location}}{location}

\item{\code{industry}}{industry}

\item{\code{company}}{company}

\item{\code{profilePicureFileHandleId}}{profilePicureFileHandleId}

\item{\code{url}}{url}

\item{\code{teamName}}{teamName}

\item{\code{notificationSettings}}{notificationSettings}

\item{\code{preferences}}{preferences}

\item{\code{createdOn}}{createdOn}

\item{\code{twoFactorAuthEnabled}}{twoFactorAuthEnabled}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserProfile in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserProfile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserProfile
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelUserProfile in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelUserProfile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelUserProfile
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelUserProfile and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelUserProfile
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelUserProfile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelUserProfile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelUserProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
