% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_v2_wiki_v2_wiki_header.R
\docType{class}
\name{OrgSagebionetworksRepoModelV2WikiV2WikiHeader}
\alias{OrgSagebionetworksRepoModelV2WikiV2WikiHeader}
\title{OrgSagebionetworksRepoModelV2WikiV2WikiHeader}
\format{
An \code{R6Class} generator object
}
\description{
A WikiHeader contains basic metadata about a WikiPage.

OrgSagebionetworksRepoModelV2WikiV2WikiHeader Class
}
\details{
Create a new OrgSagebionetworksRepoModelV2WikiV2WikiHeader
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{title}}{character [optional]}

\item{\code{parentId}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-new}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$new()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-toJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-fromJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-toJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-fromJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-validateJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-toString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-isValid}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-getInvalidFields}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-print}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$print()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-clone}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelV2WikiV2WikiHeader class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$new(
  id = NULL,
  title = NULL,
  parentId = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{title}}{title}

\item{\code{parentId}}{parentId}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelV2WikiV2WikiHeader and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelV2WikiV2WikiHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
