% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_v2_wiki_v2_wiki_history_snapshot.R
\docType{class}
\name{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot}
\alias{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot}
\title{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot}
\format{
An \code{R6Class} generator object
}
\description{
A WikiHistorySnapshot contains basic information about an update to a WikiPage.

OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot Class
}
\details{
Create a new OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-new}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$new()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-toJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-fromJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-toJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-fromJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-validateJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-toString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-isValid}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-getInvalidFields}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-print}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$print()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-clone}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$new(
  version = NULL,
  modifiedOn = NULL,
  modifiedBy = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{version}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
