% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_v2_wiki_v2_wiki_markdown_version.R
\docType{class}
\name{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion}
\alias{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion}
\title{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion}
\format{
An \code{R6Class} generator object
}
\description{
A V2WikiMarkdownVersion contains the markdown file handle id and title of a specific version of a V2WikiPage

OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion Class
}
\details{
Create a new OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{character [optional]}

\item{\code{title}}{character [optional]}

\item{\code{markdownFileHandleId}}{character [optional]}

\item{\code{attachmentFileHandleIds}}{The list of attachment file handle ids of this page for this version. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-new}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$new()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-toJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-fromJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-toJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-fromJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-validateJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-toString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-isValid}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-getInvalidFields}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-print}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$print()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-clone}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$new(
  version = NULL,
  title = NULL,
  markdownFileHandleId = NULL,
  attachmentFileHandleIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{version}

\item{\code{title}}{title}

\item{\code{markdownFileHandleId}}{markdownFileHandleId}

\item{\code{attachmentFileHandleIds}}{The list of attachment file handle ids of this page for this version.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiMarkdownVersion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
