% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_v2_wiki_v2_wiki_order_hint.R
\docType{class}
\name{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint}
\alias{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint}
\title{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint}
\format{
An \code{R6Class} generator object
}
\description{
A V2WikiOrderHint contains the order hint for the root wiki that corresponds to the given owner ID and type.

OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint Class
}
\details{
Create a new OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ownerId}}{character [optional]}

\item{\code{ownerObjectType}}{character [optional]}

\item{\code{idList}}{The list of sub wiki ids that in the order that they should be placed relative to their siblings. list(character) [optional]}

\item{\code{etag}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-new}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$new()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-toJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-fromJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-toJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-fromJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-validateJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-toString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-isValid}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-getInvalidFields}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-print}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$print()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-clone}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$new(
  ownerId = NULL,
  ownerObjectType = NULL,
  idList = NULL,
  etag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ownerId}}{ownerId}

\item{\code{ownerObjectType}}{ownerObjectType}

\item{\code{idList}}{The list of sub wiki ids that in the order that they should be placed relative to their siblings.}

\item{\code{etag}}{etag}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiOrderHint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
