% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_v2_wiki_v2_wiki_page.R
\docType{class}
\name{OrgSagebionetworksRepoModelV2WikiV2WikiPage}
\alias{OrgSagebionetworksRepoModelV2WikiV2WikiPage}
\title{OrgSagebionetworksRepoModelV2WikiV2WikiPage}
\format{
An \code{R6Class} generator object
}
\description{
The WikiPage model object represent a single user managed wiki page.

OrgSagebionetworksRepoModelV2WikiV2WikiPage Class
}
\details{
Create a new OrgSagebionetworksRepoModelV2WikiV2WikiPage
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{parentWikiId}}{character [optional]}

\item{\code{title}}{character [optional]}

\item{\code{markdownFileHandleId}}{character [optional]}

\item{\code{attachmentFileHandleIds}}{The list of attachment file handle ids of this page. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-new}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$new()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-toJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-fromJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-toJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-fromJSONString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-validateJSON}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-toString}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-isValid}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-getInvalidFields}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-print}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$print()}}
\item \href{#method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-clone}{\code{OrgSagebionetworksRepoModelV2WikiV2WikiPage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelV2WikiV2WikiPage class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$new(
  id = NULL,
  etag = NULL,
  createdOn = NULL,
  createdBy = NULL,
  modifiedOn = NULL,
  modifiedBy = NULL,
  parentWikiId = NULL,
  title = NULL,
  markdownFileHandleId = NULL,
  attachmentFileHandleIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{parentWikiId}}{parentWikiId}

\item{\code{title}}{title}

\item{\code{markdownFileHandleId}}{markdownFileHandleId}

\item{\code{attachmentFileHandleIds}}{The list of attachment file handle ids of this page.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiPage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiPage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelV2WikiV2WikiPage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelV2WikiV2WikiPage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelV2WikiV2WikiPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelV2WikiV2WikiPage and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelV2WikiV2WikiPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelV2WikiV2WikiPage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelV2WikiV2WikiPage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
