% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_verification_attachment_metadata.R
\docType{class}
\name{OrgSagebionetworksRepoModelVerificationAttachmentMetadata}
\alias{OrgSagebionetworksRepoModelVerificationAttachmentMetadata}
\title{OrgSagebionetworksRepoModelVerificationAttachmentMetadata}
\format{
An \code{R6Class} generator object
}
\description{
The metadata of a file attached to a Verification Submission

OrgSagebionetworksRepoModelVerificationAttachmentMetadata Class
}
\details{
Create a new OrgSagebionetworksRepoModelVerificationAttachmentMetadata
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fileName}}{character [optional]}

\item{\code{id}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-new}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$new()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-toJSON}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-fromJSON}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-toJSONString}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-fromJSONString}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-validateJSON}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-toString}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-isValid}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-getInvalidFields}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-print}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$print()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-clone}{\code{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelVerificationAttachmentMetadata class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$new(
  fileName = NULL,
  id = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{fileName}

\item{\code{id}}{id}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationAttachmentMetadata in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationAttachmentMetadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationAttachmentMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationAttachmentMetadata in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationAttachmentMetadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationAttachmentMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelVerificationAttachmentMetadata and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelVerificationAttachmentMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationAttachmentMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationAttachmentMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
