% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_verification_verification_paged_results.R
\docType{class}
\name{OrgSagebionetworksRepoModelVerificationVerificationPagedResults}
\alias{OrgSagebionetworksRepoModelVerificationVerificationPagedResults}
\title{OrgSagebionetworksRepoModelVerificationVerificationPagedResults}
\format{
An \code{R6Class} generator object
}
\description{
A paginated list of VerificationSubmission objects

OrgSagebionetworksRepoModelVerificationVerificationPagedResults Class
}
\details{
Create a new OrgSagebionetworksRepoModelVerificationVerificationPagedResults
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{The list of results for this page list(\link{OrgSagebionetworksRepoModelVerificationVerificationSubmission}) [optional]}

\item{\code{totalNumberOfResults}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-new}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$new()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-toJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-fromJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-toJSONString}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-fromJSONString}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-validateJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-toString}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-isValid}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-getInvalidFields}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-print}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$print()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-clone}{\code{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelVerificationVerificationPagedResults class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$new(
  results = NULL,
  totalNumberOfResults = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{The list of results for this page}

\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationVerificationPagedResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationVerificationPagedResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationVerificationPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationVerificationPagedResults in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationVerificationPagedResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationVerificationPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelVerificationVerificationPagedResults and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelVerificationVerificationPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationPagedResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationPagedResults$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
