% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_verification_verification_state.R
\docType{class}
\name{OrgSagebionetworksRepoModelVerificationVerificationState}
\alias{OrgSagebionetworksRepoModelVerificationVerificationState}
\title{OrgSagebionetworksRepoModelVerificationVerificationState}
\format{
An \code{R6Class} generator object
}
\description{
The state of a verification submission

OrgSagebionetworksRepoModelVerificationVerificationState Class
}
\details{
Create a new OrgSagebionetworksRepoModelVerificationVerificationState
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{state}}{character [optional]}

\item{\code{reason}}{character [optional]}

\item{\code{notes}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-new}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$new()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-toJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-fromJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-toJSONString}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-fromJSONString}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-validateJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-toString}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-isValid}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-getInvalidFields}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-print}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$print()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationState-clone}{\code{OrgSagebionetworksRepoModelVerificationVerificationState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelVerificationVerificationState class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$new(
  createdOn = NULL,
  createdBy = NULL,
  state = NULL,
  reason = NULL,
  notes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{state}}{state}

\item{\code{reason}}{reason}

\item{\code{notes}}{notes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationVerificationState in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationVerificationState
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationVerificationState
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationVerificationState in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationVerificationState
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationVerificationState
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelVerificationVerificationState and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelVerificationVerificationState
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
