% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_verification_verification_state_enum.R
\docType{class}
\name{OrgSagebionetworksRepoModelVerificationVerificationStateEnum}
\alias{OrgSagebionetworksRepoModelVerificationVerificationStateEnum}
\title{OrgSagebionetworksRepoModelVerificationVerificationStateEnum}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelVerificationVerificationStateEnum Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-new}{\code{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$new()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-toJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-fromJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-toJSONString}{\code{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-fromJSONString}{\code{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-clone}{\code{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelVerificationVerificationStateEnum class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationVerificationStateEnum in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationVerificationStateEnum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationVerificationStateEnum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationVerificationStateEnum in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationVerificationStateEnum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationVerificationStateEnum
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationStateEnum-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationStateEnum$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
