% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_verification_verification_submission.R
\docType{class}
\name{OrgSagebionetworksRepoModelVerificationVerificationSubmission}
\alias{OrgSagebionetworksRepoModelVerificationVerificationSubmission}
\title{OrgSagebionetworksRepoModelVerificationVerificationSubmission}
\format{
An \code{R6Class} generator object
}
\description{
User info submitted for verification by Synapse ACT

OrgSagebionetworksRepoModelVerificationVerificationSubmission Class
}
\details{
Create a new OrgSagebionetworksRepoModelVerificationVerificationSubmission
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{firstName}}{character [optional]}

\item{\code{lastName}}{character [optional]}

\item{\code{emails}}{The list of user email addresses registered to this user. list(character) [optional]}

\item{\code{notificationEmail}}{character [optional]}

\item{\code{location}}{character [optional]}

\item{\code{company}}{character [optional]}

\item{\code{orcid}}{character [optional]}

\item{\code{stateHistory}}{List of state changes the submission has passed through, ordered by time.  The last in the list contains the current state of the submission. list(\link{OrgSagebionetworksRepoModelVerificationVerificationState}) [optional]}

\item{\code{attachments}}{Metadata of Files attached to the submission. The list will be emptied after the state is APPROVED or REJECTED. list(\link{OrgSagebionetworksRepoModelVerificationAttachmentMetadata}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-new}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$new()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-toJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-fromJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-toJSONString}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-fromJSONString}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-validateJSON}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-toString}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-isValid}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-getInvalidFields}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-print}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$print()}}
\item \href{#method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-clone}{\code{OrgSagebionetworksRepoModelVerificationVerificationSubmission$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelVerificationVerificationSubmission class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$new(
  id = NULL,
  createdOn = NULL,
  createdBy = NULL,
  firstName = NULL,
  lastName = NULL,
  emails = NULL,
  notificationEmail = NULL,
  location = NULL,
  company = NULL,
  orcid = NULL,
  stateHistory = NULL,
  attachments = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{firstName}}{firstName}

\item{\code{lastName}}{lastName}

\item{\code{emails}}{The list of user email addresses registered to this user.}

\item{\code{notificationEmail}}{notificationEmail}

\item{\code{location}}{location}

\item{\code{company}}{company}

\item{\code{orcid}}{orcid}

\item{\code{stateHistory}}{List of state changes the submission has passed through, ordered by time.  The last in the list contains the current state of the submission.}

\item{\code{attachments}}{Metadata of Files attached to the submission. The list will be emptied after the state is APPROVED or REJECTED.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationVerificationSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationVerificationSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationVerificationSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVerificationVerificationSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVerificationVerificationSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVerificationVerificationSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelVerificationVerificationSubmission and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelVerificationVerificationSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVerificationVerificationSubmission-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVerificationVerificationSubmission$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
