% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_versionable.R
\docType{class}
\name{OrgSagebionetworksRepoModelVersionable}
\alias{OrgSagebionetworksRepoModelVersionable}
\title{OrgSagebionetworksRepoModelVersionable}
\format{
An \code{R6Class} generator object
}
\description{
OrgSagebionetworksRepoModelVersionable Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OrgSagebionetworksRepoModelVersionable.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelVersionable-new}{\code{OrgSagebionetworksRepoModelVersionable$new()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionable-fromJSONString}{\code{OrgSagebionetworksRepoModelVersionable$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionable-fromJSON}{\code{OrgSagebionetworksRepoModelVersionable$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionable-toJSONString}{\code{OrgSagebionetworksRepoModelVersionable$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionable-toJSON}{\code{OrgSagebionetworksRepoModelVersionable$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionable-validateJSON}{\code{OrgSagebionetworksRepoModelVersionable$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionable-toString}{\code{OrgSagebionetworksRepoModelVersionable$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionable-print}{\code{OrgSagebionetworksRepoModelVersionable$print()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionable-clone}{\code{OrgSagebionetworksRepoModelVersionable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionable-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionable-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelVersionable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionable$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "OrgSagebionetworksRepoModelACTAccessRequirement", "OrgSagebionetworksRepoModelFileEntity", "OrgSagebionetworksRepoModelLockAccessRequirement", "OrgSagebionetworksRepoModelManagedACTAccessRequirement", "OrgSagebionetworksRepoModelPostMessageContentAccessRequirement", "OrgSagebionetworksRepoModelSelfSignAccessRequirement", "OrgSagebionetworksRepoModelTableDataset", "OrgSagebionetworksRepoModelTableDatasetCollection", "OrgSagebionetworksRepoModelTableEntityView", "OrgSagebionetworksRepoModelTableMaterializedView", "OrgSagebionetworksRepoModelTableSubmissionView", "OrgSagebionetworksRepoModelTableTableEntity", "OrgSagebionetworksRepoModelTableVirtualTable", "OrgSagebionetworksRepoModelTermsOfUseAccessRequirement"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionable-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionable-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVersionable.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionable$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelVersionable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionable-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionable-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVersionable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionable$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OrgSagebionetworksRepoModelVersionable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionable-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionable-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OrgSagebionetworksRepoModelVersionable to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionable$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OrgSagebionetworksRepoModelVersionable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionable-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionable-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OrgSagebionetworksRepoModelVersionable to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionable$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OrgSagebionetworksRepoModelVersionable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionable-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionable-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OrgSagebionetworksRepoModelVersionable and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionable$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionable-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionable-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionable$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionable-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionable-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionable$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
