% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_version_info.R
\docType{class}
\name{OrgSagebionetworksRepoModelVersionInfo}
\alias{OrgSagebionetworksRepoModelVersionInfo}
\title{OrgSagebionetworksRepoModelVersionInfo}
\format{
An \code{R6Class} generator object
}
\description{
JSON schema for VersionInfo POJO

OrgSagebionetworksRepoModelVersionInfo Class
}
\details{
Create a new OrgSagebionetworksRepoModelVersionInfo
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{versionNumber}}{integer [optional]}

\item{\code{versionLabel}}{character [optional]}

\item{\code{versionComment}}{character [optional]}

\item{\code{isLatestVersion}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{contentSize}}{character [optional]}

\item{\code{contentMd5}}{character [optional]}

\item{\code{modifiedByPrincipalId}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-new}{\code{OrgSagebionetworksRepoModelVersionInfo$new()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-toJSON}{\code{OrgSagebionetworksRepoModelVersionInfo$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-fromJSON}{\code{OrgSagebionetworksRepoModelVersionInfo$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-toJSONString}{\code{OrgSagebionetworksRepoModelVersionInfo$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-fromJSONString}{\code{OrgSagebionetworksRepoModelVersionInfo$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-validateJSON}{\code{OrgSagebionetworksRepoModelVersionInfo$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-toString}{\code{OrgSagebionetworksRepoModelVersionInfo$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-isValid}{\code{OrgSagebionetworksRepoModelVersionInfo$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-getInvalidFields}{\code{OrgSagebionetworksRepoModelVersionInfo$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-print}{\code{OrgSagebionetworksRepoModelVersionInfo$print()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfo-clone}{\code{OrgSagebionetworksRepoModelVersionInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelVersionInfo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$new(
  id = NULL,
  versionNumber = NULL,
  versionLabel = NULL,
  versionComment = NULL,
  isLatestVersion = NULL,
  modifiedBy = NULL,
  contentSize = NULL,
  contentMd5 = NULL,
  modifiedByPrincipalId = NULL,
  modifiedOn = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{versionNumber}}{versionNumber}

\item{\code{versionLabel}}{versionLabel}

\item{\code{versionComment}}{versionComment}

\item{\code{isLatestVersion}}{isLatestVersion}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{contentSize}}{contentSize}

\item{\code{contentMd5}}{contentMd5}

\item{\code{modifiedByPrincipalId}}{modifiedByPrincipalId}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVersionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVersionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVersionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVersionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelVersionInfo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
