% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/org_sagebionetworks_repo_model_version_info_synapse_version_info.R
\docType{class}
\name{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo}
\alias{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo}
\title{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo}
\format{
An \code{R6Class} generator object
}
\description{
Information about the version of the stack

OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo Class
}
\details{
Create a new OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{character [optional]}

\item{\code{stackInstance}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-new}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$new()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-toJSON}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-fromJSON}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-toJSONString}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-fromJSONString}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-validateJSON}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-toString}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-isValid}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-getInvalidFields}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-print}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$print()}}
\item \href{#method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-clone}{\code{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$new(
  version = NULL,
  stackInstance = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{version}

\item{\code{stackInstance}}{stackInstance}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelVersionInfoSynapseVersionInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
