% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_sagebionetworks_repo_model_wiki_wiki_page.R
\docType{class}
\name{OrgSagebionetworksRepoModelWikiWikiPage}
\alias{OrgSagebionetworksRepoModelWikiWikiPage}
\title{OrgSagebionetworksRepoModelWikiWikiPage}
\format{
An \code{R6Class} generator object
}
\description{
The WikiPage model object represent a single user managed wiki page.

OrgSagebionetworksRepoModelWikiWikiPage Class
}
\details{
Create a new OrgSagebionetworksRepoModelWikiWikiPage
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{etag}}{character [optional]}

\item{\code{createdOn}}{character [optional]}

\item{\code{createdBy}}{character [optional]}

\item{\code{modifiedOn}}{character [optional]}

\item{\code{modifiedBy}}{character [optional]}

\item{\code{parentWikiId}}{character [optional]}

\item{\code{title}}{character [optional]}

\item{\code{markdown}}{character [optional]}

\item{\code{attachmentFileHandleIds}}{The list of attachment file handle ids of this page. list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-new}{\code{OrgSagebionetworksRepoModelWikiWikiPage$new()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-toJSON}{\code{OrgSagebionetworksRepoModelWikiWikiPage$toJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-fromJSON}{\code{OrgSagebionetworksRepoModelWikiWikiPage$fromJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-toJSONString}{\code{OrgSagebionetworksRepoModelWikiWikiPage$toJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-fromJSONString}{\code{OrgSagebionetworksRepoModelWikiWikiPage$fromJSONString()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-validateJSON}{\code{OrgSagebionetworksRepoModelWikiWikiPage$validateJSON()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-toString}{\code{OrgSagebionetworksRepoModelWikiWikiPage$toString()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-isValid}{\code{OrgSagebionetworksRepoModelWikiWikiPage$isValid()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-getInvalidFields}{\code{OrgSagebionetworksRepoModelWikiWikiPage$getInvalidFields()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-print}{\code{OrgSagebionetworksRepoModelWikiWikiPage$print()}}
\item \href{#method-OrgSagebionetworksRepoModelWikiWikiPage-clone}{\code{OrgSagebionetworksRepoModelWikiWikiPage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrgSagebionetworksRepoModelWikiWikiPage class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$new(
  id = NULL,
  etag = NULL,
  createdOn = NULL,
  createdBy = NULL,
  modifiedOn = NULL,
  modifiedBy = NULL,
  parentWikiId = NULL,
  title = NULL,
  markdown = NULL,
  attachmentFileHandleIds = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{etag}}{etag}

\item{\code{createdOn}}{createdOn}

\item{\code{createdBy}}{createdBy}

\item{\code{modifiedOn}}{modifiedOn}

\item{\code{modifiedBy}}{modifiedBy}

\item{\code{parentWikiId}}{parentWikiId}

\item{\code{title}}{title}

\item{\code{markdown}}{markdown}

\item{\code{attachmentFileHandleIds}}{The list of attachment file handle ids of this page.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelWikiWikiPage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelWikiWikiPage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelWikiWikiPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrgSagebionetworksRepoModelWikiWikiPage in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrgSagebionetworksRepoModelWikiWikiPage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrgSagebionetworksRepoModelWikiWikiPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrgSagebionetworksRepoModelWikiWikiPage and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrgSagebionetworksRepoModelWikiWikiPage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrgSagebionetworksRepoModelWikiWikiPage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrgSagebionetworksRepoModelWikiWikiPage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrgSagebionetworksRepoModelWikiWikiPage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
