% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_entity_header.R
\docType{class}
\name{PaginatedResultsOfEntityHeader}
\alias{PaginatedResultsOfEntityHeader}
\title{PaginatedResultsOfEntityHeader}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfEntityHeader Class

PaginatedResultsOfEntityHeader Class
}
\details{
Create a new PaginatedResultsOfEntityHeader
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelEntityHeader}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfEntityHeader-new}{\code{PaginatedResultsOfEntityHeader$new()}}
\item \href{#method-PaginatedResultsOfEntityHeader-toJSON}{\code{PaginatedResultsOfEntityHeader$toJSON()}}
\item \href{#method-PaginatedResultsOfEntityHeader-fromJSON}{\code{PaginatedResultsOfEntityHeader$fromJSON()}}
\item \href{#method-PaginatedResultsOfEntityHeader-toJSONString}{\code{PaginatedResultsOfEntityHeader$toJSONString()}}
\item \href{#method-PaginatedResultsOfEntityHeader-fromJSONString}{\code{PaginatedResultsOfEntityHeader$fromJSONString()}}
\item \href{#method-PaginatedResultsOfEntityHeader-validateJSON}{\code{PaginatedResultsOfEntityHeader$validateJSON()}}
\item \href{#method-PaginatedResultsOfEntityHeader-toString}{\code{PaginatedResultsOfEntityHeader$toString()}}
\item \href{#method-PaginatedResultsOfEntityHeader-isValid}{\code{PaginatedResultsOfEntityHeader$isValid()}}
\item \href{#method-PaginatedResultsOfEntityHeader-getInvalidFields}{\code{PaginatedResultsOfEntityHeader$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfEntityHeader-print}{\code{PaginatedResultsOfEntityHeader$print()}}
\item \href{#method-PaginatedResultsOfEntityHeader-clone}{\code{PaginatedResultsOfEntityHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfEntityHeader class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfEntityHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfEntityHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfEntityHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfEntityHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfEntityHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfEntityHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfEntityHeader and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfEntityHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEntityHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEntityHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEntityHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
