% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_evaluation.R
\docType{class}
\name{PaginatedResultsOfEvaluation}
\alias{PaginatedResultsOfEvaluation}
\title{PaginatedResultsOfEvaluation}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfEvaluation Class

PaginatedResultsOfEvaluation Class
}
\details{
Create a new PaginatedResultsOfEvaluation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksEvaluationModelEvaluation}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfEvaluation-new}{\code{PaginatedResultsOfEvaluation$new()}}
\item \href{#method-PaginatedResultsOfEvaluation-toJSON}{\code{PaginatedResultsOfEvaluation$toJSON()}}
\item \href{#method-PaginatedResultsOfEvaluation-fromJSON}{\code{PaginatedResultsOfEvaluation$fromJSON()}}
\item \href{#method-PaginatedResultsOfEvaluation-toJSONString}{\code{PaginatedResultsOfEvaluation$toJSONString()}}
\item \href{#method-PaginatedResultsOfEvaluation-fromJSONString}{\code{PaginatedResultsOfEvaluation$fromJSONString()}}
\item \href{#method-PaginatedResultsOfEvaluation-validateJSON}{\code{PaginatedResultsOfEvaluation$validateJSON()}}
\item \href{#method-PaginatedResultsOfEvaluation-toString}{\code{PaginatedResultsOfEvaluation$toString()}}
\item \href{#method-PaginatedResultsOfEvaluation-isValid}{\code{PaginatedResultsOfEvaluation$isValid()}}
\item \href{#method-PaginatedResultsOfEvaluation-getInvalidFields}{\code{PaginatedResultsOfEvaluation$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfEvaluation-print}{\code{PaginatedResultsOfEvaluation$print()}}
\item \href{#method-PaginatedResultsOfEvaluation-clone}{\code{PaginatedResultsOfEvaluation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfEvaluation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfEvaluation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfEvaluation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfEvaluation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfEvaluation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfEvaluation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfEvaluation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfEvaluation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfEvaluation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfEvaluation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfEvaluation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfEvaluation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
