% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_membership_invitation.R
\docType{class}
\name{PaginatedResultsOfMembershipInvitation}
\alias{PaginatedResultsOfMembershipInvitation}
\title{PaginatedResultsOfMembershipInvitation}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfMembershipInvitation Class

PaginatedResultsOfMembershipInvitation Class
}
\details{
Create a new PaginatedResultsOfMembershipInvitation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelMembershipInvitation}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfMembershipInvitation-new}{\code{PaginatedResultsOfMembershipInvitation$new()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-toJSON}{\code{PaginatedResultsOfMembershipInvitation$toJSON()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-fromJSON}{\code{PaginatedResultsOfMembershipInvitation$fromJSON()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-toJSONString}{\code{PaginatedResultsOfMembershipInvitation$toJSONString()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-fromJSONString}{\code{PaginatedResultsOfMembershipInvitation$fromJSONString()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-validateJSON}{\code{PaginatedResultsOfMembershipInvitation$validateJSON()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-toString}{\code{PaginatedResultsOfMembershipInvitation$toString()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-isValid}{\code{PaginatedResultsOfMembershipInvitation$isValid()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-getInvalidFields}{\code{PaginatedResultsOfMembershipInvitation$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-print}{\code{PaginatedResultsOfMembershipInvitation$print()}}
\item \href{#method-PaginatedResultsOfMembershipInvitation-clone}{\code{PaginatedResultsOfMembershipInvitation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfMembershipInvitation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfMembershipInvitation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfMembershipInvitation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfMembershipInvitation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfMembershipInvitation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfMembershipInvitation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfMembershipInvitation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfMembershipInvitation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfMembershipInvitation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipInvitation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipInvitation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipInvitation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
