% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_membership_request.R
\docType{class}
\name{PaginatedResultsOfMembershipRequest}
\alias{PaginatedResultsOfMembershipRequest}
\title{PaginatedResultsOfMembershipRequest}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfMembershipRequest Class

PaginatedResultsOfMembershipRequest Class
}
\details{
Create a new PaginatedResultsOfMembershipRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelMembershipRequest}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfMembershipRequest-new}{\code{PaginatedResultsOfMembershipRequest$new()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-toJSON}{\code{PaginatedResultsOfMembershipRequest$toJSON()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-fromJSON}{\code{PaginatedResultsOfMembershipRequest$fromJSON()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-toJSONString}{\code{PaginatedResultsOfMembershipRequest$toJSONString()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-fromJSONString}{\code{PaginatedResultsOfMembershipRequest$fromJSONString()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-validateJSON}{\code{PaginatedResultsOfMembershipRequest$validateJSON()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-toString}{\code{PaginatedResultsOfMembershipRequest$toString()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-isValid}{\code{PaginatedResultsOfMembershipRequest$isValid()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-getInvalidFields}{\code{PaginatedResultsOfMembershipRequest$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-print}{\code{PaginatedResultsOfMembershipRequest$print()}}
\item \href{#method-PaginatedResultsOfMembershipRequest-clone}{\code{PaginatedResultsOfMembershipRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfMembershipRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfMembershipRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfMembershipRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfMembershipRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfMembershipRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfMembershipRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfMembershipRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfMembershipRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfMembershipRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfMembershipRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfMembershipRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfMembershipRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
