% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_passing_record.R
\docType{class}
\name{PaginatedResultsOfPassingRecord}
\alias{PaginatedResultsOfPassingRecord}
\title{PaginatedResultsOfPassingRecord}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfPassingRecord Class

PaginatedResultsOfPassingRecord Class
}
\details{
Create a new PaginatedResultsOfPassingRecord
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelQuizPassingRecord}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfPassingRecord-new}{\code{PaginatedResultsOfPassingRecord$new()}}
\item \href{#method-PaginatedResultsOfPassingRecord-toJSON}{\code{PaginatedResultsOfPassingRecord$toJSON()}}
\item \href{#method-PaginatedResultsOfPassingRecord-fromJSON}{\code{PaginatedResultsOfPassingRecord$fromJSON()}}
\item \href{#method-PaginatedResultsOfPassingRecord-toJSONString}{\code{PaginatedResultsOfPassingRecord$toJSONString()}}
\item \href{#method-PaginatedResultsOfPassingRecord-fromJSONString}{\code{PaginatedResultsOfPassingRecord$fromJSONString()}}
\item \href{#method-PaginatedResultsOfPassingRecord-validateJSON}{\code{PaginatedResultsOfPassingRecord$validateJSON()}}
\item \href{#method-PaginatedResultsOfPassingRecord-toString}{\code{PaginatedResultsOfPassingRecord$toString()}}
\item \href{#method-PaginatedResultsOfPassingRecord-isValid}{\code{PaginatedResultsOfPassingRecord$isValid()}}
\item \href{#method-PaginatedResultsOfPassingRecord-getInvalidFields}{\code{PaginatedResultsOfPassingRecord$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfPassingRecord-print}{\code{PaginatedResultsOfPassingRecord$print()}}
\item \href{#method-PaginatedResultsOfPassingRecord-clone}{\code{PaginatedResultsOfPassingRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfPassingRecord class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfPassingRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfPassingRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfPassingRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfPassingRecord in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfPassingRecord
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfPassingRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfPassingRecord and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfPassingRecord
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfPassingRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfPassingRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfPassingRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
