% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_quiz_response.R
\docType{class}
\name{PaginatedResultsOfQuizResponse}
\alias{PaginatedResultsOfQuizResponse}
\title{PaginatedResultsOfQuizResponse}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfQuizResponse Class

PaginatedResultsOfQuizResponse Class
}
\details{
Create a new PaginatedResultsOfQuizResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelQuizQuizResponse}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfQuizResponse-new}{\code{PaginatedResultsOfQuizResponse$new()}}
\item \href{#method-PaginatedResultsOfQuizResponse-toJSON}{\code{PaginatedResultsOfQuizResponse$toJSON()}}
\item \href{#method-PaginatedResultsOfQuizResponse-fromJSON}{\code{PaginatedResultsOfQuizResponse$fromJSON()}}
\item \href{#method-PaginatedResultsOfQuizResponse-toJSONString}{\code{PaginatedResultsOfQuizResponse$toJSONString()}}
\item \href{#method-PaginatedResultsOfQuizResponse-fromJSONString}{\code{PaginatedResultsOfQuizResponse$fromJSONString()}}
\item \href{#method-PaginatedResultsOfQuizResponse-validateJSON}{\code{PaginatedResultsOfQuizResponse$validateJSON()}}
\item \href{#method-PaginatedResultsOfQuizResponse-toString}{\code{PaginatedResultsOfQuizResponse$toString()}}
\item \href{#method-PaginatedResultsOfQuizResponse-isValid}{\code{PaginatedResultsOfQuizResponse$isValid()}}
\item \href{#method-PaginatedResultsOfQuizResponse-getInvalidFields}{\code{PaginatedResultsOfQuizResponse$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfQuizResponse-print}{\code{PaginatedResultsOfQuizResponse$print()}}
\item \href{#method-PaginatedResultsOfQuizResponse-clone}{\code{PaginatedResultsOfQuizResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfQuizResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfQuizResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfQuizResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfQuizResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfQuizResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfQuizResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfQuizResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfQuizResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfQuizResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfQuizResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfQuizResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfQuizResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
