% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_submission.R
\docType{class}
\name{PaginatedResultsOfSubmission}
\alias{PaginatedResultsOfSubmission}
\title{PaginatedResultsOfSubmission}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfSubmission Class

PaginatedResultsOfSubmission Class
}
\details{
Create a new PaginatedResultsOfSubmission
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksEvaluationModelSubmission}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfSubmission-new}{\code{PaginatedResultsOfSubmission$new()}}
\item \href{#method-PaginatedResultsOfSubmission-toJSON}{\code{PaginatedResultsOfSubmission$toJSON()}}
\item \href{#method-PaginatedResultsOfSubmission-fromJSON}{\code{PaginatedResultsOfSubmission$fromJSON()}}
\item \href{#method-PaginatedResultsOfSubmission-toJSONString}{\code{PaginatedResultsOfSubmission$toJSONString()}}
\item \href{#method-PaginatedResultsOfSubmission-fromJSONString}{\code{PaginatedResultsOfSubmission$fromJSONString()}}
\item \href{#method-PaginatedResultsOfSubmission-validateJSON}{\code{PaginatedResultsOfSubmission$validateJSON()}}
\item \href{#method-PaginatedResultsOfSubmission-toString}{\code{PaginatedResultsOfSubmission$toString()}}
\item \href{#method-PaginatedResultsOfSubmission-isValid}{\code{PaginatedResultsOfSubmission$isValid()}}
\item \href{#method-PaginatedResultsOfSubmission-getInvalidFields}{\code{PaginatedResultsOfSubmission$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfSubmission-print}{\code{PaginatedResultsOfSubmission$print()}}
\item \href{#method-PaginatedResultsOfSubmission-clone}{\code{PaginatedResultsOfSubmission$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfSubmission class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfSubmission in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfSubmission
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfSubmission and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfSubmission
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmission-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmission-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmission$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
