% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_submission_bundle.R
\docType{class}
\name{PaginatedResultsOfSubmissionBundle}
\alias{PaginatedResultsOfSubmissionBundle}
\title{PaginatedResultsOfSubmissionBundle}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfSubmissionBundle Class

PaginatedResultsOfSubmissionBundle Class
}
\details{
Create a new PaginatedResultsOfSubmissionBundle
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksEvaluationModelSubmissionBundle}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfSubmissionBundle-new}{\code{PaginatedResultsOfSubmissionBundle$new()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-toJSON}{\code{PaginatedResultsOfSubmissionBundle$toJSON()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-fromJSON}{\code{PaginatedResultsOfSubmissionBundle$fromJSON()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-toJSONString}{\code{PaginatedResultsOfSubmissionBundle$toJSONString()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-fromJSONString}{\code{PaginatedResultsOfSubmissionBundle$fromJSONString()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-validateJSON}{\code{PaginatedResultsOfSubmissionBundle$validateJSON()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-toString}{\code{PaginatedResultsOfSubmissionBundle$toString()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-isValid}{\code{PaginatedResultsOfSubmissionBundle$isValid()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-getInvalidFields}{\code{PaginatedResultsOfSubmissionBundle$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-print}{\code{PaginatedResultsOfSubmissionBundle$print()}}
\item \href{#method-PaginatedResultsOfSubmissionBundle-clone}{\code{PaginatedResultsOfSubmissionBundle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfSubmissionBundle class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfSubmissionBundle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfSubmissionBundle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfSubmissionBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfSubmissionBundle in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfSubmissionBundle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfSubmissionBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfSubmissionBundle and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfSubmissionBundle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfSubmissionBundle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfSubmissionBundle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfSubmissionBundle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
