% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_team.R
\docType{class}
\name{PaginatedResultsOfTeam}
\alias{PaginatedResultsOfTeam}
\title{PaginatedResultsOfTeam}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfTeam Class

PaginatedResultsOfTeam Class
}
\details{
Create a new PaginatedResultsOfTeam
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelTeam}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfTeam-new}{\code{PaginatedResultsOfTeam$new()}}
\item \href{#method-PaginatedResultsOfTeam-toJSON}{\code{PaginatedResultsOfTeam$toJSON()}}
\item \href{#method-PaginatedResultsOfTeam-fromJSON}{\code{PaginatedResultsOfTeam$fromJSON()}}
\item \href{#method-PaginatedResultsOfTeam-toJSONString}{\code{PaginatedResultsOfTeam$toJSONString()}}
\item \href{#method-PaginatedResultsOfTeam-fromJSONString}{\code{PaginatedResultsOfTeam$fromJSONString()}}
\item \href{#method-PaginatedResultsOfTeam-validateJSON}{\code{PaginatedResultsOfTeam$validateJSON()}}
\item \href{#method-PaginatedResultsOfTeam-toString}{\code{PaginatedResultsOfTeam$toString()}}
\item \href{#method-PaginatedResultsOfTeam-isValid}{\code{PaginatedResultsOfTeam$isValid()}}
\item \href{#method-PaginatedResultsOfTeam-getInvalidFields}{\code{PaginatedResultsOfTeam$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfTeam-print}{\code{PaginatedResultsOfTeam$print()}}
\item \href{#method-PaginatedResultsOfTeam-clone}{\code{PaginatedResultsOfTeam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfTeam class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$new(totalNumberOfResults = NULL, results = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfTeam in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfTeam
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfTeam in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfTeam
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfTeam and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
