% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_team_member.R
\docType{class}
\name{PaginatedResultsOfTeamMember}
\alias{PaginatedResultsOfTeamMember}
\title{PaginatedResultsOfTeamMember}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfTeamMember Class

PaginatedResultsOfTeamMember Class
}
\details{
Create a new PaginatedResultsOfTeamMember
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelTeamMember}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfTeamMember-new}{\code{PaginatedResultsOfTeamMember$new()}}
\item \href{#method-PaginatedResultsOfTeamMember-toJSON}{\code{PaginatedResultsOfTeamMember$toJSON()}}
\item \href{#method-PaginatedResultsOfTeamMember-fromJSON}{\code{PaginatedResultsOfTeamMember$fromJSON()}}
\item \href{#method-PaginatedResultsOfTeamMember-toJSONString}{\code{PaginatedResultsOfTeamMember$toJSONString()}}
\item \href{#method-PaginatedResultsOfTeamMember-fromJSONString}{\code{PaginatedResultsOfTeamMember$fromJSONString()}}
\item \href{#method-PaginatedResultsOfTeamMember-validateJSON}{\code{PaginatedResultsOfTeamMember$validateJSON()}}
\item \href{#method-PaginatedResultsOfTeamMember-toString}{\code{PaginatedResultsOfTeamMember$toString()}}
\item \href{#method-PaginatedResultsOfTeamMember-isValid}{\code{PaginatedResultsOfTeamMember$isValid()}}
\item \href{#method-PaginatedResultsOfTeamMember-getInvalidFields}{\code{PaginatedResultsOfTeamMember$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfTeamMember-print}{\code{PaginatedResultsOfTeamMember$print()}}
\item \href{#method-PaginatedResultsOfTeamMember-clone}{\code{PaginatedResultsOfTeamMember$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfTeamMember class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfTeamMember in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfTeamMember
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfTeamMember in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfTeamMember
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfTeamMember and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfTeamMember-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfTeamMember-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfTeamMember$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
