% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_v2_wiki_header.R
\docType{class}
\name{PaginatedResultsOfV2WikiHeader}
\alias{PaginatedResultsOfV2WikiHeader}
\title{PaginatedResultsOfV2WikiHeader}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfV2WikiHeader Class

PaginatedResultsOfV2WikiHeader Class
}
\details{
Create a new PaginatedResultsOfV2WikiHeader
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelV2WikiV2WikiHeader}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfV2WikiHeader-new}{\code{PaginatedResultsOfV2WikiHeader$new()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-toJSON}{\code{PaginatedResultsOfV2WikiHeader$toJSON()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-fromJSON}{\code{PaginatedResultsOfV2WikiHeader$fromJSON()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-toJSONString}{\code{PaginatedResultsOfV2WikiHeader$toJSONString()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-fromJSONString}{\code{PaginatedResultsOfV2WikiHeader$fromJSONString()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-validateJSON}{\code{PaginatedResultsOfV2WikiHeader$validateJSON()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-toString}{\code{PaginatedResultsOfV2WikiHeader$toString()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-isValid}{\code{PaginatedResultsOfV2WikiHeader$isValid()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-getInvalidFields}{\code{PaginatedResultsOfV2WikiHeader$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-print}{\code{PaginatedResultsOfV2WikiHeader$print()}}
\item \href{#method-PaginatedResultsOfV2WikiHeader-clone}{\code{PaginatedResultsOfV2WikiHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfV2WikiHeader class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfV2WikiHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfV2WikiHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfV2WikiHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfV2WikiHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfV2WikiHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfV2WikiHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfV2WikiHeader and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfV2WikiHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
