% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_v2_wiki_history_snapshot.R
\docType{class}
\name{PaginatedResultsOfV2WikiHistorySnapshot}
\alias{PaginatedResultsOfV2WikiHistorySnapshot}
\title{PaginatedResultsOfV2WikiHistorySnapshot}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfV2WikiHistorySnapshot Class

PaginatedResultsOfV2WikiHistorySnapshot Class
}
\details{
Create a new PaginatedResultsOfV2WikiHistorySnapshot
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelV2WikiV2WikiHistorySnapshot}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-new}{\code{PaginatedResultsOfV2WikiHistorySnapshot$new()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-toJSON}{\code{PaginatedResultsOfV2WikiHistorySnapshot$toJSON()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-fromJSON}{\code{PaginatedResultsOfV2WikiHistorySnapshot$fromJSON()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-toJSONString}{\code{PaginatedResultsOfV2WikiHistorySnapshot$toJSONString()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-fromJSONString}{\code{PaginatedResultsOfV2WikiHistorySnapshot$fromJSONString()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-validateJSON}{\code{PaginatedResultsOfV2WikiHistorySnapshot$validateJSON()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-toString}{\code{PaginatedResultsOfV2WikiHistorySnapshot$toString()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-isValid}{\code{PaginatedResultsOfV2WikiHistorySnapshot$isValid()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-getInvalidFields}{\code{PaginatedResultsOfV2WikiHistorySnapshot$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-print}{\code{PaginatedResultsOfV2WikiHistorySnapshot$print()}}
\item \href{#method-PaginatedResultsOfV2WikiHistorySnapshot-clone}{\code{PaginatedResultsOfV2WikiHistorySnapshot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfV2WikiHistorySnapshot class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfV2WikiHistorySnapshot in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfV2WikiHistorySnapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfV2WikiHistorySnapshot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfV2WikiHistorySnapshot in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfV2WikiHistorySnapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfV2WikiHistorySnapshot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfV2WikiHistorySnapshot and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfV2WikiHistorySnapshot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfV2WikiHistorySnapshot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfV2WikiHistorySnapshot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfV2WikiHistorySnapshot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
