% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_version_info.R
\docType{class}
\name{PaginatedResultsOfVersionInfo}
\alias{PaginatedResultsOfVersionInfo}
\title{PaginatedResultsOfVersionInfo}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfVersionInfo Class

PaginatedResultsOfVersionInfo Class
}
\details{
Create a new PaginatedResultsOfVersionInfo
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelVersionInfo}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfVersionInfo-new}{\code{PaginatedResultsOfVersionInfo$new()}}
\item \href{#method-PaginatedResultsOfVersionInfo-toJSON}{\code{PaginatedResultsOfVersionInfo$toJSON()}}
\item \href{#method-PaginatedResultsOfVersionInfo-fromJSON}{\code{PaginatedResultsOfVersionInfo$fromJSON()}}
\item \href{#method-PaginatedResultsOfVersionInfo-toJSONString}{\code{PaginatedResultsOfVersionInfo$toJSONString()}}
\item \href{#method-PaginatedResultsOfVersionInfo-fromJSONString}{\code{PaginatedResultsOfVersionInfo$fromJSONString()}}
\item \href{#method-PaginatedResultsOfVersionInfo-validateJSON}{\code{PaginatedResultsOfVersionInfo$validateJSON()}}
\item \href{#method-PaginatedResultsOfVersionInfo-toString}{\code{PaginatedResultsOfVersionInfo$toString()}}
\item \href{#method-PaginatedResultsOfVersionInfo-isValid}{\code{PaginatedResultsOfVersionInfo$isValid()}}
\item \href{#method-PaginatedResultsOfVersionInfo-getInvalidFields}{\code{PaginatedResultsOfVersionInfo$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfVersionInfo-print}{\code{PaginatedResultsOfVersionInfo$print()}}
\item \href{#method-PaginatedResultsOfVersionInfo-clone}{\code{PaginatedResultsOfVersionInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfVersionInfo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfVersionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfVersionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfVersionInfo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfVersionInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfVersionInfo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfVersionInfo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfVersionInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfVersionInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfVersionInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
