% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginated_results_of_wiki_header.R
\docType{class}
\name{PaginatedResultsOfWikiHeader}
\alias{PaginatedResultsOfWikiHeader}
\title{PaginatedResultsOfWikiHeader}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResultsOfWikiHeader Class

PaginatedResultsOfWikiHeader Class
}
\details{
Create a new PaginatedResultsOfWikiHeader
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalNumberOfResults}}{integer [optional]}

\item{\code{results}}{list(\link{OrgSagebionetworksRepoModelWikiWikiHeader}) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResultsOfWikiHeader-new}{\code{PaginatedResultsOfWikiHeader$new()}}
\item \href{#method-PaginatedResultsOfWikiHeader-toJSON}{\code{PaginatedResultsOfWikiHeader$toJSON()}}
\item \href{#method-PaginatedResultsOfWikiHeader-fromJSON}{\code{PaginatedResultsOfWikiHeader$fromJSON()}}
\item \href{#method-PaginatedResultsOfWikiHeader-toJSONString}{\code{PaginatedResultsOfWikiHeader$toJSONString()}}
\item \href{#method-PaginatedResultsOfWikiHeader-fromJSONString}{\code{PaginatedResultsOfWikiHeader$fromJSONString()}}
\item \href{#method-PaginatedResultsOfWikiHeader-validateJSON}{\code{PaginatedResultsOfWikiHeader$validateJSON()}}
\item \href{#method-PaginatedResultsOfWikiHeader-toString}{\code{PaginatedResultsOfWikiHeader$toString()}}
\item \href{#method-PaginatedResultsOfWikiHeader-isValid}{\code{PaginatedResultsOfWikiHeader$isValid()}}
\item \href{#method-PaginatedResultsOfWikiHeader-getInvalidFields}{\code{PaginatedResultsOfWikiHeader$getInvalidFields()}}
\item \href{#method-PaginatedResultsOfWikiHeader-print}{\code{PaginatedResultsOfWikiHeader$print()}}
\item \href{#method-PaginatedResultsOfWikiHeader-clone}{\code{PaginatedResultsOfWikiHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResultsOfWikiHeader class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$new(
  totalNumberOfResults = NULL,
  results = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalNumberOfResults}}{totalNumberOfResults}

\item{\code{results}}{results}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfWikiHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResultsOfWikiHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfWikiHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResultsOfWikiHeader in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResultsOfWikiHeader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResultsOfWikiHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResultsOfWikiHeader and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResultsOfWikiHeader
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResultsOfWikiHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResultsOfWikiHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResultsOfWikiHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
