% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Participant.R
\docType{class}
\name{Participant-class}
\alias{Participant-class}
\alias{Participant}
\title{A Reference Class to represent consistency test participants.}
\description{
A Reference Class to represent consistency test participants.
}
\section{Fields}{

\describe{
\item{\code{id}}{A one-element character vector containing the participant's ID.
Set at class new() call.}

\item{\code{test_date}}{A one-element Date vector which specifies the date
on which the participant did the consistency test.}

\item{\code{graphemes}}{A list of \code{\link[synr]{Grapheme}} class instances.}
}}

\section{Methods}{

\describe{
\item{\code{add_grapheme(grapheme)}}{Add a passed grapheme to the participant's list
of graphemes. The grapheme's entry in
the list is named based on the grapheme's
symbol. Note that if you try to add
a grapheme with a symbol that's identical
to one of the graphemes already in the
participant's list of graphemes, the
already existing same-symbol grapheme
is overwritten.}

\item{\code{add_graphemes(grapheme_vector)}}{Go through a passed list of Grapheme instances
and add each one using the add_grapheme() method.}

\item{\code{get_consistency_scores(na.rm = FALSE)}}{Returns a list of grapheme symbols with associated consistency scores}

\item{\code{get_mean_consistency_score(na.rm = FALSE)}}{Returns the mean consistency score, with respect to
Grapheme instances, that only have valid response colors,
associated with the participant.}

\item{\code{get_mean_response_time(na.rm = FALSE)}}{Returns the mean response time, with respect to all
Grapheme instances associated with the participant.
Weights response times based on number of valid responses
that each grapheme has.}

\item{\code{get_number_all_colored_graphemes()}}{Returns the number of graphemes for which all
responses have an associated non-na color}

\item{\code{get_plot(cutoff_line = TRUE)}}{Returns a ggplot2 plot that describes this participant's
grapheme color responses and per-grapheme consistency scores.}

\item{\code{get_plot_data()}}{Returns a data frame with the following columns:
                           1. Character: grapheme (grapheme names)
                           2. Numeric: consistency_score
                           3... Character columns: color_resp<x> hold response colors (number of columns depends on number of response colors associated with each grapheme).

                           The data frame is intended to be used for plotting participant data,
                           using .get_consistency_plot(). The call will end with an error
                           if not all of the participant's graphemes have the same number
                           of color responses. This is intended.}

\item{\code{get_symbols()}}{Returns a character vector with all symbols for
graphemes associated with the participant.}

\item{\code{has_graphemes()}}{Returns TRUE if there is at least one
grapheme in the participant's graphemes list,
otherwise returns FALSE}

\item{\code{save_plot(path = NULL, file_format = "png", dpi = 300, ...)}}{Saves a ggplot2 plot that describes this participant's
grapheme color responses and per-grapheme consistency scores,
using the ggsave function.
If save_dir is not specified, the plot is saved to the current
working directory. Otherwise, the plot is saved to the specified
directory. The file is saved using the specified file_format,
e. g. JPG (see ggplot2::ggsave documentation for list of
supported formats), and the resolution specified with
the dpi argument. Apart from these, all other arguments
that ggsave accepts (e. g. 'scale') also work with this function, since
all arguments are passed on to ggsave.}

\item{\code{set_date(in_date)}}{Takes in a one-element character vector with a date
in the format 'YYYY-MM-DD' and sets the participant's
test_date to the specified date.}
}}

