% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_settings_services_api.R
\docType{class}
\name{ProjectSettingsServicesApi}
\alias{ProjectSettingsServicesApi}
\title{ProjectSettingsServices operations}
\format{
An \code{R6Class} generator object
}
\description{
ProjectSettingsServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ deleteRepoV1ProjectSettingsId } \emph{  }

\itemize{
\item \emph{ @param } id character


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1ProjectSettingsId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelProjectProjectSetting} \cr


\item status code : 200 | the ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSetting}\">,ProjectSetting,</a>, with the corresponding ID, if it exists.

\item return type : OrgSagebionetworksRepoModelProjectProjectSetting
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1ProjectSettingsProjectIdTypeType } \emph{  }

\itemize{
\item \emph{ @param } project_id character
\item \emph{ @param } type \link{OrgSagebionetworksRepoModelProjectProjectSettingsType}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelProjectProjectSetting} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelProjectProjectSetting
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1StorageLocationId } \emph{  }

\itemize{
\item \emph{ @param } id numeric
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelProjectStorageLocationSetting} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelProjectStorageLocationSetting
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1ProjectSettings } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_project_project_setting \link{OrgSagebionetworksRepoModelProjectProjectSetting}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelProjectProjectSetting} \cr


\item status code : 201 | The created ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSetting}\">,ProjectSetting,</a>,.

\item return type : OrgSagebionetworksRepoModelProjectProjectSetting
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1StorageLocation } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_project_storage_location_setting \link{OrgSagebionetworksRepoModelProjectStorageLocationSetting}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelProjectStorageLocationSetting} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelProjectStorageLocationSetting
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1ProjectSettings } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_project_project_setting \link{OrgSagebionetworksRepoModelProjectProjectSetting}


\item status code : 200 | Void


\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  deleteRepoV1ProjectSettingsId  ####################

library(synr)
var_id <- "id_example" # character | The ID of the ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSetting}\">,ProjectSetting,</a>,. This is not the ID of the project.

api_instance <- ProjectSettingsServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1ProjectSettingsId(var_id)


####################  getRepoV1ProjectSettingsId  ####################

library(synr)
var_id <- "id_example" # character | the ID of the ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSetting}\">,ProjectSetting,</a>

api_instance <- ProjectSettingsServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1ProjectSettingsId(var_iddata_file = "result.txt")
result <- api_instance$getRepoV1ProjectSettingsId(var_id)
dput(result)


####################  getRepoV1ProjectSettingsProjectIdTypeType  ####################

library(synr)
var_project_id <- "project_id_example" # character | the ID of the project
var_type <- org.sagebionetworks.repo.model.project.ProjectSettingsType$new() # OrgSagebionetworksRepoModelProjectProjectSettingsType | The ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSettingsType}\">,ProjectSettingsType,</a>, to get.

api_instance <- ProjectSettingsServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1ProjectSettingsProjectIdTypeType(var_project_id, var_typedata_file = "result.txt")
result <- api_instance$getRepoV1ProjectSettingsProjectIdTypeType(var_project_id, var_type)
dput(result)


####################  getRepoV1StorageLocationId  ####################

library(synr)
var_id <- 3.4 # numeric | the ID of the storage location setting.

api_instance <- ProjectSettingsServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1StorageLocationId(var_iddata_file = "result.txt")
result <- api_instance$getRepoV1StorageLocationId(var_id)
dput(result)


####################  postRepoV1ProjectSettings  ####################

library(synr)
var_org_sagebionetworks_repo_model_project_project_setting <- org.sagebionetworks.repo.model.project.ProjectSetting$new("concreteType_example", "id_example", "projectId_example", "settingsType_example", "etag_example", c(123)) # OrgSagebionetworksRepoModelProjectProjectSetting | 

api_instance <- ProjectSettingsServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1ProjectSettings(var_org_sagebionetworks_repo_model_project_project_settingdata_file = "result.txt")
result <- api_instance$postRepoV1ProjectSettings(var_org_sagebionetworks_repo_model_project_project_setting)
dput(result)


####################  postRepoV1StorageLocation  ####################

library(synr)
var_org_sagebionetworks_repo_model_project_storage_location_setting <- org.sagebionetworks.repo.model.project.StorageLocationSetting$new("concreteType_example", "bucket_example", "endpointUrl_example", 123, "uploadType_example", "banner_example", "description_example", "etag_example", "createdOn_example", 123, "proxyUrl_example", "secretKey_example", "benefactorId_example", "baseKey_example", "stsEnabled_example", "url_example", "supportsSubfolders_example") # OrgSagebionetworksRepoModelProjectStorageLocationSetting | 

api_instance <- ProjectSettingsServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1StorageLocation(var_org_sagebionetworks_repo_model_project_storage_location_settingdata_file = "result.txt")
result <- api_instance$postRepoV1StorageLocation(var_org_sagebionetworks_repo_model_project_storage_location_setting)
dput(result)


####################  putRepoV1ProjectSettings  ####################

library(synr)
var_org_sagebionetworks_repo_model_project_project_setting <- org.sagebionetworks.repo.model.project.ProjectSetting$new("concreteType_example", "id_example", "projectId_example", "settingsType_example", "etag_example", c(123)) # OrgSagebionetworksRepoModelProjectProjectSetting | 

api_instance <- ProjectSettingsServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$putRepoV1ProjectSettings(var_org_sagebionetworks_repo_model_project_project_setting)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectSettingsServicesApi-new}{\code{ProjectSettingsServicesApi$new()}}
\item \href{#method-ProjectSettingsServicesApi-deleteRepoV1ProjectSettingsId}{\code{ProjectSettingsServicesApi$deleteRepoV1ProjectSettingsId()}}
\item \href{#method-ProjectSettingsServicesApi-deleteRepoV1ProjectSettingsIdWithHttpInfo}{\code{ProjectSettingsServicesApi$deleteRepoV1ProjectSettingsIdWithHttpInfo()}}
\item \href{#method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsId}{\code{ProjectSettingsServicesApi$getRepoV1ProjectSettingsId()}}
\item \href{#method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsIdWithHttpInfo}{\code{ProjectSettingsServicesApi$getRepoV1ProjectSettingsIdWithHttpInfo()}}
\item \href{#method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsProjectIdTypeType}{\code{ProjectSettingsServicesApi$getRepoV1ProjectSettingsProjectIdTypeType()}}
\item \href{#method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsProjectIdTypeTypeWithHttpInfo}{\code{ProjectSettingsServicesApi$getRepoV1ProjectSettingsProjectIdTypeTypeWithHttpInfo()}}
\item \href{#method-ProjectSettingsServicesApi-getRepoV1StorageLocationId}{\code{ProjectSettingsServicesApi$getRepoV1StorageLocationId()}}
\item \href{#method-ProjectSettingsServicesApi-getRepoV1StorageLocationIdWithHttpInfo}{\code{ProjectSettingsServicesApi$getRepoV1StorageLocationIdWithHttpInfo()}}
\item \href{#method-ProjectSettingsServicesApi-postRepoV1ProjectSettings}{\code{ProjectSettingsServicesApi$postRepoV1ProjectSettings()}}
\item \href{#method-ProjectSettingsServicesApi-postRepoV1ProjectSettingsWithHttpInfo}{\code{ProjectSettingsServicesApi$postRepoV1ProjectSettingsWithHttpInfo()}}
\item \href{#method-ProjectSettingsServicesApi-postRepoV1StorageLocation}{\code{ProjectSettingsServicesApi$postRepoV1StorageLocation()}}
\item \href{#method-ProjectSettingsServicesApi-postRepoV1StorageLocationWithHttpInfo}{\code{ProjectSettingsServicesApi$postRepoV1StorageLocationWithHttpInfo()}}
\item \href{#method-ProjectSettingsServicesApi-putRepoV1ProjectSettings}{\code{ProjectSettingsServicesApi$putRepoV1ProjectSettings()}}
\item \href{#method-ProjectSettingsServicesApi-putRepoV1ProjectSettingsWithHttpInfo}{\code{ProjectSettingsServicesApi$putRepoV1ProjectSettingsWithHttpInfo()}}
\item \href{#method-ProjectSettingsServicesApi-clone}{\code{ProjectSettingsServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ProjectSettingsServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-deleteRepoV1ProjectSettingsId"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-deleteRepoV1ProjectSettingsId}{}}}
\subsection{Method \code{deleteRepoV1ProjectSettingsId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$deleteRepoV1ProjectSettingsId(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The ID of the ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSetting}\">,ProjectSetting,</a>,. This is not the ID of the project.}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-deleteRepoV1ProjectSettingsIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-deleteRepoV1ProjectSettingsIdWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1ProjectSettingsIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$deleteRepoV1ProjectSettingsIdWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The ID of the ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSetting}\">,ProjectSetting,</a>,. This is not the ID of the project.}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsId"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsId}{}}}
\subsection{Method \code{getRepoV1ProjectSettingsId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$getRepoV1ProjectSettingsId(
  id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSetting}\">,ProjectSetting,</a>}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelProjectProjectSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1ProjectSettingsIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$getRepoV1ProjectSettingsIdWithHttpInfo(
  id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSetting}\">,ProjectSetting,</a>}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelProjectProjectSetting) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsProjectIdTypeType"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsProjectIdTypeType}{}}}
\subsection{Method \code{getRepoV1ProjectSettingsProjectIdTypeType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$getRepoV1ProjectSettingsProjectIdTypeType(
  project_id,
  type,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{the ID of the project}

\item{\code{type}}{The ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSettingsType}\">,ProjectSettingsType,</a>, to get.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelProjectProjectSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsProjectIdTypeTypeWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-getRepoV1ProjectSettingsProjectIdTypeTypeWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1ProjectSettingsProjectIdTypeTypeWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$getRepoV1ProjectSettingsProjectIdTypeTypeWithHttpInfo(
  project_id,
  type,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{the ID of the project}

\item{\code{type}}{The ,<a href=\"${org.sagebionetworks.repo.model.project.ProjectSettingsType}\">,ProjectSettingsType,</a>, to get.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelProjectProjectSetting) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-getRepoV1StorageLocationId"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-getRepoV1StorageLocationId}{}}}
\subsection{Method \code{getRepoV1StorageLocationId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$getRepoV1StorageLocationId(
  id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the storage location setting.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelProjectStorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-getRepoV1StorageLocationIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-getRepoV1StorageLocationIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1StorageLocationIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$getRepoV1StorageLocationIdWithHttpInfo(
  id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the storage location setting.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelProjectStorageLocationSetting) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-postRepoV1ProjectSettings"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-postRepoV1ProjectSettings}{}}}
\subsection{Method \code{postRepoV1ProjectSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$postRepoV1ProjectSettings(
  org_sagebionetworks_repo_model_project_project_setting,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_project_project_setting}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelProjectProjectSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-postRepoV1ProjectSettingsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-postRepoV1ProjectSettingsWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1ProjectSettingsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$postRepoV1ProjectSettingsWithHttpInfo(
  org_sagebionetworks_repo_model_project_project_setting,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_project_project_setting}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelProjectProjectSetting) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-postRepoV1StorageLocation"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-postRepoV1StorageLocation}{}}}
\subsection{Method \code{postRepoV1StorageLocation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$postRepoV1StorageLocation(
  org_sagebionetworks_repo_model_project_storage_location_setting,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_project_storage_location_setting}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelProjectStorageLocationSetting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-postRepoV1StorageLocationWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-postRepoV1StorageLocationWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1StorageLocationWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$postRepoV1StorageLocationWithHttpInfo(
  org_sagebionetworks_repo_model_project_storage_location_setting,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_project_storage_location_setting}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelProjectStorageLocationSetting) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-putRepoV1ProjectSettings"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-putRepoV1ProjectSettings}{}}}
\subsection{Method \code{putRepoV1ProjectSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$putRepoV1ProjectSettings(
  org_sagebionetworks_repo_model_project_project_setting,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_project_project_setting}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-putRepoV1ProjectSettingsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-putRepoV1ProjectSettingsWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1ProjectSettingsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$putRepoV1ProjectSettingsWithHttpInfo(
  org_sagebionetworks_repo_model_project_project_setting,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_project_project_setting}}{}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSettingsServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSettingsServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSettingsServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
