% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription_services_api.R
\docType{class}
\name{SubscriptionServicesApi}
\alias{SubscriptionServicesApi}
\title{SubscriptionServices operations}
\format{
An \code{R6Class} generator object
}
\description{
SubscriptionServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ deleteRepoV1SubscriptionAll } \emph{  }

\itemize{


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ deleteRepoV1SubscriptionId } \emph{  }

\itemize{
\item \emph{ @param } id character


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1SubscriptionAll } \emph{  }

\itemize{
\item \emph{ @param } limit numeric
\item \emph{ @param } offset numeric
\item \emph{ @param } object_type \link{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType}
\item \emph{ @param } sort_by \link{OrgSagebionetworksRepoModelSubscriptionSortByType}
\item \emph{ @param } sort_direction \link{OrgSagebionetworksRepoModelSubscriptionSortDirection}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelSubscriptionSubscriptionPagedResults} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelSubscriptionSubscriptionPagedResults
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1SubscriptionId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelSubscriptionSubscription} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelSubscriptionSubscription
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1Subscription } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_subscription_topic \link{OrgSagebionetworksRepoModelSubscriptionTopic}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelSubscriptionSubscription} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelSubscriptionSubscription
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1SubscriptionAll } \emph{  }

\itemize{
\item \emph{ @param } object_type \link{OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelSubscriptionSubscription} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelSubscriptionSubscription
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1SubscriptionList } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_subscription_subscription_request \link{OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelSubscriptionSubscriptionPagedResults} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelSubscriptionSubscriptionPagedResults
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1SubscriptionSubscribers } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_subscription_topic \link{OrgSagebionetworksRepoModelSubscriptionTopic}
\item \emph{ @param } next_page_token character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelSubscriptionSubscriberPagedResults} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelSubscriptionSubscriberPagedResults
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1SubscriptionSubscribersCount } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_subscription_topic \link{OrgSagebionetworksRepoModelSubscriptionTopic}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelSubscriptionSubscriberCount} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelSubscriptionSubscriberCount
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  deleteRepoV1SubscriptionAll  ####################

library(synr)

api_instance <- SubscriptionServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1SubscriptionAll()


####################  deleteRepoV1SubscriptionId  ####################

library(synr)
var_id <- "id_example" # character | - the ID of the subscription that is created when the user subscribed to the topic

api_instance <- SubscriptionServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1SubscriptionId(var_id)


####################  getRepoV1SubscriptionAll  ####################

library(synr)
var_limit <- 3.4 # numeric | - Limits the size of the page returned. For example, a page size of 10 require limit = 10. The maximum Limit for this call is 100.
var_offset <- 3.4 # numeric | - The index of the pagination offset. For a page size of 10, the first page would be at offset = 0, and the second page would be at offset = 10.
var_object_type <- org.sagebionetworks.repo.model.subscription.SubscriptionObjectType$new() # OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType | - User can use this param to filter the results by the type of object they subscribed to.
var_sort_by <- org.sagebionetworks.repo.model.subscription.SortByType$new() # OrgSagebionetworksRepoModelSubscriptionSortByType |  (Optional)
var_sort_direction <- org.sagebionetworks.repo.model.subscription.SortDirection$new() # OrgSagebionetworksRepoModelSubscriptionSortDirection | - When provided, the results will be sorted in this direction. (Optional)

api_instance <- SubscriptionServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1SubscriptionAll(var_limit, var_offset, var_object_type, sort_by = var_sort_by, sort_direction = var_sort_directiondata_file = "result.txt")
result <- api_instance$getRepoV1SubscriptionAll(var_limit, var_offset, var_object_type, sort_by = var_sort_by, sort_direction = var_sort_direction)
dput(result)


####################  getRepoV1SubscriptionId  ####################

library(synr)
var_id <- "id_example" # character | - the ID of the subscription that is created when the user subscribed to the topic

api_instance <- SubscriptionServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1SubscriptionId(var_iddata_file = "result.txt")
result <- api_instance$getRepoV1SubscriptionId(var_id)
dput(result)


####################  postRepoV1Subscription  ####################

library(synr)
var_org_sagebionetworks_repo_model_subscription_topic <- org.sagebionetworks.repo.model.subscription.Topic$new("objectId_example", "objectType_example") # OrgSagebionetworksRepoModelSubscriptionTopic | 

api_instance <- SubscriptionServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1Subscription(var_org_sagebionetworks_repo_model_subscription_topicdata_file = "result.txt")
result <- api_instance$postRepoV1Subscription(var_org_sagebionetworks_repo_model_subscription_topic)
dput(result)


####################  postRepoV1SubscriptionAll  ####################

library(synr)
var_object_type <- org.sagebionetworks.repo.model.subscription.SubscriptionObjectType$new() # OrgSagebionetworksRepoModelSubscriptionSubscriptionObjectType | - SubscriptionObjectType to subscribe to

api_instance <- SubscriptionServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1SubscriptionAll(var_object_typedata_file = "result.txt")
result <- api_instance$postRepoV1SubscriptionAll(var_object_type)
dput(result)


####################  postRepoV1SubscriptionList  ####################

library(synr)
var_org_sagebionetworks_repo_model_subscription_subscription_request <- org.sagebionetworks.repo.model.subscription.SubscriptionRequest$new("objectType_example", c("idList_example"), "sortByType_example", "sortDirection_example") # OrgSagebionetworksRepoModelSubscriptionSubscriptionRequest | 

api_instance <- SubscriptionServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1SubscriptionList(var_org_sagebionetworks_repo_model_subscription_subscription_requestdata_file = "result.txt")
result <- api_instance$postRepoV1SubscriptionList(var_org_sagebionetworks_repo_model_subscription_subscription_request)
dput(result)


####################  postRepoV1SubscriptionSubscribers  ####################

library(synr)
var_org_sagebionetworks_repo_model_subscription_topic <- org.sagebionetworks.repo.model.subscription.Topic$new("objectId_example", "objectType_example") # OrgSagebionetworksRepoModelSubscriptionTopic | 
var_next_page_token <- "next_page_token_example" # character |  (Optional)

api_instance <- SubscriptionServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1SubscriptionSubscribers(var_org_sagebionetworks_repo_model_subscription_topic, next_page_token = var_next_page_tokendata_file = "result.txt")
result <- api_instance$postRepoV1SubscriptionSubscribers(var_org_sagebionetworks_repo_model_subscription_topic, next_page_token = var_next_page_token)
dput(result)


####################  postRepoV1SubscriptionSubscribersCount  ####################

library(synr)
var_org_sagebionetworks_repo_model_subscription_topic <- org.sagebionetworks.repo.model.subscription.Topic$new("objectId_example", "objectType_example") # OrgSagebionetworksRepoModelSubscriptionTopic | 

api_instance <- SubscriptionServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1SubscriptionSubscribersCount(var_org_sagebionetworks_repo_model_subscription_topicdata_file = "result.txt")
result <- api_instance$postRepoV1SubscriptionSubscribersCount(var_org_sagebionetworks_repo_model_subscription_topic)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SubscriptionServicesApi-new}{\code{SubscriptionServicesApi$new()}}
\item \href{#method-SubscriptionServicesApi-deleteRepoV1SubscriptionAll}{\code{SubscriptionServicesApi$deleteRepoV1SubscriptionAll()}}
\item \href{#method-SubscriptionServicesApi-deleteRepoV1SubscriptionAllWithHttpInfo}{\code{SubscriptionServicesApi$deleteRepoV1SubscriptionAllWithHttpInfo()}}
\item \href{#method-SubscriptionServicesApi-deleteRepoV1SubscriptionId}{\code{SubscriptionServicesApi$deleteRepoV1SubscriptionId()}}
\item \href{#method-SubscriptionServicesApi-deleteRepoV1SubscriptionIdWithHttpInfo}{\code{SubscriptionServicesApi$deleteRepoV1SubscriptionIdWithHttpInfo()}}
\item \href{#method-SubscriptionServicesApi-getRepoV1SubscriptionAll}{\code{SubscriptionServicesApi$getRepoV1SubscriptionAll()}}
\item \href{#method-SubscriptionServicesApi-getRepoV1SubscriptionAllWithHttpInfo}{\code{SubscriptionServicesApi$getRepoV1SubscriptionAllWithHttpInfo()}}
\item \href{#method-SubscriptionServicesApi-getRepoV1SubscriptionId}{\code{SubscriptionServicesApi$getRepoV1SubscriptionId()}}
\item \href{#method-SubscriptionServicesApi-getRepoV1SubscriptionIdWithHttpInfo}{\code{SubscriptionServicesApi$getRepoV1SubscriptionIdWithHttpInfo()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1Subscription}{\code{SubscriptionServicesApi$postRepoV1Subscription()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1SubscriptionWithHttpInfo}{\code{SubscriptionServicesApi$postRepoV1SubscriptionWithHttpInfo()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1SubscriptionAll}{\code{SubscriptionServicesApi$postRepoV1SubscriptionAll()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1SubscriptionAllWithHttpInfo}{\code{SubscriptionServicesApi$postRepoV1SubscriptionAllWithHttpInfo()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1SubscriptionList}{\code{SubscriptionServicesApi$postRepoV1SubscriptionList()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1SubscriptionListWithHttpInfo}{\code{SubscriptionServicesApi$postRepoV1SubscriptionListWithHttpInfo()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribers}{\code{SubscriptionServicesApi$postRepoV1SubscriptionSubscribers()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribersWithHttpInfo}{\code{SubscriptionServicesApi$postRepoV1SubscriptionSubscribersWithHttpInfo()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribersCount}{\code{SubscriptionServicesApi$postRepoV1SubscriptionSubscribersCount()}}
\item \href{#method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribersCountWithHttpInfo}{\code{SubscriptionServicesApi$postRepoV1SubscriptionSubscribersCountWithHttpInfo()}}
\item \href{#method-SubscriptionServicesApi-clone}{\code{SubscriptionServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SubscriptionServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-deleteRepoV1SubscriptionAll"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-deleteRepoV1SubscriptionAll}{}}}
\subsection{Method \code{deleteRepoV1SubscriptionAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$deleteRepoV1SubscriptionAll(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-deleteRepoV1SubscriptionAllWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-deleteRepoV1SubscriptionAllWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1SubscriptionAllWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$deleteRepoV1SubscriptionAllWithHttpInfo(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-deleteRepoV1SubscriptionId"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-deleteRepoV1SubscriptionId}{}}}
\subsection{Method \code{deleteRepoV1SubscriptionId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$deleteRepoV1SubscriptionId(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{- the ID of the subscription that is created when the user subscribed to the topic}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-deleteRepoV1SubscriptionIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-deleteRepoV1SubscriptionIdWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1SubscriptionIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$deleteRepoV1SubscriptionIdWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{- the ID of the subscription that is created when the user subscribed to the topic}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-getRepoV1SubscriptionAll"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-getRepoV1SubscriptionAll}{}}}
\subsection{Method \code{getRepoV1SubscriptionAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$getRepoV1SubscriptionAll(
  limit,
  offset,
  object_type,
  sort_by = NULL,
  sort_direction = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{- Limits the size of the page returned. For example, a page size of 10 require limit = 10. The maximum Limit for this call is 100.}

\item{\code{offset}}{- The index of the pagination offset. For a page size of 10, the first page would be at offset = 0, and the second page would be at offset = 10.}

\item{\code{object_type}}{- User can use this param to filter the results by the type of object they subscribed to.}

\item{\code{sort_by}}{(optional) No description}

\item{\code{sort_direction}}{(optional) - When provided, the results will be sorted in this direction.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriptionPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-getRepoV1SubscriptionAllWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-getRepoV1SubscriptionAllWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1SubscriptionAllWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$getRepoV1SubscriptionAllWithHttpInfo(
  limit,
  offset,
  object_type,
  sort_by = NULL,
  sort_direction = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{- Limits the size of the page returned. For example, a page size of 10 require limit = 10. The maximum Limit for this call is 100.}

\item{\code{offset}}{- The index of the pagination offset. For a page size of 10, the first page would be at offset = 0, and the second page would be at offset = 10.}

\item{\code{object_type}}{- User can use this param to filter the results by the type of object they subscribed to.}

\item{\code{sort_by}}{(optional) No description}

\item{\code{sort_direction}}{(optional) - When provided, the results will be sorted in this direction.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelSubscriptionSubscriptionPagedResults) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-getRepoV1SubscriptionId"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-getRepoV1SubscriptionId}{}}}
\subsection{Method \code{getRepoV1SubscriptionId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$getRepoV1SubscriptionId(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{- the ID of the subscription that is created when the user subscribed to the topic}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-getRepoV1SubscriptionIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-getRepoV1SubscriptionIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1SubscriptionIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$getRepoV1SubscriptionIdWithHttpInfo(
  id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{- the ID of the subscription that is created when the user subscribed to the topic}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelSubscriptionSubscription) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1Subscription"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1Subscription}{}}}
\subsection{Method \code{postRepoV1Subscription()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1Subscription(
  org_sagebionetworks_repo_model_subscription_topic,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_subscription_topic}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1SubscriptionWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1SubscriptionWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1SubscriptionWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1SubscriptionWithHttpInfo(
  org_sagebionetworks_repo_model_subscription_topic,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_subscription_topic}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelSubscriptionSubscription) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1SubscriptionAll"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1SubscriptionAll}{}}}
\subsection{Method \code{postRepoV1SubscriptionAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1SubscriptionAll(
  object_type,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_type}}{- SubscriptionObjectType to subscribe to}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscription
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1SubscriptionAllWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1SubscriptionAllWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1SubscriptionAllWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1SubscriptionAllWithHttpInfo(
  object_type,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_type}}{- SubscriptionObjectType to subscribe to}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelSubscriptionSubscription) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1SubscriptionList"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1SubscriptionList}{}}}
\subsection{Method \code{postRepoV1SubscriptionList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1SubscriptionList(
  org_sagebionetworks_repo_model_subscription_subscription_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_subscription_subscription_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriptionPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1SubscriptionListWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1SubscriptionListWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1SubscriptionListWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1SubscriptionListWithHttpInfo(
  org_sagebionetworks_repo_model_subscription_subscription_request,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_subscription_subscription_request}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelSubscriptionSubscriptionPagedResults) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribers"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribers}{}}}
\subsection{Method \code{postRepoV1SubscriptionSubscribers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1SubscriptionSubscribers(
  org_sagebionetworks_repo_model_subscription_topic,
  next_page_token = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_subscription_topic}}{}

\item{\code{next_page_token}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriberPagedResults
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribersWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribersWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1SubscriptionSubscribersWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1SubscriptionSubscribersWithHttpInfo(
  org_sagebionetworks_repo_model_subscription_topic,
  next_page_token = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_subscription_topic}}{}

\item{\code{next_page_token}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelSubscriptionSubscriberPagedResults) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribersCount"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribersCount}{}}}
\subsection{Method \code{postRepoV1SubscriptionSubscribersCount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1SubscriptionSubscribersCount(
  org_sagebionetworks_repo_model_subscription_topic,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_subscription_topic}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelSubscriptionSubscriberCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribersCountWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-postRepoV1SubscriptionSubscribersCountWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1SubscriptionSubscribersCountWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$postRepoV1SubscriptionSubscribersCountWithHttpInfo(
  org_sagebionetworks_repo_model_subscription_topic,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_subscription_topic}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelSubscriptionSubscriberCount) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubscriptionServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SubscriptionServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
