% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synr_exampledf_large}
\alias{synr_exampledf_large}
\title{Raw consistency test data example, long format}
\format{
A data frame with 516 rows and 4 columns:
\describe{
  \item{participant_id}{Participant ID}
  \item{trial_symbol}{Column of trial symbols/graphemes}
  \item{response_color}{Column of trial response colors}
  \item{response_time}{Column of trial response times}
}
}
\usage{
synr_exampledf_large
}
\description{
A data frame with an example of raw consistency test data that are
compatible with the synr package's `create_participantgroup' function.
The color and 'symbol' data are from five actual participants who did
a test that included all letters, digits and weekdays, with 3 trials
per grapheme. The response times are randomly generated. 
Note that response times are optional. If you don't have
them, you can still use synr - see `help(create_participantgroup_widedata)`.
}
\keyword{datasets}
