% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synr_exampledf_long_small}
\alias{synr_exampledf_long_small}
\title{Raw consistency test data example, long format (small)}
\format{
A data frame with 18 rows and 4 columns:
\describe{
  \item{participant_id}{Participant ID}
  \item{trial_symbol}{Column of trial symbols/graphemes}
  \item{response_color}{Column of trial response colors}
  \item{response_time}{Column of trial response times}
}
}
\usage{
synr_exampledf_long_small
}
\description{
A data frame with an example of raw consistency test data that are
compatible with the synr package's `create_participantgroup` function,
with completely made updata for three participants from a hypothetical 
test that included three graphemes ("A", "D", 7) and two responses
per grapheme. More graphemes and/or
responses per grapheme can be handled by the package (though participant
plots do not function correctly if there are more than three responses
per grapheme). Note that response times are optional. If you don't have
them, you can still use synr - see `help(create_participantgroup_widedata)`.
}
\keyword{datasets}
