% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_services_api.R
\docType{class}
\name{TeamServicesApi}
\alias{TeamServicesApi}
\title{TeamServices operations}
\format{
An \code{R6Class} generator object
}
\description{
TeamServicesApi
}
\details{
Synapse REST API

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: v1
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ deleteRepoV1TeamId } \emph{  }

\itemize{
\item \emph{ @param } id character


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ deleteRepoV1TeamIdMemberPrincipalId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } principal_id character


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1TeamId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelTeam} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelTeam
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1TeamIdAcl } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAccessControlList} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAccessControlList
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1TeamIdIcon } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } redirect character
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Status 200 will be returned if the 'redirect' boolean param is false

\item return type : AnyType
\item response headers :

\tabular{ll}{
}
\item status code : 307 | Status 307 will be returned if the 'redirect' boolean param is true or null


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1TeamIdIconPreview } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } redirect character
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Status 200 will be returned if the 'redirect' boolean param is false

\item return type : AnyType
\item response headers :

\tabular{ll}{
}
\item status code : 307 | Status 307 will be returned if the 'redirect' boolean param is true or null


\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1TeamIdMemberPrincipalId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } principal_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelTeamMember} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelTeamMember
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1TeamIdMemberPrincipalIdMembershipStatus } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } principal_id character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelTeamMembershipStatus} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelTeamMembershipStatus
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1TeamMembersCountId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } fragment character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelCount} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelCount
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1TeamMembersId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } fragment character
\item \emph{ @param } member_type \link{OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions}
\item \emph{ @param } limit integer
\item \emph{ @param } offset integer
\item \emph{ @returnType } \link{PaginatedResultsOfTeamMember} \cr


\item status code : 200 | Auto-generated description

\item return type : PaginatedResultsOfTeamMember
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1Teams } \emph{  }

\itemize{
\item \emph{ @param } fragment character
\item \emph{ @param } limit integer
\item \emph{ @param } offset integer
\item \emph{ @returnType } \link{PaginatedResultsOfTeam} \cr


\item status code : 200 | Auto-generated description

\item return type : PaginatedResultsOfTeam
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1UserIdTeam } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } limit integer
\item \emph{ @param } offset integer
\item \emph{ @returnType } \link{PaginatedResultsOfTeam} \cr


\item status code : 200 | Auto-generated description

\item return type : PaginatedResultsOfTeam
\item response headers :

\tabular{ll}{
}
}

\strong{ getRepoV1UserIdTeamId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } next_page_token character
\item \emph{ @param } sort \link{OrgSagebionetworksRepoModelTeamSortOrder}
\item \emph{ @param } ascending character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelPaginatedTeamIds} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelPaginatedTeamIds
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1Team } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_team \link{OrgSagebionetworksRepoModelTeam}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelTeam} \cr


\item status code : 201 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelTeam
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1TeamIdMemberList } \emph{  }

\itemize{
\item \emph{ @param } id numeric
\item \emph{ @param } org_sagebionetworks_repo_model_id_list \link{OrgSagebionetworksRepoModelIdList}
\item \emph{ @returnType } \link{ListWrapperOfTeamMember} \cr


\item status code : 200 | Auto-generated description

\item return type : ListWrapperOfTeamMember
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1TeamList } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_id_list \link{OrgSagebionetworksRepoModelIdList}
\item \emph{ @returnType } \link{ListWrapperOfTeam} \cr


\item status code : 200 | Auto-generated description

\item return type : ListWrapperOfTeam
\item response headers :

\tabular{ll}{
}
}

\strong{ postRepoV1UserIdMemberList } \emph{  }

\itemize{
\item \emph{ @param } id numeric
\item \emph{ @param } org_sagebionetworks_repo_model_id_list \link{OrgSagebionetworksRepoModelIdList}
\item \emph{ @returnType } \link{ListWrapperOfTeamMember} \cr


\item status code : 200 | Auto-generated description

\item return type : ListWrapperOfTeamMember
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1Team } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_team \link{OrgSagebionetworksRepoModelTeam}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelTeam} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelTeam
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1TeamAcl } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_access_control_list \link{OrgSagebionetworksRepoModelAccessControlList}
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelAccessControlList} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelAccessControlList
\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1TeamIdMemberPrincipalId } \emph{  }

\itemize{
\item \emph{ @param } id character
\item \emph{ @param } principal_id character
\item \emph{ @param } team_endpoint character
\item \emph{ @param } notification_unsubscribe_endpoint character


\item status code : 204 | Void


\item response headers :

\tabular{ll}{
}
}

\strong{ putRepoV1TeamMember } \emph{  }

\itemize{
\item \emph{ @param } org_sagebionetworks_repo_model_join_team_signed_token \link{OrgSagebionetworksRepoModelJoinTeamSignedToken}
\item \emph{ @param } team_endpoint character
\item \emph{ @param } notification_unsubscribe_endpoint character
\item \emph{ @returnType } \link{OrgSagebionetworksRepoModelResponseMessage} \cr


\item status code : 200 | Auto-generated description

\item return type : OrgSagebionetworksRepoModelResponseMessage
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  deleteRepoV1TeamId  ####################

library(synr)
var_id <- "id_example" # character | the ID of the Team to delete.

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1TeamId(var_id)


####################  deleteRepoV1TeamIdMemberPrincipalId  ####################

library(synr)
var_id <- "id_example" # character | the Team ID
var_principal_id <- "principal_id_example" # character | the member's principal ID

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$deleteRepoV1TeamIdMemberPrincipalId(var_id, var_principal_id)


####################  getRepoV1TeamId  ####################

library(synr)
var_id <- "id_example" # character | the ID of the Team of interest

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1TeamId(var_iddata_file = "result.txt")
result <- api_instance$getRepoV1TeamId(var_id)
dput(result)


####################  getRepoV1TeamIdAcl  ####################

library(synr)
var_id <- "id_example" # character | the ID of the Team of interest

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1TeamIdAcl(var_iddata_file = "result.txt")
result <- api_instance$getRepoV1TeamIdAcl(var_id)
dput(result)


####################  getRepoV1TeamIdIcon  ####################

library(synr)
var_id <- "id_example" # character | the ID of the Team
var_redirect <- "redirect_example" # character | if true or omitted, then redirect to the URL.  If false then simply return the URL. (Optional)

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1TeamIdIcon(var_id, redirect = var_redirectdata_file = "result.txt")
result <- api_instance$getRepoV1TeamIdIcon(var_id, redirect = var_redirect)
dput(result)


####################  getRepoV1TeamIdIconPreview  ####################

library(synr)
var_id <- "id_example" # character | the ID of the Team
var_redirect <- "redirect_example" # character | if true or omitted, then redirect to the URL.  If false then simply return the URL. (Optional)

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1TeamIdIconPreview(var_id, redirect = var_redirectdata_file = "result.txt")
result <- api_instance$getRepoV1TeamIdIconPreview(var_id, redirect = var_redirect)
dput(result)


####################  getRepoV1TeamIdMemberPrincipalId  ####################

library(synr)
var_id <- "id_example" # character | 
var_principal_id <- "principal_id_example" # character | 

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1TeamIdMemberPrincipalId(var_id, var_principal_iddata_file = "result.txt")
result <- api_instance$getRepoV1TeamIdMemberPrincipalId(var_id, var_principal_id)
dput(result)


####################  getRepoV1TeamIdMemberPrincipalIdMembershipStatus  ####################

library(synr)
var_id <- "id_example" # character | the Team ID
var_principal_id <- "principal_id_example" # character | the user ID

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1TeamIdMemberPrincipalIdMembershipStatus(var_id, var_principal_iddata_file = "result.txt")
result <- api_instance$getRepoV1TeamIdMemberPrincipalIdMembershipStatus(var_id, var_principal_id)
dput(result)


####################  getRepoV1TeamMembersCountId  ####################

library(synr)
var_id <- "id_example" # character | the id of the Team of interest
var_fragment <- "fragment_example" # character | a prefix of the user's first or last name or email address (optional) (Optional)

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1TeamMembersCountId(var_id, fragment = var_fragmentdata_file = "result.txt")
result <- api_instance$getRepoV1TeamMembersCountId(var_id, fragment = var_fragment)
dput(result)


####################  getRepoV1TeamMembersId  ####################

library(synr)
var_id <- "id_example" # character | the id of the Team of interest
var_fragment <- "fragment_example" # character | a prefix of the user's first or last name or email address (optional) (Optional)
var_member_type <- org.sagebionetworks.repo.model.TeamMemberTypeFilterOptions$new() # OrgSagebionetworksRepoModelTeamMemberTypeFilterOptions | the type of team user to retrieve (optional; default \"ALL\") (Optional)
var_limit <- 56 # integer | the maximum number of members to return (default 10, max limit 50) (Optional)
var_offset <- 56 # integer | the starting index of the returned results (default 0) (Optional)

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1TeamMembersId(var_id, fragment = var_fragment, member_type = var_member_type, limit = var_limit, offset = var_offsetdata_file = "result.txt")
result <- api_instance$getRepoV1TeamMembersId(var_id, fragment = var_fragment, member_type = var_member_type, limit = var_limit, offset = var_offset)
dput(result)


####################  getRepoV1Teams  ####################

library(synr)
var_fragment <- "fragment_example" # character | a prefix of the Team name, or a prefix of any sub-string in the name preceded by a space.    If omitted, all Teams are returned. (Optional)
var_limit <- 56 # integer | the maximum number of Teams to return (default 10, max limit 50) (Optional)
var_offset <- 56 # integer | the starting index of the returned results (default 0) (Optional)

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1Teams(fragment = var_fragment, limit = var_limit, offset = var_offsetdata_file = "result.txt")
result <- api_instance$getRepoV1Teams(fragment = var_fragment, limit = var_limit, offset = var_offset)
dput(result)


####################  getRepoV1UserIdTeam  ####################

library(synr)
var_id <- "id_example" # character | the principal ID of the user of interest.
var_limit <- 56 # integer | the maximum number of Teams to return (default 10) (Optional)
var_offset <- 56 # integer | the starting index of the returned results (default 0) (Optional)

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1UserIdTeam(var_id, limit = var_limit, offset = var_offsetdata_file = "result.txt")
result <- api_instance$getRepoV1UserIdTeam(var_id, limit = var_limit, offset = var_offset)
dput(result)


####################  getRepoV1UserIdTeamId  ####################

library(synr)
var_id <- "id_example" # character | 
var_next_page_token <- "next_page_token_example" # character | controls pagination (Optional)
var_sort <- org.sagebionetworks.repo.model.TeamSortOrder$new() # OrgSagebionetworksRepoModelTeamSortOrder | the field to sort the team IDs on. Available options ,<a href=\"${org.sagebionetworks.repo.model.TeamSortOrder}\">,TeamSortOrder,</a> (Optional)
var_ascending <- "ascending_example" # character | the direction of sort: true for ascending, and false for descending (Optional)

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$getRepoV1UserIdTeamId(var_id, next_page_token = var_next_page_token, sort = var_sort, ascending = var_ascendingdata_file = "result.txt")
result <- api_instance$getRepoV1UserIdTeamId(var_id, next_page_token = var_next_page_token, sort = var_sort, ascending = var_ascending)
dput(result)


####################  postRepoV1Team  ####################

library(synr)
var_org_sagebionetworks_repo_model_team <- org.sagebionetworks.repo.model.Team$new("id_example", "name_example", "description_example", "icon_example", "canPublicJoin_example", "canRequestMembership_example", "etag_example", "createdOn_example", "modifiedOn_example", "createdBy_example", "modifiedBy_example") # OrgSagebionetworksRepoModelTeam | 

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1Team(var_org_sagebionetworks_repo_model_teamdata_file = "result.txt")
result <- api_instance$postRepoV1Team(var_org_sagebionetworks_repo_model_team)
dput(result)


####################  postRepoV1TeamIdMemberList  ####################

library(synr)
var_id <- 3.4 # numeric | 
var_org_sagebionetworks_repo_model_id_list <- org.sagebionetworks.repo.model.IdList$new(c(123)) # OrgSagebionetworksRepoModelIdList | 

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1TeamIdMemberList(var_id, var_org_sagebionetworks_repo_model_id_listdata_file = "result.txt")
result <- api_instance$postRepoV1TeamIdMemberList(var_id, var_org_sagebionetworks_repo_model_id_list)
dput(result)


####################  postRepoV1TeamList  ####################

library(synr)
var_org_sagebionetworks_repo_model_id_list <- org.sagebionetworks.repo.model.IdList$new(c(123)) # OrgSagebionetworksRepoModelIdList | 

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1TeamList(var_org_sagebionetworks_repo_model_id_listdata_file = "result.txt")
result <- api_instance$postRepoV1TeamList(var_org_sagebionetworks_repo_model_id_list)
dput(result)


####################  postRepoV1UserIdMemberList  ####################

library(synr)
var_id <- 3.4 # numeric | user's ID
var_org_sagebionetworks_repo_model_id_list <- org.sagebionetworks.repo.model.IdList$new(c(123)) # OrgSagebionetworksRepoModelIdList | 

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$postRepoV1UserIdMemberList(var_id, var_org_sagebionetworks_repo_model_id_listdata_file = "result.txt")
result <- api_instance$postRepoV1UserIdMemberList(var_id, var_org_sagebionetworks_repo_model_id_list)
dput(result)


####################  putRepoV1Team  ####################

library(synr)
var_org_sagebionetworks_repo_model_team <- org.sagebionetworks.repo.model.Team$new("id_example", "name_example", "description_example", "icon_example", "canPublicJoin_example", "canRequestMembership_example", "etag_example", "createdOn_example", "modifiedOn_example", "createdBy_example", "modifiedBy_example") # OrgSagebionetworksRepoModelTeam | 

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1Team(var_org_sagebionetworks_repo_model_teamdata_file = "result.txt")
result <- api_instance$putRepoV1Team(var_org_sagebionetworks_repo_model_team)
dput(result)


####################  putRepoV1TeamAcl  ####################

library(synr)
var_org_sagebionetworks_repo_model_access_control_list <- org.sagebionetworks.repo.model.AccessControlList$new("id_example", "createdBy_example", "creationDate_example", "modifiedBy_example", "modifiedOn_example", "etag_example", c(org.sagebionetworks.repo.model.ResourceAccess$new(123, c("accessType_example")))) # OrgSagebionetworksRepoModelAccessControlList | 

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1TeamAcl(var_org_sagebionetworks_repo_model_access_control_listdata_file = "result.txt")
result <- api_instance$putRepoV1TeamAcl(var_org_sagebionetworks_repo_model_access_control_list)
dput(result)


####################  putRepoV1TeamIdMemberPrincipalId  ####################

library(synr)
var_id <- "id_example" # character | the ID of the Team to which the user is to be added.
var_principal_id <- "principal_id_example" # character | the ID of the user to be added to the Team.
var_team_endpoint <- "team_endpoint_example" # character | the portal prefix for the Team URL. The team ID is appended to create the complete URL. (Optional)
var_notification_unsubscribe_endpoint <- "notification_unsubscribe_endpoint_example" # character | the portal prefix for one-click email unsubscription.    A signed, serialized token is appended to create the complete URL:   ,<,ahref=\"${org.sagebionetworks.repo.model.message.NotificationSettingsSignedToken}\",>,NotificationSettingsSignedToken,</a> (Optional)

api_instance <- TeamServicesApi$new()

# Configure HTTP bearer authorization: bearerAuth
api_instance$api_client$bearer_token <- Sys.getenv("BEARER_TOKEN")

api_instance$putRepoV1TeamIdMemberPrincipalId(var_id, var_principal_id, team_endpoint = var_team_endpoint, notification_unsubscribe_endpoint = var_notification_unsubscribe_endpoint)


####################  putRepoV1TeamMember  ####################

library(synr)
var_org_sagebionetworks_repo_model_join_team_signed_token <- org.sagebionetworks.repo.model.JoinTeamSignedToken$new("org.sagebionetworks.repo.model.JoinTeamSignedToken", "hmac_example", 123, "expiresOn_example", "createdOn_example", "userId_example", "teamId_example", "memberId_example") # OrgSagebionetworksRepoModelJoinTeamSignedToken | 
var_team_endpoint <- "team_endpoint_example" # character |  (Optional)
var_notification_unsubscribe_endpoint <- "notification_unsubscribe_endpoint_example" # character |  (Optional)

api_instance <- TeamServicesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$putRepoV1TeamMember(var_org_sagebionetworks_repo_model_join_team_signed_token, team_endpoint = var_team_endpoint, notification_unsubscribe_endpoint = var_notification_unsubscribe_endpointdata_file = "result.txt")
result <- api_instance$putRepoV1TeamMember(var_org_sagebionetworks_repo_model_join_team_signed_token, team_endpoint = var_team_endpoint, notification_unsubscribe_endpoint = var_notification_unsubscribe_endpoint)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TeamServicesApi-new}{\code{TeamServicesApi$new()}}
\item \href{#method-TeamServicesApi-deleteRepoV1TeamId}{\code{TeamServicesApi$deleteRepoV1TeamId()}}
\item \href{#method-TeamServicesApi-deleteRepoV1TeamIdWithHttpInfo}{\code{TeamServicesApi$deleteRepoV1TeamIdWithHttpInfo()}}
\item \href{#method-TeamServicesApi-deleteRepoV1TeamIdMemberPrincipalId}{\code{TeamServicesApi$deleteRepoV1TeamIdMemberPrincipalId()}}
\item \href{#method-TeamServicesApi-deleteRepoV1TeamIdMemberPrincipalIdWithHttpInfo}{\code{TeamServicesApi$deleteRepoV1TeamIdMemberPrincipalIdWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamId}{\code{TeamServicesApi$getRepoV1TeamId()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdWithHttpInfo}{\code{TeamServicesApi$getRepoV1TeamIdWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdAcl}{\code{TeamServicesApi$getRepoV1TeamIdAcl()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdAclWithHttpInfo}{\code{TeamServicesApi$getRepoV1TeamIdAclWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdIcon}{\code{TeamServicesApi$getRepoV1TeamIdIcon()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdIconWithHttpInfo}{\code{TeamServicesApi$getRepoV1TeamIdIconWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdIconPreview}{\code{TeamServicesApi$getRepoV1TeamIdIconPreview()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdIconPreviewWithHttpInfo}{\code{TeamServicesApi$getRepoV1TeamIdIconPreviewWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalId}{\code{TeamServicesApi$getRepoV1TeamIdMemberPrincipalId()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalIdWithHttpInfo}{\code{TeamServicesApi$getRepoV1TeamIdMemberPrincipalIdWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalIdMembershipStatus}{\code{TeamServicesApi$getRepoV1TeamIdMemberPrincipalIdMembershipStatus()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalIdMembershipStatusWithHttpInfo}{\code{TeamServicesApi$getRepoV1TeamIdMemberPrincipalIdMembershipStatusWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamMembersCountId}{\code{TeamServicesApi$getRepoV1TeamMembersCountId()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamMembersCountIdWithHttpInfo}{\code{TeamServicesApi$getRepoV1TeamMembersCountIdWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamMembersId}{\code{TeamServicesApi$getRepoV1TeamMembersId()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamMembersIdWithHttpInfo}{\code{TeamServicesApi$getRepoV1TeamMembersIdWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1Teams}{\code{TeamServicesApi$getRepoV1Teams()}}
\item \href{#method-TeamServicesApi-getRepoV1TeamsWithHttpInfo}{\code{TeamServicesApi$getRepoV1TeamsWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1UserIdTeam}{\code{TeamServicesApi$getRepoV1UserIdTeam()}}
\item \href{#method-TeamServicesApi-getRepoV1UserIdTeamWithHttpInfo}{\code{TeamServicesApi$getRepoV1UserIdTeamWithHttpInfo()}}
\item \href{#method-TeamServicesApi-getRepoV1UserIdTeamId}{\code{TeamServicesApi$getRepoV1UserIdTeamId()}}
\item \href{#method-TeamServicesApi-getRepoV1UserIdTeamIdWithHttpInfo}{\code{TeamServicesApi$getRepoV1UserIdTeamIdWithHttpInfo()}}
\item \href{#method-TeamServicesApi-postRepoV1Team}{\code{TeamServicesApi$postRepoV1Team()}}
\item \href{#method-TeamServicesApi-postRepoV1TeamWithHttpInfo}{\code{TeamServicesApi$postRepoV1TeamWithHttpInfo()}}
\item \href{#method-TeamServicesApi-postRepoV1TeamIdMemberList}{\code{TeamServicesApi$postRepoV1TeamIdMemberList()}}
\item \href{#method-TeamServicesApi-postRepoV1TeamIdMemberListWithHttpInfo}{\code{TeamServicesApi$postRepoV1TeamIdMemberListWithHttpInfo()}}
\item \href{#method-TeamServicesApi-postRepoV1TeamList}{\code{TeamServicesApi$postRepoV1TeamList()}}
\item \href{#method-TeamServicesApi-postRepoV1TeamListWithHttpInfo}{\code{TeamServicesApi$postRepoV1TeamListWithHttpInfo()}}
\item \href{#method-TeamServicesApi-postRepoV1UserIdMemberList}{\code{TeamServicesApi$postRepoV1UserIdMemberList()}}
\item \href{#method-TeamServicesApi-postRepoV1UserIdMemberListWithHttpInfo}{\code{TeamServicesApi$postRepoV1UserIdMemberListWithHttpInfo()}}
\item \href{#method-TeamServicesApi-putRepoV1Team}{\code{TeamServicesApi$putRepoV1Team()}}
\item \href{#method-TeamServicesApi-putRepoV1TeamWithHttpInfo}{\code{TeamServicesApi$putRepoV1TeamWithHttpInfo()}}
\item \href{#method-TeamServicesApi-putRepoV1TeamAcl}{\code{TeamServicesApi$putRepoV1TeamAcl()}}
\item \href{#method-TeamServicesApi-putRepoV1TeamAclWithHttpInfo}{\code{TeamServicesApi$putRepoV1TeamAclWithHttpInfo()}}
\item \href{#method-TeamServicesApi-putRepoV1TeamIdMemberPrincipalId}{\code{TeamServicesApi$putRepoV1TeamIdMemberPrincipalId()}}
\item \href{#method-TeamServicesApi-putRepoV1TeamIdMemberPrincipalIdWithHttpInfo}{\code{TeamServicesApi$putRepoV1TeamIdMemberPrincipalIdWithHttpInfo()}}
\item \href{#method-TeamServicesApi-putRepoV1TeamMember}{\code{TeamServicesApi$putRepoV1TeamMember()}}
\item \href{#method-TeamServicesApi-putRepoV1TeamMemberWithHttpInfo}{\code{TeamServicesApi$putRepoV1TeamMemberWithHttpInfo()}}
\item \href{#method-TeamServicesApi-clone}{\code{TeamServicesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TeamServicesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-deleteRepoV1TeamId"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-deleteRepoV1TeamId}{}}}
\subsection{Method \code{deleteRepoV1TeamId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$deleteRepoV1TeamId(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team to delete.}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-deleteRepoV1TeamIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-deleteRepoV1TeamIdWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1TeamIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$deleteRepoV1TeamIdWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team to delete.}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-deleteRepoV1TeamIdMemberPrincipalId"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-deleteRepoV1TeamIdMemberPrincipalId}{}}}
\subsection{Method \code{deleteRepoV1TeamIdMemberPrincipalId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$deleteRepoV1TeamIdMemberPrincipalId(id, principal_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the Team ID}

\item{\code{principal_id}}{the member's principal ID}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-deleteRepoV1TeamIdMemberPrincipalIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-deleteRepoV1TeamIdMemberPrincipalIdWithHttpInfo}{}}}
\subsection{Method \code{deleteRepoV1TeamIdMemberPrincipalIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$deleteRepoV1TeamIdMemberPrincipalIdWithHttpInfo(
  id,
  principal_id,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the Team ID}

\item{\code{principal_id}}{the member's principal ID}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamId"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamId}{}}}
\subsection{Method \code{getRepoV1TeamId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamId(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team of interest}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1TeamIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdWithHttpInfo(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team of interest}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelTeam) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdAcl"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdAcl}{}}}
\subsection{Method \code{getRepoV1TeamIdAcl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdAcl(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team of interest}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAccessControlList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdAclWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdAclWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1TeamIdAclWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdAclWithHttpInfo(id, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team of interest}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAccessControlList) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdIcon"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdIcon}{}}}
\subsection{Method \code{getRepoV1TeamIdIcon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdIcon(id, redirect = NULL, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team}

\item{\code{redirect}}{(optional) if true or omitted, then redirect to the URL.  If false then simply return the URL.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdIconWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdIconWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1TeamIdIconWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdIconWithHttpInfo(
  id,
  redirect = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team}

\item{\code{redirect}}{(optional) if true or omitted, then redirect to the URL.  If false then simply return the URL.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdIconPreview"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdIconPreview}{}}}
\subsection{Method \code{getRepoV1TeamIdIconPreview()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdIconPreview(
  id,
  redirect = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team}

\item{\code{redirect}}{(optional) if true or omitted, then redirect to the URL.  If false then simply return the URL.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdIconPreviewWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdIconPreviewWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1TeamIdIconPreviewWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdIconPreviewWithHttpInfo(
  id,
  redirect = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team}

\item{\code{redirect}}{(optional) if true or omitted, then redirect to the URL.  If false then simply return the URL.}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalId"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalId}{}}}
\subsection{Method \code{getRepoV1TeamIdMemberPrincipalId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdMemberPrincipalId(
  id,
  principal_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{principal_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1TeamIdMemberPrincipalIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdMemberPrincipalIdWithHttpInfo(
  id,
  principal_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{principal_id}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelTeamMember) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalIdMembershipStatus"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalIdMembershipStatus}{}}}
\subsection{Method \code{getRepoV1TeamIdMemberPrincipalIdMembershipStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdMemberPrincipalIdMembershipStatus(
  id,
  principal_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the Team ID}

\item{\code{principal_id}}{the user ID}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelTeamMembershipStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalIdMembershipStatusWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamIdMemberPrincipalIdMembershipStatusWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1TeamIdMemberPrincipalIdMembershipStatusWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamIdMemberPrincipalIdMembershipStatusWithHttpInfo(
  id,
  principal_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the Team ID}

\item{\code{principal_id}}{the user ID}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelTeamMembershipStatus) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamMembersCountId"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamMembersCountId}{}}}
\subsection{Method \code{getRepoV1TeamMembersCountId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamMembersCountId(
  id,
  fragment = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the id of the Team of interest}

\item{\code{fragment}}{(optional) a prefix of the user's first or last name or email address (optional)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamMembersCountIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamMembersCountIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1TeamMembersCountIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamMembersCountIdWithHttpInfo(
  id,
  fragment = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the id of the Team of interest}

\item{\code{fragment}}{(optional) a prefix of the user's first or last name or email address (optional)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelCount) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamMembersId"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamMembersId}{}}}
\subsection{Method \code{getRepoV1TeamMembersId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamMembersId(
  id,
  fragment = NULL,
  member_type = NULL,
  limit = NULL,
  offset = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the id of the Team of interest}

\item{\code{fragment}}{(optional) a prefix of the user's first or last name or email address (optional)}

\item{\code{member_type}}{(optional) the type of team user to retrieve (optional; default \"ALL\")}

\item{\code{limit}}{(optional) the maximum number of members to return (default 10, max limit 50)}

\item{\code{offset}}{(optional) the starting index of the returned results (default 0)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PaginatedResultsOfTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamMembersIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamMembersIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1TeamMembersIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamMembersIdWithHttpInfo(
  id,
  fragment = NULL,
  member_type = NULL,
  limit = NULL,
  offset = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the id of the Team of interest}

\item{\code{fragment}}{(optional) a prefix of the user's first or last name or email address (optional)}

\item{\code{member_type}}{(optional) the type of team user to retrieve (optional; default \"ALL\")}

\item{\code{limit}}{(optional) the maximum number of members to return (default 10, max limit 50)}

\item{\code{offset}}{(optional) the starting index of the returned results (default 0)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PaginatedResultsOfTeamMember) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1Teams"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1Teams}{}}}
\subsection{Method \code{getRepoV1Teams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1Teams(
  fragment = NULL,
  limit = NULL,
  offset = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fragment}}{(optional) a prefix of the Team name, or a prefix of any sub-string in the name preceded by a space.    If omitted, all Teams are returned.}

\item{\code{limit}}{(optional) the maximum number of Teams to return (default 10, max limit 50)}

\item{\code{offset}}{(optional) the starting index of the returned results (default 0)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PaginatedResultsOfTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1TeamsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1TeamsWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1TeamsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1TeamsWithHttpInfo(
  fragment = NULL,
  limit = NULL,
  offset = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fragment}}{(optional) a prefix of the Team name, or a prefix of any sub-string in the name preceded by a space.    If omitted, all Teams are returned.}

\item{\code{limit}}{(optional) the maximum number of Teams to return (default 10, max limit 50)}

\item{\code{offset}}{(optional) the starting index of the returned results (default 0)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PaginatedResultsOfTeam) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1UserIdTeam"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1UserIdTeam}{}}}
\subsection{Method \code{getRepoV1UserIdTeam()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1UserIdTeam(
  id,
  limit = NULL,
  offset = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the principal ID of the user of interest.}

\item{\code{limit}}{(optional) the maximum number of Teams to return (default 10)}

\item{\code{offset}}{(optional) the starting index of the returned results (default 0)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PaginatedResultsOfTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1UserIdTeamWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1UserIdTeamWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1UserIdTeamWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1UserIdTeamWithHttpInfo(
  id,
  limit = NULL,
  offset = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the principal ID of the user of interest.}

\item{\code{limit}}{(optional) the maximum number of Teams to return (default 10)}

\item{\code{offset}}{(optional) the starting index of the returned results (default 0)}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (PaginatedResultsOfTeam) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1UserIdTeamId"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1UserIdTeamId}{}}}
\subsection{Method \code{getRepoV1UserIdTeamId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1UserIdTeamId(
  id,
  next_page_token = NULL,
  sort = NULL,
  ascending = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{next_page_token}}{(optional) controls pagination}

\item{\code{sort}}{(optional) the field to sort the team IDs on. Available options ,<a href=\"${org.sagebionetworks.repo.model.TeamSortOrder}\">,TeamSortOrder,</a>}

\item{\code{ascending}}{(optional) the direction of sort: true for ascending, and false for descending}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelPaginatedTeamIds
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-getRepoV1UserIdTeamIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-getRepoV1UserIdTeamIdWithHttpInfo}{}}}
\subsection{Method \code{getRepoV1UserIdTeamIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$getRepoV1UserIdTeamIdWithHttpInfo(
  id,
  next_page_token = NULL,
  sort = NULL,
  ascending = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{next_page_token}}{(optional) controls pagination}

\item{\code{sort}}{(optional) the field to sort the team IDs on. Available options ,<a href=\"${org.sagebionetworks.repo.model.TeamSortOrder}\">,TeamSortOrder,</a>}

\item{\code{ascending}}{(optional) the direction of sort: true for ascending, and false for descending}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelPaginatedTeamIds) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-postRepoV1Team"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-postRepoV1Team}{}}}
\subsection{Method \code{postRepoV1Team()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$postRepoV1Team(
  org_sagebionetworks_repo_model_team,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_team}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-postRepoV1TeamWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-postRepoV1TeamWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1TeamWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$postRepoV1TeamWithHttpInfo(
  org_sagebionetworks_repo_model_team,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_team}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelTeam) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-postRepoV1TeamIdMemberList"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-postRepoV1TeamIdMemberList}{}}}
\subsection{Method \code{postRepoV1TeamIdMemberList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$postRepoV1TeamIdMemberList(
  id,
  org_sagebionetworks_repo_model_id_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{org_sagebionetworks_repo_model_id_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ListWrapperOfTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-postRepoV1TeamIdMemberListWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-postRepoV1TeamIdMemberListWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1TeamIdMemberListWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$postRepoV1TeamIdMemberListWithHttpInfo(
  id,
  org_sagebionetworks_repo_model_id_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{}

\item{\code{org_sagebionetworks_repo_model_id_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (ListWrapperOfTeamMember) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-postRepoV1TeamList"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-postRepoV1TeamList}{}}}
\subsection{Method \code{postRepoV1TeamList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$postRepoV1TeamList(
  org_sagebionetworks_repo_model_id_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_id_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ListWrapperOfTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-postRepoV1TeamListWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-postRepoV1TeamListWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1TeamListWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$postRepoV1TeamListWithHttpInfo(
  org_sagebionetworks_repo_model_id_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_id_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (ListWrapperOfTeam) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-postRepoV1UserIdMemberList"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-postRepoV1UserIdMemberList}{}}}
\subsection{Method \code{postRepoV1UserIdMemberList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$postRepoV1UserIdMemberList(
  id,
  org_sagebionetworks_repo_model_id_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{user's ID}

\item{\code{org_sagebionetworks_repo_model_id_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ListWrapperOfTeamMember
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-postRepoV1UserIdMemberListWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-postRepoV1UserIdMemberListWithHttpInfo}{}}}
\subsection{Method \code{postRepoV1UserIdMemberListWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$postRepoV1UserIdMemberListWithHttpInfo(
  id,
  org_sagebionetworks_repo_model_id_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{user's ID}

\item{\code{org_sagebionetworks_repo_model_id_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (ListWrapperOfTeamMember) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-putRepoV1Team"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-putRepoV1Team}{}}}
\subsection{Method \code{putRepoV1Team()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$putRepoV1Team(
  org_sagebionetworks_repo_model_team,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_team}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelTeam
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-putRepoV1TeamWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-putRepoV1TeamWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1TeamWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$putRepoV1TeamWithHttpInfo(
  org_sagebionetworks_repo_model_team,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_team}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelTeam) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-putRepoV1TeamAcl"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-putRepoV1TeamAcl}{}}}
\subsection{Method \code{putRepoV1TeamAcl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$putRepoV1TeamAcl(
  org_sagebionetworks_repo_model_access_control_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_access_control_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelAccessControlList
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-putRepoV1TeamAclWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-putRepoV1TeamAclWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1TeamAclWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$putRepoV1TeamAclWithHttpInfo(
  org_sagebionetworks_repo_model_access_control_list,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_access_control_list}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelAccessControlList) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-putRepoV1TeamIdMemberPrincipalId"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-putRepoV1TeamIdMemberPrincipalId}{}}}
\subsection{Method \code{putRepoV1TeamIdMemberPrincipalId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$putRepoV1TeamIdMemberPrincipalId(
  id,
  principal_id,
  team_endpoint = NULL,
  notification_unsubscribe_endpoint = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team to which the user is to be added.}

\item{\code{principal_id}}{the ID of the user to be added to the Team.}

\item{\code{team_endpoint}}{(optional) the portal prefix for the Team URL. The team ID is appended to create the complete URL.}

\item{\code{notification_unsubscribe_endpoint}}{(optional) the portal prefix for one-click email unsubscription.    A signed, serialized token is appended to create the complete URL:   ,<,ahref=\"${org.sagebionetworks.repo.model.message.NotificationSettingsSignedToken}\",>,NotificationSettingsSignedToken,</a>}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-putRepoV1TeamIdMemberPrincipalIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-putRepoV1TeamIdMemberPrincipalIdWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1TeamIdMemberPrincipalIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$putRepoV1TeamIdMemberPrincipalIdWithHttpInfo(
  id,
  principal_id,
  team_endpoint = NULL,
  notification_unsubscribe_endpoint = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the Team to which the user is to be added.}

\item{\code{principal_id}}{the ID of the user to be added to the Team.}

\item{\code{team_endpoint}}{(optional) the portal prefix for the Team URL. The team ID is appended to create the complete URL.}

\item{\code{notification_unsubscribe_endpoint}}{(optional) the portal prefix for one-click email unsubscription.    A signed, serialized token is appended to create the complete URL:   ,<,ahref=\"${org.sagebionetworks.repo.model.message.NotificationSettingsSignedToken}\",>,NotificationSettingsSignedToken,</a>}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (void) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-putRepoV1TeamMember"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-putRepoV1TeamMember}{}}}
\subsection{Method \code{putRepoV1TeamMember()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$putRepoV1TeamMember(
  org_sagebionetworks_repo_model_join_team_signed_token,
  team_endpoint = NULL,
  notification_unsubscribe_endpoint = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_join_team_signed_token}}{}

\item{\code{team_endpoint}}{(optional) No description}

\item{\code{notification_unsubscribe_endpoint}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
OrgSagebionetworksRepoModelResponseMessage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-putRepoV1TeamMemberWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-putRepoV1TeamMemberWithHttpInfo}{}}}
\subsection{Method \code{putRepoV1TeamMemberWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$putRepoV1TeamMemberWithHttpInfo(
  org_sagebionetworks_repo_model_join_team_signed_token,
  team_endpoint = NULL,
  notification_unsubscribe_endpoint = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_sagebionetworks_repo_model_join_team_signed_token}}{}

\item{\code{team_endpoint}}{(optional) No description}

\item{\code{notification_unsubscribe_endpoint}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (OrgSagebionetworksRepoModelResponseMessage) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TeamServicesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TeamServicesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamServicesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
