% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grapheme_vector.R
\name{build_grapheme_vector}
\alias{build_grapheme_vector}
\title{build_grapheme_vector}
\usage{
build_grapheme_vector(
  language,
  use_digits,
  use_letters,
  use_weekdays,
  uppercase = TRUE
)
}
\arguments{
\item{language}{A character element. What language should be used for the vector?}

\item{use_digits}{A logical scalar. Should digits be included in the vector?}

\item{use_letters}{A logical scalar. Should letters be included in the vector?}

\item{use_weekdays}{A logical scalar. Should weekdays be included in the vector?}

\item{uppercase}{A logical scalar. Should graphemes be in uppercase? (otherwise uses all lowercase)}
}
\description{
Uses internal data constants to build a character vector
containing graphemes in accordance with specifications passed as
function arguments. Supports English and Swedish language specifications.
}
\examples{
build_grapheme_vector(language="english", use_digits=TRUE, use_letters=FALSE,
use_weekdays=TRUE, uppercase=TRUE)
}
