% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParticipantGroup.R
\docType{class}
\name{ParticipantGroup-class}
\alias{ParticipantGroup-class}
\alias{ParticipantGroup}
\title{A Reference Class to represent a group of consistency test participants.}
\description{
A Reference Class to represent a group of consistency test participants.
}
\section{Fields}{

\describe{
\item{\code{participants}}{A list of \code{\link[synr]{Participant}} class instances.}
}}

\section{Methods}{

\describe{
\item{\code{add_participant(participant)}}{Add a passed participant to the participantgroup's list
of participant The participant's entry in
the list is named based on the participant's
id. Note that if you try to add
a participant with an id that's identical
to one of the participants already in the
participantgroup's list of participants, the
already existing same-id participant
is overwritten.}

\item{\code{add_participants(participant_list)}}{Go through a passed list of Participant instances
and add each one using the add_participant() method.}

\item{\code{get_ids()}}{Returns a character vector with all ids for
participants associated with the participantgroup.}

\item{\code{get_mean_consistency_scores(na.rm = FALSE)}}{Returns a vector mean consistency scores for
participants in the group}

\item{\code{get_mean_response_times(na.rm = FALSE)}}{Returns the mean response times, with respect to
Grapheme instances associated with each participant. See
    the documentation for the Participant class for more
    information.}

\item{\code{get_numbers_all_colored_graphemes()}}{Returns the a list with number representing how many
graphemes with all-valid (non-na) response colors that each
participant has.}

\item{\code{has_participants()}}{Returns TRUE if there is at least one
participant in the participantgroup's participants list,
otherwise returns FALSE}

\item{\code{save_plots(path = NULL, file_format = "png", dpi = 300, ...)}}{Goes through all participants and for each one produces and saves
    a ggplot2 plot that describes the participant's
grapheme color responses and per-grapheme consistency scores,
using the ggsave function.
If save_dir is not specified, plots are saved to the current
working directory. Otherwise, plots are saved to the specified
directory. The file is saved using the specified file_format,
e. g. JPG (see ggplot2::ggsave documentation for list of
supported formats), and the resolution specified with
the dpi argument. Apart from these, all other arguments
that ggsave accepts (e. g. 'scale') also work with this function, since
all arguments are passed on to ggsave.}
}}

