% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synr.R
\docType{package}
\name{synr-package}
\alias{synr}
\alias{synr-package}
\title{synr: Explore and Process Synesthesia Consistency Test Data}
\description{
synr helps you work with data resulting from grapheme-color consistency tests for synesthesia.

To learn more about synr, start with the vignettes:
\verb{browseVignettes(package = "synr")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://datalowe.github.io/synr/}
  \item \url{https://github.com/datalowe/synr}
}

}
\author{
\strong{Maintainer}: Lowe Wilsson \email{datalowe@posteo.de}

}
